Highest Posterior Density intervals for objects of class mtar

Highest Posterior Density intervals for objects of class mtar

Usage

# S3 method for class 'mtar'
HPDinterval(obj, prob = 0.95, ...)

Arguments

obj

an object of class mtar generated by a call to the function mtar().

prob

a numeric scalar in the interval \((0,1)\) giving the target probability content of the intervals. By default, prob is set to 0.95.

...

Optional additional arguments for methods. None are used at present.

See also

Examples

# \donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
             subset={Date<="2015-12-07"}, dist="Student-t",
             ars=ars(nregim=3,p=c(1,1,2)), n.burnin=1000, n.sim=2000,
             n.thin=2, ssvs=TRUE)
coda::HPDinterval(fit1)
#> 
#> Probability = 0.95
#> 
#> Thresholds:
#>                  lower      upper
#> Threshold.1 -0.0105467 -0.0053141
#> Threshold.2  0.0067634  0.0079366
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                             lower      upper
#> COLCAP:(Intercept)     -0.0093466 -0.0044662
#> BOVESPA:(Intercept)    -0.0171372 -0.0101973
#> COLCAP:COLCAP.lag(1)    0.0604813  0.4595918
#> BOVESPA:COLCAP.lag(1)  -0.2905249  0.1583518
#> COLCAP:BOVESPA.lag(1)  -0.0063395  0.2392283
#> BOVESPA:BOVESPA.lag(1) -0.0806571  0.2936382
#> 
#> 
#> Scale parameter:
#>                      lower      upper
#> COLCAP.COLCAP   5.7269e-05 9.5435e-05
#> COLCAP.BOVESPA  2.1088e-05 5.7450e-05
#> BOVESPA.BOVESPA 1.0771e-04 1.8162e-04
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                              lower      upper
#> COLCAP:(Intercept)     -0.00052473 0.00049450
#> BOVESPA:(Intercept)    -0.00146854 0.00018636
#> COLCAP:COLCAP.lag(1)    0.01496205 0.13806387
#> BOVESPA:COLCAP.lag(1)  -0.01518846 0.14768619
#> COLCAP:BOVESPA.lag(1)   0.03610111 0.10737915
#> BOVESPA:BOVESPA.lag(1) -0.09018070 0.01712991
#> 
#> 
#> Scale parameter:
#>                      lower      upper
#> COLCAP.COLCAP   3.6230e-05 4.6794e-05
#> COLCAP.BOVESPA  8.1152e-06 1.7354e-05
#> BOVESPA.BOVESPA 7.5937e-05 9.8350e-05
#> 
#> 
#> Regime 3
#> 
#> 
#> Autoregressive coefficients:
#>                             lower     upper
#> COLCAP:(Intercept)      0.0037811 0.0059328
#> BOVESPA:(Intercept)     0.0096097 0.0132295
#> COLCAP:COLCAP.lag(1)   -0.0738601 0.1621739
#> BOVESPA:COLCAP.lag(1)  -0.0107754 0.3390709
#> COLCAP:BOVESPA.lag(1)  -0.0144497 0.1532598
#> BOVESPA:BOVESPA.lag(1) -0.2529857 0.0198287
#> COLCAP:COLCAP.lag(2)   -0.0460785 0.1684784
#> BOVESPA:COLCAP.lag(2)  -0.2476527 0.0945643
#> COLCAP:BOVESPA.lag(2)  -0.1434067 0.0030170
#> BOVESPA:BOVESPA.lag(2) -0.1692394 0.0687651
#> 
#> 
#> Scale parameter:
#>                      lower      upper
#> COLCAP.COLCAP   4.3265e-05 6.8451e-05
#> COLCAP.BOVESPA  1.3898e-05 3.9686e-05
#> BOVESPA.BOVESPA 1.0671e-04 1.6467e-04
#> 
#> 
#> Extra parameter:
#>     lower  upper
#> nu 4.7195 7.5517

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, row.names=Date,
             subset={Date<="2009-02-13"}, dist="Laplace",
             ars=ars(nregim=3,p=5), n.burnin=1000, n.sim=2000, n.thin=2)
coda::HPDinterval(fit2)
#> 
#> Probability = 0.95
#> 
#> Thresholds:
#>               lower   upper
#> Threshold.1  3.0789  3.9707
#> Threshold.2 10.0004 10.0154
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                             lower      upper
#> Bedon:(Intercept)       1.1160201  1.5651419
#> LaPlata:(Intercept)     2.7945679  4.0672872
#> Bedon:Bedon.lag(1)      0.4877313  0.6415489
#> LaPlata:Bedon.lag(1)   -0.0430582  0.3677093
#> Bedon:LaPlata.lag(1)    0.0184300  0.0734901
#> LaPlata:LaPlata.lag(1)  0.5625058  0.7184208
#> Bedon:Bedon.lag(2)     -0.0181463  0.1198864
#> LaPlata:Bedon.lag(2)   -0.2282692  0.1387220
#> Bedon:LaPlata.lag(2)   -0.0469403  0.0027720
#> LaPlata:LaPlata.lag(2) -0.1395688 -0.0044446
#> Bedon:Bedon.lag(3)     -0.0301124  0.0937705
#> LaPlata:Bedon.lag(3)   -0.1336340  0.1657376
#> Bedon:LaPlata.lag(3)   -0.0162033  0.0255501
#> LaPlata:LaPlata.lag(3)  0.0082618  0.1164385
#> Bedon:Bedon.lag(4)     -0.0230290  0.0975958
#> LaPlata:Bedon.lag(4)   -0.2699160  0.0626935
#> Bedon:LaPlata.lag(4)   -0.0304046  0.0028671
#> LaPlata:LaPlata.lag(4) -0.0406204  0.0619381
#> Bedon:Bedon.lag(5)      0.0279773  0.1298280
#> LaPlata:Bedon.lag(5)    0.0159901  0.2712195
#> Bedon:LaPlata.lag(5)   -0.0194786  0.0061966
#> LaPlata:LaPlata.lag(5) -0.0218077  0.0653388
#> 
#> 
#> Scale parameter:
#>                   lower   upper
#> Bedon.Bedon     0.27335 0.40019
#> Bedon.LaPlata   0.25788 0.49196
#> LaPlata.LaPlata 1.92787 2.84566
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                             lower     upper
#> Bedon:(Intercept)       1.2205244  2.866194
#> LaPlata:(Intercept)     4.8176837  9.239897
#> Bedon:Bedon.lag(1)      0.4950651  0.675152
#> LaPlata:Bedon.lag(1)   -0.0734034  0.386079
#> Bedon:LaPlata.lag(1)   -0.0034366  0.050697
#> LaPlata:LaPlata.lag(1)  0.4499684  0.597393
#> Bedon:Bedon.lag(2)     -0.0312621  0.213392
#> LaPlata:Bedon.lag(2)   -0.2528789  0.248898
#> Bedon:LaPlata.lag(2)   -0.0560719  0.012144
#> LaPlata:LaPlata.lag(2) -0.0463682  0.105041
#> Bedon:Bedon.lag(3)     -0.1477841  0.076553
#> LaPlata:Bedon.lag(3)   -0.2969824  0.154529
#> Bedon:LaPlata.lag(3)   -0.0367606  0.024793
#> LaPlata:LaPlata.lag(3) -0.0269240  0.112720
#> Bedon:Bedon.lag(4)     -0.0236022  0.216805
#> LaPlata:Bedon.lag(4)   -0.0237688  0.482530
#> Bedon:LaPlata.lag(4)   -0.0247901  0.042682
#> LaPlata:LaPlata.lag(4) -0.1256113  0.035177
#> Bedon:Bedon.lag(5)     -0.0591167  0.109126
#> LaPlata:Bedon.lag(5)   -0.4621428 -0.071275
#> Bedon:LaPlata.lag(5)   -0.0260272  0.033221
#> LaPlata:LaPlata.lag(5)  0.0449784  0.197214
#> 
#> 
#> Scale parameter:
#>                   lower  upper
#> Bedon.Bedon     0.88666 1.3085
#> Bedon.LaPlata   0.95006 1.6848
#> LaPlata.LaPlata 5.25300 7.7639
#> 
#> 
#> Regime 3
#> 
#> 
#> Autoregressive coefficients:
#>                             lower     upper
#> Bedon:(Intercept)       4.0842459  7.185747
#> LaPlata:(Intercept)    11.4364966 22.560302
#> Bedon:Bedon.lag(1)      0.2990350  0.642322
#> LaPlata:Bedon.lag(1)    0.0658349  1.082438
#> Bedon:LaPlata.lag(1)    0.0065802  0.078881
#> LaPlata:LaPlata.lag(1)  0.1983719  0.449526
#> Bedon:Bedon.lag(2)     -0.0683105  0.225472
#> LaPlata:Bedon.lag(2)   -1.0826811  0.011055
#> Bedon:LaPlata.lag(2)   -0.0372042  0.029155
#> LaPlata:LaPlata.lag(2) -0.0032045  0.259856
#> Bedon:Bedon.lag(3)     -0.2216860  0.035352
#> LaPlata:Bedon.lag(3)   -1.0152121 -0.133005
#> Bedon:LaPlata.lag(3)   -0.0032350  0.071016
#> LaPlata:LaPlata.lag(3)  0.1257305  0.434277
#> Bedon:Bedon.lag(4)     -0.1512500  0.144840
#> LaPlata:Bedon.lag(4)   -0.5194604  0.584387
#> Bedon:LaPlata.lag(4)   -0.0363866  0.045406
#> LaPlata:LaPlata.lag(4) -0.1683937  0.142857
#> Bedon:Bedon.lag(5)      0.0452300  0.323623
#> LaPlata:Bedon.lag(5)   -0.2050171  0.762937
#> Bedon:LaPlata.lag(5)   -0.0470294  0.022292
#> LaPlata:LaPlata.lag(5) -0.0699537  0.199694
#> 
#> 
#> Scale parameter:
#>                   lower   upper
#> Bedon.Bedon      2.2152  3.3723
#> Bedon.LaPlata    5.4652  9.0536
#> LaPlata.LaPlata 34.7138 51.9353

###### Example 3: Temperature, precipitation, and two river flows in Iceland
data(iceland.rf)
fit3 <- mtar(~ Jokulsa + Vatnsdalsa | Temperature | Precipitation,
             data=iceland.rf, subset={Date<="1974-11-06"}, row.names=Date,
             ars=ars(nregim=2,p=15,q=4,d=2), n.burnin=1000, n.sim=2000,
             n.thin=2, dist="Slash")
coda::HPDinterval(fit3)
#> 
#> Probability = 0.95
#> 
#> Thresholds:
#>              lower  upper
#> Threshold.1 1.0861 1.1992
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                                       lower       upper
#> Jokulsa:(Intercept)              2.88382503  4.44732602
#> Vatnsdalsa:(Intercept)           0.43187726  1.23252479
#> Jokulsa:Jokulsa.lag( 1)          0.76000694  0.91630629
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.10359938 -0.02755573
#> Jokulsa:Vatnsdalsa.lag( 1)       0.09432201  0.32264150
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.07781821  1.24815183
#> Jokulsa:Jokulsa.lag( 2)         -0.09711439 -0.00020152
#> Vatnsdalsa:Jokulsa.lag( 2)       0.02100004  0.08033333
#> Jokulsa:Vatnsdalsa.lag( 2)      -0.31503295 -0.04996223
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.38669222 -0.20906830
#> Jokulsa:Jokulsa.lag( 3)         -0.03302854  0.04352367
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.04889499 -0.00198104
#> Jokulsa:Vatnsdalsa.lag( 3)      -0.03502837  0.10300355
#> Vatnsdalsa:Vatnsdalsa.lag( 3)   -0.01977713  0.08694671
#> Jokulsa:Jokulsa.lag( 4)         -0.05018181  0.04111967
#> Vatnsdalsa:Jokulsa.lag( 4)      -0.01210339  0.03960628
#> Jokulsa:Vatnsdalsa.lag( 4)      -0.07223401  0.08447178
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -0.05357632  0.05856575
#> Jokulsa:Jokulsa.lag( 5)         -0.05465539  0.05046748
#> Vatnsdalsa:Jokulsa.lag( 5)      -0.02238151  0.03538820
#> Jokulsa:Vatnsdalsa.lag( 5)      -0.11248248  0.02840708
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -0.06551680  0.02613639
#> Jokulsa:Jokulsa.lag( 6)         -0.02912973  0.07583977
#> Vatnsdalsa:Jokulsa.lag( 6)      -0.02283292  0.03503461
#> Jokulsa:Vatnsdalsa.lag( 6)      -0.08780229  0.03852471
#> Vatnsdalsa:Vatnsdalsa.lag( 6)   -0.04117395  0.04203699
#> Jokulsa:Jokulsa.lag( 7)         -0.04991157  0.05342262
#> Vatnsdalsa:Jokulsa.lag( 7)      -0.02971443  0.02634622
#> Jokulsa:Vatnsdalsa.lag( 7)      -0.04147282  0.07971818
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -0.02565624  0.04581258
#> Jokulsa:Jokulsa.lag( 8)         -0.05184395  0.05086101
#> Vatnsdalsa:Jokulsa.lag( 8)      -0.03581333  0.01390658
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.06452424  0.05639897
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -0.02979028  0.04231896
#> Jokulsa:Jokulsa.lag( 9)         -0.05686980  0.03384085
#> Vatnsdalsa:Jokulsa.lag( 9)      -0.00896816  0.04817773
#> Jokulsa:Vatnsdalsa.lag( 9)      -0.06775950  0.06784114
#> Vatnsdalsa:Vatnsdalsa.lag( 9)   -0.04507373  0.03498430
#> Jokulsa:Jokulsa.lag(10)         -0.00523168  0.06580287
#> Vatnsdalsa:Jokulsa.lag(10)      -0.03802815  0.00732220
#> Jokulsa:Vatnsdalsa.lag(10)      -0.03617874  0.08936170
#> Vatnsdalsa:Vatnsdalsa.lag(10)   -0.02344705  0.05990774
#> Jokulsa:Jokulsa.lag(11)         -0.04115176  0.01597498
#> Vatnsdalsa:Jokulsa.lag(11)      -0.00787191  0.02828408
#> Jokulsa:Vatnsdalsa.lag(11)      -0.06787615  0.03600412
#> Vatnsdalsa:Vatnsdalsa.lag(11)   -0.04464346  0.03175820
#> Jokulsa:Jokulsa.lag(12)         -0.02097614  0.03699410
#> Vatnsdalsa:Jokulsa.lag(12)      -0.02746569  0.00974474
#> Jokulsa:Vatnsdalsa.lag(12)      -0.04153324  0.05730462
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.03861288  0.03093955
#> Jokulsa:Jokulsa.lag(13)         -0.04977446  0.01704202
#> Vatnsdalsa:Jokulsa.lag(13)      -0.01766610  0.02210310
#> Jokulsa:Vatnsdalsa.lag(13)      -0.06370271  0.04302112
#> Vatnsdalsa:Vatnsdalsa.lag(13)   -0.06100834  0.01951933
#> Jokulsa:Jokulsa.lag(14)         -0.02220951  0.03446611
#> Vatnsdalsa:Jokulsa.lag(14)      -0.02253331  0.01209167
#> Jokulsa:Vatnsdalsa.lag(14)      -0.04748785  0.05861120
#> Vatnsdalsa:Vatnsdalsa.lag(14)   -0.00178360  0.07059931
#> Jokulsa:Jokulsa.lag(15)         -0.00684027  0.04821752
#> Vatnsdalsa:Jokulsa.lag(15)      -0.01293825  0.01610465
#> Jokulsa:Vatnsdalsa.lag(15)      -0.04913044  0.02864937
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -0.02041337  0.03164976
#> Jokulsa:Precipitation.lag(1)    -0.01089299  0.02524450
#> Vatnsdalsa:Precipitation.lag(1) -0.00498399  0.01634566
#> Jokulsa:Precipitation.lag(2)    -0.00862971  0.01878436
#> Vatnsdalsa:Precipitation.lag(2) -0.00812234  0.00871486
#> Jokulsa:Precipitation.lag(3)    -0.02244260 -0.00110221
#> Vatnsdalsa:Precipitation.lag(3) -0.01197880  0.00360246
#> Jokulsa:Precipitation.lag(4)     0.00609342  0.03238467
#> Vatnsdalsa:Precipitation.lag(4) -0.00534757  0.01291966
#> Jokulsa:Temperature.lag(1)       0.00081042  0.04282116
#> Vatnsdalsa:Temperature.lag(1)   -0.01139348  0.01543358
#> Jokulsa:Temperature.lag(2)      -0.06260580 -0.02011095
#> Vatnsdalsa:Temperature.lag(2)   -0.02608801  0.00075713
#> 
#> 
#> Scale parameter:
#>                           lower    upper
#> Jokulsa.Jokulsa       0.0437689 0.082602
#> Jokulsa.Vatnsdalsa    0.0056372 0.016817
#> Vatnsdalsa.Vatnsdalsa 0.0197654 0.036438
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                                      lower       upper
#> Jokulsa:(Intercept)             -1.6918970  1.04485636
#> Vatnsdalsa:(Intercept)           0.2998765  0.66295773
#> Jokulsa:Jokulsa.lag( 1)          0.9404280  1.07922779
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.0096740  0.00486718
#> Jokulsa:Vatnsdalsa.lag( 1)       0.4687186  1.34202903
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.1174301  1.25145416
#> Jokulsa:Jokulsa.lag( 2)         -0.2969496 -0.04337666
#> Vatnsdalsa:Jokulsa.lag( 2)      -0.0013701  0.02243668
#> Jokulsa:Vatnsdalsa.lag( 2)      -1.0723204  0.30694462
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.4429921 -0.23983504
#> Jokulsa:Jokulsa.lag( 3)         -0.1016859  0.12815447
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.0240219  0.00012616
#> Jokulsa:Vatnsdalsa.lag( 3)      -0.6940183  0.65306950
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    0.1090014  0.28229128
#> Jokulsa:Jokulsa.lag( 4)         -0.1572271  0.00683034
#> Vatnsdalsa:Jokulsa.lag( 4)      -0.0030215  0.01478683
#> Jokulsa:Vatnsdalsa.lag( 4)      -0.5802067  0.30909470
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -0.1588649 -0.00228283
#> Jokulsa:Jokulsa.lag( 5)         -0.0411497  0.11918496
#> Vatnsdalsa:Jokulsa.lag( 5)      -0.0151608  0.00343887
#> Jokulsa:Vatnsdalsa.lag( 5)      -0.5859201  0.60041157
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -0.0813984  0.11187237
#> Jokulsa:Jokulsa.lag( 6)         -0.1039489  0.02395885
#> Vatnsdalsa:Jokulsa.lag( 6)      -0.0033997  0.01250092
#> Jokulsa:Vatnsdalsa.lag( 6)      -0.4671615  0.70241051
#> Vatnsdalsa:Vatnsdalsa.lag( 6)   -0.0859096  0.11095210
#> Jokulsa:Jokulsa.lag( 7)         -0.0554288  0.05943931
#> Vatnsdalsa:Jokulsa.lag( 7)      -0.0142107  0.00207168
#> Jokulsa:Vatnsdalsa.lag( 7)      -0.3400239  0.55882117
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -0.1236017  0.00966018
#> Jokulsa:Jokulsa.lag( 8)         -0.0461301  0.06837444
#> Vatnsdalsa:Jokulsa.lag( 8)      -0.0031991  0.01148060
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.6437366  0.23491536
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -0.1071862  0.01773855
#> Jokulsa:Jokulsa.lag( 9)         -0.0251946  0.10184730
#> Vatnsdalsa:Jokulsa.lag( 9)      -0.0099757  0.00567991
#> Jokulsa:Vatnsdalsa.lag( 9)      -0.2960697  0.54928753
#> Vatnsdalsa:Vatnsdalsa.lag( 9)    0.0212982  0.15841235
#> Jokulsa:Jokulsa.lag(10)         -0.0973212  0.06136879
#> Vatnsdalsa:Jokulsa.lag(10)      -0.0050251  0.01369390
#> Jokulsa:Vatnsdalsa.lag(10)      -0.3851555  0.38465339
#> Vatnsdalsa:Vatnsdalsa.lag(10)   -0.1347294 -0.01752026
#> Jokulsa:Jokulsa.lag(11)         -0.0773338  0.07746450
#> Vatnsdalsa:Jokulsa.lag(11)      -0.0168276  0.00133286
#> Jokulsa:Vatnsdalsa.lag(11)      -0.4949327  0.47928834
#> Vatnsdalsa:Vatnsdalsa.lag(11)    0.0124115  0.14668298
#> Jokulsa:Jokulsa.lag(12)         -0.0795674  0.06518172
#> Vatnsdalsa:Jokulsa.lag(12)       0.0014610  0.01815014
#> Jokulsa:Vatnsdalsa.lag(12)      -0.4718757  0.45873115
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.1363640 -0.02219283
#> Jokulsa:Jokulsa.lag(13)         -0.0768279  0.08128056
#> Vatnsdalsa:Jokulsa.lag(13)      -0.0140588  0.00155564
#> Jokulsa:Vatnsdalsa.lag(13)      -0.0916343  0.93959370
#> Vatnsdalsa:Vatnsdalsa.lag(13)    0.0860228  0.22033316
#> Jokulsa:Jokulsa.lag(14)         -0.0842848  0.07550527
#> Vatnsdalsa:Jokulsa.lag(14)      -0.0109699  0.00553780
#> Jokulsa:Vatnsdalsa.lag(14)      -0.4266958  0.63449900
#> Vatnsdalsa:Vatnsdalsa.lag(14)   -0.1314771  0.02220386
#> Jokulsa:Jokulsa.lag(15)          0.0017429  0.09240085
#> Vatnsdalsa:Jokulsa.lag(15)      -0.0043250  0.00690267
#> Jokulsa:Vatnsdalsa.lag(15)      -0.7521815 -0.13878258
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -0.0684182  0.02632012
#> Jokulsa:Precipitation.lag(1)    -0.1882165 -0.04203210
#> Vatnsdalsa:Precipitation.lag(1) -0.0123536  0.00761347
#> Jokulsa:Precipitation.lag(2)    -0.0969530  0.15720893
#> Vatnsdalsa:Precipitation.lag(2) -0.0139397  0.01235777
#> Jokulsa:Precipitation.lag(3)    -0.0147974  0.11894979
#> Vatnsdalsa:Precipitation.lag(3) -0.0048789  0.01526502
#> Jokulsa:Precipitation.lag(4)    -0.0390826  0.09290304
#> Vatnsdalsa:Precipitation.lag(4) -0.0058609  0.01132458
#> Jokulsa:Temperature.lag(1)       0.9413910  1.32975123
#> Vatnsdalsa:Temperature.lag(1)   -0.0065553  0.04217064
#> Jokulsa:Temperature.lag(2)      -0.7733226 -0.37379654
#> Vatnsdalsa:Temperature.lag(2)   -0.0481004  0.00051169
#> 
#> 
#> Scale parameter:
#>                          lower    upper
#> Jokulsa.Jokulsa       0.901087 1.713881
#> Jokulsa.Vatnsdalsa    0.023817 0.070197
#> Vatnsdalsa.Vatnsdalsa 0.015432 0.030313
#> 
#> 
#> Extra parameter:
#>      lower   upper
#> nu 0.71691 0.88855

###### Example 4: U.S. stock returns
data(US.returns)
fit4 <- mtar(~ CCR | dVIX, data=US.returns, subset={Date<="2025-11-28"},
             row.names=Date, ars=ars(nregim=2,p=3,d=3), n.burnin=1000,
             n.sim=2000, n.thin=2, dist="Student-t")
coda::HPDinterval(fit4)
#> 
#> Probability = 0.95
#> 
#> Thresholds:
#>               lower  upper
#> Threshold.1 0.90027 1.2317
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                       lower      upper
#> CCR:(Intercept)  0.06697037  0.1168107
#> CCR:CCR.lag(1)  -0.08064096 -0.0213740
#> CCR:CCR.lag(2)  -0.08060333  0.0062743
#> CCR:CCR.lag(3)  -0.07945977  0.0165688
#> CCR:dVIX.lag(1) -0.06652391 -0.0106839
#> CCR:dVIX.lag(2) -0.05291829  0.0039770
#> CCR:dVIX.lag(3)  0.00052384  0.0329808
#> 
#> 
#> Scale parameter:
#>           lower  upper
#> CCR.CCR 0.31557 0.3797
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                     lower      upper
#> CCR:(Intercept) -0.084002  0.1832318
#> CCR:CCR.lag(1)  -0.176610 -0.0040544
#> CCR:CCR.lag(2)  -0.152975  0.1405693
#> CCR:CCR.lag(3)  -0.067007  0.1739317
#> CCR:dVIX.lag(1) -0.089884  0.1291549
#> CCR:dVIX.lag(2) -0.068848  0.0758559
#> CCR:dVIX.lag(3) -0.011633  0.0842883
#> 
#> 
#> Scale parameter:
#>           lower  upper
#> CCR.CCR 0.64887 0.9336
#> 
#> 
#> Extra parameter:
#>     lower  upper
#> nu 2.2205 2.6419
# }