Coercion of mtar objects to mcmc objects

This method converts an object of class mtar into a list of mcmc objects, each corresponding to a Markov chain produced during Bayesian estimation.

Usage

# S3 method for class 'mtar'
as.mcmc(x, ...)

Arguments

x

an object of class mtar obtained from a call to mtar().

...

additional arguments passed to specific coercion methods.

Value

A list of mcmc objects containing the posterior simulation draws generated by the mtar() routine.

See also

Examples

# \donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
             subset={Date<="2015-12-07"}, dist="Student-t",
             ars=ars(nregim=3,p=c(1,1,2)), n.burnin=1000, n.sim=2000,
             n.thin=2, ssvs=TRUE)
fit1.mcmc <- coda::as.mcmc(fit1)
summary(fit1.mcmc)
#> 
#> 
#>  Iterations = 1001:4999
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 2000
#> 
#> 
#> Thresholds:
#>                   Mean        Sd   Sd(Mean)       2.5%        25%        50%
#> Threshold.1 -0.0098449 0.0011715 0.00052681 -0.0105261 -0.0103634 -0.0102402
#> Threshold.2  0.0082971 0.0012641 0.00066556  0.0072055  0.0075823  0.0076891
#>                    75%      97.5%
#> Threshold.1 -0.0100040 -0.0063585
#> Threshold.2  0.0083318  0.0109856
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean        Sd   Sd(Mean)       2.5%        25%
#> COLCAP:(Intercept)     -0.0079028 0.0010506 0.00023674 -0.0096521 -0.0086074
#> BOVESPA:(Intercept)    -0.0150632 0.0014960 0.00036080 -0.0176469 -0.0160426
#> COLCAP:COLCAP.lag(1)    0.2923132 0.1081331 0.00875649  0.0831473  0.2161677
#> BOVESPA:COLCAP.lag(1)  -0.0967465 0.1222021 0.00287223 -0.3466001 -0.1787800
#> COLCAP:BOVESPA.lag(1)   0.1282406 0.0703248 0.00323090 -0.0038321  0.0796096
#> BOVESPA:BOVESPA.lag(1)  0.1440970 0.0966307 0.00912543 -0.0472085  0.0802670
#>                               50%        75%      97.5%
#> COLCAP:(Intercept)     -0.0079964 -0.0073214 -0.0054361
#> BOVESPA:(Intercept)    -0.0152131 -0.0143596 -0.0114482
#> COLCAP:COLCAP.lag(1)    0.2942815  0.3684427  0.5001423
#> BOVESPA:COLCAP.lag(1)  -0.0929852 -0.0128736  0.1435152
#> COLCAP:BOVESPA.lag(1)   0.1247854  0.1751368  0.2710462
#> BOVESPA:BOVESPA.lag(1)  0.1449485  0.2112339  0.3251170
#> 
#> 
#> Scale parameter:
#>                       Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP.COLCAP   7.9390e-05 1.1598e-05 4.1422e-07 5.9563e-05 7.1189e-05
#> COLCAP.BOVESPA  3.9983e-05 1.0819e-05 2.6946e-07 2.0614e-05 3.2754e-05
#> BOVESPA.BOVESPA 1.4725e-04 2.1799e-05 7.6582e-07 1.0800e-04 1.3223e-04
#>                        50%        75%      97.5%
#> COLCAP.COLCAP   7.8048e-05 8.6628e-05 1.0486e-04
#> COLCAP.BOVESPA  3.9379e-05 4.6417e-05 6.3421e-05
#> BOVESPA.BOVESPA 1.4586e-04 1.6088e-04 1.9285e-04
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                               Mean         Sd   Sd(Mean)        2.5%
#> COLCAP:(Intercept)     -7.5126e-05 0.00024749 2.6554e-05 -0.00055789
#> BOVESPA:(Intercept)    -8.4127e-04 0.00041449 6.5217e-05 -0.00162733
#> COLCAP:COLCAP.lag(1)    7.3853e-02 0.02913853 1.0968e-03  0.01649344
#> BOVESPA:COLCAP.lag(1)   5.2231e-02 0.04122283 9.5146e-04 -0.03461979
#> COLCAP:BOVESPA.lag(1)   7.3471e-02 0.01750560 5.2436e-04  0.03973755
#> BOVESPA:BOVESPA.lag(1) -3.6483e-02 0.02642267 6.2151e-04 -0.08759459
#>                                25%         50%         75%      97.5%
#> COLCAP:(Intercept)     -0.00023889 -7.5304e-05  8.9653e-05 4.0643e-04
#> BOVESPA:(Intercept)    -0.00112968 -8.5649e-04 -5.6214e-04 8.3225e-06
#> COLCAP:COLCAP.lag(1)    0.05471679  7.3430e-02  9.3155e-02 1.3140e-01
#> BOVESPA:COLCAP.lag(1)   0.02524044  5.3528e-02  8.0341e-02 1.3054e-01
#> COLCAP:BOVESPA.lag(1)   0.06191821  7.3594e-02  8.4858e-02 1.0682e-01
#> BOVESPA:BOVESPA.lag(1) -0.05460091 -3.7032e-02 -1.8475e-02 1.5374e-02
#> 
#> 
#> Scale parameter:
#>                       Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP.COLCAP   4.1899e-05 2.6695e-06 1.4911e-07 3.7087e-05 4.0041e-05
#> COLCAP.BOVESPA  1.3424e-05 2.3092e-06 9.9519e-08 9.0679e-06 1.1804e-05
#> BOVESPA.BOVESPA 8.9673e-05 5.9974e-06 3.2091e-07 7.8501e-05 8.5377e-05
#>                        50%        75%      97.5%
#> COLCAP.COLCAP   4.1719e-05 4.3705e-05 4.7439e-05
#> COLCAP.BOVESPA  1.3383e-05 1.5047e-05 1.8004e-05
#> BOVESPA.BOVESPA 8.9558e-05 9.3507e-05 1.0203e-04
#> 
#> 
#> Regime 3
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP:(Intercept)      0.0052565 0.00092046 0.00037004  0.0039343  0.0046487
#> BOVESPA:(Intercept)     0.0122583 0.00180407 0.00075012  0.0098673  0.0110593
#> COLCAP:COLCAP.lag(1)    0.0591146 0.06759904 0.00427588 -0.0690938  0.0148925
#> BOVESPA:COLCAP.lag(1)   0.1638565 0.09752149 0.00230941 -0.0438568  0.1011873
#> COLCAP:BOVESPA.lag(1)   0.0515688 0.05986209 0.01553129 -0.0876924  0.0193007
#> BOVESPA:BOVESPA.lag(1) -0.1367765 0.08341116 0.01353941 -0.3159181 -0.1891504
#> COLCAP:COLCAP.lag(2)    0.0612289 0.05865579 0.00140913 -0.0541204  0.0221912
#> BOVESPA:COLCAP.lag(2)  -0.0717404 0.09272151 0.00222727 -0.2663071 -0.1310096
#> COLCAP:BOVESPA.lag(2)  -0.0695398 0.04063881 0.00161832 -0.1475786 -0.0956989
#> BOVESPA:BOVESPA.lag(2) -0.0491019 0.06500672 0.00198179 -0.1748092 -0.0911222
#>                               50%        75%     97.5%
#> COLCAP:(Intercept)      0.0050619  0.0056379 0.0076471
#> BOVESPA:(Intercept)     0.0117787  0.0127490 0.0167875
#> COLCAP:COLCAP.lag(1)    0.0550996  0.1022533 0.1940790
#> BOVESPA:COLCAP.lag(1)   0.1669419  0.2273925 0.3514258
#> COLCAP:BOVESPA.lag(1)   0.0599761  0.0919152 0.1521219
#> BOVESPA:BOVESPA.lag(1) -0.1300983 -0.0788036 0.0122758
#> COLCAP:COLCAP.lag(2)    0.0619907  0.1001013 0.1754402
#> BOVESPA:COLCAP.lag(2)  -0.0715174 -0.0081490 0.1089878
#> COLCAP:BOVESPA.lag(2)  -0.0699931 -0.0445970 0.0147283
#> BOVESPA:BOVESPA.lag(2) -0.0500568 -0.0061299 0.0779731
#> 
#> 
#> Scale parameter:
#>                       Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP.COLCAP   5.6123e-05 7.3089e-06 3.6417e-07 4.3544e-05 5.0933e-05
#> COLCAP.BOVESPA  2.5193e-05 7.3258e-06 3.8538e-07 1.0968e-05 2.0491e-05
#> BOVESPA.BOVESPA 1.3475e-04 1.6273e-05 5.6930e-07 1.0588e-04 1.2339e-04
#>                        50%        75%      97.5%
#> COLCAP.COLCAP   5.5506e-05 6.0458e-05 7.2601e-05
#> COLCAP.BOVESPA  2.5021e-05 2.9907e-05 3.9813e-05
#> BOVESPA.BOVESPA 1.3345e-04 1.4516e-04 1.6820e-04
#> 
#> 
#> Extra parameter:
#>      Mean      Sd Sd(Mean)   2.5%    25%    50%    75%  97.5%
#> nu 5.9504 0.75148 0.063583 4.7195 5.4349 5.8563 6.3757 7.7183
#plot(fit1.mcmc)

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, row.names=Date,
             subset={Date<="2009-02-13"}, dist="Laplace",
             ars=ars(nregim=3,p=5), n.burnin=1000, n.sim=2000, n.thin=2)
fit2.mcmc <- coda::as.mcmc(fit2)
summary(fit2.mcmc)
#> 
#> 
#>  Iterations = 1001:4999
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 2000
#> 
#> 
#> Thresholds:
#>                Mean       Sd Sd(Mean)    2.5%     25%     50%     75%   97.5%
#> Threshold.1  3.5169 0.275950 0.078702  3.0049  3.3456  3.6052  3.7128  3.9961
#> Threshold.2 10.0174 0.040726 0.007638 10.0006 10.0057 10.0096 10.0130 10.1661
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean        Sd   Sd(Mean)      2.5%        25%
#> Bedon:(Intercept)       1.3193505 0.1127258 0.00342768  1.091589  1.2453818
#> LaPlata:(Intercept)     3.4322447 0.3094529 0.00935441  2.824335  3.2390263
#> Bedon:Bedon.lag(1)      0.5654524 0.0392776 0.00148966  0.487752  0.5381575
#> LaPlata:Bedon.lag(1)    0.1525932 0.1071023 0.00349469 -0.063273  0.0810367
#> Bedon:LaPlata.lag(1)    0.0464725 0.0143816 0.00046756  0.016894  0.0368451
#> LaPlata:LaPlata.lag(1)  0.6352614 0.0408356 0.00132547  0.557048  0.6067606
#> Bedon:Bedon.lag(2)      0.0496703 0.0352578 0.00123998 -0.018793  0.0263078
#> LaPlata:Bedon.lag(2)   -0.0491709 0.0942782 0.00311613 -0.229293 -0.1157552
#> Bedon:LaPlata.lag(2)   -0.0208420 0.0123129 0.00046460 -0.045180 -0.0289578
#> LaPlata:LaPlata.lag(2) -0.0707247 0.0336987 0.00103659 -0.137925 -0.0940502
#> Bedon:Bedon.lag(3)      0.0258304 0.0301477 0.00093152 -0.032381  0.0052270
#> LaPlata:Bedon.lag(3)    0.0250960 0.0765949 0.00230602 -0.128934 -0.0252818
#> Bedon:LaPlata.lag(3)    0.0038293 0.0107795 0.00036739 -0.017589 -0.0035446
#> LaPlata:LaPlata.lag(3)  0.0666464 0.0266457 0.00082592  0.012597  0.0498963
#> Bedon:Bedon.lag(4)      0.0342831 0.0317293 0.00106109 -0.023746  0.0118680
#> LaPlata:Bedon.lag(4)   -0.0925667 0.0832175 0.00271606 -0.260258 -0.1471590
#> Bedon:LaPlata.lag(4)   -0.0152013 0.0086460 0.00024766 -0.032029 -0.0210999
#> LaPlata:LaPlata.lag(4)  0.0107927 0.0268718 0.00091630 -0.041990 -0.0081599
#> Bedon:Bedon.lag(5)      0.0829414 0.0272569 0.00104854  0.028109  0.0645736
#> LaPlata:Bedon.lag(5)    0.1510782 0.0647932 0.00224619  0.025501  0.1086216
#> Bedon:LaPlata.lag(5)   -0.0069683 0.0068379 0.00020462 -0.021130 -0.0113918
#> LaPlata:LaPlata.lag(5)  0.0235731 0.0225161 0.00081583 -0.020216  0.0071216
#>                               50%        75%      97.5%
#> Bedon:(Intercept)       1.3248057  1.3941986  1.5299681
#> LaPlata:(Intercept)     3.4323680  3.6178742  4.0789336
#> Bedon:Bedon.lag(1)      0.5653634  0.5931639  0.6407875
#> LaPlata:Bedon.lag(1)    0.1521169  0.2254229  0.3600735
#> Bedon:LaPlata.lag(1)    0.0466488  0.0564812  0.0734752
#> LaPlata:LaPlata.lag(1)  0.6348666  0.6623958  0.7161397
#> Bedon:Bedon.lag(2)      0.0504077  0.0724835  0.1209331
#> LaPlata:Bedon.lag(2)   -0.0529062  0.0118278  0.1450655
#> Bedon:LaPlata.lag(2)   -0.0208870 -0.0125562  0.0040785
#> LaPlata:LaPlata.lag(2) -0.0713336 -0.0474290 -0.0063825
#> Bedon:Bedon.lag(3)      0.0257802  0.0452802  0.0837502
#> LaPlata:Bedon.lag(3)    0.0280878  0.0743802  0.1773239
#> Bedon:LaPlata.lag(3)    0.0038075  0.0116295  0.0235991
#> LaPlata:LaPlata.lag(3)  0.0676308  0.0839746  0.1184909
#> Bedon:Bedon.lag(4)      0.0324143  0.0552927  0.1004718
#> LaPlata:Bedon.lag(4)   -0.0923528 -0.0358403  0.0675584
#> Bedon:LaPlata.lag(4)   -0.0153807 -0.0092597  0.0016551
#> LaPlata:LaPlata.lag(4)  0.0106618  0.0297189  0.0605281
#> Bedon:Bedon.lag(5)      0.0837074  0.1016159  0.1350252
#> LaPlata:Bedon.lag(5)    0.1499995  0.1930781  0.2824901
#> Bedon:LaPlata.lag(5)   -0.0069778 -0.0024039  0.0061734
#> LaPlata:LaPlata.lag(5)  0.0240982  0.0398701  0.0656260
#> 
#> 
#> Scale parameter:
#>                    Mean       Sd   Sd(Mean)    2.5%     25%     50%     75%
#> Bedon.Bedon     0.32948 0.032951 0.00094499 0.27225 0.30533 0.32766 0.35112
#> Bedon.LaPlata   0.37264 0.059529 0.00154449 0.25898 0.33065 0.37058 0.41153
#> LaPlata.LaPlata 2.34122 0.241069 0.00717522 1.90429 2.16696 2.33092 2.49763
#>                   97.5%
#> Bedon.Bedon     0.39781
#> Bedon.LaPlata   0.49513
#> LaPlata.LaPlata 2.84737
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean       Sd   Sd(Mean)       2.5%        25%
#> Bedon:(Intercept)       2.1248357 0.433812 0.01594273  1.3203677  1.8225357
#> LaPlata:(Intercept)     6.9721550 1.158641 0.04817659  4.7924220  6.1705467
#> Bedon:Bedon.lag(1)      0.5857975 0.046542 0.00156780  0.4914769  0.5553107
#> LaPlata:Bedon.lag(1)    0.1484477 0.121418 0.00433772 -0.0873430  0.0661141
#> Bedon:LaPlata.lag(1)    0.0215759 0.013997 0.00043336 -0.0062756  0.0122296
#> LaPlata:LaPlata.lag(1)  0.5246416 0.040380 0.00138053  0.4449998  0.4975889
#> Bedon:Bedon.lag(2)      0.0968346 0.062458 0.00222859 -0.0281986  0.0553548
#> LaPlata:Bedon.lag(2)   -0.0050875 0.127543 0.00397489 -0.2531959 -0.0884551
#> Bedon:LaPlata.lag(2)   -0.0197639 0.017655 0.00070874 -0.0550329 -0.0322484
#> LaPlata:LaPlata.lag(2)  0.0346105 0.038116 0.00134852 -0.0391805  0.0086679
#> Bedon:Bedon.lag(3)     -0.0355025 0.055798 0.00200557 -0.1438275 -0.0736019
#> LaPlata:Bedon.lag(3)   -0.0686076 0.115490 0.00344201 -0.3005243 -0.1446480
#> Bedon:LaPlata.lag(3)   -0.0075018 0.015102 0.00055778 -0.0365888 -0.0183420
#> LaPlata:LaPlata.lag(3)  0.0432505 0.034932 0.00115495 -0.0222825  0.0184738
#> Bedon:Bedon.lag(4)      0.1007043 0.061447 0.00231018 -0.0314489  0.0618790
#> LaPlata:Bedon.lag(4)    0.2250122 0.134024 0.00481510 -0.0469479  0.1339963
#> Bedon:LaPlata.lag(4)    0.0078918 0.016938 0.00057483 -0.0253919 -0.0040928
#> LaPlata:LaPlata.lag(4) -0.0445658 0.042874 0.00156655 -0.1262748 -0.0731985
#> Bedon:Bedon.lag(5)      0.0286323 0.044379 0.00145013 -0.0574487 -0.0015386
#> LaPlata:Bedon.lag(5)   -0.2660897 0.101892 0.00324029 -0.4719620 -0.3317998
#> Bedon:LaPlata.lag(5)    0.0043741 0.015095 0.00047785 -0.0247761 -0.0058911
#> LaPlata:LaPlata.lag(5)  0.1189203 0.039701 0.00132341  0.0446283  0.0923020
#>                               50%        75%     97.5%
#> Bedon:(Intercept)       2.1126306  2.4112586  3.013849
#> LaPlata:(Intercept)     6.9264405  7.7959945  9.252277
#> Bedon:Bedon.lag(1)      0.5863151  0.6172421  0.673891
#> LaPlata:Bedon.lag(1)    0.1478055  0.2299792  0.383866
#> Bedon:LaPlata.lag(1)    0.0212062  0.0304706  0.049908
#> LaPlata:LaPlata.lag(1)  0.5263025  0.5495580  0.605835
#> Bedon:Bedon.lag(2)      0.0962872  0.1397555  0.218304
#> LaPlata:Bedon.lag(2)   -0.0053993  0.0807182  0.245496
#> Bedon:LaPlata.lag(2)   -0.0195516 -0.0073651  0.014943
#> LaPlata:LaPlata.lag(2)  0.0341961  0.0604619  0.110174
#> Bedon:Bedon.lag(3)     -0.0359905  0.0028343  0.070898
#> LaPlata:Bedon.lag(3)   -0.0657353  0.0099675  0.157753
#> Bedon:LaPlata.lag(3)   -0.0072439  0.0029430  0.022706
#> LaPlata:LaPlata.lag(3)  0.0435174  0.0678853  0.108091
#> Bedon:Bedon.lag(4)      0.1057199  0.1429476  0.213172
#> LaPlata:Bedon.lag(4)    0.2255667  0.3175097  0.474039
#> Bedon:LaPlata.lag(4)    0.0080264  0.0196265  0.040411
#> LaPlata:LaPlata.lag(4) -0.0439764 -0.0152325  0.038360
#> Bedon:Bedon.lag(5)      0.0283911  0.0581242  0.119007
#> LaPlata:Bedon.lag(5)   -0.2668852 -0.2001445 -0.061854
#> Bedon:LaPlata.lag(5)    0.0044585  0.0143826  0.033627
#> LaPlata:LaPlata.lag(5)  0.1176350  0.1463065  0.198984
#> 
#> 
#> Scale parameter:
#>                   Mean      Sd  Sd(Mean)    2.5%    25%    50%    75%  97.5%
#> Bedon.Bedon     1.0917 0.10989 0.0031224 0.90046 1.0161 1.0854 1.1591 1.3224
#> Bedon.LaPlata   1.3378 0.18953 0.0050169 0.99617 1.2121 1.3259 1.4472 1.7611
#> LaPlata.LaPlata 6.5653 0.65153 0.0187734 5.43950 6.1262 6.5100 6.9947 7.9488
#> 
#> 
#> Regime 3
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean       Sd   Sd(Mean)        2.5%       25%
#> Bedon:(Intercept)       5.5830454 0.824365 0.02906227  4.07666477  5.042027
#> LaPlata:(Intercept)    16.9596700 2.923085 0.09915233 11.23090115 15.019441
#> Bedon:Bedon.lag(1)      0.4692892 0.085077 0.00349707  0.30111379  0.409509
#> LaPlata:Bedon.lag(1)    0.5534638 0.273717 0.00874211  0.01595357  0.362196
#> Bedon:LaPlata.lag(1)    0.0436224 0.017583 0.00063955  0.00940404  0.031510
#> LaPlata:LaPlata.lag(1)  0.3278131 0.065969 0.00213724  0.20291718  0.282661
#> Bedon:Bedon.lag(2)      0.0821862 0.076510 0.00260224 -0.06788039  0.028246
#> LaPlata:Bedon.lag(2)   -0.5787238 0.283517 0.00934453 -1.13021658 -0.756440
#> Bedon:LaPlata.lag(2)   -0.0023041 0.017222 0.00051380 -0.03537751 -0.014163
#> LaPlata:LaPlata.lag(2)  0.1196095 0.067657 0.00179833 -0.00093319  0.072329
#> Bedon:Bedon.lag(3)     -0.0907051 0.070148 0.00231994 -0.22911261 -0.133972
#> LaPlata:Bedon.lag(3)   -0.5896026 0.242658 0.00679066 -1.06130349 -0.751876
#> Bedon:LaPlata.lag(3)    0.0332560 0.019923 0.00064264 -0.00810273  0.020382
#> LaPlata:LaPlata.lag(3)  0.2829787 0.081023 0.00278231  0.12810257  0.227992
#> Bedon:Bedon.lag(4)      0.0032741 0.077181 0.00241851 -0.15665447 -0.046440
#> LaPlata:Bedon.lag(4)    0.0491419 0.292325 0.00843977 -0.48079563 -0.154549
#> Bedon:LaPlata.lag(4)    0.0048796 0.021392 0.00071762 -0.03485629 -0.010244
#> LaPlata:LaPlata.lag(4) -0.0067282 0.079384 0.00251477 -0.15815998 -0.060438
#> Bedon:Bedon.lag(5)      0.1795482 0.070612 0.00244786  0.04380886  0.130663
#> LaPlata:Bedon.lag(5)    0.2871194 0.252720 0.00829308 -0.21277982  0.116953
#> Bedon:LaPlata.lag(5)   -0.0144314 0.018154 0.00056473 -0.05004996 -0.026415
#> LaPlata:LaPlata.lag(5)  0.0625642 0.067793 0.00203951 -0.06776611  0.015713
#>                               50%        75%     97.5%
#> Bedon:(Intercept)       5.5542254  6.1088358  7.265874
#> LaPlata:(Intercept)    16.8891221 18.8314828 22.959471
#> Bedon:Bedon.lag(1)      0.4692433  0.5305947  0.638091
#> LaPlata:Bedon.lag(1)    0.5602344  0.7474092  1.059001
#> Bedon:LaPlata.lag(1)    0.0432053  0.0553705  0.078159
#> LaPlata:LaPlata.lag(1)  0.3258042  0.3716032  0.465209
#> Bedon:Bedon.lag(2)      0.0818624  0.1342441  0.230435
#> LaPlata:Bedon.lag(2)   -0.5807391 -0.3906873 -0.024407
#> Bedon:LaPlata.lag(2)   -0.0026046  0.0089355  0.033261
#> LaPlata:LaPlata.lag(2)  0.1184670  0.1609828  0.258237
#> Bedon:Bedon.lag(3)     -0.0930971 -0.0438894  0.048658
#> LaPlata:Bedon.lag(3)   -0.5977694 -0.4349132 -0.090942
#> Bedon:LaPlata.lag(3)    0.0336576  0.0466224  0.070432
#> LaPlata:LaPlata.lag(3)  0.2825212  0.3386578  0.441414
#> Bedon:Bedon.lag(4)      0.0035138  0.0574571  0.148541
#> LaPlata:Bedon.lag(4)    0.0340193  0.2413648  0.656148
#> Bedon:LaPlata.lag(4)    0.0040957  0.0193318  0.047315
#> LaPlata:LaPlata.lag(4) -0.0077711  0.0471909  0.153957
#> Bedon:Bedon.lag(5)      0.1800547  0.2263282  0.320344
#> LaPlata:Bedon.lag(5)    0.2855695  0.4639078  0.751301
#> Bedon:LaPlata.lag(5)   -0.0148944 -0.0024857  0.021949
#> LaPlata:LaPlata.lag(5)  0.0615175  0.1080645  0.197106
#> 
#> 
#> Scale parameter:
#>                    Mean      Sd  Sd(Mean)    2.5%     25%     50%     75%
#> Bedon.Bedon      2.7942 0.28656 0.0081903  2.2867  2.5910  2.7762  2.9791
#> Bedon.LaPlata    7.2195 0.91846 0.0251570  5.5873  6.5893  7.1497  7.8045
#> LaPlata.LaPlata 43.2357 4.47214 0.1334651 35.1717 40.0455 42.9166 46.0941
#>                  97.5%
#> Bedon.Bedon      3.399
#> Bedon.LaPlata    9.176
#> LaPlata.LaPlata 53.121
#plot(fit2.mcmc)

###### Example 3: Temperature, precipitation, and two river flows in Iceland
data(iceland.rf)
fit3 <- mtar(~ Jokulsa + Vatnsdalsa | Temperature | Precipitation,
             data=iceland.rf, subset={Date<="1974-11-06"}, row.names=Date,
             ars=ars(nregim=2,p=15,q=4,d=2), n.burnin=1000, n.sim=2000,
             n.thin=2, dist="Slash")
fit3.mcmc <- coda::as.mcmc(fit3)
summary(fit3.mcmc)
#> 
#> 
#>  Iterations = 1001:4999
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 2000
#> 
#> 
#> Thresholds:
#>               Mean       Sd  Sd(Mean)   2.5%   25%    50%    75%  97.5%
#> Threshold.1 1.1419 0.040504 0.0031006 1.0262 1.119 1.1442 1.1752 1.1972
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                                        Mean        Sd   Sd(Mean)        2.5%
#> Jokulsa:(Intercept)              3.65970327 0.3685327 0.01843645  3.02013667
#> Vatnsdalsa:(Intercept)           0.81635571 0.2136963 0.00943817  0.41535895
#> Jokulsa:Jokulsa.lag( 1)          0.85076239 0.0383996 0.00200603  0.76100722
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.06248223 0.0214721 0.00098903 -0.11060679
#> Jokulsa:Vatnsdalsa.lag( 1)       0.20167814 0.0584023 0.00215077  0.09096903
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.16125945 0.0398512 0.00166427  1.08562479
#> Jokulsa:Jokulsa.lag( 2)         -0.05175005 0.0240827 0.00087061 -0.09613883
#> Vatnsdalsa:Jokulsa.lag( 2)       0.04938036 0.0163876 0.00063135  0.01755511
#> Jokulsa:Vatnsdalsa.lag( 2)      -0.16638805 0.0659073 0.00281869 -0.31408926
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.29138303 0.0433337 0.00202204 -0.38537402
#> Jokulsa:Jokulsa.lag( 3)          0.00452993 0.0189912 0.00069504 -0.03318693
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.02423354 0.0121120 0.00037222 -0.04782829
#> Jokulsa:Vatnsdalsa.lag( 3)       0.03233082 0.0336965 0.00114127 -0.02983403
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    0.02802005 0.0244679 0.00086333 -0.01498165
#> Jokulsa:Jokulsa.lag( 4)          0.00054753 0.0241686 0.00098672 -0.04371830
#> Vatnsdalsa:Jokulsa.lag( 4)       0.01478949 0.0131981 0.00043210 -0.01148967
#> Jokulsa:Vatnsdalsa.lag( 4)       0.00886005 0.0391511 0.00152350 -0.06776966
#> Vatnsdalsa:Vatnsdalsa.lag( 4)    0.00180707 0.0278828 0.00102155 -0.05731536
#> Jokulsa:Jokulsa.lag( 5)          0.00032386 0.0282466 0.00132360 -0.05892373
#> Vatnsdalsa:Jokulsa.lag( 5)       0.00505470 0.0149772 0.00046329 -0.02294182
#> Jokulsa:Vatnsdalsa.lag( 5)      -0.03618646 0.0322306 0.00103663 -0.09875869
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -0.01531263 0.0225617 0.00067044 -0.06230404
#> Jokulsa:Jokulsa.lag( 6)          0.02383594 0.0268975 0.00129346 -0.03096304
#> Vatnsdalsa:Jokulsa.lag( 6)       0.00580623 0.0146450 0.00059377 -0.02193416
#> Jokulsa:Vatnsdalsa.lag( 6)      -0.02362550 0.0310742 0.00088857 -0.08489459
#> Vatnsdalsa:Vatnsdalsa.lag( 6)   -0.00102337 0.0199383 0.00058411 -0.04020115
#> Jokulsa:Jokulsa.lag( 7)         -0.00131642 0.0257554 0.00095790 -0.05269526
#> Vatnsdalsa:Jokulsa.lag( 7)      -0.00248750 0.0151322 0.00052765 -0.03295646
#> Jokulsa:Vatnsdalsa.lag( 7)       0.01699562 0.0300519 0.00099759 -0.04167217
#> Vatnsdalsa:Vatnsdalsa.lag( 7)    0.00862158 0.0184969 0.00046878 -0.02821050
#> Jokulsa:Jokulsa.lag( 8)         -0.00060472 0.0256137 0.00109432 -0.05303024
#> Vatnsdalsa:Jokulsa.lag( 8)      -0.00936050 0.0134334 0.00039015 -0.03561358
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.00414907 0.0302320 0.00085666 -0.06861560
#> Vatnsdalsa:Vatnsdalsa.lag( 8)    0.00662861 0.0191633 0.00048462 -0.03202822
#> Jokulsa:Jokulsa.lag( 9)         -0.00782426 0.0234812 0.00099880 -0.05268124
#> Vatnsdalsa:Jokulsa.lag( 9)       0.01988174 0.0153061 0.00065805 -0.00933481
#> Jokulsa:Vatnsdalsa.lag( 9)      -0.00940073 0.0352332 0.00125736 -0.07032987
#> Vatnsdalsa:Vatnsdalsa.lag( 9)   -0.00160190 0.0210576 0.00063231 -0.04643243
#> Jokulsa:Jokulsa.lag(10)          0.02685252 0.0181540 0.00064076 -0.01174231
#> Vatnsdalsa:Jokulsa.lag(10)      -0.01413663 0.0118946 0.00046436 -0.03741677
#> Jokulsa:Vatnsdalsa.lag(10)       0.02441274 0.0316305 0.00095363 -0.03646064
#> Vatnsdalsa:Vatnsdalsa.lag(10)    0.01938332 0.0221634 0.00074138 -0.01744038
#> Jokulsa:Jokulsa.lag(11)         -0.01488697 0.0147225 0.00046629 -0.04428974
#> Vatnsdalsa:Jokulsa.lag(11)       0.00927270 0.0095233 0.00025044 -0.00900358
#> Jokulsa:Vatnsdalsa.lag(11)      -0.01623645 0.0267783 0.00074449 -0.06903284
#> Vatnsdalsa:Vatnsdalsa.lag(11)   -0.00847983 0.0190343 0.00059103 -0.04439180
#> Jokulsa:Jokulsa.lag(12)          0.00966869 0.0147481 0.00045403 -0.02037815
#> Vatnsdalsa:Jokulsa.lag(12)      -0.00877958 0.0099189 0.00026348 -0.02831679
#> Jokulsa:Vatnsdalsa.lag(12)       0.00910919 0.0249039 0.00055687 -0.03992355
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.00180034 0.0177812 0.00045662 -0.03718982
#> Jokulsa:Jokulsa.lag(13)         -0.01752059 0.0173781 0.00075403 -0.05172978
#> Vatnsdalsa:Jokulsa.lag(13)       0.00280654 0.0104714 0.00030969 -0.01744440
#> Jokulsa:Vatnsdalsa.lag(13)      -0.01431133 0.0268185 0.00066114 -0.06809299
#> Vatnsdalsa:Vatnsdalsa.lag(13)   -0.02034149 0.0203906 0.00067301 -0.06289639
#> Jokulsa:Jokulsa.lag(14)          0.00550140 0.0139884 0.00047824 -0.02184265
#> Vatnsdalsa:Jokulsa.lag(14)      -0.00476804 0.0093008 0.00027454 -0.02233069
#> Jokulsa:Vatnsdalsa.lag(14)       0.00548054 0.0274834 0.00085843 -0.04642736
#> Vatnsdalsa:Vatnsdalsa.lag(14)    0.03468555 0.0189027 0.00054542  0.00147735
#> Jokulsa:Jokulsa.lag(15)          0.02030438 0.0142358 0.00066201 -0.00677105
#> Vatnsdalsa:Jokulsa.lag(15)       0.00160019 0.0074165 0.00029414 -0.01296412
#> Jokulsa:Vatnsdalsa.lag(15)      -0.01236817 0.0201900 0.00061494 -0.05410096
#> Vatnsdalsa:Vatnsdalsa.lag(15)    0.00569622 0.0144500 0.00047214 -0.02297260
#> Jokulsa:Precipitation.lag(1)     0.00707155 0.0092733 0.00032717 -0.01058566
#> Vatnsdalsa:Precipitation.lag(1)  0.00507279 0.0055616 0.00016362 -0.00623816
#> Jokulsa:Precipitation.lag(2)     0.00610265 0.0068675 0.00018534 -0.00773482
#> Vatnsdalsa:Precipitation.lag(2)  0.00057339 0.0043238 0.00012375 -0.00765144
#> Jokulsa:Precipitation.lag(3)    -0.01184315 0.0055156 0.00016003 -0.02208719
#> Vatnsdalsa:Precipitation.lag(3) -0.00358979 0.0039729 0.00010562 -0.01166866
#> Jokulsa:Precipitation.lag(4)     0.01882990 0.0066018 0.00017990  0.00585795
#> Vatnsdalsa:Precipitation.lag(4)  0.00434721 0.0047687 0.00014828 -0.00526423
#> Jokulsa:Temperature.lag(1)       0.02192199 0.0107932 0.00035133  0.00094334
#> Vatnsdalsa:Temperature.lag(1)    0.00192107 0.0068839 0.00022455 -0.01134931
#> Jokulsa:Temperature.lag(2)      -0.03841055 0.0108050 0.00036467 -0.05934805
#> Vatnsdalsa:Temperature.lag(2)   -0.01178138 0.0068497 0.00020765 -0.02477174
#>                                         25%         50%         75%       97.5%
#> Jokulsa:(Intercept)              3.39421014  3.63825208  3.88227496  4.46838352
#> Vatnsdalsa:(Intercept)           0.67130980  0.80909825  0.95429222  1.27849548
#> Jokulsa:Jokulsa.lag( 1)          0.82805348  0.85553531  0.88033138  0.90927505
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.07682018 -0.06015946 -0.04594494 -0.02845092
#> Jokulsa:Vatnsdalsa.lag( 1)       0.16018415  0.20006295  0.24030339  0.32114949
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.13340208  1.16114403  1.18795634  1.24237353
#> Jokulsa:Jokulsa.lag( 2)         -0.06889009 -0.05267400 -0.03487831 -0.00213243
#> Vatnsdalsa:Jokulsa.lag( 2)       0.03850063  0.04936986  0.06059690  0.08078985
#> Jokulsa:Vatnsdalsa.lag( 2)      -0.20608226 -0.15863705 -0.12052487 -0.05376382
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.31837704 -0.28675434 -0.26149357 -0.21800983
#> Jokulsa:Jokulsa.lag( 3)         -0.00751184  0.00502058  0.01755418  0.04101720
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.03260491 -0.02438240 -0.01624486  0.00067676
#> Jokulsa:Vatnsdalsa.lag( 3)       0.01022403  0.03138347  0.05252608  0.10129611
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    0.01154510  0.02707386  0.04288956  0.08074326
#> Jokulsa:Jokulsa.lag( 4)         -0.01529162 -0.00080136  0.01617930  0.04954630
#> Vatnsdalsa:Jokulsa.lag( 4)       0.00597310  0.01508556  0.02399173  0.03996582
#> Jokulsa:Vatnsdalsa.lag( 4)      -0.01740525  0.01065497  0.03456296  0.08398309
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -0.01576242  0.00316781  0.02025114  0.05387090
#> Jokulsa:Jokulsa.lag( 5)         -0.01817741  0.00165590  0.02056222  0.05038365
#> Vatnsdalsa:Jokulsa.lag( 5)      -0.00552994  0.00489255  0.01492402  0.03488434
#> Jokulsa:Vatnsdalsa.lag( 5)      -0.05686960 -0.03648871 -0.01428431  0.02493251
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -0.03009417 -0.01524295 -0.00026647  0.02763718
#> Jokulsa:Jokulsa.lag( 6)          0.00612065  0.02492963  0.04234628  0.07438845
#> Vatnsdalsa:Jokulsa.lag( 6)      -0.00402200  0.00611671  0.01549765  0.03385233
#> Jokulsa:Vatnsdalsa.lag( 6)      -0.04364990 -0.02369234 -0.00337761  0.03608127
#> Vatnsdalsa:Vatnsdalsa.lag( 6)   -0.01412121 -0.00090867  0.01231164  0.03654395
#> Jokulsa:Jokulsa.lag( 7)         -0.01914214 -0.00232780  0.01584525  0.04936255
#> Vatnsdalsa:Jokulsa.lag( 7)      -0.01279096 -0.00245413  0.00778217  0.02679468
#> Jokulsa:Vatnsdalsa.lag( 7)      -0.00266540  0.01738421  0.03731466  0.07535361
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -0.00312987  0.00794622  0.02027544  0.04554144
#> Jokulsa:Jokulsa.lag( 8)         -0.01701342  0.00042033  0.01777709  0.04666896
#> Vatnsdalsa:Jokulsa.lag( 8)      -0.01860493 -0.00939764 -0.00039686  0.01678914
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.02325020 -0.00293270  0.01587965  0.05212455
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -0.00563436  0.00639618  0.01871844  0.04402768
#> Jokulsa:Jokulsa.lag( 9)         -0.02369230 -0.00851603  0.00877385  0.03697713
#> Vatnsdalsa:Jokulsa.lag( 9)       0.00959755  0.01960036  0.03004107  0.05066946
#> Jokulsa:Vatnsdalsa.lag( 9)      -0.03416156 -0.01286309  0.01201875  0.06800673
#> Vatnsdalsa:Vatnsdalsa.lag( 9)   -0.01483045 -0.00077236  0.01258637  0.03752767
#> Jokulsa:Jokulsa.lag(10)          0.01571294  0.02760138  0.03867577  0.06098396
#> Vatnsdalsa:Jokulsa.lag(10)      -0.02187684 -0.01408014 -0.00599184  0.00771067
#> Jokulsa:Vatnsdalsa.lag(10)       0.00419637  0.02441398  0.04414059  0.08838577
#> Vatnsdalsa:Vatnsdalsa.lag(10)    0.00338994  0.01714036  0.03382580  0.06706840
#> Jokulsa:Jokulsa.lag(11)         -0.02429236 -0.01486997 -0.00502416  0.01415052
#> Vatnsdalsa:Jokulsa.lag(11)       0.00290933  0.00906451  0.01587872  0.02779064
#> Jokulsa:Vatnsdalsa.lag(11)      -0.03379595 -0.01566798  0.00131931  0.03470410
#> Vatnsdalsa:Vatnsdalsa.lag(11)   -0.02155907 -0.00858224  0.00361941  0.03102663
#> Jokulsa:Jokulsa.lag(12)         -0.00011176  0.00951557  0.01970642  0.03741619
#> Vatnsdalsa:Jokulsa.lag(12)      -0.01533233 -0.00882205 -0.00233619  0.01045810
#> Jokulsa:Vatnsdalsa.lag(12)      -0.00701196  0.00882073  0.02532889  0.05813379
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.01414883 -0.00148197  0.01085450  0.03220172
#> Jokulsa:Jokulsa.lag(13)         -0.02975541 -0.01769960 -0.00520727  0.01699433
#> Vatnsdalsa:Jokulsa.lag(13)      -0.00404625  0.00252311  0.00909304  0.02511390
#> Jokulsa:Vatnsdalsa.lag(13)      -0.03148409 -0.01388268  0.00320738  0.03720466
#> Vatnsdalsa:Vatnsdalsa.lag(13)   -0.03299980 -0.01916063 -0.00658872  0.01669792
#> Jokulsa:Jokulsa.lag(14)         -0.00375577  0.00583804  0.01472294  0.03251486
#> Vatnsdalsa:Jokulsa.lag(14)      -0.01082666 -0.00501068  0.00127420  0.01445377
#> Jokulsa:Vatnsdalsa.lag(14)      -0.01247061  0.00476369  0.02290109  0.06281577
#> Vatnsdalsa:Vatnsdalsa.lag(14)    0.02172434  0.03310344  0.04631760  0.07591186
#> Jokulsa:Jokulsa.lag(15)          0.01017920  0.02043276  0.02984539  0.04876827
#> Vatnsdalsa:Jokulsa.lag(15)      -0.00339784  0.00169217  0.00662763  0.01634978
#> Jokulsa:Vatnsdalsa.lag(15)      -0.02537263 -0.01175959  0.00089489  0.02595415
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -0.00384832  0.00570760  0.01487129  0.03468596
#> Jokulsa:Precipitation.lag(1)     0.00085359  0.00696410  0.01326309  0.02600197
#> Vatnsdalsa:Precipitation.lag(1)  0.00148398  0.00526960  0.00877977  0.01565190
#> Jokulsa:Precipitation.lag(2)     0.00155253  0.00636855  0.01072261  0.01892911
#> Vatnsdalsa:Precipitation.lag(2) -0.00226771  0.00044221  0.00333785  0.00904441
#> Jokulsa:Precipitation.lag(3)    -0.01567653 -0.01207901 -0.00823635 -0.00074934
#> Vatnsdalsa:Precipitation.lag(3) -0.00619775 -0.00355415 -0.00092241  0.00405725
#> Jokulsa:Precipitation.lag(4)     0.01432122  0.01895945  0.02312273  0.03163175
#> Vatnsdalsa:Precipitation.lag(4)  0.00117499  0.00443698  0.00753685  0.01326168
#> Jokulsa:Temperature.lag(1)       0.01493087  0.02145933  0.02904935  0.04361013
#> Vatnsdalsa:Temperature.lag(1)   -0.00281637  0.00188770  0.00631317  0.01574102
#> Jokulsa:Temperature.lag(2)      -0.04570998 -0.03827206 -0.03100159 -0.01794393
#> Vatnsdalsa:Temperature.lag(2)   -0.01641307 -0.01173078 -0.00682528  0.00152518
#> 
#> 
#> Scale parameter:
#>                           Mean        Sd   Sd(Mean)      2.5%       25%
#> Jokulsa.Jokulsa       0.063210 0.0093660 0.00068806 0.0460995 0.0565193
#> Jokulsa.Vatnsdalsa    0.010718 0.0027414 0.00012944 0.0057622 0.0088562
#> Vatnsdalsa.Vatnsdalsa 0.027782 0.0041867 0.00029634 0.0204039 0.0248023
#>                            50%      75%    97.5%
#> Jokulsa.Jokulsa       0.062700 0.069428 0.082292
#> Jokulsa.Vatnsdalsa    0.010520 0.012431 0.016740
#> Vatnsdalsa.Vatnsdalsa 0.027375 0.030579 0.036520
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                                        Mean        Sd   Sd(Mean)       2.5%
#> Jokulsa:(Intercept)             -0.27650304 0.6902683 2.4577e-02 -1.5895371
#> Vatnsdalsa:(Intercept)           0.48695242 0.0951766 3.2498e-03  0.2993410
#> Jokulsa:Jokulsa.lag( 1)          1.01359023 0.0362666 1.3023e-03  0.9464870
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.00284150 0.0036920 1.0013e-04 -0.0099035
#> Jokulsa:Vatnsdalsa.lag( 1)       0.91413419 0.2455212 1.1080e-02  0.4598712
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.18565629 0.0356777 1.5947e-03  1.1191950
#> Jokulsa:Jokulsa.lag( 2)         -0.16802519 0.0673327 3.0845e-03 -0.3065176
#> Vatnsdalsa:Jokulsa.lag( 2)       0.00991045 0.0062727 2.0500e-04 -0.0021910
#> Jokulsa:Vatnsdalsa.lag( 2)      -0.39115855 0.3660734 1.7255e-02 -1.0888052
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.34380128 0.0529917 2.6432e-03 -0.4437385
#> Jokulsa:Jokulsa.lag( 3)          0.00867313 0.0567463 2.3495e-03 -0.0985986
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.01208427 0.0063602 2.8579e-04 -0.0243201
#> Jokulsa:Vatnsdalsa.lag( 3)       0.03147738 0.3455714 1.6847e-02 -0.7051387
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    0.18969252 0.0470677 2.1099e-03  0.0865238
#> Jokulsa:Jokulsa.lag( 4)         -0.07334962 0.0413224 1.3823e-03 -0.1592459
#> Vatnsdalsa:Jokulsa.lag( 4)       0.00602328 0.0047029 1.4154e-04 -0.0036774
#> Jokulsa:Vatnsdalsa.lag( 4)      -0.16214860 0.2205611 6.6369e-03 -0.5786047
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -0.08943569 0.0424512 2.1447e-03 -0.1596081
#> Jokulsa:Jokulsa.lag( 5)          0.03857330 0.0402814 1.6408e-03 -0.0368181
#> Vatnsdalsa:Jokulsa.lag( 5)      -0.00514256 0.0046898 1.8556e-04 -0.0140709
#> Jokulsa:Vatnsdalsa.lag( 5)       0.00057357 0.2997092 1.1861e-02 -0.6344070
#> Vatnsdalsa:Vatnsdalsa.lag( 5)    0.01226254 0.0502266 2.6848e-03 -0.0854721
#> Jokulsa:Jokulsa.lag( 6)         -0.04044167 0.0338560 1.3526e-03 -0.1092773
#> Vatnsdalsa:Jokulsa.lag( 6)       0.00408966 0.0041770 1.5068e-04 -0.0038591
#> Jokulsa:Vatnsdalsa.lag( 6)       0.06390337 0.3003472 1.1161e-02 -0.4943770
#> Vatnsdalsa:Vatnsdalsa.lag( 6)    0.02098000 0.0508198 3.1490e-03 -0.0784843
#> Jokulsa:Jokulsa.lag( 7)          0.00066603 0.0301272 8.6515e-04 -0.0571213
#> Vatnsdalsa:Jokulsa.lag( 7)      -0.00579756 0.0040137 1.1360e-04 -0.0137843
#> Jokulsa:Vatnsdalsa.lag( 7)       0.09929750 0.2370051 8.2346e-03 -0.3604246
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -0.05588032 0.0336821 1.5134e-03 -0.1170104
#> Jokulsa:Jokulsa.lag( 8)          0.01578162 0.0297254 1.0455e-03 -0.0430993
#> Vatnsdalsa:Jokulsa.lag( 8)       0.00422085 0.0037493 9.8398e-05 -0.0030605
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.21765000 0.2307434 9.8145e-03 -0.6819812
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -0.04473480 0.0317800 1.2118e-03 -0.1090791
#> Jokulsa:Jokulsa.lag( 9)          0.03959881 0.0321918 9.7583e-04 -0.0261328
#> Vatnsdalsa:Jokulsa.lag( 9)      -0.00198340 0.0041877 1.2121e-04 -0.0103337
#> Jokulsa:Vatnsdalsa.lag( 9)       0.14275367 0.2256002 8.6910e-03 -0.2993979
#> Vatnsdalsa:Vatnsdalsa.lag( 9)    0.09013145 0.0366182 1.7749e-03  0.0142147
#> Jokulsa:Jokulsa.lag(10)         -0.02060622 0.0412269 1.3916e-03 -0.0925238
#> Vatnsdalsa:Jokulsa.lag(10)       0.00351031 0.0049034 1.5312e-04 -0.0055721
#> Jokulsa:Vatnsdalsa.lag(10)       0.00735458 0.2026042 9.5361e-03 -0.3999714
#> Vatnsdalsa:Vatnsdalsa.lag(10)   -0.07331538 0.0303883 1.6240e-03 -0.1275909
#> Jokulsa:Jokulsa.lag(11)         -0.00642610 0.0391801 1.2208e-03 -0.0849767
#> Vatnsdalsa:Jokulsa.lag(11)      -0.00755268 0.0046562 1.3652e-04 -0.0168367
#> Jokulsa:Vatnsdalsa.lag(11)      -0.02961008 0.2592531 1.1394e-02 -0.5281137
#> Vatnsdalsa:Vatnsdalsa.lag(11)    0.08334799 0.0356181 1.6225e-03  0.0083751
#> Jokulsa:Jokulsa.lag(12)         -0.00254781 0.0374224 1.1456e-03 -0.0785282
#> Vatnsdalsa:Jokulsa.lag(12)       0.00963627 0.0044395 1.2482e-04  0.0012621
#> Jokulsa:Vatnsdalsa.lag(12)       0.01952914 0.2434052 1.0531e-02 -0.4621226
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.08312785 0.0298128 1.2256e-03 -0.1385296
#> Jokulsa:Jokulsa.lag(13)         -0.00321743 0.0415389 1.7069e-03 -0.0797270
#> Vatnsdalsa:Jokulsa.lag(13)      -0.00628240 0.0042057 1.1210e-04 -0.0145516
#> Jokulsa:Vatnsdalsa.lag(13)       0.40961896 0.2607744 1.2219e-02 -0.1333710
#> Vatnsdalsa:Vatnsdalsa.lag(13)    0.14859278 0.0357562 1.5596e-03  0.0810085
#> Jokulsa:Jokulsa.lag(14)         -0.00502626 0.0413345 1.7734e-03 -0.0900907
#> Vatnsdalsa:Jokulsa.lag(14)      -0.00188549 0.0043715 1.3634e-04 -0.0105103
#> Jokulsa:Vatnsdalsa.lag(14)       0.13146938 0.2840282 1.3483e-02 -0.4458599
#> Vatnsdalsa:Vatnsdalsa.lag(14)   -0.04892762 0.0407836 1.9402e-03 -0.1312982
#> Jokulsa:Jokulsa.lag(15)          0.04654496 0.0231349 7.7720e-04  0.0022002
#> Vatnsdalsa:Jokulsa.lag(15)       0.00121854 0.0029373 9.3522e-05 -0.0043902
#> Jokulsa:Vatnsdalsa.lag(15)      -0.43689889 0.1674265 6.9591e-03 -0.7426968
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -0.02117698 0.0240782 1.1289e-03 -0.0660075
#> Jokulsa:Precipitation.lag(1)    -0.11473365 0.0393979 1.0677e-03 -0.1932005
#> Vatnsdalsa:Precipitation.lag(1) -0.00304512 0.0050793 1.4238e-04 -0.0128008
#> Jokulsa:Precipitation.lag(2)     0.03393762 0.0665266 2.5634e-03 -0.0967797
#> Vatnsdalsa:Precipitation.lag(2) -0.00221211 0.0071104 2.7760e-04 -0.0159124
#> Jokulsa:Precipitation.lag(3)     0.04993820 0.0327443 9.1259e-04 -0.0134584
#> Vatnsdalsa:Precipitation.lag(3)  0.00577239 0.0051483 1.9265e-04 -0.0039944
#> Jokulsa:Precipitation.lag(4)     0.02836750 0.0341169 1.0655e-03 -0.0384604
#> Vatnsdalsa:Precipitation.lag(4)  0.00281257 0.0043504 1.5392e-04 -0.0061918
#> Jokulsa:Temperature.lag(1)       1.12472301 0.0966009 2.6527e-03  0.9356289
#> Vatnsdalsa:Temperature.lag(1)    0.02007816 0.0127318 3.9958e-04 -0.0042711
#> Jokulsa:Temperature.lag(2)      -0.56778234 0.1041674 3.4089e-03 -0.7686833
#> Vatnsdalsa:Temperature.lag(2)   -0.02454704 0.0132406 4.3129e-04 -0.0496818
#>                                         25%         50%         75%       97.5%
#> Jokulsa:(Intercept)             -7.4370e-01 -0.26529751  0.19014861  1.05548381
#> Vatnsdalsa:(Intercept)           4.2481e-01  0.48974693  0.54855116  0.68383230
#> Jokulsa:Jokulsa.lag( 1)          9.9058e-01  1.01129165  1.03435025  1.08850144
#> Vatnsdalsa:Jokulsa.lag( 1)      -5.4407e-03 -0.00279260 -0.00034900  0.00437219
#> Jokulsa:Vatnsdalsa.lag( 1)       7.5009e-01  0.89910492  1.07461893  1.41052061
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.1623e+00  1.18439283  1.20851783  1.26083392
#> Jokulsa:Jokulsa.lag( 2)         -2.1053e-01 -0.16529127 -0.12207704 -0.04225220
#> Vatnsdalsa:Jokulsa.lag( 2)       5.4813e-03  0.00982553  0.01426767  0.02216289
#> Jokulsa:Vatnsdalsa.lag( 2)      -6.3534e-01 -0.39229033 -0.15555186  0.37515028
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -3.7748e-01 -0.34674605 -0.31193134 -0.22916246
#> Jokulsa:Jokulsa.lag( 3)         -3.1126e-02  0.00754907  0.04690640  0.12218555
#> Vatnsdalsa:Jokulsa.lag( 3)      -1.6587e-02 -0.01190373 -0.00770868  0.00031689
#> Jokulsa:Vatnsdalsa.lag( 3)      -1.9192e-01  0.05814698  0.26157025  0.66397121
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    1.6070e-01  0.19104236  0.22256639  0.27165182
#> Jokulsa:Jokulsa.lag( 4)         -1.0017e-01 -0.07296466 -0.04399549  0.00502006
#> Vatnsdalsa:Jokulsa.lag( 4)       2.9163e-03  0.00598145  0.00921910  0.01539248
#> Jokulsa:Vatnsdalsa.lag( 4)      -3.0580e-01 -0.17293935 -0.02078963  0.29137767
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -1.1747e-01 -0.09576346 -0.06725812  0.01272078
#> Jokulsa:Jokulsa.lag( 5)          1.0115e-02  0.03735981  0.06445848  0.12253263
#> Vatnsdalsa:Jokulsa.lag( 5)      -8.5010e-03 -0.00521479 -0.00197368  0.00435894
#> Jokulsa:Vatnsdalsa.lag( 5)      -1.9137e-01  0.01085344  0.20159116  0.56242992
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -2.2708e-02  0.01233373  0.04601132  0.11570726
#> Jokulsa:Jokulsa.lag( 6)         -6.2221e-02 -0.03934415 -0.01818015  0.02418110
#> Vatnsdalsa:Jokulsa.lag( 6)       1.2184e-03  0.00400121  0.00696576  0.01217458
#> Jokulsa:Vatnsdalsa.lag( 6)      -1.4237e-01  0.05629678  0.25525977  0.67922615
#> Vatnsdalsa:Vatnsdalsa.lag( 6)   -1.2790e-02  0.02176101  0.05499990  0.11805002
#> Jokulsa:Jokulsa.lag( 7)         -1.9553e-02  0.00063103  0.02100353  0.05985336
#> Vatnsdalsa:Jokulsa.lag( 7)      -8.4569e-03 -0.00574456 -0.00319019  0.00203935
#> Jokulsa:Vatnsdalsa.lag( 7)      -6.4111e-02  0.09094277  0.25737478  0.58035475
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -7.9172e-02 -0.05722714 -0.03503114  0.01540115
#> Jokulsa:Jokulsa.lag( 8)         -2.7662e-03  0.01600047  0.03609640  0.07319234
#> Vatnsdalsa:Jokulsa.lag( 8)       1.6785e-03  0.00418367  0.00675206  0.01186245
#> Jokulsa:Vatnsdalsa.lag( 8)      -3.7605e-01 -0.20950901 -0.05943613  0.22083172
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -6.5316e-02 -0.04411448 -0.02361240  0.01673790
#> Jokulsa:Jokulsa.lag( 9)          1.8725e-02  0.04022083  0.06178536  0.10314562
#> Vatnsdalsa:Jokulsa.lag( 9)      -4.7552e-03 -0.00194978  0.00073560  0.00614905
#> Jokulsa:Vatnsdalsa.lag( 9)      -1.2469e-02  0.13827188  0.29891598  0.58245077
#> Vatnsdalsa:Vatnsdalsa.lag( 9)    6.7096e-02  0.09005011  0.11463113  0.16010023
#> Jokulsa:Jokulsa.lag(10)         -5.0145e-02 -0.02254873  0.00415465  0.06625740
#> Vatnsdalsa:Jokulsa.lag(10)       1.8922e-04  0.00337835  0.00664360  0.01329505
#> Jokulsa:Vatnsdalsa.lag(10)      -1.2880e-01  0.01304113  0.14794072  0.39079770
#> Vatnsdalsa:Vatnsdalsa.lag(10)   -9.5396e-02 -0.07454838 -0.05395358 -0.01099736
#> Jokulsa:Jokulsa.lag(11)         -3.1264e-02 -0.00574947  0.01859778  0.07127905
#> Vatnsdalsa:Jokulsa.lag(11)      -1.0638e-02 -0.00747556 -0.00436729  0.00147085
#> Jokulsa:Vatnsdalsa.lag(11)      -2.0371e-01 -0.04161159  0.13399147  0.51523458
#> Vatnsdalsa:Vatnsdalsa.lag(11)    6.0591e-02  0.08613034  0.10921483  0.14732647
#> Jokulsa:Jokulsa.lag(12)         -2.8244e-02 -0.00208441  0.02320352  0.06861666
#> Vatnsdalsa:Jokulsa.lag(12)       6.5954e-03  0.00961350  0.01251786  0.01881159
#> Jokulsa:Vatnsdalsa.lag(12)      -1.4362e-01  0.01987253  0.17472857  0.49290729
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -1.0303e-01 -0.08491177 -0.06330259 -0.02252096
#> Jokulsa:Jokulsa.lag(13)         -3.1637e-02 -0.00463267  0.02315597  0.08449124
#> Vatnsdalsa:Jokulsa.lag(13)      -9.0035e-03 -0.00622468 -0.00354221  0.00188924
#> Jokulsa:Vatnsdalsa.lag(13)       2.3591e-01  0.41454073  0.59011615  0.90159103
#> Vatnsdalsa:Vatnsdalsa.lag(13)    1.2415e-01  0.14828540  0.17279994  0.21993643
#> Jokulsa:Jokulsa.lag(14)         -3.2833e-02 -0.00357137  0.02324108  0.07280139
#> Vatnsdalsa:Jokulsa.lag(14)      -4.8565e-03 -0.00176284  0.00111782  0.00647517
#> Jokulsa:Vatnsdalsa.lag(14)      -5.6757e-02  0.13948385  0.33191735  0.64766979
#> Vatnsdalsa:Vatnsdalsa.lag(14)   -7.7112e-02 -0.04752902 -0.02126807  0.02641642
#> Jokulsa:Jokulsa.lag(15)          3.0900e-02  0.04616779  0.06163355  0.09209432
#> Vatnsdalsa:Jokulsa.lag(15)      -7.9892e-04  0.00110319  0.00316447  0.00703642
#> Jokulsa:Vatnsdalsa.lag(15)      -5.5547e-01 -0.44331883 -0.32332344 -0.10325310
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -3.7673e-02 -0.02140315 -0.00537020  0.02561595
#> Jokulsa:Precipitation.lag(1)    -1.4144e-01 -0.11471854 -0.08997907 -0.03635178
#> Vatnsdalsa:Precipitation.lag(1) -6.4868e-03 -0.00321868  0.00044773  0.00701640
#> Jokulsa:Precipitation.lag(2)    -1.1970e-02  0.03659491  0.07983539  0.16189060
#> Vatnsdalsa:Precipitation.lag(2) -6.9493e-03 -0.00232410  0.00268838  0.01199792
#> Jokulsa:Precipitation.lag(3)     2.7858e-02  0.05025959  0.07192131  0.11480811
#> Vatnsdalsa:Precipitation.lag(3)  2.2640e-03  0.00557047  0.00926362  0.01638989
#> Jokulsa:Precipitation.lag(4)     6.1272e-03  0.02848527  0.05108057  0.09541855
#> Vatnsdalsa:Precipitation.lag(4)  2.6813e-05  0.00278012  0.00567318  0.01134830
#> Jokulsa:Temperature.lag(1)       1.0575e+00  1.12631627  1.19142963  1.31069565
#> Vatnsdalsa:Temperature.lag(1)    1.1096e-02  0.02005208  0.02866069  0.04424128
#> Jokulsa:Temperature.lag(2)      -6.3809e-01 -0.56969406 -0.49861349 -0.36217982
#> Vatnsdalsa:Temperature.lag(2)   -3.3884e-02 -0.02421556 -0.01503664  0.00026380
#> 
#> 
#> Scale parameter:
#>                           Mean        Sd   Sd(Mean)     2.5%      25%      50%
#> Jokulsa.Jokulsa       1.314155 0.2162755 0.01851702 0.928579 1.168538 1.297836
#> Jokulsa.Vatnsdalsa    0.045079 0.0117487 0.00063880 0.024724 0.036888 0.044240
#> Vatnsdalsa.Vatnsdalsa 0.022451 0.0038981 0.00032142 0.015632 0.019740 0.022205
#>                            75%    97.5%
#> Jokulsa.Jokulsa       1.451223 1.773745
#> Jokulsa.Vatnsdalsa    0.052428 0.069826
#> Vatnsdalsa.Vatnsdalsa 0.024743 0.031406
#> 
#> 
#> Extra parameter:
#>       Mean       Sd  Sd(Mean)    2.5%     25%     50%     75%   97.5%
#> nu 0.80818 0.043963 0.0034236 0.72767 0.77829 0.80706 0.83676 0.89561
#plot(fit3.mcmc)

###### Example 4: U.S. stock returns
data(US.returns)
fit4 <- mtar(~ CCR | dVIX, data=US.returns, subset={Date<="2025-11-28"},
             row.names=Date, ars=ars(nregim=2,p=3,d=3), n.burnin=1000,
             n.sim=2000, n.thin=2, dist="Student-t")
fit4.mcmc <- coda::as.mcmc(fit4)
summary(fit4.mcmc)
#> 
#> 
#>  Iterations = 1001:4999
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 2000
#> 
#> 
#> Thresholds:
#>               Mean     Sd Sd(Mean)   2.5%    25%    50%    75% 97.5%
#> Threshold.1 3.6224 1.3718   1.1269 1.4552 2.0963 3.4684 5.0162 5.302
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                      Mean        Sd   Sd(Mean)        2.5%       25%       50%
#> CCR:(Intercept)  0.091417 0.0115467 0.00030387  0.06834603  0.083750  0.091535
#> CCR:CCR.lag(1)  -0.049677 0.0125381 0.00036915 -0.07342392 -0.058257 -0.050133
#> CCR:CCR.lag(2)  -0.031245 0.0226139 0.00126700 -0.07615302 -0.046697 -0.031817
#> CCR:CCR.lag(3)  -0.027274 0.0227332 0.00068850 -0.07280925 -0.042638 -0.026898
#> CCR:dVIX.lag(1) -0.030392 0.0144065 0.00067601 -0.05761527 -0.040086 -0.030795
#> CCR:dVIX.lag(2) -0.023860 0.0140320 0.00046772 -0.05080213 -0.033365 -0.023838
#> CCR:dVIX.lag(3)  0.017496 0.0083194 0.00056570  0.00060204  0.011970  0.017417
#>                       75%      97.5%
#> CCR:(Intercept)  0.099157  0.1137541
#> CCR:CCR.lag(1)  -0.041382 -0.0234052
#> CCR:CCR.lag(2)  -0.015035  0.0132932
#> CCR:CCR.lag(3)  -0.012192  0.0175700
#> CCR:dVIX.lag(1) -0.020643 -0.0012841
#> CCR:dVIX.lag(2) -0.014417  0.0035304
#> CCR:dVIX.lag(3)  0.023112  0.0342567
#> 
#> 
#> Scale parameter:
#>            Mean       Sd  Sd(Mean)    2.5%     25%   50%     75%   97.5%
#> CCR.CCR 0.38227 0.018551 0.0017324 0.34457 0.36984 0.382 0.39508 0.41874
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                      Mean       Sd Sd(Mean)       2.5%       25%      50%
#> CCR:(Intercept) -0.815398 0.754408 0.345860 -2.4735882 -1.318450 -0.68373
#> CCR:CCR.lag(1)  -0.369900 0.194982 0.090700 -0.7296702 -0.517233 -0.37724
#> CCR:CCR.lag(2)  -0.404532 0.372507 0.212614 -1.1543479 -0.688622 -0.37659
#> CCR:CCR.lag(3)   0.489812 0.290179 0.158833 -0.0056435  0.228900  0.53127
#> CCR:dVIX.lag(1) -0.155861 0.168212 0.051993 -0.5262304 -0.272299 -0.12811
#> CCR:dVIX.lag(2)  0.286653 0.194610 0.105326 -0.0364396  0.099816  0.31125
#> CCR:dVIX.lag(3)  0.046886 0.061873 0.003851 -0.0978021  0.014386  0.05334
#>                       75%     97.5%
#> CCR:(Intercept) -0.168161  0.164679
#> CCR:CCR.lag(1)  -0.205950 -0.038134
#> CCR:CCR.lag(2)  -0.075043  0.142502
#> CCR:CCR.lag(3)   0.700302  1.016868
#> CCR:dVIX.lag(1) -0.027835  0.112903
#> CCR:dVIX.lag(2)  0.430842  0.635824
#> CCR:dVIX.lag(3)  0.084423  0.157245
#> 
#> 
#> Scale parameter:
#>           Mean      Sd Sd(Mean)    2.5%   25%    50%   75%  97.5%
#> CCR.CCR 1.9093 0.88033   0.3525 0.86626 1.122 1.7382 2.477 3.9766
#> 
#> 
#> Extra parameter:
#>      Mean      Sd  Sd(Mean)   2.5%    25%    50%    75%  97.5%
#> nu 2.4641 0.10898 0.0069373 2.2597 2.3969 2.4655 2.5341 2.6811
#plot(fit4.mcmc)
# }