Forecasting for multivariate TAR models

Computes forecasts from a fitted multivariate Threshold Autoregressive (TAR) model.

Usage

# S3 method for class 'mtar'
predict(
  object,
  ...,
  newdata,
  n.ahead = NULL,
  row.names,
  credible = 0.95,
  out.of.sample = FALSE
)

Arguments

object

An object of class mtar obtained from a call to mtar().

...

Additional arguments that may affect the prediction method.

newdata

An optional data.frame containing future values of the threshold series (if included in the fitted model), the exogenous series (if included in the fitted model), and, when out.of.sample = TRUE, the realized values of the output series.

n.ahead

A positive integer specifying the number of steps ahead to forecast.

row.names

An optional variable in newdata specifying labels for the time

credible

An optional numeric value in \((0,1)\) specifying the level of the required credible intervals. By default, credible is set to 0.95.

out.of.sample

An optional logical indicator. If TRUE then the log-score, Energy-Score (ES), Absolute Error (AE), Absolute Percentage Error (APE), Squared Error (SE), are computed as measures of predictive accuracy. In this case, newdata must include the observed values of the output series.

Value

A list containing the forecast summaries and, when requested, measures of predictive accuracy.

ypred	a matrix with the results of the forecasting,

summary	a matrix with the mean and credible intervals of the forecasting,

References

Nieto, F.H. (2005) Modeling Bivariate Threshold Autoregressive Processes in the Presence of Missing Data. Communications in Statistics - Theory and Methods, 34, 905-930.

Romero, L.V. and Calderon, S.A. (2021) Bayesian estimation of a multivariate TAR model when the noise process follows a Student-t distribution. Communications in Statistics - Theory and Methods, 50, 2508-2530.

Calderon, S.A. and Nieto, F.H. (2017) Bayesian analysis of multivariate threshold autoregressive models with missing data. Communications in Statistics - Theory and Methods, 46, 296-318.

Karlsson, S. (2013) Chapter 15-Forecasting with Bayesian Vector Autoregression. In Elliott, G. and Timmermann, A. Handbook of Economic Forecasting, Volume 2, 791–89, Elsevier.

Vanegas, L.H. and Calderón, S.A. and Rondón, L.M. (2025) Bayesian estimation of a multivariate tar model when the noise process distribution belongs to the class of gaussian variance mixtures. International Journal of Forecasting.

Examples

# \donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
             subset={Date<="2015-12-07"}, dist="Student-t",
             ars=ars(nregim=3,p=c(1,1,2)), n.burnin=1000, n.sim=2000,
             n.thin=2)
p1 <- predict(fit1, newdata=subset(returns,Date>"2015-12-07"), n.ahead=75,
              credible=0.8, out.of.sample=TRUE)
with(p1,summary)
#>      COLCAP.Mean COLCAP.HDI_Low COLCAP.HDI_high  BOVESPA.Mean BOVESPA.HDI_Low
#> 1  -1.352518e-02   -0.026482794   -0.0002722152 -0.0131591992    -0.030937508
#> 2  -1.981115e-03   -0.010783472    0.0086400532 -0.0011269559    -0.014270290
#> 3  -8.307425e-03   -0.023187848    0.0044240212 -0.0144542463    -0.030288641
#> 4  -1.747232e-03   -0.011392964    0.0076248214 -0.0008312935    -0.014473353
#> 5   5.655981e-03   -0.005448066    0.0160717986  0.0122826673    -0.004425186
#> 6   6.149061e-03   -0.003520897    0.0188319782  0.0108759330    -0.007293894
#> 7  -4.881357e-03   -0.019113314    0.0082134841 -0.0139690250    -0.032157556
#> 8  -1.040217e-02   -0.023788856    0.0045808035 -0.0158166521    -0.035166518
#> 9   3.393165e-03   -0.008470709    0.0145800361  0.0108420639    -0.006884879
#> 10  6.057568e-03   -0.004276137    0.0178893394  0.0119588088    -0.002732819
#> 11  5.732276e-03   -0.005755945    0.0158086366  0.0111217869    -0.004624859
#> 12  1.246929e-03   -0.008224437    0.0107648621 -0.0014306393    -0.013704836
#> 13 -3.636750e-06   -0.009146298    0.0099377577 -0.0009981431    -0.015510028
#> 14  5.210836e-03   -0.004644794    0.0161695421  0.0112503820    -0.005758236
#> 15  3.297650e-04   -0.010110604    0.0101585622 -0.0024078883    -0.016806008
#> 16 -7.969688e-03   -0.020127920    0.0072170202 -0.0151884094    -0.032454522
#> 17 -1.644531e-03   -0.011289957    0.0079411375 -0.0009227632    -0.014507430
#> 18 -8.030669e-03   -0.021183168    0.0055140315 -0.0146252863    -0.031761754
#> 19 -1.171934e-02   -0.026536444    0.0022644912 -0.0165055938    -0.036668786
#> 20 -1.350317e-02   -0.027571713    0.0021401254 -0.0164070435    -0.033936451
#> 21  3.605212e-03   -0.006644173    0.0147545711  0.0121568975    -0.005504977
#> 22 -5.272325e-03   -0.018432356    0.0099529123 -0.0131850740    -0.029630653
#> 23  3.272787e-03   -0.007276436    0.0142741553  0.0110374844    -0.006480355
#> 24 -5.779528e-03   -0.018550053    0.0082223188 -0.0138517788    -0.029817174
#> 25 -1.561321e-03   -0.010929795    0.0078075355 -0.0004131582    -0.013724433
#> 26 -3.414718e-04   -0.009959469    0.0091585373 -0.0009841357    -0.013770939
#> 27 -7.929321e-03   -0.020018204    0.0067930966 -0.0154180797    -0.032994183
#> 28 -1.977491e-03   -0.011841610    0.0079638221 -0.0009331428    -0.014300291
#> 29  5.136258e-03   -0.005923867    0.0155951584  0.0123046620    -0.006682115
#> 30 -4.857743e-03   -0.017729262    0.0095589270 -0.0129159245    -0.031249304
#> 31  3.312645e-03   -0.007698260    0.0137165332  0.0112641997    -0.005819308
#> 32 -5.151293e-03   -0.018687118    0.0082748037 -0.0129668942    -0.029958082
#> 33 -1.278819e-03   -0.011084748    0.0080868944 -0.0010188556    -0.015796969
#> 34  5.402140e-03   -0.006240593    0.0167646666  0.0128283661    -0.004884687
#> 35  1.140464e-03   -0.008525567    0.0104411812 -0.0012203555    -0.013976946
#> 36 -7.582673e-03   -0.020836893    0.0063684651 -0.0147069457    -0.031758019
#> 37 -1.745177e-03   -0.011159459    0.0073988844 -0.0003409811    -0.015610225
#> 38 -2.779041e-04   -0.010553515    0.0086096084 -0.0008624164    -0.014450462
#> 39 -7.379061e-03   -0.021461674    0.0049159255 -0.0143788696    -0.031304969
#> 40 -1.148473e-02   -0.025595440    0.0029089438 -0.0159564461    -0.033301446
#> 41 -2.204462e-03   -0.011813281    0.0072261266 -0.0004714068    -0.014513905
#> 42 -1.317263e-04   -0.009895684    0.0098546915 -0.0009475002    -0.014227861
#> 43 -7.847076e-03   -0.021802981    0.0052731124 -0.0148720594    -0.033254193
#> 44  4.204162e-03   -0.006905034    0.0153231366  0.0118720506    -0.005655982
#> 45  1.670392e-03   -0.007796957    0.0102974131 -0.0009626484    -0.016002439
#> 46  1.493157e-04   -0.010008788    0.0083469077 -0.0005064720    -0.014071356
#> 47  5.205729e-03   -0.006598682    0.0153613686  0.0118173134    -0.006848600
#> 48  1.083226e-03   -0.008697041    0.0104068464 -0.0009765894    -0.015077723
#> 49 -8.882921e-05   -0.010589717    0.0091977955 -0.0009039849    -0.014688809
#> 50  4.756967e-03   -0.006155038    0.0146167202  0.0114698182    -0.005746142
#> 51 -4.912315e-03   -0.020042318    0.0075575283 -0.0136130265    -0.030958441
#> 52 -1.308053e-03   -0.011130258    0.0079337899 -0.0006485917    -0.013585058
#> 53  5.727443e-03   -0.005091183    0.0167787995  0.0119576898    -0.004301983
#> 54  1.402401e-03   -0.008073072    0.0104961662 -0.0008306037    -0.015147507
#> 55 -1.149174e-03   -0.011040493    0.0099017182 -0.0033562872    -0.017872400
#> 56  4.755292e-03   -0.007695858    0.0144551628  0.0117187223    -0.004383531
#> 57  1.004812e-03   -0.008380693    0.0104149208 -0.0012791483    -0.015409465
#> 58 -6.301328e-05   -0.008790684    0.0097880689 -0.0004157720    -0.012400120
#> 59  2.066962e-04   -0.009360639    0.0094464615 -0.0007465883    -0.014594849
#> 60 -3.640390e-04   -0.010889880    0.0079340664 -0.0009792470    -0.015899825
#> 61 -8.021476e-03   -0.020868952    0.0052797870 -0.0147353062    -0.034947452
#> 62 -1.571962e-03   -0.011262191    0.0085575090 -0.0006091310    -0.014222075
#> 63 -1.600678e-05   -0.009324496    0.0094907221 -0.0005486316    -0.014096548
#> 64  5.150862e-03   -0.005039554    0.0163606915  0.0114405864    -0.006079244
#> 65  9.585312e-04   -0.010367301    0.0098618672 -0.0009877862    -0.012867147
#> 66  2.457755e-05   -0.009336189    0.0093505671 -0.0006853472    -0.014280143
#> 67 -1.302217e-04   -0.010374146    0.0090270935 -0.0005913684    -0.015782902
#> 68 -7.176580e-05   -0.009442415    0.0087801078 -0.0004167674    -0.013866639
#> 69 -8.936240e-05   -0.009604787    0.0094340885 -0.0013099498    -0.014800088
#> 70  2.092524e-05   -0.009485403    0.0097055182 -0.0008341875    -0.014605422
#> 71 -1.236254e-03   -0.011791206    0.0083743462 -0.0025986220    -0.016323878
#> 72 -4.967410e-04   -0.009393540    0.0088791312 -0.0010260083    -0.014811469
#> 73  5.019935e-03   -0.005960756    0.0162520549  0.0122422089    -0.005972813
#> 74  1.330317e-03   -0.007165246    0.0111995120 -0.0012056276    -0.014849785
#> 75  2.885821e-04   -0.009162038    0.0096582925 -0.0001881108    -0.014992880
#>    BOVESPA.HDI_high
#> 1      0.0042320764
#> 2      0.0129013968
#> 3      0.0055947486
#> 4      0.0134990004
#> 5      0.0294564949
#> 6      0.0274497732
#> 7      0.0032840476
#> 8      0.0018285663
#> 9      0.0281578419
#> 10     0.0304953639
#> 11     0.0289769854
#> 12     0.0130940977
#> 13     0.0123809856
#> 14     0.0283562372
#> 15     0.0116142332
#> 16     0.0035067699
#> 17     0.0127377257
#> 18     0.0036331696
#> 19    -0.0005939915
#> 20     0.0026290058
#> 21     0.0285809529
#> 22     0.0067421377
#> 23     0.0270298379
#> 24     0.0062618871
#> 25     0.0134756249
#> 26     0.0131758924
#> 27     0.0017895954
#> 28     0.0129161781
#> 29     0.0283660734
#> 30     0.0056153420
#> 31     0.0266374135
#> 32     0.0064766566
#> 33     0.0118495853
#> 34     0.0290632335
#> 35     0.0134461629
#> 36     0.0034021196
#> 37     0.0117066321
#> 38     0.0126991694
#> 39     0.0039690925
#> 40     0.0019385240
#> 41     0.0134662522
#> 42     0.0129170599
#> 43     0.0023909913
#> 44     0.0293139059
#> 45     0.0116073280
#> 46     0.0129525625
#> 47     0.0270605285
#> 48     0.0120632322
#> 49     0.0124865246
#> 50     0.0272294172
#> 51     0.0047002116
#> 52     0.0133227068
#> 53     0.0298055235
#> 54     0.0121813015
#> 55     0.0114981486
#> 56     0.0280143018
#> 57     0.0121786282
#> 58     0.0130728473
#> 59     0.0122343485
#> 60     0.0111088077
#> 61     0.0016356168
#> 62     0.0124154971
#> 63     0.0133449636
#> 64     0.0278993733
#> 65     0.0133004077
#> 66     0.0122436498
#> 67     0.0120956769
#> 68     0.0128175056
#> 69     0.0136739671
#> 70     0.0128399427
#> 71     0.0129480132
#> 72     0.0110659082
#> 73     0.0281329917
#> 74     0.0129459881
#> 75     0.0124259919
with(p1,cbind(LS,ES,APE,CR))
#>     Log.Score Energy.Score  COLCAP.APE BOVESPA.APE COLCAP.CoverageRate
#> 1  -3.4562102  0.036271032  232.693213  167.873088                   0
#> 2  -3.7706987  0.019721754  111.761660   89.170307                   0
#> 3  -7.0800299  0.011385651   72.798922   78.508770                   1
#> 4  -6.9925001  0.011160112  202.826366   92.741319                   1
#> 5  -3.7709093  0.020692704   77.198004  339.744817                   0
#> 6  -3.9216295  0.020227787   75.967189  240.918717                   0
#> 7  -5.4968500  0.022053892  150.061732  356.703817                   0
#> 8  -5.3533195  0.017671214 1279.399624   47.800846                   1
#> 9  -4.9183801  0.025980098  144.189923  166.448828                   1
#> 10 -6.7336789  0.012176557  369.490130   92.243254                   1
#> 11 -3.7637343  0.021533073   79.676977   10.803725                   0
#> 12 -7.2066478  0.009234547  123.100941         Inf                   1
#> 13 -7.7851554  0.007810153   99.835491         Inf                   1
#> 14 -6.3448737  0.015501917  250.974321  545.541751                   1
#> 15 -7.2486883  0.009561940   93.048495   65.544737                   1
#> 16 -5.9705859  0.017650083   62.088595   46.299765                   0
#> 17 -5.8498132  0.014387099  114.722286  114.019856                   0
#> 18 -7.1927786  0.010384752   30.262882    4.723556                   1
#> 19 -6.5257663  0.014579205   44.484243   36.893631                   1
#> 20 -6.4340835  0.016004728  100.820172  708.390424                   1
#> 21 -2.6382468  0.040790327  120.615834  144.339074                   0
#> 22 -4.5780558  0.020805957  132.965872    9.160686                   0
#> 23 -3.1262508  0.024110412   88.610367   22.092608                   0
#> 24 -5.5299710  0.018850955   74.598604   41.924285                   0
#> 25 -3.6921603  0.024046643   93.067601   97.498772                   0
#> 26 -4.0974826  0.019218731  101.721516  131.229591                   0
#> 27 -5.6912322  0.016666246  197.014222   41.809288                   0
#> 28 -5.2371357  0.015560997  114.288101  149.088396                   0
#> 29 -4.4389896  0.018643030   78.211540   48.363989                   0
#> 30 -5.9116655  0.015317482   54.761418         Inf                   1
#> 31 -6.7080398  0.012602222   46.224279         Inf                   1
#> 32 -3.6991695  0.034056571  183.429639  155.968561                   1
#> 33 -6.1322825  0.013418117  112.438551  115.455496                   0
#> 34 -4.2235363  0.029053646   34.894573   71.456965                   1
#> 35 -7.5442261  0.008886358   76.139325  130.118687                   1
#> 36 -3.8558937  0.031927778   40.493559   70.513462                   1
#> 37 -3.8736612  0.026726153  113.750436  101.342778                   0
#> 38 -3.6412488  0.029281242  104.079682  102.812103                   1
#> 39 -6.1496073  0.015715370  237.856257  154.873883                   0
#> 40 -6.1353803  0.017825016  543.815887         Inf                   1
#> 41 -7.4396762  0.008726475  167.198908         Inf                   1
#> 42 -7.7087581  0.007929115  104.485651         Inf                   1
#> 43 -6.1095077  0.014601297  392.790116   44.005494                   1
#> 44 -7.1875207  0.009816583   51.893419    4.095963                   1
#> 45 -7.3462109  0.009600641  137.678157  113.501648                   1
#> 46 -4.3483999  0.021158181  101.817313  102.400725                   1
#> 47 -3.8631740  0.021595799   81.015571   28.575215                   0
#> 48 -7.7767543  0.007724456   47.554274   73.509126                   1
#> 49 -7.8082368  0.007841460   88.568699  157.054692                   1
#> 50 -4.3403420  0.025792008  308.190643   71.259509                   1
#> 51 -7.2157328  0.010275393    9.867276   18.241174                   1
#> 52 -7.1760117  0.010419074    9.352741   93.712597                   1
#> 53 -5.8743386  0.016353662   40.849398  355.292075                   1
#> 54 -7.2478891  0.009492965   73.377885   88.240398                   1
#> 55 -3.6988027  0.028736034   42.979701  111.792655                   1
#> 56 -6.0843270  0.017660221   47.727490   61.652304                   1
#> 57 -4.5560562  0.021944177   93.968002  107.377372                   0
#> 58 -0.7616944  0.048966200  100.431040  100.832189                   0
#> 59 -2.1983794  0.036764264  146.965192  101.899861                   1
#> 60 -6.6497570  0.011308944  103.847414  129.861137                   0
#> 61 -6.5282991  0.013723952   16.785824  403.330924                   1
#> 62 -7.1455838  0.010104957  224.170568   93.186975                   1
#> 63 -5.7394970  0.017302331  100.653288  102.974214                   1
#> 64 -7.0511658  0.011782501  105.572999  739.883084                   1
#> 65 -6.4614011  0.014239657  129.471864   93.692818                   1
#> 66 -2.9709634  0.033695056  100.241642   98.106719                   0
#> 67 -6.5065952  0.013340535   82.200219  104.429686                   1
#> 68  0.7263472  0.061244839  100.805811  100.652488                   0
#> 69 -6.2809264  0.011787182  100.836384   32.785675                   0
#> 70 -7.3488700  0.009230623   99.613468  121.720686                   1
#> 71 -4.6295565  0.021947944  139.481809   90.089427                   1
#> 72 -3.5489049  0.024306011   95.829108  104.491385                   0
#> 73 -6.6210491  0.012628174  213.517456   97.092166                   1
#> 74 -5.5890563  0.014180371   91.340479  165.719657                   0
#> 75 -3.5057230  0.024107935   97.677346   99.201796                   0
#>    BOVESPA.CoverageRate
#> 1                     0
#> 2                     1
#> 3                     1
#> 4                     1
#> 5                     1
#> 6                     1
#> 7                     0
#> 8                     1
#> 9                     0
#> 10                    1
#> 11                    1
#> 12                    1
#> 13                    1
#> 14                    1
#> 15                    1
#> 16                    1
#> 17                    1
#> 18                    1
#> 19                    1
#> 20                    1
#> 21                    0
#> 22                    1
#> 23                    1
#> 24                    1
#> 25                    0
#> 26                    1
#> 27                    1
#> 28                    1
#> 29                    1
#> 30                    1
#> 31                    1
#> 32                    0
#> 33                    1
#> 34                    0
#> 35                    1
#> 36                    0
#> 37                    0
#> 38                    0
#> 39                    1
#> 40                    1
#> 41                    1
#> 42                    1
#> 43                    1
#> 44                    1
#> 45                    1
#> 46                    0
#> 47                    1
#> 48                    1
#> 49                    1
#> 50                    0
#> 51                    1
#> 52                    1
#> 53                    0
#> 54                    1
#> 55                    0
#> 56                    0
#> 57                    0
#> 58                    0
#> 59                    0
#> 60                    1
#> 61                    1
#> 62                    1
#> 63                    0
#> 64                    1
#> 65                    0
#> 66                    0
#> 67                    0
#> 68                    0
#> 69                    1
#> 70                    1
#> 71                    0
#> 72                    0
#> 73                    1
#> 74                    1
#> 75                    0
plot(p1,last=100)

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, row.names=Date,
             subset={Date<="2009-02-13"}, dist="Laplace",
             ars=ars(nregim=3,p=5), n.burnin=1000, n.sim=2000, n.thin=2)
p2 <- predict(fit2, newdata=subset(riverflows,Date>"2009-02-13"), n.ahead=60,
              credible=0.8, out.of.sample=TRUE)
with(p2,summary)
#>    Bedon.Mean Bedon.HDI_Low Bedon.HDI_high LaPlata.Mean LaPlata.HDI_Low
#> 1   13.668129      8.308054      18.487113     31.52380       13.034267
#> 2   15.407844      9.040602      21.768478     35.43275       10.562703
#> 3   12.966312      7.877511      17.793349     29.16281       11.651909
#> 4   11.269151      7.203884      15.028312     24.55496       12.446761
#> 5   11.103251      7.146349      15.857072     24.45268       12.957108
#> 6   10.147824      6.874025      14.416100     22.50846       12.435445
#> 7    9.528080      5.799581      12.658264     21.17441       12.747532
#> 8    8.996827      5.921143      12.553043     20.01963       11.939404
#> 9    8.520733      5.474291      11.654445     18.90342       10.904412
#> 10   8.121719      5.209591      11.354045     18.28649       10.684750
#> 11   7.745801      4.699976      10.539960     17.45294       10.495502
#> 12   7.453628      4.737079      10.524844     16.70866        9.955815
#> 13   7.204614      4.095389       9.716183     16.09438        8.497108
#> 14   7.981417      4.098855      11.881241     18.31993        7.203814
#> 15   7.555613      4.555427      11.051980     17.18292        9.079893
#> 16   8.114788      4.495161      12.506672     18.80462        9.462940
#> 17   8.413582      3.933578      13.018955     19.84977        8.887150
#> 18  12.141063      6.169479      18.213359     29.59514        6.714238
#> 19  14.394179      7.545736      20.813744     35.20050       12.344156
#> 20  12.773091      6.886169      18.962573     29.59588       12.094909
#> 21  11.780080      6.654596      17.387572     27.28210       14.337188
#> 22  10.201098      5.633371      14.478209     23.49535       13.097938
#> 23   9.397152      5.219739      13.049389     21.34983       11.927034
#> 24   8.986229      5.619009      12.834627     20.51261       11.850105
#> 25   8.559351      5.151531      11.944360     19.44313        9.853437
#> 26   8.185416      5.147902      11.731974     18.47636       10.555618
#> 27   8.850512      5.069177      13.277195     19.89026        8.925877
#> 28   8.187654      4.345045      11.527801     18.37664        9.141540
#> 29   7.687675      4.688165      11.372675     17.13873        9.300677
#> 30   8.202993      3.757084      12.483037     18.71463        8.614455
#> 31   8.648435      3.802000      13.025014     19.79841        9.315701
#> 32   8.941498      4.172676      13.790714     20.46114        9.943936
#> 33  12.459043      7.356389      19.863919     30.46893        8.477810
#> 34  14.448220      7.725652      21.960874     34.72203       13.433000
#> 35  12.803883      7.619829      19.933539     29.54913       13.288097
#> 36  15.130149      7.932823      22.260468     34.99280        8.856406
#> 37  16.500592      8.425998      23.554986     38.59745       10.601501
#> 38  17.969203     10.181108      25.570592     40.79578       16.072095
#> 39  15.660997      9.809669      22.793982     34.51413       15.294139
#> 40  14.504806      8.910199      20.982804     31.23907       15.837444
#> 41  16.761738      9.064417      23.422992     37.44858       11.177536
#> 42  15.524895      9.300843      22.325584     33.10299       15.803718
#> 43  18.241111     10.760901      26.174018     39.63903       13.210701
#> 44  19.289133     12.276136      27.929854     42.48169       15.522988
#> 45  17.136394     10.808393      24.409720     35.96042       17.463003
#> 46  14.494259      9.366483      19.839535     30.46804       17.337234
#> 47  12.770733      8.317832      17.071468     26.92352       16.269775
#> 48  11.719181      7.851493      15.822301     24.75837       14.655153
#> 49  11.036682      7.554867      14.799222     23.48324       13.611506
#> 50  10.349534      6.904154      14.142884     22.22174       13.996648
#> 51   9.674710      6.284333      13.059144     21.00426       12.663427
#> 52  13.857345      8.804443      20.674739     32.39543       11.736702
#> 53  16.021594      9.870631      23.256492     38.10757       11.725162
#> 54  14.227997      8.117374      20.153394     31.78535       12.334941
#> 55  12.688242      7.194655      18.116363     28.68723       13.700799
#> 56  12.091595      6.733683      17.135129     27.19418       13.596877
#> 57  15.409072      9.407123      22.509376     35.18375        9.339949
#> 58  14.163546      8.647541      20.402708     30.88826       13.959850
#> 59  12.177748      7.494093      16.551746     26.15388       14.700042
#> 60  10.908723      6.751815      14.710174     23.58933       13.211422
#>    LaPlata.HDI_high
#> 1          52.74340
#> 2          57.63613
#> 3          43.00075
#> 4          34.44295
#> 5          34.80100
#> 6          31.71847
#> 7          30.13000
#> 8          28.60337
#> 9          26.70259
#> 10         25.55281
#> 11         25.24127
#> 12         24.37520
#> 13         22.81194
#> 14         26.19461
#> 15         25.19492
#> 16         28.20473
#> 17         30.13505
#> 18         49.84246
#> 19         60.38053
#> 20         45.23383
#> 21         42.18337
#> 22         35.45414
#> 23         31.38427
#> 24         30.04228
#> 25         27.61062
#> 26         26.65143
#> 27         29.47671
#> 28         26.82866
#> 29         25.18379
#> 30         29.11310
#> 31         29.76845
#> 32         31.27456
#> 33         52.44405
#> 34         61.33641
#> 35         47.32763
#> 36         58.64960
#> 37         64.50856
#> 38         67.64345
#> 39         52.87747
#> 40         46.61187
#> 41         62.08389
#> 42         51.12520
#> 43         64.26461
#> 44         68.31293
#> 45         54.94902
#> 46         44.71179
#> 47         38.58759
#> 48         34.72910
#> 49         32.15489
#> 50         31.49827
#> 51         29.16644
#> 52         54.40309
#> 53         58.93299
#> 54         46.83923
#> 55         42.79895
#> 56         39.36558
#> 57         56.52677
#> 58         47.72394
#> 59         38.76176
#> 60         32.77781
with(p2,cbind(LS,ES,APE,CR))
#>    Log.Score Energy.Score   Bedon.APE LaPlata.APE Bedon.CoverageRate
#> 1   6.716409     9.821660 20.31803579   15.098834                  1
#> 2   8.257321    25.535935  6.90124197   44.920333                  1
#> 3   5.767070     7.553173 11.77855256   10.844366                  1
#> 4   4.574487     5.209667  1.76700662    8.640548                  1
#> 5   6.851829    10.766406  2.99861728   33.006361                  1
#> 6   4.665943     4.996668  7.57901547   12.993962                  1
#> 7   5.313793     6.974854 10.02757469   26.452222                  1
#> 8   4.714162     5.179978 18.72785449   18.220483                  1
#> 9   4.353489     3.833778  5.84761984    6.695849                  1
#> 10  4.887182     4.693245  2.78054975   18.798892                  1
#> 11  4.962801     4.228918 16.58339763   14.946672                  1
#> 12  4.981493     3.807065 34.44494281    5.172209                  1
#> 13  5.350013     4.308616 31.30333121   16.043942                  1
#> 14  6.379644     6.460672 46.98741838   22.307327                  1
#> 15  4.256433     3.865276  0.31350429    6.614586                  1
#> 16  4.642357     4.758004  4.65292159   12.400595                  1
#> 17  5.210089     5.633984 23.51289006   11.345374                  1
#> 18  8.430215    32.034122 39.23391947   54.517989                  0
#> 19  6.775762    12.269322 15.98854945   18.780569                  1
#> 20  5.609313     8.419987 24.01059665   18.288879                  1
#> 21  5.470372     7.370713 21.15684470   21.958442                  1
#> 22  4.931146     5.628197  7.04195142   17.008720                  1
#> 23  4.813063     5.145514  2.58899589   20.893728                  1
#> 24  4.813652     5.166753  1.37894076   25.459383                  1
#> 25  4.358422     3.996583  0.95636281    2.548148                  1
#> 26  5.120486     6.338019  9.91177108   27.172411                  1
#> 27  5.053684     5.061394 20.71074498    7.357878                  1
#> 28  6.882617     7.148336 41.85124438   27.278816                  1
#> 29  4.361091     4.062222  0.07806871   13.265556                  1
#> 30  4.661683     5.096682 11.66470642   11.305157                  1
#> 31  4.923182     5.084763  7.40742193    7.223927                  1
#> 32  6.000505     6.109412 61.28243071    7.624665                  1
#> 33  6.758907    16.452961 19.25442001   36.470130                  1
#> 34  6.276760    11.299976 20.08728062   10.266829                  1
#> 35  5.478558     7.894735  1.88335491   13.957315                  1
#> 36  7.570785    13.505202 33.34736239   15.458931                  0
#> 37  6.006448    11.679024  8.73566556    6.920560                  1
#> 38  6.433044    14.392315 18.80687172   22.925024                  1
#> 39  5.889348    10.501491 10.04596766   20.711859                  1
#> 40  6.638322    10.450304  8.56890760   24.924123                  1
#> 41  6.374051    11.302349 16.48184991    4.511907                  1
#> 42  6.817904    11.394730  8.41407447   25.004547                  1
#> 43  7.517561    20.526724  5.68194762   36.271650                  1
#> 44  6.433561    12.651094 22.70441134   13.254310                  1
#> 45  6.161646     8.956069 25.26603538    6.397858                  1
#> 46  5.263164     6.652808  1.61743170   11.655628                  1
#> 47  5.600332     7.543037  3.54431229   22.276225                  1
#> 48  5.102049     5.516435  3.16180345   14.861170                  1
#> 49  4.596921     4.455508  0.51622804    8.196870                  1
#> 50  5.061223     5.416339  0.48091689   18.452321                  1
#> 51  4.305118     3.926872  2.31291791    7.164568                  1
#> 52  6.010507    10.894788 33.24370572   28.247927                  1
#> 53  6.729709    11.170479 21.00901827    9.868578                  1
#> 54  6.114894     8.262287 25.24645606    4.976539                  1
#> 55  5.769955     7.249327 13.59213733   11.459169                  1
#> 56  5.256016     6.249172 14.17936386    8.042015                  1
#> 57  7.301949    12.307587 34.45961496   17.660302                  1
#> 58  5.559909     7.745774 13.79279008    5.528726                  1
#> 59  5.092459     5.587653  6.26306933    4.478164                  1
#> 60  4.720533     4.779421  1.45688234    8.639296                  1
#>    LaPlata.CoverageRate
#> 1                     1
#> 2                     0
#> 3                     1
#> 4                     1
#> 5                     0
#> 6                     1
#> 7                     1
#> 8                     1
#> 9                     1
#> 10                    1
#> 11                    1
#> 12                    1
#> 13                    1
#> 14                    1
#> 15                    1
#> 16                    1
#> 17                    1
#> 18                    0
#> 19                    1
#> 20                    1
#> 21                    1
#> 22                    1
#> 23                    1
#> 24                    1
#> 25                    1
#> 26                    1
#> 27                    1
#> 28                    1
#> 29                    1
#> 30                    1
#> 31                    1
#> 32                    1
#> 33                    1
#> 34                    1
#> 35                    1
#> 36                    1
#> 37                    1
#> 38                    1
#> 39                    1
#> 40                    1
#> 41                    1
#> 42                    1
#> 43                    1
#> 44                    1
#> 45                    1
#> 46                    1
#> 47                    1
#> 48                    1
#> 49                    1
#> 50                    1
#> 51                    1
#> 52                    1
#> 53                    1
#> 54                    1
#> 55                    1
#> 56                    1
#> 57                    1
#> 58                    1
#> 59                    1
#> 60                    1
plot(p2,last=100)

###### Example 3: Temperature, precipitation, and two river flows in Iceland
data(iceland.rf)
fit3 <- mtar(~ Jokulsa + Vatnsdalsa | Temperature | Precipitation,
             data=iceland.rf, subset={Date<="1974-11-06"}, row.names=Date,
             ars=ars(nregim=2,p=15,q=4,d=2), n.burnin=1000, n.sim=2000,
             n.thin=2, dist="Slash")
p3 <- predict(fit3, newdata=subset(iceland.rf,Date>"1974-11-06"), n.ahead=55,
              credible=0.8, out.of.sample=TRUE)
with(p3,summary)
#>    Jokulsa.Mean Jokulsa.HDI_Low Jokulsa.HDI_high Vatnsdalsa.Mean
#> 1     33.106016       25.311128         28.61341      12.8012775
#> 2     48.473006       23.597231         29.98225      -3.2966712
#> 3     40.834338       22.910011         31.95979      -9.6071849
#> 4    -27.671576       21.593184         33.39997     -62.6143251
#> 5    -34.802008       20.428551         34.13869     -63.5077855
#> 6    -33.298235       20.617026         35.93647     -55.1678948
#> 7    -26.090443       17.112141         34.79726     -43.8942977
#> 8    -21.519565       18.073698         36.48763     -18.9074680
#> 9     -8.213603       17.041353         36.96428     -10.3426130
#> 10    -4.968105       17.000401         37.90228      -6.6016257
#> 11     2.103727       14.614077         37.15933      -4.9475513
#> 12     6.707205       13.264909         36.54061      -5.0113138
#> 13     7.531198       13.371293         36.76634      -3.2463334
#> 14     6.146723       15.391240         40.35836      -1.6394191
#> 15     7.197421       15.168738         40.26090      -1.9437599
#> 16     5.087235       14.418998         41.31107       0.4310260
#> 17    14.534380       14.145663         41.40842      -1.1707147
#> 18    18.885246       13.243267         41.68655       0.6941420
#> 19    20.440936       11.210289         39.03942       0.8031345
#> 20    18.751000       12.214179         40.70716       6.6122418
#> 21    19.899707        8.948683         37.33851       6.4400095
#> 22    17.292939       11.947259         40.62643       6.5862118
#> 23     3.990757       11.562323         41.48012     -25.0811715
#> 24    -2.446414       10.189218         40.73062     -34.3517606
#> 25     3.234156       11.616632         40.76223     -25.8785421
#> 26    13.875933       11.041729         38.89810      35.9452147
#> 27    20.584076       10.468555         40.57379      47.7338704
#> 28    37.320855       11.123727         41.07422      36.1393705
#> 29    37.947937        8.968725         39.44769      29.2601488
#> 30    41.521137       10.933007         41.96963      23.1785410
#> 31    39.128702       12.423874         44.41999      17.1385879
#> 32    22.063336       11.213033         42.70304       4.7170357
#> 33    18.739650        8.557326         40.16097       2.4092727
#> 34    18.529201        9.923592         42.53654       3.7357484
#> 35    17.301151        9.664347         42.94330      11.7179995
#> 36    22.735522       10.436516         43.53326      14.5736331
#> 37    23.505166        9.584710         42.65053      12.9121802
#> 38    24.366718        5.496211         41.37973      10.9428932
#> 39    25.529033        7.876647         43.92559       7.1506433
#> 40    43.276048        6.691434         41.37431      16.8837594
#> 41    41.372060        6.877863         41.49077      18.1019575
#> 42    47.339785        5.498409         40.76758      18.4821397
#> 43    40.821416        8.115475         43.39844      19.7407151
#> 44    36.772435        4.465235         41.38065      20.6067831
#> 45    36.251762        6.399772         42.01415      19.4628312
#> 46    32.779584        8.573102         42.35945      18.1359765
#> 47    30.249396       11.941193         45.64998      14.6803406
#> 48    30.436952       10.629786         42.94327      13.1538404
#> 49    32.147004        9.086195         42.44668      15.9904093
#> 50    27.220489       10.906445         45.21386      14.2099994
#> 51    33.035153        7.628047         44.33764       9.2967859
#> 52    24.888913        6.248752         45.31872       3.8290580
#> 53    22.862961        6.509939         45.36833       3.8878805
#> 54    22.910932        7.914704         46.21162       4.7596642
#> 55    24.697030        8.774027         46.34353       5.7627378
#>    Vatnsdalsa.HDI_Low Vatnsdalsa.HDI_high
#> 1           6.7166422            8.897364
#> 2           4.8886043            9.691262
#> 3           4.0445192           11.414959
#> 4           3.8503158           13.689746
#> 5           1.0360331           13.181068
#> 6           0.1660210           13.406685
#> 7          -1.0954869           13.907640
#> 8          -0.4826257           15.463808
#> 9          -1.4292419           16.265037
#> 10         -2.9782980           16.467012
#> 11         -3.5414583           16.998534
#> 12         -5.4654259           16.077073
#> 13         -4.8553234           16.679091
#> 14         -4.1527118           19.134454
#> 15         -3.7930952           19.224272
#> 16         -5.9007325           18.038536
#> 17         -7.6216895           18.666751
#> 18         -7.3314600           20.019150
#> 19         -6.9732527           22.511754
#> 20         -8.2785647           22.246880
#> 21         -9.9242514           21.665834
#> 22         -9.5817289           22.285881
#> 23         -7.9931074           24.102715
#> 24         -7.2261542           25.915435
#> 25         -8.0156157           24.996255
#> 26         -9.9901823           23.501428
#> 27         -9.5000564           25.554308
#> 28         -8.1024452           27.758283
#> 29         -9.2849496           25.837130
#> 30        -10.1252064           25.599500
#> 31        -12.3040549           24.815481
#> 32        -13.7624051           23.779299
#> 33        -12.2421893           25.943095
#> 34        -13.7124991           23.995162
#> 35        -14.2100501           22.757457
#> 36        -10.6853076           26.976801
#> 37        -14.3967439           24.793139
#> 38        -13.8331477           26.774897
#> 39        -15.5182158           26.526432
#> 40        -12.1970808           29.553505
#> 41        -11.0702424           29.826521
#> 42        -15.5166835           27.063799
#> 43        -14.4585886           28.323267
#> 44        -14.6713935           29.197026
#> 45        -13.9878000           29.696651
#> 46        -13.9662981           29.448952
#> 47        -15.5909443           28.013935
#> 48        -12.8436577           31.628778
#> 49        -15.8614550           27.447848
#> 50        -16.0653456           29.186920
#> 51        -21.8041281           25.939400
#> 52        -18.0927135           31.986099
#> 53        -20.1397621           30.259175
#> 54        -19.1843903           29.993548
#> 55        -15.2928602           33.547638
with(p3,cbind(LS,ES,APE,CR))
#>    Log.Score Energy.Score Jokulsa.APE Vatnsdalsa.APE Jokulsa.CoverageRate
#> 1   5.213987     8.656414  14.1586773      85.525761                    0
#> 2   5.461300    24.783977  60.5066424     145.159880                    0
#> 3   4.300153    28.056969  43.7828804     252.494998                    1
#> 4   5.097243    69.933708 199.5380433    1313.455912                    1
#> 5   4.278481    71.740889 222.5422799    1141.111238                    1
#> 6   3.873442    66.348951 224.7124902    1099.418384                    1
#> 7   4.587386    58.298659 195.5693866     979.645245                    1
#> 8   4.529259    67.758384 178.8262455     454.072435                    1
#> 9   3.999854    62.029584 130.7625564     293.681892                    1
#> 10  3.870825    56.042564 118.9622343     223.625950                    1
#> 11  3.959287    50.011124  91.9705058     195.882778                    1
#> 12  3.832120    45.053901  74.3999808     187.763815                    1
#> 13  3.943438    42.748087  71.2549706     158.810387                    1
#> 14  4.088973    41.051905  76.0827900     129.699621                    1
#> 15  4.604348    38.276969  70.7421908     137.669766                    1
#> 16  4.558014    37.981149  79.3201827      91.057553                    1
#> 17  3.870223    39.631257  44.5252654     121.208599                    1
#> 18  3.834087    37.319668  29.2687414      88.620624                    1
#> 19  4.030539    36.024287  23.4421885      86.387551                    1
#> 20  4.226752    38.802525  29.7715338      15.801082                    1
#> 21  4.182992    38.522746  24.0469204      16.666839                    1
#> 22  4.559392    38.546495  33.9964143      27.639764                    1
#> 23  4.339873    55.564046  84.4717641     602.628687                    1
#> 24  4.760170    60.812778 109.5191209     866.780371                    1
#> 25  4.638151    60.484275  86.8530249     636.899214                    1
#> 26  4.275355    84.832342  44.9367741     596.612689                    1
#> 27  4.252750    87.774367  19.9063206     825.075008                    1
#> 28  4.319150    86.257014  48.0986316     600.375397                    1
#> 29  4.491278    79.301045  47.6573433     467.057147                    1
#> 30  4.554078    71.498525  58.4776214     349.196531                    1
#> 31  4.556817    65.540868  55.2726253     255.572362                    1
#> 32  4.513369    72.855199  15.7887933       1.441628                    1
#> 33  4.766419    66.480255  25.6363103      36.930035                    1
#> 34  4.804696    59.536828  26.4714249       9.764531                    1
#> 35  4.556151    57.386325  31.3446380     151.999989                    1
#> 36  4.781069    55.219694   7.5791806     213.411465                    1
#> 37  5.008093    54.014689   4.4505435     199.586548                    1
#> 38  5.136284    49.940651   0.9483026     153.895433                    1
#> 39  4.944870    46.406315   3.7765556      38.578358                    1
#> 40  4.802097    53.561915  75.9188952     227.204639                    1
#> 41  5.031889    51.606391  68.1791063     250.813129                    1
#> 42  4.772697    54.363251  92.4381519     258.181003                    1
#> 43  4.667293    53.968432  65.9407152     282.571997                    1
#> 44  4.843244    51.737053  49.4814446     312.961586                    1
#> 45  5.070776    48.385967  47.3648850     318.555510                    1
#> 46  4.894496    46.493018  33.2503431     251.472412                    1
#> 47  4.820055    47.404884  22.9650248     184.502725                    1
#> 48  4.876475    44.932412  23.7274487     154.919388                    1
#> 49  4.683277    43.642197  30.6788793     209.891654                    1
#> 50  4.559835    44.181690  10.6523935     175.387585                    1
#> 51  4.765551    46.521933  34.2892410      80.170269                    1
#> 52  5.033485    49.362115   1.1744417      25.793449                    1
#> 53  4.937663    46.314349   7.0611344      24.653478                    1
#> 54  4.904479    43.042016   6.8661287      10.867711                    1
#> 55  4.440471    39.748098   3.9026057       7.916438                    1
#>    Vatnsdalsa.CoverageRate
#> 1                        1
#> 2                        1
#> 3                        1
#> 4                        1
#> 5                        1
#> 6                        1
#> 7                        1
#> 8                        1
#> 9                        1
#> 10                       1
#> 11                       1
#> 12                       1
#> 13                       1
#> 14                       1
#> 15                       1
#> 16                       1
#> 17                       1
#> 18                       1
#> 19                       1
#> 20                       1
#> 21                       1
#> 22                       1
#> 23                       1
#> 24                       1
#> 25                       1
#> 26                       1
#> 27                       1
#> 28                       1
#> 29                       1
#> 30                       1
#> 31                       1
#> 32                       1
#> 33                       1
#> 34                       1
#> 35                       1
#> 36                       1
#> 37                       1
#> 38                       1
#> 39                       1
#> 40                       1
#> 41                       1
#> 42                       1
#> 43                       1
#> 44                       1
#> 45                       1
#> 46                       1
#> 47                       1
#> 48                       1
#> 49                       1
#> 50                       1
#> 51                       1
#> 52                       1
#> 53                       1
#> 54                       1
#> 55                       1
plot(p3,last=100)

###### Example 4: U.S. stock returns
data(US.returns)
fit4 <- mtar(~ CCR | dVIX, data=US.returns, subset={Date<="2025-11-28"},
             row.names=Date, ars=ars(nregim=2,p=3,d=3), n.burnin=1000,
             n.sim=2000, n.thin=2, dist="Student-t")
p4 <- predict(fit4, newdata=subset(US.returns,Date>"2025-11-28"),n.ahead=100,
              credible=0.8, out.of.sample=TRUE)
with(p4,summary)
#>         CCR.Mean CCR.HDI_Low CCR.HDI_high
#> 1    0.054198088  -0.9701386    1.0587927
#> 2    0.089099454  -0.9652197    1.0943493
#> 3   -0.001549539  -1.0059625    1.0014837
#> 4    0.057693805  -0.9747121    1.0136290
#> 5    0.148877446  -0.9769467    1.0116988
#> 6    0.093775786  -0.8858311    1.1599319
#> 7    0.103224454  -0.9365096    1.0525292
#> 8    0.007994990  -0.9355058    1.1327576
#> 9    0.055450478  -0.9443388    1.0842710
#> 10   0.183182805  -0.8956913    1.1537167
#> 11   0.097386330  -0.7994374    1.1246712
#> 12   0.017480110  -1.0712409    0.9913675
#> 13   0.066899240  -1.0839440    0.9897383
#> 14   0.080831699  -0.9761198    1.0674727
#> 15   0.078956282  -1.1098680    0.9720447
#> 16   0.122361646  -0.8746186    1.1832682
#> 17   0.197766308  -0.7619499    1.1981078
#> 18   0.114934322  -0.8153500    1.1979100
#> 19   0.056680508  -0.9841628    1.0399123
#> 20   0.102372851  -0.9072236    1.0931639
#> 21   0.057409033  -0.9603829    1.0959361
#> 22   0.051345729  -1.0350537    0.9976347
#> 23   0.044189161  -0.9061821    1.0733410
#> 24   0.061306821  -0.9656510    1.1070603
#> 25   0.092786440  -0.9710928    1.0349543
#> 26   0.089696727  -0.9099258    1.1190126
#> 27   0.080994720  -0.9085513    1.0521863
#> 28   0.025489580  -1.0399485    1.0208185
#> 29   0.101811498  -0.9277166    1.0883491
#> 30   0.143546982  -0.8378124    1.1888072
#> 31   0.084172769  -0.7874223    1.1738334
#> 32   0.029979482  -0.9539165    1.0649308
#> 33   0.043139051  -1.0165516    0.9521885
#> 34   0.172235831  -0.8552912    1.2726370
#> 35   0.094420857  -0.8870822    1.0701563
#> 36   0.208244423  -1.6802804    2.1407979
#> 37   0.217172816  -0.7636550    1.2781334
#> 38   0.254006874  -0.8063573    1.3634413
#> 39  -0.036781456  -1.0386606    1.1602372
#> 40   0.055367523  -0.9850429    1.0574955
#> 41   0.062875309  -0.8855299    1.0430417
#> 42   0.083913288  -0.9228063    1.1485323
#> 43  -0.011420648  -0.9232161    1.1029629
#> 44   0.032494637  -0.9425304    1.2032396
#> 45   0.129628808  -0.9188253    1.1786931
#> 46   0.311839661  -1.7414433    2.0942329
#> 47   0.052979456  -1.0757560    0.9721164
#> 48   0.287122802  -1.5718821    2.4737909
#> 49   0.274232109  -0.7226321    1.3213215
#> 50   0.167431635  -0.9618710    1.1044384
#> 51   0.022202831  -1.1158356    1.0025455
#> 52   0.062314500  -1.0561171    1.1250970
#> 53   0.263535210  -1.9222503    2.2212784
#> 54   0.062821489  -0.8925157    1.1543111
#> 55   0.069933040  -1.0827647    1.1707149
#> 56   0.084518187  -0.9095060    1.0649086
#> 57   0.034224069  -0.9782502    1.0751539
#> 58   0.145739777  -0.9299383    1.1303202
#> 59   0.416906201  -1.7552005    2.2301428
#> 60   0.121847464  -0.9798897    1.0583585
#> 61   0.250122406  -0.7882545    1.1762363
#> 62   0.047755132  -1.0233807    1.0873296
#> 63  -0.101339214  -1.0296021    1.0933332
#> 64   0.022486773  -1.4261050    1.4014485
#> 65   0.145563720  -2.0768208    2.0114287
#> 66   0.208360075  -0.7971965    1.2765875
#> 67   0.520655237  -1.7960959    2.2990815
#> 68  -0.141644183  -2.5070382    1.5860414
#> 69   0.308793442  -0.7930173    1.2104956
#> 70   0.222071245  -0.7314277    1.3616814
#> 71   0.005721482  -1.0213546    1.0068585
#> 72   0.346797433  -1.7463119    2.1982388
#> 73  -0.006980366  -1.0275735    1.0075975
#> 74   0.314912566  -0.6841842    1.4982260
#> 75   0.124071210  -0.8377069    1.2155033
#> 76   0.284629240  -1.5778724    2.4077965
#> 77   0.125548295  -1.0071044    1.0081606
#> 78   0.360493761  -1.6793138    2.2460407
#> 79   0.073697869  -1.0031636    1.0723223
#> 80   0.066948997  -0.8794953    1.0725613
#> 81   0.192213173  -0.8836572    1.2093108
#> 82   0.382006956  -1.7738152    2.2069168
#> 83  -0.049473313  -1.9524202    2.1018183
#> 84   0.121383488  -0.9201677    1.1301205
#> 85   0.449228698  -0.5605154    1.5148692
#> 86   0.137883076  -1.0183715    1.0739651
#> 87   0.032980507  -0.9982398    1.1024936
#> 88   0.068984348  -0.9553385    1.0741132
#> 89   0.119608978  -1.5560113    1.5359604
#> 90   0.334710247  -0.7309236    1.3379186
#> 91   0.198974020  -0.8828116    1.2644419
#> 92   0.008084194  -1.0500263    1.0361447
#> 93   0.062902606  -1.0236601    1.0570292
#> 94   0.136536228  -0.8767471    1.1930781
#> 95   0.101227825  -0.9336732    1.1059223
#> 96   0.036956538  -0.9086187    1.0494114
#> 97   0.064072802  -0.9391502    1.1288520
#> 98   0.063565195  -0.9900079    1.1079006
#> 99   0.030044471  -1.0690513    0.9894717
#> 100  0.154954920  -0.8014870    1.1064331
with(p4,cbind(LS,ES,APE,CR))
#>     Log.Score Energy.Score    CCR.APE CCR.CoverageRate
#> 1   1.1047645    0.6356654  110.15411                1
#> 2   0.5462159    0.4608460   63.69492                1
#> 3   0.6538700    0.4902708  100.52079                1
#> 4   0.5013662    0.4294784   46.56759                1
#> 5   0.5097445    0.4332795   23.05281                1
#> 6   0.8358752    0.5504941  126.92157                1
#> 7   0.5618847    0.4606419  217.73626                1
#> 8   1.2125343    0.6749869   98.81148                1
#> 9   0.5510491    0.4432830   73.30541                1
#> 10  2.2511620    1.1342603  117.08646                0
#> 11  0.6676749    0.4755175  160.95096                1
#> 12  0.6184898    0.4941009  107.32377                1
#> 13  2.1515602    1.0806642  105.73754                0
#> 14  1.2675510    0.7130733   89.77204                1
#> 15  1.5442204    0.7848069   91.00661                1
#> 16  0.9757898    0.5993090   80.92819                1
#> 17  0.6130874    0.4731575   56.44006                1
#> 18  0.5784063    0.4616364   64.26547                1
#> 19  0.5137288    0.4522142  286.18593                1
#> 20  0.8488845    0.5429137  129.26439                1
#> 21  0.5854574    0.4610401  141.70306                1
#> 22  1.4002234    0.7397944  106.95286                1
#> 23  0.5271780    0.4543331   76.65510                1
#> 24  1.0608999    0.6394408   90.32112                1
#> 25  0.9584393    0.6081175   84.98015                1
#> 26  0.8031465    0.5527638  126.02995                1
#> 27  0.5110832    0.4275993  957.69858                1
#> 28  1.0882499    0.6441716   96.05097                1
#> 29  0.5151015    0.4444722   35.41340                1
#> 30  0.7092024    0.5117710  173.95378                1
#> 31  1.0601775    0.6191495  115.74025                1
#> 32  0.6188484    0.4995855   88.36465                1
#> 33  0.5207329    0.4407263  167.14835                1
#> 34  3.7956150    2.0453580  108.26369                0
#> 35  1.9253131    0.9808042   91.80451                0
#> 36  1.2349617    0.8414011   61.94714                1
#> 37  0.5834448    0.4594139  564.44118                1
#> 38  0.6600690    0.5064127   49.13328                1
#> 39  0.7956470    0.5553987  109.02929                1
#> 40  0.5077296    0.4607875  777.84565                1
#> 41  0.5807651    0.4461033  148.61000                1
#> 42  0.9472602    0.5995491  119.46410                1
#> 43  0.9113847    0.5837186  102.12408                1
#> 44  1.6127955    0.8431076  103.85030                1
#> 45  1.1933596    0.6731641  125.49077                1
#> 46  2.0944648    1.4716113  125.29804                1
#> 47  3.4091843    1.7599208   97.28371                0
#> 48  1.2158415    0.9243230   38.65048                1
#> 49  1.1167005    0.6457988  182.86921                1
#> 50  0.5706463    0.4727229 3518.38275                1
#> 51  2.8031995    1.4189252  101.40659                0
#> 52  0.5048895    0.4660065   24.89333                1
#> 53  1.1837448    0.9073466  155.67380                1
#> 54  0.9269331    0.5734301   88.68216                1
#> 55  0.8089836    0.5355578  124.74519                1
#> 56  1.0500849    0.6370379   87.77899                1
#> 57  1.9795721    1.0000127  103.27817                0
#> 58  1.1485132    0.6550728   80.88035                1
#> 59  1.2946423    0.9382591   48.55186                1
#> 60  1.1774434    0.6943237  122.64813                1
#> 61  1.1330150    0.6602837  157.51530                1
#> 62  0.5059083    0.4533797   19.91505                1
#> 63  1.6662069    0.8411928   89.32021                1
#> 64  1.3642771    0.8415838   97.08951                1
#> 65  1.3900602    1.0181087  125.70570                1
#> 66  2.7931766    1.3920475  115.58913                0
#> 67  1.2531079    0.9094463   37.04168                1
#> 68  1.2614616    0.8967701   33.72953                1
#> 69  0.7617401    0.5112778  468.52103                1
#> 70  3.0888031    1.5788658  114.47203                0
#> 71  1.1872932    0.6891942  100.94142                1
#> 72  1.4322366    0.9843008   65.59521                1
#> 73  0.5930433    0.4702340  102.80206                1
#> 74  3.0067645    1.5114672  122.98455                0
#> 75  0.8360488    0.5478231  145.07437                1
#> 76  2.2517073    1.6806322  118.65945                1
#> 77  1.9004627    0.9465644   88.97549                0
#> 78  1.4579027    1.0122711  196.14158                1
#> 79  0.9006381    0.5747043   86.36368                1
#> 80  3.1974747    1.6712546  103.81270                0
#> 81  3.2536902    1.7059084  111.39837                0
#> 82  1.4315511    1.0118810  196.62324                1
#> 83  3.3365169    2.6079620  101.72291                0
#> 84  1.1197455    0.6607213   83.00659                1
#> 85  0.7162096    0.5127876  301.01459                1
#> 86  0.6810714    0.4941415   68.78349                1
#> 87  0.5095276    0.4507696   56.54497                1
#> 88  4.0629766    2.1850270   97.21523                0
#> 89  1.2068169    0.7987596   80.55471                1
#> 90  0.8661189    0.5580152  393.82021                1
#> 91  1.4387731    0.7819788   80.34215                1
#> 92  2.0927797    1.0444607   99.30987                0
#> 93  1.3301728    0.7109589   92.08185                1
#> 94  0.5382463    0.4557379   47.61932                1
#> 95  2.0173077    1.0125270   91.54216                0
#> 96  0.6847619    0.4849281  115.54620                1
#> 97  1.3374712    0.7380174  110.06105                1
#> 98  1.8802533    0.9707032   93.89133                1
#> 99  0.8450245    0.5572679  107.25461                1
#> 100 1.1117318    0.6377634   79.11639                1
plot(p4,last=100)
# }