Highest Posterior Density intervals for objects of class mtar

Highest Posterior Density intervals for objects of class mtar

Usage

# S3 method for class 'mtar'
HPDinterval(obj, prob = 0.95, ...)

Arguments

obj

an object of class mtar generated by a call to the function mtar().

prob

a numeric scalar in the interval \((0,1)\) giving the target probability content of the intervals. By default, prob is set to 0.95.

...

Optional additional arguments for methods. None are used at present.

See also

Examples

# \donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
             subset={Date<="2015-12-07"}, dist="Student-t",
             ars=ars(nregim=3,p=c(1,1,2)), n.burnin=1000, n.sim=2000,
             n.thin=2, ssvs=TRUE)
coda::HPDinterval(fit1)
#> 
#> Probability = 0.95
#> 
#> Thresholds:
#>                  lower      upper
#> Threshold.1 -0.0105420 -0.0049244
#> Threshold.2  0.0095292  0.0110304
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                             lower      upper
#> COLCAP:(Intercept)     -0.0097865 -0.0046935
#> BOVESPA:(Intercept)    -0.0175248 -0.0104906
#> COLCAP:COLCAP.lag(1)    0.0714910  0.4955079
#> BOVESPA:COLCAP.lag(1)  -0.3362183  0.1521801
#> COLCAP:BOVESPA.lag(1)  -0.0064880  0.2706587
#> BOVESPA:BOVESPA.lag(1) -0.0491111  0.3261144
#> 
#> 
#> Scale parameter:
#>                      lower      upper
#> COLCAP.COLCAP   5.7862e-05 1.0095e-04
#> COLCAP.BOVESPA  1.9654e-05 5.9056e-05
#> BOVESPA.BOVESPA 1.0877e-04 1.8919e-04
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                              lower      upper
#> COLCAP:(Intercept)     -0.00033615 0.00060734
#> BOVESPA:(Intercept)    -0.00106372 0.00047654
#> COLCAP:COLCAP.lag(1)    0.00818074 0.11625723
#> BOVESPA:COLCAP.lag(1)  -0.02070697 0.13605005
#> COLCAP:BOVESPA.lag(1)   0.04220101 0.11025610
#> BOVESPA:BOVESPA.lag(1) -0.09427521 0.01043800
#> 
#> 
#> Scale parameter:
#>                      lower      upper
#> COLCAP.COLCAP   3.7111e-05 4.7470e-05
#> COLCAP.BOVESPA  9.7003e-06 1.8399e-05
#> BOVESPA.BOVESPA 8.0153e-05 1.0202e-04
#> 
#> 
#> Regime 3
#> 
#> 
#> Autoregressive coefficients:
#>                             lower      upper
#> COLCAP:(Intercept)      0.0050632  0.0082017
#> BOVESPA:(Intercept)     0.0127336  0.0177791
#> COLCAP:COLCAP.lag(1)   -0.0704008  0.2227840
#> BOVESPA:COLCAP.lag(1)  -0.0635717  0.3455599
#> COLCAP:BOVESPA.lag(1)  -0.1371853  0.0996818
#> BOVESPA:BOVESPA.lag(1) -0.3955939 -0.0604130
#> COLCAP:COLCAP.lag(2)   -0.0755365  0.2143032
#> BOVESPA:COLCAP.lag(2)  -0.2657331  0.1382123
#> COLCAP:BOVESPA.lag(2)  -0.1589451  0.0413391
#> BOVESPA:BOVESPA.lag(2) -0.2036336  0.1165574
#> 
#> 
#> Scale parameter:
#>                      lower      upper
#> COLCAP.COLCAP   4.3426e-05 7.5928e-05
#> COLCAP.BOVESPA  6.4743e-06 3.8754e-05
#> BOVESPA.BOVESPA 9.7007e-05 1.6814e-04
#> 
#> 
#> Extra parameter:
#>     lower  upper
#> nu 4.5431 7.3165

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, row.names=Date,
             subset={Date<="2009-02-13"}, dist="Laplace",
             ars=ars(nregim=3,p=5), n.burnin=1000, n.sim=2000, n.thin=2)
coda::HPDinterval(fit2)
#> 
#> Probability = 0.95
#> 
#> Thresholds:
#>               lower   upper
#> Threshold.1  3.0287  3.8155
#> Threshold.2 10.0000 10.0149
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                            lower      upper
#> Bedon:(Intercept)       1.108891 1.53216226
#> LaPlata:(Intercept)     2.860044 4.04402149
#> Bedon:Bedon.lag(1)      0.482920 0.63969522
#> LaPlata:Bedon.lag(1)   -0.063305 0.34888959
#> Bedon:LaPlata.lag(1)    0.016203 0.07358646
#> LaPlata:LaPlata.lag(1)  0.554588 0.70845898
#> Bedon:Bedon.lag(2)     -0.019105 0.11604248
#> LaPlata:Bedon.lag(2)   -0.237660 0.13904147
#> Bedon:LaPlata.lag(2)   -0.045098 0.00503449
#> LaPlata:LaPlata.lag(2) -0.136166 0.00093774
#> Bedon:Bedon.lag(3)     -0.032472 0.09042262
#> LaPlata:Bedon.lag(3)   -0.139796 0.16376737
#> Bedon:LaPlata.lag(3)   -0.017717 0.02469308
#> LaPlata:LaPlata.lag(3)  0.012149 0.11990613
#> Bedon:Bedon.lag(4)     -0.021334 0.10134620
#> LaPlata:Bedon.lag(4)   -0.259139 0.05819055
#> Bedon:LaPlata.lag(4)   -0.032318 0.00155077
#> LaPlata:LaPlata.lag(4) -0.045775 0.05922443
#> Bedon:Bedon.lag(5)      0.031486 0.13375885
#> LaPlata:Bedon.lag(5)    0.027727 0.27832650
#> Bedon:LaPlata.lag(5)   -0.019542 0.00777224
#> LaPlata:LaPlata.lag(5) -0.017927 0.06704365
#> 
#> 
#> Scale parameter:
#>                   lower   upper
#> Bedon.Bedon     0.26328 0.38948
#> Bedon.LaPlata   0.25262 0.48966
#> LaPlata.LaPlata 1.89676 2.77788
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                             lower     upper
#> Bedon:(Intercept)       1.3194596  2.916638
#> LaPlata:(Intercept)     4.8508130  8.910132
#> Bedon:Bedon.lag(1)      0.4953451  0.671637
#> LaPlata:Bedon.lag(1)   -0.1002882  0.363021
#> Bedon:LaPlata.lag(1)   -0.0077359  0.047793
#> LaPlata:LaPlata.lag(1)  0.4452416  0.599067
#> Bedon:Bedon.lag(2)     -0.0290426  0.213174
#> LaPlata:Bedon.lag(2)   -0.2339122  0.253258
#> Bedon:LaPlata.lag(2)   -0.0543574  0.014471
#> LaPlata:LaPlata.lag(2) -0.0356174  0.108142
#> Bedon:Bedon.lag(3)     -0.1465706  0.067465
#> LaPlata:Bedon.lag(3)   -0.2800328  0.165803
#> Bedon:LaPlata.lag(3)   -0.0396706  0.021091
#> LaPlata:LaPlata.lag(3) -0.0282009  0.111385
#> Bedon:Bedon.lag(4)     -0.0039448  0.221469
#> LaPlata:Bedon.lag(4)   -0.0148561  0.490533
#> Bedon:LaPlata.lag(4)   -0.0293677  0.040315
#> LaPlata:LaPlata.lag(4) -0.1298770  0.038449
#> Bedon:Bedon.lag(5)     -0.0629416  0.097919
#> LaPlata:Bedon.lag(5)   -0.4876191 -0.092673
#> Bedon:LaPlata.lag(5)   -0.0259730  0.031472
#> LaPlata:LaPlata.lag(5)  0.0437830  0.197992
#> 
#> 
#> Scale parameter:
#>                   lower  upper
#> Bedon.Bedon     0.87758 1.3019
#> Bedon.LaPlata   0.97158 1.7093
#> LaPlata.LaPlata 5.29540 7.8042
#> 
#> 
#> Regime 3
#> 
#> 
#> Autoregressive coefficients:
#>                             lower     upper
#> Bedon:(Intercept)       3.9894134  7.294559
#> LaPlata:(Intercept)    11.1902226 23.020733
#> Bedon:Bedon.lag(1)      0.3082183  0.643302
#> LaPlata:Bedon.lag(1)    0.0618211  1.074497
#> Bedon:LaPlata.lag(1)    0.0093636  0.076327
#> LaPlata:LaPlata.lag(1)  0.2114032  0.464734
#> Bedon:Bedon.lag(2)     -0.0640751  0.223012
#> LaPlata:Bedon.lag(2)   -1.1002530 -0.019625
#> Bedon:LaPlata.lag(2)   -0.0375335  0.031024
#> LaPlata:LaPlata.lag(2) -0.0098455  0.260123
#> Bedon:Bedon.lag(3)     -0.2215945  0.040573
#> LaPlata:Bedon.lag(3)   -1.0350627 -0.113238
#> Bedon:LaPlata.lag(3)   -0.0021404  0.073491
#> LaPlata:LaPlata.lag(3)  0.1369960  0.447739
#> Bedon:Bedon.lag(4)     -0.1471890  0.137477
#> LaPlata:Bedon.lag(4)   -0.5163643  0.626348
#> Bedon:LaPlata.lag(4)   -0.0369771  0.046311
#> LaPlata:LaPlata.lag(4) -0.1695370  0.140870
#> Bedon:Bedon.lag(5)      0.0371031  0.309065
#> LaPlata:Bedon.lag(5)   -0.1978915  0.788498
#> Bedon:LaPlata.lag(5)   -0.0475675  0.022435
#> LaPlata:LaPlata.lag(5) -0.0688179  0.196537
#> 
#> 
#> Scale parameter:
#>                   lower   upper
#> Bedon.Bedon      2.2380  3.3653
#> Bedon.LaPlata    5.3841  8.9079
#> LaPlata.LaPlata 33.7646 52.2459

###### Example 3: Temperature, precipitation, and two river flows in Iceland
data(iceland.rf)
fit3 <- mtar(~ Jokulsa + Vatnsdalsa | Temperature | Precipitation,
             data=iceland.rf, subset={Date<="1974-11-06"}, row.names=Date,
             ars=ars(nregim=2,p=15,q=4,d=2), n.burnin=1000, n.sim=2000,
             n.thin=2, dist="Slash")
coda::HPDinterval(fit3)
#> 
#> Probability = 0.95
#> 
#> Thresholds:
#>              lower  upper
#> Threshold.1 1.1008 1.2955
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                                       lower       upper
#> Jokulsa:(Intercept)              3.01216112  4.44910666
#> Vatnsdalsa:(Intercept)           0.44668018  1.22930933
#> Jokulsa:Jokulsa.lag( 1)          0.77506393  0.91634416
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.10235444 -0.02407218
#> Jokulsa:Vatnsdalsa.lag( 1)       0.09998825  0.32709513
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.08389466  1.24401168
#> Jokulsa:Jokulsa.lag( 2)         -0.09951992 -0.00281991
#> Vatnsdalsa:Jokulsa.lag( 2)       0.01689143  0.08022687
#> Jokulsa:Vatnsdalsa.lag( 2)      -0.30479455 -0.04746716
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.38602776 -0.21802254
#> Jokulsa:Jokulsa.lag( 3)         -0.03149461  0.04139062
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.04881667 -0.00140318
#> Jokulsa:Vatnsdalsa.lag( 3)      -0.03449313  0.09827377
#> Vatnsdalsa:Vatnsdalsa.lag( 3)   -0.01611032  0.08225878
#> Jokulsa:Jokulsa.lag( 4)         -0.04499217  0.04141711
#> Vatnsdalsa:Jokulsa.lag( 4)      -0.01165285  0.04086384
#> Jokulsa:Vatnsdalsa.lag( 4)      -0.06259357  0.08832019
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -0.05280651  0.05646114
#> Jokulsa:Jokulsa.lag( 5)         -0.05363857  0.04941220
#> Vatnsdalsa:Jokulsa.lag( 5)      -0.02085180  0.03710989
#> Jokulsa:Vatnsdalsa.lag( 5)      -0.10744124  0.02592505
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -0.06351926  0.02606703
#> Jokulsa:Jokulsa.lag( 6)         -0.02955569  0.07324768
#> Vatnsdalsa:Jokulsa.lag( 6)      -0.02457565  0.03424973
#> Jokulsa:Vatnsdalsa.lag( 6)      -0.07929339  0.04304337
#> Vatnsdalsa:Vatnsdalsa.lag( 6)   -0.03622490  0.04064777
#> Jokulsa:Jokulsa.lag( 7)         -0.04782558  0.05946626
#> Vatnsdalsa:Jokulsa.lag( 7)      -0.03190665  0.02715151
#> Jokulsa:Vatnsdalsa.lag( 7)      -0.04203596  0.07437360
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -0.02836016  0.04507614
#> Jokulsa:Jokulsa.lag( 8)         -0.05281753  0.05013904
#> Vatnsdalsa:Jokulsa.lag( 8)      -0.03412357  0.01741990
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.06173290  0.05484411
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -0.02826708  0.04573722
#> Jokulsa:Jokulsa.lag( 9)         -0.05338480  0.03580947
#> Vatnsdalsa:Jokulsa.lag( 9)      -0.00761687  0.05052029
#> Jokulsa:Vatnsdalsa.lag( 9)      -0.07209249  0.05646874
#> Vatnsdalsa:Vatnsdalsa.lag( 9)   -0.04564142  0.03617924
#> Jokulsa:Jokulsa.lag(10)         -0.00742436  0.06156446
#> Vatnsdalsa:Jokulsa.lag(10)      -0.03706139  0.00776789
#> Jokulsa:Vatnsdalsa.lag(10)      -0.03125298  0.09160990
#> Vatnsdalsa:Vatnsdalsa.lag(10)   -0.02026314  0.06519512
#> Jokulsa:Jokulsa.lag(11)         -0.04241130  0.01764877
#> Vatnsdalsa:Jokulsa.lag(11)      -0.01023831  0.02735578
#> Jokulsa:Vatnsdalsa.lag(11)      -0.07431870  0.03723784
#> Vatnsdalsa:Vatnsdalsa.lag(11)   -0.04663509  0.03196626
#> Jokulsa:Jokulsa.lag(12)         -0.01950480  0.03958562
#> Vatnsdalsa:Jokulsa.lag(12)      -0.02673506  0.01147259
#> Jokulsa:Vatnsdalsa.lag(12)      -0.04277132  0.06048036
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.03745907  0.03180228
#> Jokulsa:Jokulsa.lag(13)         -0.05358290  0.01440206
#> Vatnsdalsa:Jokulsa.lag(13)      -0.01621208  0.02280159
#> Jokulsa:Vatnsdalsa.lag(13)      -0.06514952  0.03727700
#> Vatnsdalsa:Vatnsdalsa.lag(13)   -0.06104591  0.01884027
#> Jokulsa:Jokulsa.lag(14)         -0.02202848  0.03331771
#> Vatnsdalsa:Jokulsa.lag(14)      -0.02354132  0.01275104
#> Jokulsa:Vatnsdalsa.lag(14)      -0.04550223  0.06128400
#> Vatnsdalsa:Vatnsdalsa.lag(14)   -0.00351401  0.07006321
#> Jokulsa:Jokulsa.lag(15)         -0.00781469  0.04756129
#> Vatnsdalsa:Jokulsa.lag(15)      -0.01388953  0.01514017
#> Jokulsa:Vatnsdalsa.lag(15)      -0.05336082  0.02749317
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -0.02156065  0.03381161
#> Jokulsa:Precipitation.lag(1)    -0.01121017  0.02457475
#> Vatnsdalsa:Precipitation.lag(1) -0.00525892  0.01730374
#> Jokulsa:Precipitation.lag(2)    -0.00878198  0.01835785
#> Vatnsdalsa:Precipitation.lag(2) -0.00888620  0.00843741
#> Jokulsa:Precipitation.lag(3)    -0.02269249 -0.00078006
#> Vatnsdalsa:Precipitation.lag(3) -0.01191608  0.00341215
#> Jokulsa:Precipitation.lag(4)     0.00631820  0.03176934
#> Vatnsdalsa:Precipitation.lag(4) -0.00493681  0.01360282
#> Jokulsa:Temperature.lag(1)       0.00067557  0.04277496
#> Vatnsdalsa:Temperature.lag(1)   -0.01051131  0.01609708
#> Jokulsa:Temperature.lag(2)      -0.05885582 -0.01694403
#> Vatnsdalsa:Temperature.lag(2)   -0.02557699  0.00072988
#> 
#> 
#> Scale parameter:
#>                           lower    upper
#> Jokulsa.Jokulsa       0.0487295 0.086925
#> Jokulsa.Vatnsdalsa    0.0057223 0.016686
#> Vatnsdalsa.Vatnsdalsa 0.0205724 0.036612
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                                      lower       upper
#> Jokulsa:(Intercept)             -1.5723999  1.17426201
#> Vatnsdalsa:(Intercept)           0.3176978  0.68149696
#> Jokulsa:Jokulsa.lag( 1)          0.9468790  1.08749796
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.0096881  0.00479818
#> Jokulsa:Vatnsdalsa.lag( 1)       0.4146019  1.35134422
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.1046851  1.24589024
#> Jokulsa:Jokulsa.lag( 2)         -0.3015312 -0.05110430
#> Vatnsdalsa:Jokulsa.lag( 2)      -0.0023161  0.02199606
#> Jokulsa:Vatnsdalsa.lag( 2)      -1.0766025  0.30767417
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.4416007 -0.25139746
#> Jokulsa:Jokulsa.lag( 3)         -0.0953003  0.12478641
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.0241543  0.00054486
#> Jokulsa:Vatnsdalsa.lag( 3)      -0.6221723  0.68931177
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    0.1058766  0.27674337
#> Jokulsa:Jokulsa.lag( 4)         -0.1603535  0.01032929
#> Vatnsdalsa:Jokulsa.lag( 4)      -0.0037103  0.01476837
#> Jokulsa:Vatnsdalsa.lag( 4)      -0.6506232  0.29530082
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -0.1592454 -0.00126648
#> Jokulsa:Jokulsa.lag( 5)         -0.0385360  0.11645162
#> Vatnsdalsa:Jokulsa.lag( 5)      -0.0149819  0.00371398
#> Jokulsa:Vatnsdalsa.lag( 5)      -0.6203686  0.62775905
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -0.0857835  0.10574123
#> Jokulsa:Jokulsa.lag( 6)         -0.1052252  0.01974520
#> Vatnsdalsa:Jokulsa.lag( 6)      -0.0047613  0.01193621
#> Jokulsa:Vatnsdalsa.lag( 6)      -0.4951375  0.71809928
#> Vatnsdalsa:Vatnsdalsa.lag( 6)   -0.0707261  0.12346937
#> Jokulsa:Jokulsa.lag( 7)         -0.0602005  0.05638895
#> Vatnsdalsa:Jokulsa.lag( 7)      -0.0140332  0.00274295
#> Jokulsa:Vatnsdalsa.lag( 7)      -0.3470259  0.54083718
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -0.1184494  0.00963466
#> Jokulsa:Jokulsa.lag( 8)         -0.0407069  0.07351912
#> Vatnsdalsa:Jokulsa.lag( 8)      -0.0037274  0.01195866
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.6947072  0.21922124
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -0.1042657  0.02067240
#> Jokulsa:Jokulsa.lag( 9)         -0.0199993  0.10376144
#> Vatnsdalsa:Jokulsa.lag( 9)      -0.0098786  0.00672259
#> Jokulsa:Vatnsdalsa.lag( 9)      -0.2992893  0.55651690
#> Vatnsdalsa:Vatnsdalsa.lag( 9)    0.0174975  0.16059199
#> Jokulsa:Jokulsa.lag(10)         -0.0984319  0.05595441
#> Vatnsdalsa:Jokulsa.lag(10)      -0.0060387  0.01295489
#> Jokulsa:Vatnsdalsa.lag(10)      -0.4075627  0.36245157
#> Vatnsdalsa:Vatnsdalsa.lag(10)   -0.1261248 -0.01086358
#> Jokulsa:Jokulsa.lag(11)         -0.0790049  0.07680963
#> Vatnsdalsa:Jokulsa.lag(11)      -0.0158582  0.00231287
#> Jokulsa:Vatnsdalsa.lag(11)      -0.5615513  0.45983671
#> Vatnsdalsa:Vatnsdalsa.lag(11)    0.0076393  0.14697657
#> Jokulsa:Jokulsa.lag(12)         -0.0788102  0.06739931
#> Vatnsdalsa:Jokulsa.lag(12)       0.0010255  0.01800121
#> Jokulsa:Vatnsdalsa.lag(12)      -0.4693305  0.46357053
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.1430989 -0.02646569
#> Jokulsa:Jokulsa.lag(13)         -0.0890793  0.07088979
#> Vatnsdalsa:Jokulsa.lag(13)      -0.0143842  0.00160358
#> Jokulsa:Vatnsdalsa.lag(13)      -0.0823032  0.90049477
#> Vatnsdalsa:Vatnsdalsa.lag(13)    0.0725195  0.21115161
#> Jokulsa:Jokulsa.lag(14)         -0.0856602  0.06853469
#> Vatnsdalsa:Jokulsa.lag(14)      -0.0105884  0.00626477
#> Jokulsa:Vatnsdalsa.lag(14)      -0.3887712  0.63335348
#> Vatnsdalsa:Vatnsdalsa.lag(14)   -0.1308672  0.02422233
#> Jokulsa:Jokulsa.lag(15)          0.0040650  0.09222334
#> Vatnsdalsa:Jokulsa.lag(15)      -0.0045390  0.00647517
#> Jokulsa:Vatnsdalsa.lag(15)      -0.7344939 -0.11163292
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -0.0644018  0.03515953
#> Jokulsa:Precipitation.lag(1)    -0.1922928 -0.03924551
#> Vatnsdalsa:Precipitation.lag(1) -0.0128229  0.00751237
#> Jokulsa:Precipitation.lag(2)    -0.0992820  0.15022712
#> Vatnsdalsa:Precipitation.lag(2) -0.0157662  0.01159919
#> Jokulsa:Precipitation.lag(3)    -0.0162778  0.11219349
#> Vatnsdalsa:Precipitation.lag(3) -0.0034313  0.01637317
#> Jokulsa:Precipitation.lag(4)    -0.0394901  0.09785084
#> Vatnsdalsa:Precipitation.lag(4) -0.0059753  0.01150767
#> Jokulsa:Temperature.lag(1)       0.9292779  1.31060648
#> Vatnsdalsa:Temperature.lag(1)   -0.0028265  0.04685542
#> Jokulsa:Temperature.lag(2)      -0.7683949 -0.35371213
#> Vatnsdalsa:Temperature.lag(2)   -0.0536449 -0.00166153
#> 
#> 
#> Scale parameter:
#>                          lower    upper
#> Jokulsa.Jokulsa       0.959662 1.781023
#> Jokulsa.Vatnsdalsa    0.024084 0.069135
#> Vatnsdalsa.Vatnsdalsa 0.015911 0.030668
#> 
#> 
#> Extra parameter:
#>      lower   upper
#> nu 0.72757 0.89162

###### Example 4: U.S. stock returns
data(US.returns)
fit4 <- mtar(~ CCR | dVIX, data=US.returns, subset={Date<="2025-11-28"},
             row.names=Date, ars=ars(nregim=2,p=3,d=3), n.burnin=1000,
             n.sim=2000, n.thin=2, dist="Student-t")
coda::HPDinterval(fit4)
#> 
#> Probability = 0.95
#> 
#> Thresholds:
#>              lower  upper
#> Threshold.1 1.1487 2.2008
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                      lower       upper
#> CCR:(Intercept)  0.0677208  0.11502896
#> CCR:CCR.lag(1)  -0.0780972 -0.02240717
#> CCR:CCR.lag(2)  -0.0826999  0.00020929
#> CCR:CCR.lag(3)  -0.0664100  0.01755325
#> CCR:dVIX.lag(1) -0.0627957 -0.00958405
#> CCR:dVIX.lag(2) -0.0486701  0.00520388
#> CCR:dVIX.lag(3) -0.0011461  0.03140187
#> 
#> 
#> Scale parameter:
#>           lower   upper
#> CCR.CCR 0.33502 0.40059
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                      lower     upper
#> CCR:(Intercept) -0.4090450 0.2460139
#> CCR:CCR.lag(1)  -0.3216915 0.0067886
#> CCR:CCR.lag(2)  -0.2320086 0.1935572
#> CCR:CCR.lag(3)  -0.0481479 0.3456160
#> CCR:dVIX.lag(1) -0.1451481 0.1306261
#> CCR:dVIX.lag(2) -0.0804351 0.1579721
#> CCR:dVIX.lag(3) -0.0019516 0.1317174
#> 
#> 
#> Scale parameter:
#>           lower  upper
#> CCR.CCR 0.73511 1.2924
#> 
#> 
#> Extra parameter:
#>     lower  upper
#> nu 2.2401 2.6517
# }