Coercion of mtar objects to mcmc objects

This method converts an object of class mtar into a list of mcmc objects, each corresponding to a Markov chain produced during Bayesian estimation.

Usage

# S3 method for class 'mtar'
as.mcmc(x, ...)

Arguments

x

an object of class mtar obtained from a call to mtar().

...

additional arguments passed to specific coercion methods.

Value

A list of mcmc objects containing the posterior simulation draws generated by the mtar() routine.

See also

Examples

# \donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
             subset={Date<="2015-12-07"}, dist="Student-t",
             ars=ars(nregim=3,p=c(1,1,2)), n.burnin=1000, n.sim=2000,
             n.thin=2, ssvs=TRUE)
fit1.mcmc <- coda::as.mcmc(fit1)
summary(fit1.mcmc)
#> 
#> 
#>  Iterations = 1001:4999
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 2000
#> 
#> 
#> Thresholds:
#>                   Mean         Sd  Sd(Mean)       2.5%       25%        50%
#> Threshold.1 -0.0101589 0.00069974 0.0001697 -0.0106760 -0.010497 -0.0103215
#> Threshold.2  0.0095336 0.00145886 0.0013751  0.0075593  0.007743  0.0097311
#>                   75%      97.5%
#> Threshold.1 -0.010045 -0.0078729
#> Threshold.2  0.010903  0.0110353
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                             Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP:(Intercept)     -0.008057 0.00091589 0.00008155 -0.0097084 -0.0086718
#> COLCAP:COLCAP.lag(1)    0.299391 0.10516081 0.00282932  0.0785726  0.2326173
#> COLCAP:BOVESPA.lag(1)   0.130708 0.07093882 0.00247929 -0.0052205  0.0810115
#> BOVESPA:(Intercept)    -0.015261 0.00124969 0.00010732 -0.0175112 -0.0160771
#> BOVESPA:COLCAP.lag(1)  -0.101359 0.12312909 0.00293981 -0.3313455 -0.1856808
#> BOVESPA:BOVESPA.lag(1)  0.153742 0.09237879 0.00270247 -0.0268622  0.0908962
#>                               50%        75%      97.5%
#> COLCAP:(Intercept)     -0.0080916 -0.0074938 -0.0060299
#> COLCAP:COLCAP.lag(1)    0.2993499  0.3733060  0.5009554
#> COLCAP:BOVESPA.lag(1)   0.1304052  0.1769775  0.2683950
#> BOVESPA:(Intercept)    -0.0153258 -0.0145095 -0.0124424
#> BOVESPA:COLCAP.lag(1)  -0.1030126 -0.0166265  0.1373350
#> BOVESPA:BOVESPA.lag(1)  0.1544796  0.2164015  0.3335205
#> 
#> 
#> Scale parameter:
#>                       Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP.COLCAP   8.0064e-05 1.1409e-05 3.1039e-07 6.0509e-05 7.2159e-05
#> COLCAP.BOVESPA  3.9890e-05 1.0535e-05 2.6015e-07 2.1320e-05 3.2467e-05
#> BOVESPA.BOVESPA 1.4590e-04 2.1585e-05 7.8373e-07 1.0903e-04 1.3089e-04
#>                        50%        75%      97.5%
#> COLCAP.COLCAP   7.9314e-05 8.6537e-05 1.0518e-04
#> COLCAP.BOVESPA  3.9316e-05 4.6576e-05 6.2513e-05
#> BOVESPA.BOVESPA 1.4433e-04 1.5960e-04 1.9350e-04
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                               Mean         Sd   Sd(Mean)        2.5%
#> COLCAP:(Intercept)      0.00001823 0.00024403 2.5771e-05 -0.00045551
#> COLCAP:COLCAP.lag(1)    0.06879896 0.02961210 1.1323e-03  0.01169162
#> COLCAP:BOVESPA.lag(1)   0.07559660 0.01781444 5.9594e-04  0.04022774
#> BOVESPA:(Intercept)    -0.00066070 0.00039346 8.1600e-05 -0.00142813
#> BOVESPA:COLCAP.lag(1)   0.05227204 0.04032652 1.0632e-03 -0.02696929
#> BOVESPA:BOVESPA.lag(1) -0.03894794 0.02654350 6.4531e-04 -0.09076353
#>                                25%         50%         75%      97.5%
#> COLCAP:(Intercept)     -0.00014283  2.0624e-05  0.00018436 5.0088e-04
#> COLCAP:COLCAP.lag(1)    0.04880248  6.8406e-02  0.08841167 1.2655e-01
#> COLCAP:BOVESPA.lag(1)   0.06319044  7.5875e-02  0.08776550 1.1059e-01
#> BOVESPA:(Intercept)    -0.00093138 -6.4425e-04 -0.00040060 8.3544e-05
#> BOVESPA:COLCAP.lag(1)   0.02626675  5.2025e-02  0.07823156 1.3278e-01
#> BOVESPA:BOVESPA.lag(1) -0.05720061 -3.8763e-02 -0.02114904 1.2094e-02
#> 
#> 
#> Scale parameter:
#>                       Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP.COLCAP   4.2022e-05 2.5901e-06 1.0909e-07 3.6748e-05 4.0281e-05
#> COLCAP.BOVESPA  1.3904e-05 2.2225e-06 7.3745e-08 9.7078e-06 1.2405e-05
#> BOVESPA.BOVESPA 9.0438e-05 5.5865e-06 3.6094e-07 8.0003e-05 8.6541e-05
#>                        50%        75%      97.5%
#> COLCAP.COLCAP   4.1976e-05 4.3700e-05 4.7269e-05
#> COLCAP.BOVESPA  1.3868e-05 1.5406e-05 1.8492e-05
#> BOVESPA.BOVESPA 9.0276e-05 9.4205e-05 1.0149e-04
#> 
#> 
#> Regime 3
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean        Sd   Sd(Mean)       2.5%        25%
#> COLCAP:(Intercept)      0.0058897 0.0010418 0.00052202  0.0041176  0.0050728
#> COLCAP:COLCAP.lag(1)    0.0654752 0.0724501 0.00259055 -0.0692502  0.0171677
#> COLCAP:BOVESPA.lag(1)   0.0199495 0.0684536 0.02057968 -0.1184985 -0.0263710
#> COLCAP:COLCAP.lag(2)    0.0696672 0.0694223 0.00474183 -0.0618251  0.0206742
#> COLCAP:BOVESPA.lag(2)  -0.0692949 0.0496381 0.00462381 -0.1657904 -0.1022962
#> BOVESPA:(Intercept)     0.0136651 0.0020539 0.00149569  0.0102386  0.0119022
#> BOVESPA:COLCAP.lag(1)   0.1582281 0.1023490 0.00250665 -0.0432890  0.0924246
#> BOVESPA:BOVESPA.lag(1) -0.1699800 0.0945461 0.01888720 -0.3589328 -0.2346333
#> BOVESPA:COLCAP.lag(2)  -0.0585460 0.1065381 0.00350170 -0.2653610 -0.1329735
#> BOVESPA:BOVESPA.lag(2) -0.0530023 0.0773159 0.00703171 -0.2055121 -0.1044594
#>                               50%         75%     97.5%
#> COLCAP:(Intercept)      0.0058098  6.6896e-03 0.0078994
#> COLCAP:COLCAP.lag(1)    0.0606849  1.1407e-01 0.2110387
#> COLCAP:BOVESPA.lag(1)   0.0241347  7.0093e-02 0.1430249
#> COLCAP:COLCAP.lag(2)    0.0688236  1.1384e-01 0.2146507
#> COLCAP:BOVESPA.lag(2)  -0.0708959 -3.7057e-02 0.0319110
#> BOVESPA:(Intercept)     0.0136780  1.5406e-02 0.0172490
#> BOVESPA:COLCAP.lag(1)   0.1562439  2.2596e-01 0.3648904
#> BOVESPA:BOVESPA.lag(1) -0.1637307 -1.0638e-01 0.0090830
#> BOVESPA:COLCAP.lag(2)  -0.0604384  1.4415e-02 0.1525360
#> BOVESPA:BOVESPA.lag(2) -0.0533851 -9.8014e-05 0.1019390
#> 
#> 
#> Scale parameter:
#>                       Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP.COLCAP   5.7644e-05 8.1169e-06 4.4589e-07 4.4046e-05 5.1727e-05
#> COLCAP.BOVESPA  2.3825e-05 8.0050e-06 6.7997e-07 8.6921e-06 1.8570e-05
#> BOVESPA.BOVESPA 1.3419e-04 1.7919e-05 4.9682e-07 1.0320e-04 1.2192e-04
#>                        50%        75%      97.5%
#> COLCAP.COLCAP   5.6899e-05 6.2463e-05 7.5482e-05
#> COLCAP.BOVESPA  2.3762e-05 2.9026e-05 4.0611e-05
#> BOVESPA.BOVESPA 1.3280e-04 1.4502e-04 1.7468e-04
#> 
#> 
#> Extra parameter:
#>      Mean     Sd Sd(Mean)   2.5%    25%    50%    75%  97.5%
#> nu 5.8937 0.6827 0.051235 4.6901 5.3859 5.8367 6.3194 7.4049
#plot(fit1.mcmc)

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, row.names=Date,
             subset={Date<="2009-02-13"}, dist="Laplace",
             ars=ars(nregim=3,p=5), n.burnin=1000, n.sim=2000, n.thin=2)
fit2.mcmc <- coda::as.mcmc(fit2)
summary(fit2.mcmc)
#> 
#> 
#>  Iterations = 1001:4999
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 2000
#> 
#> 
#> Thresholds:
#>                Mean       Sd   Sd(Mean)    2.5%     25%     50%     75%   97.5%
#> Threshold.1  3.3716 0.189472 0.03758168  3.0944  3.2181  3.3314  3.5061  3.6988
#> Threshold.2 10.0087 0.009892 0.00071107 10.0011 10.0038 10.0090 10.0124 10.0166
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean        Sd   Sd(Mean)       2.5%        25%
#> Bedon:(Intercept)       1.3244528 0.1103459 0.00320215  1.1052100  1.2526895
#> Bedon:Bedon.lag(1)      0.5627382 0.0390357 0.00131585  0.4831818  0.5368784
#> Bedon:LaPlata.lag(1)    0.0470131 0.0148443 0.00046157  0.0163573  0.0376313
#> Bedon:Bedon.lag(2)      0.0509047 0.0361121 0.00117251 -0.0205603  0.0272716
#> Bedon:LaPlata.lag(2)   -0.0213700 0.0126407 0.00043956 -0.0464701 -0.0295998
#> Bedon:Bedon.lag(3)      0.0286610 0.0312549 0.00103200 -0.0308281  0.0071364
#> Bedon:LaPlata.lag(3)    0.0030908 0.0112715 0.00037344 -0.0197082 -0.0047712
#> Bedon:Bedon.lag(4)      0.0336914 0.0316308 0.00109269 -0.0266598  0.0124165
#> Bedon:LaPlata.lag(4)   -0.0149756 0.0087284 0.00024845 -0.0318335 -0.0208009
#> Bedon:Bedon.lag(5)      0.0829710 0.0265156 0.00101662  0.0314746  0.0648580
#> Bedon:LaPlata.lag(5)   -0.0070203 0.0067696 0.00020499 -0.0202293 -0.0117757
#> LaPlata:(Intercept)     3.4644140 0.3079626 0.00922825  2.8802981  3.2583113
#> LaPlata:Bedon.lag(1)    0.1409616 0.1061126 0.00357759 -0.0615315  0.0670645
#> LaPlata:LaPlata.lag(1)  0.6368182 0.0402192 0.00136980  0.5595890  0.6082662
#> LaPlata:Bedon.lag(2)   -0.0395043 0.0929195 0.00295397 -0.2066937 -0.1044203
#> LaPlata:LaPlata.lag(2) -0.0729759 0.0343050 0.00107731 -0.1399159 -0.0963679
#> LaPlata:Bedon.lag(3)    0.0329106 0.0792311 0.00268402 -0.1247031 -0.0187641
#> LaPlata:LaPlata.lag(3)  0.0655060 0.0277672 0.00084240  0.0086508  0.0478279
#> LaPlata:Bedon.lag(4)   -0.1018056 0.0850024 0.00309801 -0.2704314 -0.1594722
#> LaPlata:LaPlata.lag(4)  0.0122715 0.0274356 0.00093798 -0.0428939 -0.0057089
#> LaPlata:Bedon.lag(5)    0.1520586 0.0639587 0.00222085  0.0269308  0.1090330
#> LaPlata:LaPlata.lag(5)  0.0233208 0.0225335 0.00082675 -0.0231201  0.0078568
#>                               50%        75%      97.5%
#> Bedon:(Intercept)       1.3209127  1.3960972  1.5551897
#> Bedon:Bedon.lag(1)      0.5637176  0.5898609  0.6360506
#> Bedon:LaPlata.lag(1)    0.0475284  0.0571285  0.0750042
#> Bedon:Bedon.lag(2)      0.0498182  0.0738744  0.1250183
#> Bedon:LaPlata.lag(2)   -0.0212799 -0.0130899  0.0031829
#> Bedon:Bedon.lag(3)      0.0280114  0.0500294  0.0908224
#> Bedon:LaPlata.lag(3)    0.0033993  0.0110162  0.0237728
#> Bedon:Bedon.lag(4)      0.0322682  0.0535865  0.0992529
#> Bedon:LaPlata.lag(4)   -0.0148641 -0.0092604  0.0025176
#> Bedon:Bedon.lag(5)      0.0829496  0.1020709  0.1350920
#> Bedon:LaPlata.lag(5)   -0.0070011 -0.0025662  0.0063919
#> LaPlata:(Intercept)     3.4532297  3.6522761  4.1024573
#> LaPlata:Bedon.lag(1)    0.1412666  0.2141034  0.3371317
#> LaPlata:LaPlata.lag(1)  0.6372904  0.6642141  0.7125159
#> LaPlata:Bedon.lag(2)   -0.0418004  0.0251359  0.1423103
#> LaPlata:LaPlata.lag(2) -0.0742485 -0.0504970 -0.0025322
#> LaPlata:Bedon.lag(3)    0.0337283  0.0851432  0.1900152
#> LaPlata:LaPlata.lag(3)  0.0660375  0.0838053  0.1178193
#> LaPlata:Bedon.lag(4)   -0.1037964 -0.0435607  0.0632860
#> LaPlata:LaPlata.lag(4)  0.0134829  0.0315637  0.0621205
#> LaPlata:Bedon.lag(5)    0.1530846  0.1967536  0.2779692
#> LaPlata:LaPlata.lag(5)  0.0238665  0.0388265  0.0646078
#> 
#> 
#> Scale parameter:
#>                    Mean       Sd  Sd(Mean)    2.5%     25%     50%     75%
#> Bedon.Bedon     0.32986 0.032767 0.0010158 0.27171 0.30692 0.32778 0.35021
#> Bedon.LaPlata   0.36946 0.058849 0.0015581 0.26149 0.32885 0.36741 0.40822
#> LaPlata.LaPlata 2.31975 0.235505 0.0068961 1.89983 2.16273 2.30599 2.46497
#>                   97.5%
#> Bedon.Bedon     0.39997
#> Bedon.LaPlata   0.49251
#> LaPlata.LaPlata 2.81999
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean       Sd   Sd(Mean)       2.5%         25%
#> Bedon:(Intercept)       2.1648722 0.440263 0.01643416  1.3148125  1.8553e+00
#> Bedon:Bedon.lag(1)      0.5860844 0.045921 0.00144176  0.4971243  5.5617e-01
#> Bedon:LaPlata.lag(1)    0.0217075 0.014062 0.00045238 -0.0039856  1.2252e-02
#> Bedon:Bedon.lag(2)      0.0966552 0.063061 0.00244278 -0.0247199  5.3454e-02
#> Bedon:LaPlata.lag(2)   -0.0198063 0.017936 0.00063798 -0.0558333 -3.1841e-02
#> Bedon:Bedon.lag(3)     -0.0375235 0.057134 0.00205342 -0.1518498 -7.5556e-02
#> Bedon:LaPlata.lag(3)   -0.0068435 0.015707 0.00055699 -0.0356985 -1.8192e-02
#> Bedon:Bedon.lag(4)      0.1021946 0.059500 0.00219405 -0.0200166  6.4161e-02
#> Bedon:LaPlata.lag(4)    0.0064100 0.017352 0.00063171 -0.0276715 -4.5828e-03
#> Bedon:Bedon.lag(5)      0.0275118 0.042639 0.00135706 -0.0569605 -5.7801e-05
#> Bedon:LaPlata.lag(5)    0.0047099 0.014948 0.00046016 -0.0240483 -4.9075e-03
#> LaPlata:(Intercept)     7.0249609 1.064089 0.03875091  4.9114711  6.2908e+00
#> LaPlata:Bedon.lag(1)    0.1499684 0.118898 0.00405315 -0.0692066  6.9277e-02
#> LaPlata:LaPlata.lag(1)  0.5269207 0.039972 0.00147251  0.4497349  5.0079e-01
#> LaPlata:Bedon.lag(2)   -0.0038704 0.126488 0.00411288 -0.2533531 -8.9645e-02
#> LaPlata:LaPlata.lag(2)  0.0319723 0.038426 0.00120724 -0.0442783  6.7302e-03
#> LaPlata:Bedon.lag(3)   -0.0683723 0.116761 0.00343150 -0.3007189 -1.4521e-01
#> LaPlata:LaPlata.lag(3)  0.0441851 0.036417 0.00121640 -0.0279538  1.9736e-02
#> LaPlata:Bedon.lag(4)    0.2268666 0.135550 0.00476316 -0.0342364  1.3658e-01
#> LaPlata:LaPlata.lag(4) -0.0477873 0.042586 0.00156336 -0.1270650 -7.8403e-02
#> LaPlata:Bedon.lag(5)   -0.2708921 0.106208 0.00352915 -0.4833299 -3.3957e-01
#> LaPlata:LaPlata.lag(5)  0.1198760 0.039881 0.00139207  0.0460560  9.2207e-02
#>                               50%        75%     97.5%
#> Bedon:(Intercept)       2.1457576  2.4650701  3.065611
#> Bedon:Bedon.lag(1)      0.5852149  0.6169302  0.676536
#> Bedon:LaPlata.lag(1)    0.0208385  0.0307748  0.051048
#> Bedon:Bedon.lag(2)      0.0963276  0.1396565  0.222820
#> Bedon:LaPlata.lag(2)   -0.0194381 -0.0069476  0.013233
#> Bedon:Bedon.lag(3)     -0.0377007  0.0037890  0.065227
#> Bedon:LaPlata.lag(3)   -0.0069053  0.0036561  0.024922
#> Bedon:Bedon.lag(4)      0.1024485  0.1414443  0.214630
#> Bedon:LaPlata.lag(4)    0.0062665  0.0187604  0.039889
#> Bedon:Bedon.lag(5)      0.0278751  0.0569080  0.110964
#> Bedon:LaPlata.lag(5)    0.0044568  0.0146321  0.034902
#> LaPlata:(Intercept)     7.0177598  7.7336366  9.089502
#> LaPlata:Bedon.lag(1)    0.1469023  0.2285406  0.399860
#> LaPlata:LaPlata.lag(1)  0.5273473  0.5533241  0.605898
#> LaPlata:Bedon.lag(2)   -0.0021746  0.0810614  0.249936
#> LaPlata:LaPlata.lag(2)  0.0324451  0.0575803  0.105648
#> LaPlata:Bedon.lag(3)   -0.0690653  0.0081795  0.159617
#> LaPlata:LaPlata.lag(3)  0.0445908  0.0693498  0.113130
#> LaPlata:Bedon.lag(4)    0.2277741  0.3253375  0.468355
#> LaPlata:LaPlata.lag(4) -0.0480892 -0.0181504  0.037541
#> LaPlata:Bedon.lag(5)   -0.2704452 -0.2001764 -0.064723
#> LaPlata:LaPlata.lag(5)  0.1188704  0.1466345  0.199168
#> 
#> 
#> Scale parameter:
#>                   Mean      Sd  Sd(Mean)    2.5%    25%    50%    75%  97.5%
#> Bedon.Bedon     1.0862 0.10510 0.0031385 0.88963 1.0148 1.0803 1.1580 1.2981
#> Bedon.LaPlata   1.3159 0.18708 0.0051741 0.97441 1.1837 1.3087 1.4369 1.7053
#> LaPlata.LaPlata 6.4683 0.64866 0.0202091 5.30782 6.0216 6.4154 6.8989 7.7786
#> 
#> 
#> Regime 3
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean       Sd   Sd(Mean)       2.5%       25%
#> Bedon:(Intercept)       5.5550222 0.835562 0.03072968  4.0255199  4.970857
#> Bedon:Bedon.lag(1)      0.4715347 0.087081 0.00376192  0.2882416  0.415628
#> Bedon:LaPlata.lag(1)    0.0445755 0.017291 0.00063534  0.0097177  0.033022
#> Bedon:Bedon.lag(2)      0.0828094 0.077888 0.00255373 -0.0672880  0.030182
#> Bedon:LaPlata.lag(2)   -0.0035119 0.017292 0.00053192 -0.0357354 -0.015306
#> Bedon:Bedon.lag(3)     -0.0936635 0.067842 0.00214221 -0.2240345 -0.140984
#> Bedon:LaPlata.lag(3)    0.0330091 0.019718 0.00062627 -0.0063559  0.019819
#> Bedon:Bedon.lag(4)      0.0051736 0.077117 0.00245497 -0.1522201 -0.044493
#> Bedon:LaPlata.lag(4)    0.0041103 0.021927 0.00077562 -0.0380974 -0.010730
#> Bedon:Bedon.lag(5)      0.1816836 0.071318 0.00249161  0.0439980  0.132522
#> Bedon:LaPlata.lag(5)   -0.0146580 0.018288 0.00054679 -0.0495327 -0.026932
#> LaPlata:(Intercept)    16.9766225 2.995892 0.10114299 11.2868551 14.887548
#> LaPlata:Bedon.lag(1)    0.5453561 0.262593 0.00806628  0.0320716  0.366246
#> LaPlata:LaPlata.lag(1)  0.3271254 0.063186 0.00202170  0.2055057  0.283793
#> LaPlata:Bedon.lag(2)   -0.5660067 0.289487 0.00852248 -1.1393877 -0.762372
#> LaPlata:LaPlata.lag(2)  0.1174032 0.066227 0.00200242 -0.0080023  0.072913
#> LaPlata:Bedon.lag(3)   -0.5790286 0.231359 0.00667679 -1.0293763 -0.735900
#> LaPlata:LaPlata.lag(3)  0.2825450 0.077940 0.00268522  0.1293920  0.230718
#> LaPlata:Bedon.lag(4)    0.0416098 0.297988 0.00919217 -0.5223999 -0.163652
#> LaPlata:LaPlata.lag(4) -0.0079300 0.082181 0.00267876 -0.1610883 -0.062514
#> LaPlata:Bedon.lag(5)    0.2814524 0.249240 0.00749996 -0.2009808  0.109330
#> LaPlata:LaPlata.lag(5)  0.0649178 0.066878 0.00208951 -0.0672037  0.020186
#>                               50%        75%     97.5%
#> Bedon:(Intercept)       5.5458850  6.0960230  7.214441
#> Bedon:Bedon.lag(1)      0.4742436  0.5302936  0.640888
#> Bedon:LaPlata.lag(1)    0.0444803  0.0556490  0.079913
#> Bedon:Bedon.lag(2)      0.0851528  0.1355145  0.233384
#> Bedon:LaPlata.lag(2)   -0.0041908  0.0073252  0.032087
#> Bedon:Bedon.lag(3)     -0.0943056 -0.0476730  0.047104
#> Bedon:LaPlata.lag(3)    0.0331220  0.0460773  0.071207
#> Bedon:Bedon.lag(4)      0.0084873  0.0577857  0.146271
#> Bedon:LaPlata.lag(4)    0.0035633  0.0184435  0.048419
#> Bedon:Bedon.lag(5)      0.1818226  0.2296683  0.318798
#> Bedon:LaPlata.lag(5)   -0.0151685 -0.0027350  0.022639
#> LaPlata:(Intercept)    16.9904649 19.0738297 22.857655
#> LaPlata:Bedon.lag(1)    0.5475919  0.7197221  1.068015
#> LaPlata:LaPlata.lag(1)  0.3262288  0.3711018  0.453802
#> LaPlata:Bedon.lag(2)   -0.5617587 -0.3715052 -0.016327
#> LaPlata:LaPlata.lag(2)  0.1157398  0.1620768  0.252986
#> LaPlata:Bedon.lag(3)   -0.5736432 -0.4306504 -0.133631
#> LaPlata:LaPlata.lag(3)  0.2824097  0.3364467  0.435876
#> LaPlata:Bedon.lag(4)    0.0343437  0.2430612  0.643896
#> LaPlata:LaPlata.lag(4) -0.0103942  0.0449317  0.162760
#> LaPlata:Bedon.lag(5)    0.2818948  0.4532395  0.763928
#> LaPlata:LaPlata.lag(5)  0.0638691  0.1083971  0.198409
#> 
#> 
#> Scale parameter:
#>                    Mean      Sd  Sd(Mean)   2.5%     25%     50%     75%
#> Bedon.Bedon      2.7873 0.29816 0.0096167  2.278  2.5821  2.7706  2.9697
#> Bedon.LaPlata    7.1934 0.91561 0.0254491  5.605  6.5375  7.1455  7.7540
#> LaPlata.LaPlata 43.1276 4.55554 0.1347389 35.141 39.9423 42.7743 46.0169
#>                   97.5%
#> Bedon.Bedon      3.4126
#> Bedon.LaPlata    9.1595
#> LaPlata.LaPlata 52.8495
#plot(fit2.mcmc)

###### Example 3: Temperature, precipitation, and two river flows in Iceland
data(iceland.rf)
fit3 <- mtar(~ Jokulsa + Vatnsdalsa | Temperature | Precipitation,
             data=iceland.rf, subset={Date<="1974-11-06"}, row.names=Date,
             ars=ars(nregim=2,p=15,q=4,d=2), n.burnin=1000, n.sim=2000,
             n.thin=2, dist="Slash")
fit3.mcmc <- coda::as.mcmc(fit3)
summary(fit3.mcmc)
#> 
#> 
#>  Iterations = 1001:4999
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 2000
#> 
#> 
#> Thresholds:
#>             Mean       Sd  Sd(Mean)   2.5%    25%   50%    75%  97.5%
#> threshold 1.1386 0.052649 0.0060595 1.0112 1.1204 1.145 1.1683 1.1972
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                                        Mean        Sd   Sd(Mean)        2.5%
#> Jokulsa:(Intercept)              3.6453e+00 0.3520020 0.01701765  3.00589632
#> Jokulsa:Jokulsa.lag( 1)          8.5047e-01 0.0370368 0.00198459  0.77217109
#> Jokulsa:Vatnsdalsa.lag( 1)       2.0624e-01 0.0583850 0.00221108  0.09492076
#> Jokulsa:Jokulsa.lag( 2)         -5.0926e-02 0.0244854 0.00089469 -0.09728321
#> Jokulsa:Vatnsdalsa.lag( 2)      -1.7181e-01 0.0679877 0.00311423 -0.32290904
#> Jokulsa:Jokulsa.lag( 3)          4.5755e-03 0.0185595 0.00068568 -0.03178605
#> Jokulsa:Vatnsdalsa.lag( 3)       3.2837e-02 0.0342003 0.00110318 -0.03169090
#> Jokulsa:Jokulsa.lag( 4)          4.2491e-05 0.0243022 0.00104418 -0.04400388
#> Jokulsa:Vatnsdalsa.lag( 4)       9.3857e-03 0.0386241 0.00142275 -0.06685350
#> Jokulsa:Jokulsa.lag( 5)          1.4494e-03 0.0282243 0.00134944 -0.05695638
#> Jokulsa:Vatnsdalsa.lag( 5)      -3.7197e-02 0.0334297 0.00104763 -0.10527720
#> Jokulsa:Jokulsa.lag( 6)          2.3008e-02 0.0263573 0.00120600 -0.03055907
#> Jokulsa:Vatnsdalsa.lag( 6)      -2.2402e-02 0.0302176 0.00085215 -0.08159326
#> Jokulsa:Jokulsa.lag( 7)         -1.7494e-03 0.0253504 0.00094947 -0.05214902
#> Jokulsa:Vatnsdalsa.lag( 7)       1.6392e-02 0.0297434 0.00074968 -0.04231044
#> Jokulsa:Jokulsa.lag( 8)          6.5825e-04 0.0256414 0.00104031 -0.05209577
#> Jokulsa:Vatnsdalsa.lag( 8)      -2.6851e-03 0.0300225 0.00083823 -0.06329791
#> Jokulsa:Jokulsa.lag( 9)         -8.6836e-03 0.0230430 0.00089141 -0.05240424
#> Jokulsa:Vatnsdalsa.lag( 9)      -1.0334e-02 0.0338699 0.00122626 -0.07095006
#> Jokulsa:Jokulsa.lag(10)          2.7019e-02 0.0174953 0.00056966 -0.01014846
#> Jokulsa:Vatnsdalsa.lag(10)       2.4182e-02 0.0306571 0.00087665 -0.03592046
#> Jokulsa:Jokulsa.lag(11)         -1.4885e-02 0.0150770 0.00040831 -0.04505572
#> Jokulsa:Vatnsdalsa.lag(11)      -1.6509e-02 0.0292179 0.00082358 -0.07561458
#> Jokulsa:Jokulsa.lag(12)          9.4747e-03 0.0150832 0.00040075 -0.02034212
#> Jokulsa:Vatnsdalsa.lag(12)       8.9427e-03 0.0257704 0.00064092 -0.04140748
#> Jokulsa:Jokulsa.lag(13)         -1.6984e-02 0.0182395 0.00083269 -0.05167006
#> Jokulsa:Vatnsdalsa.lag(13)      -1.4003e-02 0.0278598 0.00071102 -0.07085565
#> Jokulsa:Jokulsa.lag(14)          5.4445e-03 0.0143221 0.00043859 -0.02370430
#> Jokulsa:Vatnsdalsa.lag(14)       5.1893e-03 0.0271128 0.00074495 -0.04711329
#> Jokulsa:Jokulsa.lag(15)          1.9971e-02 0.0146916 0.00067665 -0.00832191
#> Jokulsa:Vatnsdalsa.lag(15)      -1.1809e-02 0.0199583 0.00065009 -0.05217344
#> Jokulsa:Precipitation.lag(1)     7.3637e-03 0.0091445 0.00034314 -0.01000702
#> Jokulsa:Precipitation.lag(2)     6.0775e-03 0.0068989 0.00021061 -0.00806726
#> Jokulsa:Precipitation.lag(3)    -1.2134e-02 0.0054809 0.00018062 -0.02252268
#> Jokulsa:Precipitation.lag(4)     1.9168e-02 0.0067323 0.00021147  0.00634306
#> Jokulsa:Temperature.lag(1)       2.2101e-02 0.0108236 0.00031535  0.00111240
#> Jokulsa:Temperature.lag(2)      -3.8695e-02 0.0109137 0.00034242 -0.06024009
#> Vatnsdalsa:(Intercept)           8.2184e-01 0.2067065 0.00932940  0.42889440
#> Vatnsdalsa:Jokulsa.lag( 1)      -6.3107e-02 0.0206784 0.00092772 -0.10659126
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.1645e+00 0.0401658 0.00158231  1.08697412
#> Vatnsdalsa:Jokulsa.lag( 2)       4.9965e-02 0.0155658 0.00057223  0.01872326
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -2.9498e-01 0.0433569 0.00194746 -0.38983018
#> Vatnsdalsa:Jokulsa.lag( 3)      -2.4053e-02 0.0120144 0.00041679 -0.04755875
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    2.8245e-02 0.0247497 0.00089562 -0.01593657
#> Vatnsdalsa:Jokulsa.lag( 4)       1.4705e-02 0.0130781 0.00042813 -0.01091260
#> Vatnsdalsa:Vatnsdalsa.lag( 4)    1.9071e-03 0.0273263 0.00101311 -0.05540838
#> Vatnsdalsa:Jokulsa.lag( 5)       5.2754e-03 0.0150520 0.00048827 -0.02174433
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -1.5866e-02 0.0229959 0.00071670 -0.06447352
#> Vatnsdalsa:Jokulsa.lag( 6)       5.7175e-03 0.0147217 0.00054967 -0.02280600
#> Vatnsdalsa:Vatnsdalsa.lag( 6)    2.6376e-04 0.0195968 0.00052051 -0.03989479
#> Vatnsdalsa:Jokulsa.lag( 7)      -3.0093e-03 0.0149581 0.00048051 -0.03140486
#> Vatnsdalsa:Vatnsdalsa.lag( 7)    8.0908e-03 0.0175522 0.00044052 -0.02479084
#> Vatnsdalsa:Jokulsa.lag( 8)      -9.1601e-03 0.0132044 0.00040645 -0.03519181
#> Vatnsdalsa:Vatnsdalsa.lag( 8)    6.5742e-03 0.0185052 0.00050567 -0.02952938
#> Vatnsdalsa:Jokulsa.lag( 9)       2.0329e-02 0.0151033 0.00059165 -0.00863684
#> Vatnsdalsa:Vatnsdalsa.lag( 9)   -1.9471e-03 0.0204536 0.00065876 -0.04736964
#> Vatnsdalsa:Jokulsa.lag(10)      -1.5184e-02 0.0118265 0.00037574 -0.03935958
#> Vatnsdalsa:Vatnsdalsa.lag(10)    1.9523e-02 0.0227416 0.00075096 -0.02145983
#> Vatnsdalsa:Jokulsa.lag(11)       9.8592e-03 0.0096379 0.00025026 -0.00887165
#> Vatnsdalsa:Vatnsdalsa.lag(11)   -7.1777e-03 0.0196753 0.00056216 -0.04472734
#> Vatnsdalsa:Jokulsa.lag(12)      -9.0315e-03 0.0100855 0.00027772 -0.02805763
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -2.9309e-03 0.0177348 0.00044166 -0.03911232
#> Vatnsdalsa:Jokulsa.lag(13)       3.0575e-03 0.0103450 0.00031546 -0.01658155
#> Vatnsdalsa:Vatnsdalsa.lag(13)   -1.9569e-02 0.0204571 0.00065198 -0.06254845
#> Vatnsdalsa:Jokulsa.lag(14)      -4.5392e-03 0.0091466 0.00024610 -0.02211711
#> Vatnsdalsa:Vatnsdalsa.lag(14)    3.3972e-02 0.0183457 0.00051004  0.00022658
#> Vatnsdalsa:Jokulsa.lag(15)       1.4122e-03 0.0075272 0.00030496 -0.01318021
#> Vatnsdalsa:Vatnsdalsa.lag(15)    5.2932e-03 0.0141626 0.00044257 -0.02322284
#> Vatnsdalsa:Precipitation.lag(1)  5.1819e-03 0.0055240 0.00015291 -0.00593959
#> Vatnsdalsa:Precipitation.lag(2)  4.7831e-04 0.0045001 0.00013038 -0.00803138
#> Vatnsdalsa:Precipitation.lag(3) -3.9295e-03 0.0039427 0.00013335 -0.01157798
#> Vatnsdalsa:Precipitation.lag(4)  4.6408e-03 0.0047749 0.00015914 -0.00512651
#> Vatnsdalsa:Temperature.lag(1)    2.0501e-03 0.0069766 0.00021191 -0.01143515
#> Vatnsdalsa:Temperature.lag(2)   -1.1762e-02 0.0067240 0.00019046 -0.02457357
#>                                         25%         50%         75%       97.5%
#> Jokulsa:(Intercept)              3.39708019  3.62810370  3.87965992  4.39450865
#> Jokulsa:Jokulsa.lag( 1)          0.82527812  0.85611163  0.87863085  0.90955466
#> Jokulsa:Vatnsdalsa.lag( 1)       0.16655167  0.20501929  0.24456810  0.32493280
#> Jokulsa:Jokulsa.lag( 2)         -0.06817406 -0.05120724 -0.03420919 -0.00272222
#> Jokulsa:Vatnsdalsa.lag( 2)      -0.21288811 -0.16512991 -0.12158610 -0.05652177
#> Jokulsa:Jokulsa.lag( 3)         -0.00800539  0.00503123  0.01739540  0.03981156
#> Jokulsa:Vatnsdalsa.lag( 3)       0.01038765  0.03122345  0.05359151  0.10482772
#> Jokulsa:Jokulsa.lag( 4)         -0.01728811 -0.00132318  0.01507079  0.05059049
#> Jokulsa:Vatnsdalsa.lag( 4)      -0.01612235  0.00866084  0.03603586  0.08245870
#> Jokulsa:Jokulsa.lag( 5)         -0.01649241  0.00361758  0.02084187  0.05031803
#> Jokulsa:Vatnsdalsa.lag( 5)      -0.05859619 -0.03598653 -0.01601452  0.02897573
#> Jokulsa:Jokulsa.lag( 6)          0.00496075  0.02355272  0.04174719  0.07134566
#> Jokulsa:Vatnsdalsa.lag( 6)      -0.04148934 -0.02279264 -0.00218967  0.03518622
#> Jokulsa:Jokulsa.lag( 7)         -0.01910149 -0.00159716  0.01480348  0.05036984
#> Jokulsa:Vatnsdalsa.lag( 7)      -0.00282860  0.01580560  0.03555808  0.07693437
#> Jokulsa:Jokulsa.lag( 8)         -0.01603019  0.00104483  0.01874912  0.04761604
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.02255210 -0.00129847  0.01719594  0.05446171
#> Jokulsa:Jokulsa.lag( 9)         -0.02405928 -0.00832449  0.00671838  0.03685804
#> Jokulsa:Vatnsdalsa.lag( 9)      -0.03271258 -0.01354712  0.00998009  0.06467347
#> Jokulsa:Jokulsa.lag(10)          0.01566695  0.02817305  0.03858792  0.05862569
#> Jokulsa:Vatnsdalsa.lag(10)       0.00464249  0.02426870  0.04313043  0.08897412
#> Jokulsa:Jokulsa.lag(11)         -0.02461527 -0.01551488 -0.00477020  0.01421942
#> Jokulsa:Vatnsdalsa.lag(11)      -0.03463205 -0.01614379  0.00257730  0.03973918
#> Jokulsa:Jokulsa.lag(12)         -0.00089483  0.00928984  0.01965545  0.03927357
#> Jokulsa:Vatnsdalsa.lag(12)      -0.00804580  0.00827293  0.02567986  0.06069560
#> Jokulsa:Jokulsa.lag(13)         -0.02966354 -0.01767486 -0.00392760  0.01816022
#> Jokulsa:Vatnsdalsa.lag(13)      -0.03172206 -0.01472997  0.00470231  0.04151807
#> Jokulsa:Jokulsa.lag(14)         -0.00361236  0.00578003  0.01516138  0.03197176
#> Jokulsa:Vatnsdalsa.lag(14)      -0.01340242  0.00480004  0.02374562  0.05865363
#> Jokulsa:Jokulsa.lag(15)          0.00994103  0.01964075  0.02971033  0.04962176
#> Jokulsa:Vatnsdalsa.lag(15)      -0.02523204 -0.01140807  0.00203287  0.02717161
#> Jokulsa:Precipitation.lag(1)     0.00108191  0.00753804  0.01364956  0.02466328
#> Jokulsa:Precipitation.lag(2)     0.00194936  0.00632775  0.01086029  0.01863353
#> Jokulsa:Precipitation.lag(3)    -0.01584052 -0.01247997 -0.00867130 -0.00089269
#> Jokulsa:Precipitation.lag(4)     0.01464563  0.01924885  0.02372782  0.03250955
#> Jokulsa:Temperature.lag(1)       0.01471614  0.02215607  0.02939191  0.04433218
#> Jokulsa:Temperature.lag(2)      -0.04598437 -0.03861993 -0.03123401 -0.01749362
#> Vatnsdalsa:(Intercept)           0.67982678  0.81446414  0.96126000  1.23821499
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.07663942 -0.06179942 -0.04739748 -0.02796685
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.13706255  1.16565456  1.19218844  1.24297135
#> Vatnsdalsa:Jokulsa.lag( 2)       0.03940728  0.05008815  0.06087371  0.07986030
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.32257171 -0.29085483 -0.26448399 -0.21905775
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.03242368 -0.02428036 -0.01599785 -0.00018945
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    0.01120642  0.02658391  0.04412237  0.07808964
#> Vatnsdalsa:Jokulsa.lag( 4)       0.00564909  0.01473687  0.02385734  0.04013299
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -0.01588658  0.00264505  0.01986744  0.05387273
#> Vatnsdalsa:Jokulsa.lag( 5)      -0.00492721  0.00481273  0.01522904  0.03645302
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -0.03039187 -0.01516725 -0.00032289  0.02698258
#> Vatnsdalsa:Jokulsa.lag( 6)      -0.00388236  0.00584355  0.01606918  0.03463158
#> Vatnsdalsa:Vatnsdalsa.lag( 6)   -0.01186182  0.00051446  0.01310291  0.03873348
#> Vatnsdalsa:Jokulsa.lag( 7)      -0.01317091 -0.00325712  0.00660161  0.02794579
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -0.00377621  0.00785849  0.01995007  0.04227343
#> Vatnsdalsa:Jokulsa.lag( 8)      -0.01813887 -0.00911958 -0.00016263  0.01642219
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -0.00584260  0.00641939  0.01873821  0.04474950
#> Vatnsdalsa:Jokulsa.lag( 9)       0.00974794  0.02025723  0.03053784  0.04999908
#> Vatnsdalsa:Vatnsdalsa.lag( 9)   -0.01460146 -0.00107256  0.01227841  0.03416928
#> Vatnsdalsa:Jokulsa.lag(10)      -0.02238456 -0.01480653 -0.00734084  0.00672517
#> Vatnsdalsa:Vatnsdalsa.lag(10)    0.00366297  0.01805010  0.03337072  0.06881818
#> Vatnsdalsa:Jokulsa.lag(11)       0.00350186  0.00994153  0.01620672  0.02940448
#> Vatnsdalsa:Vatnsdalsa.lag(11)   -0.02061984 -0.00780812  0.00551297  0.03216434
#> Vatnsdalsa:Jokulsa.lag(12)      -0.01591823 -0.00921394 -0.00243518  0.01087513
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.01415951 -0.00234988  0.00904091  0.03110558
#> Vatnsdalsa:Jokulsa.lag(13)      -0.00403256  0.00283193  0.01006928  0.02313380
#> Vatnsdalsa:Vatnsdalsa.lag(13)   -0.03300406 -0.01869446 -0.00520053  0.01738530
#> Vatnsdalsa:Jokulsa.lag(14)      -0.01072032 -0.00460689  0.00141530  0.01404826
#> Vatnsdalsa:Vatnsdalsa.lag(14)    0.02188713  0.03301661  0.04567634  0.07167534
#> Vatnsdalsa:Jokulsa.lag(15)      -0.00377073  0.00144224  0.00651181  0.01613835
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -0.00437306  0.00555071  0.01448907  0.03239281
#> Vatnsdalsa:Precipitation.lag(1)  0.00156934  0.00527703  0.00894406  0.01609397
#> Vatnsdalsa:Precipitation.lag(2) -0.00249283  0.00034685  0.00322825  0.00999337
#> Vatnsdalsa:Precipitation.lag(3) -0.00659270 -0.00384345 -0.00123371  0.00371380
#> Vatnsdalsa:Precipitation.lag(4)  0.00167531  0.00481408  0.00776515  0.01359144
#> Vatnsdalsa:Temperature.lag(1)   -0.00258199  0.00208135  0.00670523  0.01545152
#> Vatnsdalsa:Temperature.lag(2)   -0.01628668 -0.01191500 -0.00718334  0.00165808
#> 
#> 
#> Scale parameter:
#>                           Mean        Sd   Sd(Mean)      2.5%       25%
#> Jokulsa.Jokulsa       0.063841 0.0094220 0.00072336 0.0480841 0.0569612
#> Jokulsa.Vatnsdalsa    0.010768 0.0027921 0.00013214 0.0060564 0.0088029
#> Vatnsdalsa.Vatnsdalsa 0.027942 0.0042399 0.00031880 0.0206479 0.0248310
#>                            50%      75%    97.5%
#> Jokulsa.Jokulsa       0.062986 0.069923 0.084767
#> Jokulsa.Vatnsdalsa    0.010552 0.012473 0.016831
#> Vatnsdalsa.Vatnsdalsa 0.027483 0.030692 0.037425
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                                        Mean        Sd   Sd(Mean)        2.5%
#> Jokulsa:(Intercept)             -2.6634e-01 0.6899361 2.7030e-02 -1.61083120
#> Jokulsa:Jokulsa.lag( 1)          1.0133e+00 0.0358855 1.3201e-03  0.95289268
#> Jokulsa:Vatnsdalsa.lag( 1)       9.1330e-01 0.2310747 9.7526e-03  0.48487679
#> Jokulsa:Jokulsa.lag( 2)         -1.6312e-01 0.0678981 3.1569e-03 -0.30185098
#> Jokulsa:Vatnsdalsa.lag( 2)      -4.2708e-01 0.3501138 1.3557e-02 -1.08668910
#> Jokulsa:Jokulsa.lag( 3)          3.7061e-03 0.0582615 2.4053e-03 -0.10278597
#> Jokulsa:Vatnsdalsa.lag( 3)       7.3732e-02 0.3368247 1.3494e-02 -0.66330345
#> Jokulsa:Jokulsa.lag( 4)         -7.1529e-02 0.0417722 1.3665e-03 -0.15609611
#> Jokulsa:Vatnsdalsa.lag( 4)      -1.7940e-01 0.2295687 7.8622e-03 -0.60217130
#> Jokulsa:Jokulsa.lag( 5)          3.8465e-02 0.0407771 1.6348e-03 -0.04080151
#> Jokulsa:Vatnsdalsa.lag( 5)       1.4401e-02 0.3068499 1.3446e-02 -0.64022813
#> Jokulsa:Jokulsa.lag( 6)         -4.0166e-02 0.0332718 1.2186e-03 -0.10488161
#> Jokulsa:Vatnsdalsa.lag( 6)       5.9651e-02 0.3089013 1.1923e-02 -0.53150444
#> Jokulsa:Jokulsa.lag( 7)         -2.8537e-04 0.0297368 8.1967e-04 -0.05888056
#> Jokulsa:Vatnsdalsa.lag( 7)       1.0856e-01 0.2351079 8.9051e-03 -0.35641099
#> Jokulsa:Jokulsa.lag( 8)          1.6663e-02 0.0294704 8.6830e-04 -0.04265392
#> Jokulsa:Vatnsdalsa.lag( 8)      -2.2249e-01 0.2306677 8.8670e-03 -0.67316324
#> Jokulsa:Jokulsa.lag( 9)          4.0000e-02 0.0316933 9.7783e-04 -0.02332182
#> Jokulsa:Vatnsdalsa.lag( 9)       1.3537e-01 0.2244226 8.1938e-03 -0.30241140
#> Jokulsa:Jokulsa.lag(10)         -2.2266e-02 0.0413066 1.4119e-03 -0.09899195
#> Jokulsa:Vatnsdalsa.lag(10)       5.1462e-05 0.2080615 1.1234e-02 -0.43247323
#> Jokulsa:Jokulsa.lag(11)         -5.8875e-03 0.0394207 1.2488e-03 -0.08598797
#> Jokulsa:Vatnsdalsa.lag(11)      -1.6285e-02 0.2569680 1.6357e-02 -0.49532058
#> Jokulsa:Jokulsa.lag(12)         -2.4376e-03 0.0373984 1.0849e-03 -0.07793742
#> Jokulsa:Vatnsdalsa.lag(12)       9.8251e-03 0.2391477 8.7936e-03 -0.48230235
#> Jokulsa:Jokulsa.lag(13)         -5.9975e-04 0.0412243 1.6754e-03 -0.07438782
#> Jokulsa:Vatnsdalsa.lag(13)       4.0785e-01 0.2558755 1.0471e-02 -0.11574643
#> Jokulsa:Jokulsa.lag(14)         -8.5933e-03 0.0412067 1.7946e-03 -0.08841502
#> Jokulsa:Vatnsdalsa.lag(14)       1.3275e-01 0.2744349 1.1529e-02 -0.43022933
#> Jokulsa:Jokulsa.lag(15)          4.7871e-02 0.0235455 7.4835e-04  0.00121913
#> Jokulsa:Vatnsdalsa.lag(15)      -4.3381e-01 0.1619985 5.5707e-03 -0.72636722
#> Jokulsa:Precipitation.lag(1)    -1.1378e-01 0.0378993 1.1196e-03 -0.18698752
#> Jokulsa:Precipitation.lag(2)     3.3086e-02 0.0657382 2.6700e-03 -0.09986422
#> Jokulsa:Precipitation.lag(3)     4.8308e-02 0.0341770 8.6652e-04 -0.01695345
#> Jokulsa:Precipitation.lag(4)     2.8971e-02 0.0347795 1.2584e-03 -0.04287306
#> Jokulsa:Temperature.lag(1)       1.1257e+00 0.0954609 2.6339e-03  0.94733225
#> Jokulsa:Temperature.lag(2)      -5.7283e-01 0.1038110 3.3477e-03 -0.77720718
#> Vatnsdalsa:(Intercept)           4.8874e-01 0.0943002 3.7971e-03  0.30336877
#> Vatnsdalsa:Jokulsa.lag( 1)      -2.6874e-03 0.0037976 1.0235e-04 -0.01032611
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.1853e+00 0.0333410 1.4066e-03  1.12412282
#> Vatnsdalsa:Jokulsa.lag( 2)       1.0089e-02 0.0060824 2.3241e-04 -0.00159485
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -3.4898e-01 0.0483691 2.2384e-03 -0.43899864
#> Vatnsdalsa:Jokulsa.lag( 3)      -1.2230e-02 0.0061648 2.7925e-04 -0.02460720
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    1.9542e-01 0.0432820 1.8682e-03  0.10367450
#> Vatnsdalsa:Jokulsa.lag( 4)       6.0053e-03 0.0046233 1.1741e-04 -0.00283493
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -9.3184e-02 0.0410927 2.2232e-03 -0.16097478
#> Vatnsdalsa:Jokulsa.lag( 5)      -5.1829e-03 0.0046382 1.8381e-04 -0.01401053
#> Vatnsdalsa:Vatnsdalsa.lag( 5)    1.3156e-02 0.0519646 2.9790e-03 -0.08739550
#> Vatnsdalsa:Jokulsa.lag( 6)       4.0256e-03 0.0040468 1.3875e-04 -0.00408057
#> Vatnsdalsa:Vatnsdalsa.lag( 6)    2.3311e-02 0.0532043 3.6232e-03 -0.07792985
#> Vatnsdalsa:Jokulsa.lag( 7)      -5.5791e-03 0.0040252 1.2001e-04 -0.01364294
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -5.6517e-02 0.0329130 1.8487e-03 -0.11924633
#> Vatnsdalsa:Jokulsa.lag( 8)       4.1251e-03 0.0038328 1.1054e-04 -0.00333229
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -4.3819e-02 0.0348763 1.6284e-03 -0.11262682
#> Vatnsdalsa:Jokulsa.lag( 9)      -1.9784e-03 0.0042656 1.3948e-04 -0.01067617
#> Vatnsdalsa:Vatnsdalsa.lag( 9)    8.8889e-02 0.0394806 2.0432e-03  0.01016572
#> Vatnsdalsa:Jokulsa.lag(10)       3.3997e-03 0.0050642 1.6885e-04 -0.00625931
#> Vatnsdalsa:Vatnsdalsa.lag(10)   -7.3863e-02 0.0310885 1.6390e-03 -0.12900457
#> Vatnsdalsa:Jokulsa.lag(11)      -7.5691e-03 0.0048349 1.4152e-04 -0.01715604
#> Vatnsdalsa:Vatnsdalsa.lag(11)    8.4241e-02 0.0364402 1.7380e-03  0.00432740
#> Vatnsdalsa:Jokulsa.lag(12)       9.8027e-03 0.0045407 1.4031e-04  0.00090535
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -8.4062e-02 0.0299243 1.3124e-03 -0.14427539
#> Vatnsdalsa:Jokulsa.lag(13)      -6.4752e-03 0.0043925 1.3862e-04 -0.01524706
#> Vatnsdalsa:Vatnsdalsa.lag(13)    1.4968e-01 0.0349412 1.4289e-03  0.08220235
#> Vatnsdalsa:Jokulsa.lag(14)      -1.8062e-03 0.0043734 1.3537e-04 -0.01046295
#> Vatnsdalsa:Vatnsdalsa.lag(14)   -5.0697e-02 0.0401393 1.7390e-03 -0.13568899
#> Vatnsdalsa:Jokulsa.lag(15)       1.2788e-03 0.0028570 8.5488e-05 -0.00408101
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -2.0000e-02 0.0243717 9.6315e-04 -0.06550437
#> Vatnsdalsa:Precipitation.lag(1) -2.9260e-03 0.0052535 1.4765e-04 -0.01345889
#> Vatnsdalsa:Precipitation.lag(2) -1.4552e-03 0.0070884 2.3370e-04 -0.01544198
#> Vatnsdalsa:Precipitation.lag(3)  5.4234e-03 0.0052492 1.8516e-04 -0.00462744
#> Vatnsdalsa:Precipitation.lag(4)  2.9794e-03 0.0043805 1.4884e-04 -0.00599487
#> Vatnsdalsa:Temperature.lag(1)    2.0339e-02 0.0124111 3.2834e-04 -0.00399768
#> Vatnsdalsa:Temperature.lag(2)   -2.5314e-02 0.0126361 3.6945e-04 -0.05092318
#>                                         25%         50%         75%       97.5%
#> Jokulsa:(Intercept)             -6.9754e-01 -0.24771975  1.7363e-01  1.05905214
#> Jokulsa:Jokulsa.lag( 1)          9.8976e-01  1.01132032  1.0337e+00  1.08943387
#> Jokulsa:Vatnsdalsa.lag( 1)       7.4861e-01  0.91133630  1.0734e+00  1.36438724
#> Jokulsa:Jokulsa.lag( 2)         -2.0882e-01 -0.15894651 -1.1663e-01 -0.03553909
#> Jokulsa:Vatnsdalsa.lag( 2)      -6.6525e-01 -0.43801664 -1.9924e-01  0.28356311
#> Jokulsa:Jokulsa.lag( 3)         -3.6804e-02  0.00151827  4.2524e-02  0.12525995
#> Jokulsa:Vatnsdalsa.lag( 3)      -1.4576e-01  0.09119071  3.0376e-01  0.68340429
#> Jokulsa:Jokulsa.lag( 4)         -9.8436e-02 -0.07056991 -4.3670e-02  0.00888772
#> Jokulsa:Vatnsdalsa.lag( 4)      -3.3019e-01 -0.19374359 -3.6800e-02  0.29669210
#> Jokulsa:Jokulsa.lag( 5)          1.0725e-02  0.03640131  6.4112e-02  0.12174006
#> Jokulsa:Vatnsdalsa.lag( 5)      -1.7378e-01  0.03184291  2.3126e-01  0.55457578
#> Jokulsa:Jokulsa.lag( 6)         -6.1813e-02 -0.03974496 -1.7915e-02  0.02604221
#> Jokulsa:Vatnsdalsa.lag( 6)      -1.4932e-01  0.03786251  2.4796e-01  0.71307286
#> Jokulsa:Jokulsa.lag( 7)         -2.0656e-02 -0.00039647  1.9796e-02  0.05821777
#> Jokulsa:Vatnsdalsa.lag( 7)      -4.9583e-02  0.10565419  2.6404e-01  0.57355890
#> Jokulsa:Jokulsa.lag( 8)         -3.5695e-03  0.01720561  3.7113e-02  0.07140658
#> Jokulsa:Vatnsdalsa.lag( 8)      -3.7418e-01 -0.21821077 -7.1260e-02  0.21641318
#> Jokulsa:Jokulsa.lag( 9)          1.9566e-02  0.04023229  6.1703e-02  0.09929369
#> Jokulsa:Vatnsdalsa.lag( 9)      -9.9971e-03  0.12749910  2.7737e-01  0.58027425
#> Jokulsa:Jokulsa.lag(10)         -5.1196e-02 -0.02317646  5.1502e-03  0.06211869
#> Jokulsa:Vatnsdalsa.lag(10)      -1.3277e-01  0.01391533  1.4206e-01  0.38078068
#> Jokulsa:Jokulsa.lag(11)         -3.0974e-02 -0.00515970  2.0782e-02  0.06965107
#> Jokulsa:Vatnsdalsa.lag(11)      -1.8502e-01 -0.02515474  1.4030e-01  0.52498947
#> Jokulsa:Jokulsa.lag(12)         -2.6787e-02 -0.00240560  2.2966e-02  0.06934007
#> Jokulsa:Vatnsdalsa.lag(12)      -1.5095e-01  0.01746315  1.6969e-01  0.47455344
#> Jokulsa:Jokulsa.lag(13)         -2.9208e-02 -0.00307427  2.6395e-02  0.08555623
#> Jokulsa:Vatnsdalsa.lag(13)       2.5020e-01  0.41361748  5.8265e-01  0.90149881
#> Jokulsa:Jokulsa.lag(14)         -3.7174e-02 -0.00824152  2.0536e-02  0.07397532
#> Jokulsa:Vatnsdalsa.lag(14)      -4.5061e-02  0.14916723  3.2264e-01  0.64136597
#> Jokulsa:Jokulsa.lag(15)          3.2628e-02  0.04766502  6.3267e-02  0.09461067
#> Jokulsa:Vatnsdalsa.lag(15)      -5.4708e-01 -0.44068743 -3.3007e-01 -0.10453274
#> Jokulsa:Precipitation.lag(1)    -1.3967e-01 -0.11363373 -8.8892e-02 -0.03862135
#> Jokulsa:Precipitation.lag(2)    -1.1286e-02  0.03552609  7.8829e-02  0.15118704
#> Jokulsa:Precipitation.lag(3)     2.5597e-02  0.04803027  7.0716e-02  0.11702669
#> Jokulsa:Precipitation.lag(4)     5.5981e-03  0.02957650  5.3172e-02  0.09425543
#> Jokulsa:Temperature.lag(1)       1.0604e+00  1.12348975  1.1890e+00  1.31707964
#> Jokulsa:Temperature.lag(2)      -6.3847e-01 -0.57034266 -5.0405e-01 -0.36163188
#> Vatnsdalsa:(Intercept)           4.2806e-01  0.48565679  5.5403e-01  0.67219568
#> Vatnsdalsa:Jokulsa.lag( 1)      -5.2519e-03 -0.00276910 -4.7617e-05  0.00443313
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.1622e+00  1.18353707  1.2058e+00  1.25738213
#> Vatnsdalsa:Jokulsa.lag( 2)       5.9138e-03  0.01008251  1.4048e-02  0.02228399
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -3.8060e-01 -0.35188664 -3.2047e-01 -0.24483746
#> Vatnsdalsa:Jokulsa.lag( 3)      -1.6456e-02 -0.01199891 -7.9471e-03 -0.00080793
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    1.7115e-01  0.19762672  2.2321e-01  0.27537291
#> Vatnsdalsa:Jokulsa.lag( 4)       2.8453e-03  0.00597634  9.2871e-03  0.01506453
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -1.2039e-01 -0.09828702 -7.3062e-02  0.00331726
#> Vatnsdalsa:Jokulsa.lag( 5)      -8.3996e-03 -0.00536437 -1.9336e-03  0.00413806
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -2.3768e-02  0.01318641  5.1425e-02  0.11272333
#> Vatnsdalsa:Jokulsa.lag( 6)       1.3883e-03  0.00406036  6.6072e-03  0.01185681
#> Vatnsdalsa:Vatnsdalsa.lag( 6)   -1.4316e-02  0.01992537  5.9827e-02  0.12497216
#> Vatnsdalsa:Jokulsa.lag( 7)      -8.1715e-03 -0.00548689 -2.9193e-03  0.00184426
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -7.8911e-02 -0.05822696 -3.5635e-02  0.01185956
#> Vatnsdalsa:Jokulsa.lag( 8)       1.6092e-03  0.00401537  6.4256e-03  0.01196438
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -6.7042e-02 -0.04545162 -2.1166e-02  0.02794665
#> Vatnsdalsa:Jokulsa.lag( 9)      -4.8475e-03 -0.00192797  9.5677e-04  0.00597804
#> Vatnsdalsa:Vatnsdalsa.lag( 9)    6.2816e-02  0.08962618  1.1516e-01  0.16182473
#> Vatnsdalsa:Jokulsa.lag(10)      -4.9425e-05  0.00320829  6.7883e-03  0.01377056
#> Vatnsdalsa:Vatnsdalsa.lag(10)   -9.5731e-02 -0.07607525 -5.3407e-02 -0.00881501
#> Vatnsdalsa:Jokulsa.lag(11)      -1.0818e-02 -0.00757247 -4.3226e-03  0.00183991
#> Vatnsdalsa:Vatnsdalsa.lag(11)    6.1636e-02  0.08646571  1.0950e-01  0.14810452
#> Vatnsdalsa:Jokulsa.lag(12)       6.6509e-03  0.00976126  1.2849e-02  0.01881002
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -1.0400e-01 -0.08391380 -6.3975e-02 -0.02593476
#> Vatnsdalsa:Jokulsa.lag(13)      -9.3854e-03 -0.00632600 -3.4152e-03  0.00165677
#> Vatnsdalsa:Vatnsdalsa.lag(13)    1.2606e-01  0.14925796  1.7262e-01  0.21812611
#> Vatnsdalsa:Jokulsa.lag(14)      -4.7183e-03 -0.00176024  1.1439e-03  0.00664288
#> Vatnsdalsa:Vatnsdalsa.lag(14)   -7.6343e-02 -0.04947386 -2.2638e-02  0.02308749
#> Vatnsdalsa:Jokulsa.lag(15)      -7.0251e-04  0.00117359  3.1687e-03  0.00734694
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -3.6640e-02 -0.02066875 -3.9701e-03  0.03010690
#> Vatnsdalsa:Precipitation.lag(1) -6.3396e-03 -0.00294388  5.0784e-04  0.00788813
#> Vatnsdalsa:Precipitation.lag(2) -6.2940e-03 -0.00133113  3.2548e-03  0.01284514
#> Vatnsdalsa:Precipitation.lag(3)  1.7616e-03  0.00528462  9.0604e-03  0.01569673
#> Vatnsdalsa:Precipitation.lag(4)  1.1199e-04  0.00316071  5.8287e-03  0.01153929
#> Vatnsdalsa:Temperature.lag(1)    1.2001e-02  0.02040362  2.8704e-02  0.04508347
#> Vatnsdalsa:Temperature.lag(2)   -3.3680e-02 -0.02505871 -1.6916e-02 -0.00092964
#> 
#> 
#> Scale parameter:
#>                           Mean        Sd   Sd(Mean)     2.5%      25%      50%
#> Jokulsa.Jokulsa       1.313088 0.2081044 0.01634820 0.949891 1.169092 1.293302
#> Jokulsa.Vatnsdalsa    0.045220 0.0117358 0.00062115 0.024760 0.037140 0.044202
#> Vatnsdalsa.Vatnsdalsa 0.022742 0.0036684 0.00026456 0.016116 0.020215 0.022565
#>                            75%    97.5%
#> Jokulsa.Jokulsa       1.439655 1.769184
#> Jokulsa.Vatnsdalsa    0.052637 0.069916
#> Vatnsdalsa.Vatnsdalsa 0.024913 0.030634
#> 
#> 
#> Extra parameter:
#>       Mean       Sd  Sd(Mean)    2.5%     25%     50%    75%   97.5%
#> nu 0.80922 0.040607 0.0030615 0.73465 0.78071 0.80819 0.8349 0.89432
#plot(fit3.mcmc)

###### Example 4: U.S. stock returns
data(US.returns)
fit4 <- mtar(~ CCR | dVIX, data=US.returns, subset={Date<="2025-11-28"},
             row.names=Date, ars=ars(nregim=2,p=3,d=3), n.burnin=1000,
             n.sim=2000, n.thin=2, dist="Student-t")
fit4.mcmc <- coda::as.mcmc(fit4)
summary(fit4.mcmc)
#> 
#> 
#>  Iterations = 1001:4999
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 2000
#> 
#> 
#> Thresholds:
#>             Mean      Sd Sd(Mean)   2.5%    25%    50%    75%  97.5%
#> threshold 1.4222 0.12925 0.026537 1.1562 1.3566 1.4431 1.5103 1.6507
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                      Mean        Sd   Sd(Mean)        2.5%       25%       50%
#> CCR:(Intercept)  0.090004 0.0119877 0.00031519  0.06690167  0.081970  0.089886
#> CCR:CCR.lag(1)  -0.049704 0.0145615 0.00043454 -0.07781327 -0.059552 -0.049459
#> CCR:CCR.lag(2)  -0.038782 0.0216978 0.00056343 -0.08136907 -0.053113 -0.038906
#> CCR:CCR.lag(3)  -0.019775 0.0231391 0.00072356 -0.06350973 -0.035481 -0.019907
#> CCR:dVIX.lag(1) -0.034107 0.0139458 0.00038726 -0.06180283 -0.043635 -0.034198
#> CCR:dVIX.lag(2) -0.018733 0.0142745 0.00042040 -0.04556743 -0.028205 -0.018818
#> CCR:dVIX.lag(3)  0.016551 0.0083545 0.00028667  0.00028107  0.010721  0.016615
#>                        75%      97.5%
#> CCR:(Intercept)  0.0978430  0.1144790
#> CCR:CCR.lag(1)  -0.0401604 -0.0215670
#> CCR:CCR.lag(2)  -0.0238736  0.0036861
#> CCR:CCR.lag(3)  -0.0045244  0.0261261
#> CCR:dVIX.lag(1) -0.0246382 -0.0067232
#> CCR:dVIX.lag(2) -0.0092704  0.0110885
#> CCR:dVIX.lag(3)  0.0221363  0.0331622
#> 
#> 
#> Scale parameter:
#>            Mean       Sd   Sd(Mean)    2.5%     25%     50%     75% 97.5%
#> CCR.CCR 0.36035 0.015263 0.00079117 0.33173 0.34973 0.35963 0.37104 0.392
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                       Mean       Sd  Sd(Mean)      2.5%        25%        50%
#> CCR:(Intercept)  0.0850813 0.109453 0.0037307 -0.130376  0.0097787  0.0846968
#> CCR:CCR.lag(1)  -0.0723153 0.059752 0.0020843 -0.193442 -0.1131221 -0.0729988
#> CCR:CCR.lag(2)   0.0074577 0.092662 0.0038946 -0.177386 -0.0538180  0.0071382
#> CCR:CCR.lag(3)   0.0502517 0.073575 0.0021689 -0.095158  0.0022875  0.0515444
#> CCR:dVIX.lag(1) -0.0053496 0.059236 0.0020494 -0.119548 -0.0449160 -0.0065261
#> CCR:dVIX.lag(2) -0.0057799 0.043672 0.0015151 -0.090987 -0.0354109 -0.0050868
#> CCR:dVIX.lag(3)  0.0419153 0.029103 0.0015882 -0.014531  0.0226075  0.0420801
#>                       75%    97.5%
#> CCR:(Intercept)  0.156937 0.301647
#> CCR:CCR.lag(1)  -0.031618 0.045193
#> CCR:CCR.lag(2)   0.069053 0.187837
#> CCR:CCR.lag(3)   0.099268 0.194784
#> CCR:dVIX.lag(1)  0.034319 0.109851
#> CCR:dVIX.lag(2)  0.023780 0.080399
#> CCR:dVIX.lag(3)  0.061903 0.097129
#> 
#> 
#> Scale parameter:
#>            Mean     Sd  Sd(Mean)    2.5%     25%     50%     75%  97.5%
#> CCR.CCR 0.89638 0.0936 0.0050363 0.72502 0.83347 0.89331 0.95781 1.0846
#> 
#> 
#> Extra parameter:
#>      Mean      Sd  Sd(Mean)   2.5%    25%    50%    75%  97.5%
#> nu 2.4405 0.11263 0.0073787 2.2303 2.3651 2.4361 2.5145 2.6713
#plot(fit4.mcmc)
# }