Coercion of mtar objects to mcmc objects

This method converts an object of class mtar into a list of mcmc objects, each corresponding to a Markov chain produced during Bayesian estimation.

Usage

# S3 method for class 'mtar'
as.mcmc(x, ...)

Arguments

x

an object of class mtar obtained from a call to mtar().

...

additional arguments passed to specific coercion methods.

Value

A list of mcmc objects containing the posterior simulation draws generated by the mtar() routine.

See also

Examples

# \donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
             subset={Date<="2015-12-07"}, dist="Student-t",
             ars=ars(nregim=3,p=c(1,1,2)), n.burnin=100, n.sim=200,
             n.thin=2, ssvs=TRUE)
fit1.mcmc <- coda::as.mcmc(fit1)
summary(fit1.mcmc)
#> 
#> 
#>  Iterations = 101:499
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 200
#> 
#> 
#> Thresholds:
#>                   Mean         Sd   Sd(Mean)       2.5%        25%        50%
#> Threshold.1 -0.0015704 0.00040343 9.8811e-05 -0.0033046 -0.0014981 -0.0014981
#> Threshold.2  0.0081640 0.00034864 2.3824e-04  0.0075809  0.0079687  0.0083086
#>                    75%      97.5%
#> Threshold.1 -0.0014600 -0.0012082
#> Threshold.2  0.0083086  0.0085583
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP:(Intercept)     -0.0030169 0.00045708 6.0248e-05 -0.0038826 -0.0032846
#> COLCAP:COLCAP.lag(1)    0.1379850 0.04778368 3.3788e-03  0.0385074  0.1051673
#> COLCAP:BOVESPA.lag(1)   0.0709223 0.03018100 2.1341e-03  0.0139327  0.0497841
#> BOVESPA:(Intercept)    -0.0075388 0.00066615 8.9857e-05 -0.0090556 -0.0079743
#> BOVESPA:COLCAP.lag(1)   0.0105890 0.06276100 4.0632e-03 -0.1028637 -0.0326674
#> BOVESPA:BOVESPA.lag(1)  0.0030930 0.04454147 3.1496e-03 -0.0691770 -0.0261490
#>                               50%        75%      97.5%
#> COLCAP:(Intercept)     -0.0029683 -0.0027279 -0.0021661
#> COLCAP:COLCAP.lag(1)    0.1380977  0.1695964  0.2385282
#> COLCAP:BOVESPA.lag(1)   0.0698567  0.0928912  0.1277778
#> BOVESPA:(Intercept)    -0.0075011 -0.0070768 -0.0064423
#> BOVESPA:COLCAP.lag(1)   0.0092444  0.0471468  0.1432526
#> BOVESPA:BOVESPA.lag(1)  0.0025078  0.0300333  0.0967621
#> 
#> 
#> Scale parameter:
#>                       Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP.COLCAP   6.0519e-05 5.3116e-06 4.5795e-07 5.2299e-05 5.7003e-05
#> COLCAP.BOVESPA  3.2890e-05 4.5431e-06 4.0919e-07 2.3591e-05 2.9893e-05
#> BOVESPA.BOVESPA 1.2630e-04 1.0582e-05 1.0337e-06 1.0839e-04 1.1896e-04
#>                        50%        75%      97.5%
#> COLCAP.COLCAP   5.9594e-05 6.3688e-05 7.1338e-05
#> COLCAP.BOVESPA  3.2947e-05 3.5462e-05 4.1463e-05
#> BOVESPA.BOVESPA 1.2559e-04 1.3267e-04 1.4793e-04
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                               Mean         Sd   Sd(Mean)        2.5%
#> COLCAP:(Intercept)      0.00047208 0.00030375 2.1478e-05 -1.4768e-04
#> COLCAP:COLCAP.lag(1)    0.06575099 0.03841514 2.7164e-03 -8.0012e-03
#> COLCAP:BOVESPA.lag(1)   0.08401157 0.02207752 2.1772e-03  4.2321e-02
#> BOVESPA:(Intercept)     0.00076591 0.00041088 4.2179e-05  8.0515e-05
#> BOVESPA:COLCAP.lag(1)   0.04110792 0.04686585 3.8492e-03 -4.1290e-02
#> BOVESPA:BOVESPA.lag(1) -0.04721785 0.03477892 2.4592e-03 -1.0351e-01
#>                                25%         50%         75%     97.5%
#> COLCAP:(Intercept)      0.00026578  0.00044830  0.00066066 0.0010856
#> COLCAP:COLCAP.lag(1)    0.03910855  0.06657204  0.09601748 0.1343307
#> COLCAP:BOVESPA.lag(1)   0.07024464  0.08315921  0.09682998 0.1309042
#> BOVESPA:(Intercept)     0.00047792  0.00080724  0.00105204 0.0015451
#> BOVESPA:COLCAP.lag(1)   0.01084602  0.04411217  0.07139244 0.1209733
#> BOVESPA:BOVESPA.lag(1) -0.07014447 -0.04601177 -0.02383935 0.0156685
#> 
#> 
#> Scale parameter:
#>                       Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP.COLCAP   3.9501e-05 2.8600e-06 3.9760e-07 3.3919e-05 3.7706e-05
#> COLCAP.BOVESPA  1.0531e-05 2.3379e-06 1.6532e-07 6.0204e-06 8.9819e-06
#> BOVESPA.BOVESPA 7.7062e-05 5.7915e-06 5.7723e-07 6.7093e-05 7.3042e-05
#>                        50%        75%      97.5%
#> COLCAP.COLCAP   3.9482e-05 4.1306e-05 4.4597e-05
#> COLCAP.BOVESPA  1.0442e-05 1.2047e-05 1.5188e-05
#> BOVESPA.BOVESPA 7.7021e-05 8.0850e-05 8.8926e-05
#> 
#> 
#> Regime 3
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP:(Intercept)      0.0048316 0.00056953 4.0272e-05  0.0036192  0.0044708
#> COLCAP:COLCAP.lag(1)    0.0501918 0.06350180 4.0503e-03 -0.0665759  0.0037104
#> COLCAP:BOVESPA.lag(1)   0.0709977 0.04505180 3.7821e-03 -0.0203482  0.0380490
#> COLCAP:COLCAP.lag(2)    0.0667071 0.05996957 5.3049e-03 -0.0615165  0.0268626
#> COLCAP:BOVESPA.lag(2)  -0.0819840 0.03992233 3.1725e-03 -0.1524631 -0.1092179
#> BOVESPA:(Intercept)     0.0120410 0.00093558 6.6156e-05  0.0104884  0.0113879
#> BOVESPA:COLCAP.lag(1)   0.1518670 0.09649154 6.8230e-03 -0.0290288  0.0856866
#> BOVESPA:BOVESPA.lag(1) -0.1342812 0.07334714 6.8892e-03 -0.2703567 -0.1825595
#> BOVESPA:COLCAP.lag(2)  -0.0767417 0.09999033 7.0704e-03 -0.2766744 -0.1383482
#> BOVESPA:BOVESPA.lag(2) -0.0499785 0.06070038 4.2922e-03 -0.1668229 -0.0885379
#>                               50%        75%      97.5%
#> COLCAP:(Intercept)      0.0048251  0.0051938  0.0059027
#> COLCAP:COLCAP.lag(1)    0.0528018  0.0996909  0.1582552
#> COLCAP:BOVESPA.lag(1)   0.0694680  0.0987029  0.1585834
#> COLCAP:COLCAP.lag(2)    0.0703653  0.1062830  0.1904573
#> COLCAP:BOVESPA.lag(2)  -0.0847855 -0.0578492  0.0012511
#> BOVESPA:(Intercept)     0.0119666  0.0126036  0.0140753
#> BOVESPA:COLCAP.lag(1)   0.1527960  0.2101642  0.3372482
#> BOVESPA:BOVESPA.lag(1) -0.1350334 -0.0865545 -0.0033522
#> BOVESPA:COLCAP.lag(2)  -0.0820886 -0.0202161  0.1442965
#> BOVESPA:BOVESPA.lag(2) -0.0520063 -0.0153912  0.0765832
#> 
#> 
#> Scale parameter:
#>                       Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP.COLCAP   5.5386e-05 6.3559e-06 5.0306e-07 4.4120e-05 5.1316e-05
#> COLCAP.BOVESPA  2.6945e-05 7.4021e-06 5.2341e-07 1.2617e-05 2.1985e-05
#> BOVESPA.BOVESPA 1.3758e-04 1.4258e-05 9.7806e-07 1.1405e-04 1.2676e-04
#>                        50%        75%      97.5%
#> COLCAP.COLCAP   5.4779e-05 5.9453e-05 6.7880e-05
#> COLCAP.BOVESPA  2.7232e-05 3.1970e-05 4.0672e-05
#> BOVESPA.BOVESPA 1.3747e-04 1.4676e-04 1.6641e-04
#> 
#> 
#> Extra parameter:
#>      Mean      Sd Sd(Mean)  2.5%    25%    50%    75%  97.5%
#> nu 5.4435 0.58722  0.14681 4.572 4.9351 5.4104 5.9494 6.5347
#plot(fit1.mcmc)

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, row.names=Date,
             subset={Date<="2009-02-13"}, dist="Laplace",
             ars=ars(nregim=3,p=5), n.burnin=100, n.sim=200, n.thin=2)
fit2.mcmc <- coda::as.mcmc(fit2)
summary(fit2.mcmc)
#> 
#> 
#>  Iterations = 101:499
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 200
#> 
#> 
#> Thresholds:
#>                Mean        Sd  Sd(Mean) 2.5%     25%     50%     75%  97.5%
#> Threshold.1  3.1732 0.0873641 0.0386706    3  3.1335  3.1582  3.2604  3.262
#> Threshold.2 10.0039 0.0037908 0.0011869   10 10.0022 10.0030 10.0041 10.016
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean        Sd   Sd(Mean)       2.5%        25%
#> Bedon:(Intercept)       1.3220352 0.1019647 0.00997124  1.1324093  1.2560170
#> Bedon:Bedon.lag(1)      0.5644270 0.0398889 0.00357450  0.4966624  0.5369192
#> Bedon:LaPlata.lag(1)    0.0461370 0.0144712 0.00134675  0.0195937  0.0368419
#> Bedon:Bedon.lag(2)      0.0511096 0.0330614 0.00319497 -0.0070626  0.0250264
#> Bedon:LaPlata.lag(2)   -0.0209183 0.0133878 0.00122535 -0.0484938 -0.0296542
#> Bedon:Bedon.lag(3)      0.0258411 0.0309318 0.00303575 -0.0273461  0.0040119
#> Bedon:LaPlata.lag(3)    0.0045734 0.0113169 0.00131921 -0.0189427 -0.0037912
#> Bedon:Bedon.lag(4)      0.0381240 0.0303154 0.00307585 -0.0201999  0.0181501
#> Bedon:LaPlata.lag(4)   -0.0169986 0.0085230 0.00071043 -0.0338814 -0.0222784
#> Bedon:Bedon.lag(5)      0.0830955 0.0256308 0.00287444  0.0342763  0.0655201
#> Bedon:LaPlata.lag(5)   -0.0073456 0.0072185 0.00053602 -0.0218788 -0.0118504
#> LaPlata:(Intercept)     3.4197730 0.2955877 0.02921470  2.8819234  3.2245152
#> LaPlata:Bedon.lag(1)    0.1658265 0.1191039 0.01235077 -0.0228704  0.0821870
#> LaPlata:LaPlata.lag(1)  0.6344990 0.0414172 0.00456503  0.5510638  0.6053851
#> LaPlata:Bedon.lag(2)   -0.0605884 0.0917623 0.00929824 -0.2593677 -0.1093337
#> LaPlata:LaPlata.lag(2) -0.0666999 0.0309407 0.00282934 -0.1282098 -0.0881902
#> LaPlata:Bedon.lag(3)    0.0152486 0.0777125 0.00756713 -0.1433863 -0.0330052
#> LaPlata:LaPlata.lag(3)  0.0675917 0.0245413 0.00202024  0.0155268  0.0512849
#> LaPlata:Bedon.lag(4)   -0.0775384 0.0791501 0.00538009 -0.2370006 -0.1307437
#> LaPlata:LaPlata.lag(4)  0.0084946 0.0248169 0.00244424 -0.0423478 -0.0064076
#> LaPlata:Bedon.lag(5)    0.1509150 0.0690208 0.00697014  0.0204504  0.1063218
#> LaPlata:LaPlata.lag(5)  0.0206459 0.0231018 0.00188276 -0.0257146  0.0062717
#>                               50%         75%       97.5%
#> Bedon:(Intercept)       1.3193449  1.38983559  1.52390479
#> Bedon:Bedon.lag(1)      0.5617613  0.59030894  0.64038296
#> Bedon:LaPlata.lag(1)    0.0467336  0.05668171  0.07265074
#> Bedon:Bedon.lag(2)      0.0528638  0.07180600  0.11165873
#> Bedon:LaPlata.lag(2)   -0.0206298 -0.01207780  0.00341941
#> Bedon:Bedon.lag(3)      0.0264811  0.04412203  0.09697826
#> Bedon:LaPlata.lag(3)    0.0057017  0.01310019  0.02310418
#> Bedon:Bedon.lag(4)      0.0366130  0.05787241  0.09357313
#> Bedon:LaPlata.lag(4)   -0.0169424 -0.01188768 -0.00056407
#> Bedon:Bedon.lag(5)      0.0810464  0.10048991  0.13098070
#> Bedon:LaPlata.lag(5)   -0.0069528 -0.00192396  0.00464513
#> LaPlata:(Intercept)     3.3841450  3.62383387  4.03530157
#> LaPlata:Bedon.lag(1)    0.1536810  0.23651315  0.44120670
#> LaPlata:LaPlata.lag(1)  0.6368828  0.66064996  0.71251308
#> LaPlata:Bedon.lag(2)   -0.0518992 -0.00017928  0.11456641
#> LaPlata:LaPlata.lag(2) -0.0691976 -0.04514089 -0.01592881
#> LaPlata:Bedon.lag(3)    0.0119832  0.07361810  0.15385043
#> LaPlata:LaPlata.lag(3)  0.0669601  0.08527495  0.11606844
#> LaPlata:Bedon.lag(4)   -0.0718978 -0.01642066  0.06646769
#> LaPlata:LaPlata.lag(4)  0.0063328  0.02459172  0.05813333
#> LaPlata:Bedon.lag(5)    0.1519332  0.19540224  0.28102278
#> LaPlata:LaPlata.lag(5)  0.0237321  0.03585088  0.06225368
#> 
#> 
#> Scale parameter:
#>                    Mean       Sd  Sd(Mean)    2.5%     25%     50%     75%
#> Bedon.Bedon     0.32897 0.032817 0.0033532 0.26469 0.30520 0.33104 0.35093
#> Bedon.LaPlata   0.37021 0.061205 0.0056077 0.24885 0.32948 0.36249 0.41220
#> LaPlata.LaPlata 2.36148 0.237068 0.0217740 1.91811 2.19162 2.35726 2.50242
#>                   97.5%
#> Bedon.Bedon     0.39122
#> Bedon.LaPlata   0.48933
#> LaPlata.LaPlata 2.85497
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean       Sd  Sd(Mean)       2.5%        25%
#> Bedon:(Intercept)       2.1644994 0.398818 0.0466455  1.4351968  1.8347245
#> Bedon:Bedon.lag(1)      0.5837136 0.044359 0.0047131  0.5001564  0.5556680
#> Bedon:LaPlata.lag(1)    0.0213059 0.012905 0.0011851 -0.0028246  0.0124902
#> Bedon:Bedon.lag(2)      0.1083160 0.067232 0.0091411 -0.0103649  0.0620049
#> Bedon:LaPlata.lag(2)   -0.0220019 0.018519 0.0023911 -0.0587318 -0.0336609
#> Bedon:Bedon.lag(3)     -0.0437056 0.058162 0.0074676 -0.1551532 -0.0822192
#> Bedon:LaPlata.lag(3)   -0.0059995 0.015523 0.0017575 -0.0349683 -0.0163380
#> Bedon:Bedon.lag(4)      0.1017669 0.065360 0.0070905 -0.0355455  0.0599450
#> Bedon:LaPlata.lag(4)    0.0081045 0.018510 0.0020420 -0.0249613 -0.0046025
#> Bedon:Bedon.lag(5)      0.0313367 0.046792 0.0052689 -0.0492157 -0.0024524
#> Bedon:LaPlata.lag(5)    0.0013371 0.016225 0.0020723 -0.0289993 -0.0101728
#> LaPlata:(Intercept)     6.9813682 1.046652 0.1146345  4.7413583  6.3285208
#> LaPlata:Bedon.lag(1)    0.1541159 0.122894 0.0123725 -0.0752362  0.0738841
#> LaPlata:LaPlata.lag(1)  0.5242683 0.037182 0.0034497  0.4348442  0.5026668
#> LaPlata:Bedon.lag(2)    0.0071711 0.140142 0.0128820 -0.2566729 -0.0832243
#> LaPlata:LaPlata.lag(2)  0.0323805 0.040127 0.0040926 -0.0552131  0.0075749
#> LaPlata:Bedon.lag(3)   -0.0878083 0.118786 0.0113659 -0.3176517 -0.1791852
#> LaPlata:LaPlata.lag(3)  0.0468602 0.040180 0.0043793 -0.0262030  0.0159439
#> LaPlata:Bedon.lag(4)    0.2304865 0.142478 0.0162891 -0.0568971  0.1292867
#> LaPlata:LaPlata.lag(4) -0.0473946 0.042799 0.0048099 -0.1397173 -0.0743306
#> LaPlata:Bedon.lag(5)   -0.2627784 0.115809 0.0128146 -0.4872729 -0.3377605
#> LaPlata:LaPlata.lag(5)  0.1150192 0.040094 0.0052946  0.0376415  0.0896651
#>                               50%        75%     97.5%
#> Bedon:(Intercept)       2.1882110  2.4364178  2.937074
#> Bedon:Bedon.lag(1)      0.5881853  0.6119048  0.659125
#> Bedon:LaPlata.lag(1)    0.0209474  0.0299056  0.048608
#> Bedon:Bedon.lag(2)      0.1080549  0.1517038  0.246557
#> Bedon:LaPlata.lag(2)   -0.0207264 -0.0092543  0.010298
#> Bedon:Bedon.lag(3)     -0.0500201 -0.0019374  0.069355
#> Bedon:LaPlata.lag(3)   -0.0058171  0.0052364  0.024829
#> Bedon:Bedon.lag(4)      0.1055841  0.1530379  0.209215
#> Bedon:LaPlata.lag(4)    0.0099305  0.0210612  0.042913
#> Bedon:Bedon.lag(5)      0.0280046  0.0647158  0.118514
#> Bedon:LaPlata.lag(5)    0.0033343  0.0123017  0.033190
#> LaPlata:(Intercept)     6.9566536  7.7243513  8.957696
#> LaPlata:Bedon.lag(1)    0.1510620  0.2297829  0.401707
#> LaPlata:LaPlata.lag(1)  0.5257864  0.5493561  0.588391
#> LaPlata:Bedon.lag(2)   -0.0069121  0.0906941  0.275620
#> LaPlata:LaPlata.lag(2)  0.0320918  0.0604918  0.107861
#> LaPlata:Bedon.lag(3)   -0.0967152  0.0051373  0.135667
#> LaPlata:LaPlata.lag(3)  0.0474588  0.0743120  0.124892
#> LaPlata:Bedon.lag(4)    0.2305306  0.3287845  0.480218
#> LaPlata:LaPlata.lag(4) -0.0446981 -0.0158014  0.031027
#> LaPlata:Bedon.lag(5)   -0.2732255 -0.1757297 -0.032771
#> LaPlata:LaPlata.lag(5)  0.1135194  0.1421402  0.201402
#> 
#> 
#> Scale parameter:
#>                   Mean      Sd Sd(Mean)   2.5%    25%    50%    75%  97.5%
#> Bedon.Bedon     1.0914 0.10418 0.009171 0.9250 1.0151 1.0840 1.1650 1.3179
#> Bedon.LaPlata   1.3525 0.18341 0.015995 1.0255 1.2501 1.3447 1.4372 1.7766
#> LaPlata.LaPlata 6.4701 0.65224 0.074437 5.2483 6.1028 6.4497 6.8191 7.9005
#> 
#> 
#> Regime 3
#> 
#> 
#> Autoregressive coefficients:
#>                               Mean       Sd  Sd(Mean)      2.5%       25%
#> Bedon:(Intercept)       5.43489327 0.792165 0.1129349  3.973462  4.848253
#> Bedon:Bedon.lag(1)      0.44105252 0.088884 0.0137359  0.276248  0.383303
#> Bedon:LaPlata.lag(1)    0.04915687 0.018792 0.0022812  0.013816  0.037309
#> Bedon:Bedon.lag(2)      0.09393121 0.075235 0.0082793 -0.049845  0.036351
#> Bedon:LaPlata.lag(2)    0.00058550 0.016844 0.0015321 -0.031256 -0.012083
#> Bedon:Bedon.lag(3)     -0.07796135 0.074368 0.0080351 -0.200000 -0.130305
#> Bedon:LaPlata.lag(3)    0.03157646 0.021299 0.0021680 -0.010818  0.018874
#> Bedon:Bedon.lag(4)     -0.00997218 0.075222 0.0067056 -0.190477 -0.055538
#> Bedon:LaPlata.lag(4)    0.00488445 0.022999 0.0023602 -0.035288 -0.010301
#> Bedon:Bedon.lag(5)      0.19220428 0.070943 0.0073142  0.044252  0.145924
#> Bedon:LaPlata.lag(5)   -0.01522527 0.018869 0.0018250 -0.050251 -0.027795
#> LaPlata:(Intercept)    16.51613226 2.662362 0.2875837 11.409488 14.758404
#> LaPlata:Bedon.lag(1)    0.53289887 0.237879 0.0238649  0.094524  0.373218
#> LaPlata:LaPlata.lag(1)  0.33174085 0.062258 0.0062800  0.228951  0.290229
#> LaPlata:Bedon.lag(2)   -0.57238325 0.301387 0.0353196 -1.161505 -0.800660
#> LaPlata:LaPlata.lag(2)  0.13148865 0.078820 0.0085432 -0.018202  0.078115
#> LaPlata:Bedon.lag(3)   -0.55545600 0.205062 0.0171049 -0.945387 -0.681142
#> LaPlata:LaPlata.lag(3)  0.28286882 0.084823 0.0111445  0.122333  0.232554
#> LaPlata:Bedon.lag(4)    0.00654002 0.259495 0.0251034 -0.502643 -0.174299
#> LaPlata:LaPlata.lag(4) -0.00086377 0.086347 0.0091904 -0.151037 -0.061161
#> LaPlata:Bedon.lag(5)    0.29339389 0.255124 0.0255321 -0.167696  0.103311
#> LaPlata:LaPlata.lag(5)  0.06174044 0.072050 0.0075010 -0.081232  0.017478
#>                                50%        75%     97.5%
#> Bedon:(Intercept)       5.3880e+00  5.9942814  6.868993
#> Bedon:Bedon.lag(1)      4.3404e-01  0.4992573  0.620091
#> Bedon:LaPlata.lag(1)    4.9423e-02  0.0602581  0.085969
#> Bedon:Bedon.lag(2)      9.6868e-02  0.1506527  0.239009
#> Bedon:LaPlata.lag(2)    1.6022e-05  0.0120068  0.032021
#> Bedon:Bedon.lag(3)     -8.4505e-02 -0.0294757  0.062463
#> Bedon:LaPlata.lag(3)    3.0905e-02  0.0469503  0.066773
#> Bedon:Bedon.lag(4)     -5.6713e-03  0.0385835  0.130510
#> Bedon:LaPlata.lag(4)    1.9916e-03  0.0185954  0.053832
#> Bedon:Bedon.lag(5)      1.9645e-01  0.2373123  0.320461
#> Bedon:LaPlata.lag(5)   -1.6325e-02 -0.0013157  0.023405
#> LaPlata:(Intercept)     1.6729e+01 18.4930213 21.449124
#> LaPlata:Bedon.lag(1)    5.2098e-01  0.6930274  0.979990
#> LaPlata:LaPlata.lag(1)  3.3142e-01  0.3693390  0.468746
#> LaPlata:Bedon.lag(2)   -5.4587e-01 -0.3775337 -0.034162
#> LaPlata:LaPlata.lag(2)  1.3075e-01  0.1806401  0.297310
#> LaPlata:Bedon.lag(3)   -5.6636e-01 -0.4445288 -0.132281
#> LaPlata:LaPlata.lag(3)  2.7882e-01  0.3346020  0.446564
#> LaPlata:Bedon.lag(4)    1.3708e-02  0.1950273  0.472808
#> LaPlata:LaPlata.lag(4) -1.0976e-02  0.0649603  0.168107
#> LaPlata:Bedon.lag(5)    2.7811e-01  0.4434454  0.815670
#> LaPlata:LaPlata.lag(5)  6.7064e-02  0.1125624  0.198426
#> 
#> 
#> Scale parameter:
#>                    Mean      Sd Sd(Mean)    2.5%     25%     50%     75%
#> Bedon.Bedon      2.7790 0.31959 0.029415  2.2222  2.5464  2.7723  2.9927
#> Bedon.LaPlata    7.2322 1.03432 0.075986  5.5060  6.5100  7.2172  7.9008
#> LaPlata.LaPlata 43.4289 5.13693 0.421517 34.9021 39.8833 43.2229 46.5876
#>                   97.5%
#> Bedon.Bedon      3.3728
#> Bedon.LaPlata    9.2041
#> LaPlata.LaPlata 54.2245
#plot(fit2.mcmc)

###### Example 3: Temperature, precipitation, and two river flows in Iceland
data(iceland.rf)
fit3 <- mtar(~ Jokulsa + Vatnsdalsa | Temperature | Precipitation,
             data=iceland.rf, subset={Date<="1974-11-06"}, row.names=Date,
             ars=ars(nregim=2,p=15,q=4,d=2), n.burnin=100, n.sim=200,
             n.thin=2, dist="Slash")
fit3.mcmc <- coda::as.mcmc(fit3)
summary(fit3.mcmc)
#> 
#> 
#>  Iterations = 101:499
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 200
#> 
#> 
#> Thresholds:
#>              Mean       Sd  Sd(Mean)    2.5%     25%     50%     75%   97.5%
#> threshold 0.65232 0.027328 0.0044054 0.60048 0.63312 0.65157 0.66809 0.69811
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                                        Mean        Sd   Sd(Mean)        2.5%
#> Jokulsa:(Intercept)              3.31597764 0.2871775 0.03469329  2.76868312
#> Jokulsa:Jokulsa.lag( 1)          0.87843276 0.0252165 0.00246289  0.82215512
#> Jokulsa:Vatnsdalsa.lag( 1)       0.27128499 0.0621493 0.00712250  0.15607110
#> Jokulsa:Jokulsa.lag( 2)         -0.05305278 0.0244442 0.00256905 -0.09669210
#> Jokulsa:Vatnsdalsa.lag( 2)      -0.27928567 0.0757836 0.00851909 -0.40752109
#> Jokulsa:Jokulsa.lag( 3)         -0.00791600 0.0212430 0.00206864 -0.05220684
#> Jokulsa:Vatnsdalsa.lag( 3)       0.09569996 0.0554141 0.00547826 -0.00451994
#> Jokulsa:Jokulsa.lag( 4)          0.00092482 0.0263768 0.00303191 -0.04963529
#> Jokulsa:Vatnsdalsa.lag( 4)      -0.01171005 0.0520963 0.00615562 -0.11458218
#> Jokulsa:Jokulsa.lag( 5)         -0.00655746 0.0342798 0.00549308 -0.08558367
#> Jokulsa:Vatnsdalsa.lag( 5)      -0.02805058 0.0365001 0.00341355 -0.10575934
#> Jokulsa:Jokulsa.lag( 6)          0.01579666 0.0328781 0.00544708 -0.04564263
#> Jokulsa:Vatnsdalsa.lag( 6)      -0.01847398 0.0427215 0.00483873 -0.09461617
#> Jokulsa:Jokulsa.lag( 7)          0.01302036 0.0323574 0.00455537 -0.04487751
#> Jokulsa:Vatnsdalsa.lag( 7)       0.01601320 0.0369696 0.00448957 -0.06061974
#> Jokulsa:Jokulsa.lag( 8)         -0.00442467 0.0263675 0.00389063 -0.05348855
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.00739734 0.0285379 0.00230664 -0.06977037
#> Jokulsa:Jokulsa.lag( 9)          0.00029872 0.0240756 0.00313311 -0.04424579
#> Jokulsa:Vatnsdalsa.lag( 9)      -0.01606116 0.0294089 0.00234438 -0.07478509
#> Jokulsa:Jokulsa.lag(10)          0.02073854 0.0177734 0.00182056 -0.01307767
#> Jokulsa:Vatnsdalsa.lag(10)       0.02788672 0.0309371 0.00277291 -0.02715460
#> Jokulsa:Jokulsa.lag(11)         -0.01338678 0.0158734 0.00156818 -0.04239032
#> Jokulsa:Vatnsdalsa.lag(11)      -0.01277985 0.0310554 0.00299041 -0.07515962
#> Jokulsa:Jokulsa.lag(12)          0.00975873 0.0152897 0.00265604 -0.02138844
#> Jokulsa:Vatnsdalsa.lag(12)       0.00486086 0.0245341 0.00217854 -0.04819315
#> Jokulsa:Jokulsa.lag(13)         -0.01615101 0.0175667 0.00187855 -0.04553038
#> Jokulsa:Vatnsdalsa.lag(13)      -0.01670698 0.0281234 0.00229562 -0.07024558
#> Jokulsa:Jokulsa.lag(14)          0.00281484 0.0127212 0.00103524 -0.02154219
#> Jokulsa:Vatnsdalsa.lag(14)       0.01210680 0.0286163 0.00248367 -0.03776025
#> Jokulsa:Jokulsa.lag(15)          0.02018663 0.0153499 0.00198863 -0.00658457
#> Jokulsa:Vatnsdalsa.lag(15)      -0.01706744 0.0201641 0.00144652 -0.06311213
#> Jokulsa:Precipitation.lag(1)     0.00720397 0.0093912 0.00090175 -0.00969978
#> Jokulsa:Precipitation.lag(2)     0.00536919 0.0074016 0.00078606 -0.01099511
#> Jokulsa:Precipitation.lag(3)    -0.01101393 0.0066423 0.00074502 -0.02244599
#> Jokulsa:Precipitation.lag(4)     0.01926884 0.0066558 0.00061227  0.00567249
#> Jokulsa:Temperature.lag(1)       0.01430541 0.0100089 0.00085436 -0.00690994
#> Jokulsa:Temperature.lag(2)      -0.03066675 0.0102335 0.00097245 -0.04950191
#> Vatnsdalsa:(Intercept)           0.67613054 0.1826218 0.01932549  0.37058830
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.04840149 0.0173080 0.00186478 -0.08902244
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.16203336 0.0492668 0.00719242  1.05953099
#> Vatnsdalsa:Jokulsa.lag( 2)       0.05021856 0.0158600 0.00159906  0.01750448
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.32451671 0.0566654 0.00769845 -0.43725347
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.03306981 0.0138196 0.00149220 -0.06368043
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    0.07298983 0.0417645 0.00606781  0.00092183
#> Vatnsdalsa:Jokulsa.lag( 4)       0.01233849 0.0135364 0.00136540 -0.01312566
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -0.01715311 0.0396802 0.00755002 -0.09604661
#> Vatnsdalsa:Jokulsa.lag( 5)       0.01226298 0.0153467 0.00170826 -0.01508391
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -0.00696600 0.0228154 0.00199281 -0.05027279
#> Vatnsdalsa:Jokulsa.lag( 6)      -0.00036501 0.0150581 0.00151662 -0.02651218
#> Vatnsdalsa:Vatnsdalsa.lag( 6)   -0.01863454 0.0265513 0.00281639 -0.07143266
#> Vatnsdalsa:Jokulsa.lag( 7)       0.00477518 0.0171099 0.00179167 -0.02684659
#> Vatnsdalsa:Vatnsdalsa.lag( 7)    0.01521973 0.0196911 0.00166309 -0.02180639
#> Vatnsdalsa:Jokulsa.lag( 8)      -0.01901088 0.0140311 0.00145637 -0.04468151
#> Vatnsdalsa:Vatnsdalsa.lag( 8)    0.01192925 0.0163905 0.00111430 -0.02074988
#> Vatnsdalsa:Jokulsa.lag( 9)       0.02494118 0.0163093 0.00193865 -0.00929839
#> Vatnsdalsa:Vatnsdalsa.lag( 9)    0.00042155 0.0214096 0.00141958 -0.04301097
#> Vatnsdalsa:Jokulsa.lag(10)      -0.01916468 0.0132790 0.00133364 -0.04225202
#> Vatnsdalsa:Vatnsdalsa.lag(10)    0.01439392 0.0203332 0.00201160 -0.02343691
#> Vatnsdalsa:Jokulsa.lag(11)       0.01045173 0.0095431 0.00080738 -0.00595994
#> Vatnsdalsa:Vatnsdalsa.lag(11)    0.00029073 0.0201635 0.00191400 -0.03512382
#> Vatnsdalsa:Jokulsa.lag(12)      -0.00908408 0.0095442 0.00094801 -0.02921852
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.00656747 0.0191316 0.00167606 -0.04271149
#> Vatnsdalsa:Jokulsa.lag(13)       0.00503675 0.0112929 0.00067962 -0.01511542
#> Vatnsdalsa:Vatnsdalsa.lag(13)   -0.02924035 0.0227996 0.00244518 -0.07777448
#> Vatnsdalsa:Jokulsa.lag(14)      -0.00893364 0.0098198 0.00083943 -0.02718676
#> Vatnsdalsa:Vatnsdalsa.lag(14)    0.04468357 0.0189786 0.00167683  0.00844713
#> Vatnsdalsa:Jokulsa.lag(15)       0.00587526 0.0073374 0.00061758 -0.00970388
#> Vatnsdalsa:Vatnsdalsa.lag(15)    0.00421870 0.0137238 0.00163275 -0.01950085
#> Vatnsdalsa:Precipitation.lag(1)  0.00148228 0.0060874 0.00059070 -0.00956911
#> Vatnsdalsa:Precipitation.lag(2)  0.00157304 0.0048284 0.00045087 -0.00833025
#> Vatnsdalsa:Precipitation.lag(3) -0.00311641 0.0045309 0.00038898 -0.01149717
#> Vatnsdalsa:Precipitation.lag(4)  0.00551756 0.0050606 0.00047619 -0.00501884
#> Vatnsdalsa:Temperature.lag(1)   -0.00078532 0.0074171 0.00063958 -0.01422193
#> Vatnsdalsa:Temperature.lag(2)   -0.01087030 0.0073456 0.00065072 -0.02346811
#>                                         25%         50%         75%      97.5%
#> Jokulsa:(Intercept)              3.12231148  3.30101303  3.53137228  3.8721758
#> Jokulsa:Jokulsa.lag( 1)          0.86326007  0.88242963  0.89706698  0.9196946
#> Jokulsa:Vatnsdalsa.lag( 1)       0.22254039  0.27492930  0.31269493  0.3809861
#> Jokulsa:Jokulsa.lag( 2)         -0.07206913 -0.05285699 -0.03552445 -0.0081747
#> Jokulsa:Vatnsdalsa.lag( 2)      -0.32931808 -0.28435550 -0.22882288 -0.1397838
#> Jokulsa:Jokulsa.lag( 3)         -0.02285969 -0.00682311  0.00824460  0.0252809
#> Jokulsa:Vatnsdalsa.lag( 3)       0.05821994  0.09231383  0.12854861  0.2078557
#> Jokulsa:Jokulsa.lag( 4)         -0.01452615 -0.00113705  0.01605690  0.0547075
#> Jokulsa:Vatnsdalsa.lag( 4)      -0.04461858 -0.00813783  0.02423514  0.0755009
#> Jokulsa:Jokulsa.lag( 5)         -0.03031126  0.00113063  0.01884167  0.0436288
#> Jokulsa:Vatnsdalsa.lag( 5)      -0.05184657 -0.02803598 -0.00220505  0.0330428
#> Jokulsa:Jokulsa.lag( 6)         -0.00453377  0.01691859  0.03899802  0.0762470
#> Jokulsa:Vatnsdalsa.lag( 6)      -0.04848342 -0.02342130  0.00599146  0.0664988
#> Jokulsa:Jokulsa.lag( 7)         -0.00659669  0.01253450  0.02788577  0.0868097
#> Jokulsa:Vatnsdalsa.lag( 7)      -0.00708888  0.01546060  0.03934807  0.0896537
#> Jokulsa:Jokulsa.lag( 8)         -0.02435563 -0.00444482  0.01405274  0.0467138
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.02397892 -0.00609819  0.01184229  0.0448415
#> Jokulsa:Jokulsa.lag( 9)         -0.01636045  0.00040013  0.01680796  0.0438961
#> Jokulsa:Vatnsdalsa.lag( 9)      -0.03817038 -0.01556654  0.00221981  0.0404895
#> Jokulsa:Jokulsa.lag(10)          0.00871838  0.02094046  0.03273050  0.0526801
#> Jokulsa:Vatnsdalsa.lag(10)       0.00827776  0.02715546  0.04482136  0.0944934
#> Jokulsa:Jokulsa.lag(11)         -0.02349630 -0.01266258 -0.00308101  0.0152275
#> Jokulsa:Vatnsdalsa.lag(11)      -0.03217477 -0.01352083  0.01036128  0.0450952
#> Jokulsa:Jokulsa.lag(12)         -0.00102263  0.01000994  0.01955431  0.0387991
#> Jokulsa:Vatnsdalsa.lag(12)      -0.01218773  0.00684929  0.02189509  0.0506587
#> Jokulsa:Jokulsa.lag(13)         -0.02830944 -0.01718883 -0.00494659  0.0183093
#> Jokulsa:Vatnsdalsa.lag(13)      -0.03350273 -0.01597760 -0.00093886  0.0372722
#> Jokulsa:Jokulsa.lag(14)         -0.00609618  0.00325941  0.01154438  0.0280586
#> Jokulsa:Vatnsdalsa.lag(14)      -0.00678864  0.01097799  0.03240348  0.0613643
#> Jokulsa:Jokulsa.lag(15)          0.00883958  0.01984634  0.03193084  0.0473305
#> Jokulsa:Vatnsdalsa.lag(15)      -0.02953211 -0.01658250 -0.00397636  0.0209181
#> Jokulsa:Precipitation.lag(1)     0.00095259  0.00660540  0.01318193  0.0254017
#> Jokulsa:Precipitation.lag(2)     0.00093908  0.00564685  0.01004121  0.0175878
#> Jokulsa:Precipitation.lag(3)    -0.01515191 -0.01112775 -0.00843129  0.0042003
#> Jokulsa:Precipitation.lag(4)     0.01498232  0.01902798  0.02276670  0.0340430
#> Jokulsa:Temperature.lag(1)       0.00819654  0.01467357  0.02022756  0.0356471
#> Jokulsa:Temperature.lag(2)      -0.03730619 -0.03059441 -0.02446071 -0.0090603
#> Vatnsdalsa:(Intercept)           0.54726028  0.67785964  0.79783740  1.0196602
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.05551316 -0.04554439 -0.03711891 -0.0218726
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.12510854  1.16004846  1.20138356  1.2508062
#> Vatnsdalsa:Jokulsa.lag( 2)       0.04067402  0.05016266  0.05998600  0.0783185
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.36049303 -0.32170952 -0.28508412 -0.2071491
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.04096099 -0.03222109 -0.02371316 -0.0090646
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    0.04126434  0.06904742  0.10241248  0.1587052
#> Vatnsdalsa:Jokulsa.lag( 4)       0.00406181  0.01266433  0.02157321  0.0390034
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -0.03797660 -0.01188478  0.01131310  0.0465951
#> Vatnsdalsa:Jokulsa.lag( 5)       0.00105373  0.01258275  0.02177761  0.0451613
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -0.02069304 -0.00710148  0.00681591  0.0364391
#> Vatnsdalsa:Jokulsa.lag( 6)      -0.01012070 -0.00058244  0.00817367  0.0298478
#> Vatnsdalsa:Vatnsdalsa.lag( 6)   -0.03678722 -0.01975764 -0.00021429  0.0347481
#> Vatnsdalsa:Jokulsa.lag( 7)      -0.00629512  0.00346347  0.01743851  0.0378433
#> Vatnsdalsa:Vatnsdalsa.lag( 7)    0.00151939  0.01708982  0.02762018  0.0537267
#> Vatnsdalsa:Jokulsa.lag( 8)      -0.02874132 -0.01942884 -0.00992715  0.0108024
#> Vatnsdalsa:Vatnsdalsa.lag( 8)    0.00182199  0.01020030  0.02263950  0.0410879
#> Vatnsdalsa:Jokulsa.lag( 9)       0.01620498  0.02474582  0.03386324  0.0576012
#> Vatnsdalsa:Vatnsdalsa.lag( 9)   -0.01287279  0.00191435  0.01313528  0.0388053
#> Vatnsdalsa:Jokulsa.lag(10)      -0.02858801 -0.01890829 -0.00961878  0.0077041
#> Vatnsdalsa:Vatnsdalsa.lag(10)    0.00100803  0.01330751  0.02812216  0.0519328
#> Vatnsdalsa:Jokulsa.lag(11)       0.00398990  0.01042696  0.01744604  0.0291770
#> Vatnsdalsa:Vatnsdalsa.lag(11)   -0.01385915 -0.00168640  0.01318210  0.0446145
#> Vatnsdalsa:Jokulsa.lag(12)      -0.01471688 -0.00966183 -0.00248428  0.0087769
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.01944270 -0.00687746  0.00606090  0.0303696
#> Vatnsdalsa:Jokulsa.lag(13)      -0.00264651  0.00408684  0.01238322  0.0275737
#> Vatnsdalsa:Vatnsdalsa.lag(13)   -0.04130232 -0.02671993 -0.01483622  0.0117232
#> Vatnsdalsa:Jokulsa.lag(14)      -0.01503859 -0.00913798 -0.00279858  0.0104319
#> Vatnsdalsa:Vatnsdalsa.lag(14)    0.03177305  0.04343631  0.05723413  0.0849746
#> Vatnsdalsa:Jokulsa.lag(15)       0.00161289  0.00593869  0.01059201  0.0209084
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -0.00653464  0.00401509  0.01292930  0.0288367
#> Vatnsdalsa:Precipitation.lag(1) -0.00254236  0.00145048  0.00563201  0.0148789
#> Vatnsdalsa:Precipitation.lag(2) -0.00103942  0.00150793  0.00437348  0.0116442
#> Vatnsdalsa:Precipitation.lag(3) -0.00655266 -0.00307808  0.00034654  0.0062559
#> Vatnsdalsa:Precipitation.lag(4)  0.00215300  0.00588805  0.00878507  0.0150593
#> Vatnsdalsa:Temperature.lag(1)   -0.00562816 -0.00044218  0.00373010  0.0137513
#> Vatnsdalsa:Temperature.lag(2)   -0.01556561 -0.01107156 -0.00602543  0.0035362
#> 
#> 
#> Scale parameter:
#>                            Mean        Sd   Sd(Mean)      2.5%       25%
#> Jokulsa.Jokulsa       0.0609655 0.0081282 0.00130525 0.0453816 0.0558892
#> Jokulsa.Vatnsdalsa    0.0098107 0.0025155 0.00024954 0.0047626 0.0080441
#> Vatnsdalsa.Vatnsdalsa 0.0261827 0.0036640 0.00058811 0.0189619 0.0236421
#>                            50%      75%    97.5%
#> Jokulsa.Jokulsa       0.060905 0.067420 0.076514
#> Jokulsa.Vatnsdalsa    0.009740 0.011594 0.014222
#> Vatnsdalsa.Vatnsdalsa 0.026119 0.028736 0.034244
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                                        Mean        Sd   Sd(Mean)       2.5%
#> Jokulsa:(Intercept)             -0.20156603 0.6585899 0.06819094 -1.6078330
#> Jokulsa:Jokulsa.lag( 1)          1.00486612 0.0365779 0.00415236  0.9375927
#> Jokulsa:Vatnsdalsa.lag( 1)       0.91918336 0.1744176 0.01636035  0.5766093
#> Jokulsa:Jokulsa.lag( 2)         -0.14743850 0.0670949 0.00957265 -0.2918509
#> Jokulsa:Vatnsdalsa.lag( 2)      -0.42959032 0.2394453 0.02558654 -0.8586939
#> Jokulsa:Jokulsa.lag( 3)         -0.00083315 0.0579116 0.00967668 -0.1029937
#> Jokulsa:Vatnsdalsa.lag( 3)       0.00808555 0.2514929 0.02908830 -0.3957245
#> Jokulsa:Jokulsa.lag( 4)         -0.07654166 0.0420374 0.00378485 -0.1597099
#> Jokulsa:Vatnsdalsa.lag( 4)      -0.18755176 0.2128000 0.02092972 -0.6172107
#> Jokulsa:Jokulsa.lag( 5)          0.04378063 0.0387434 0.00426458 -0.0316252
#> Jokulsa:Vatnsdalsa.lag( 5)       0.25365555 0.2748655 0.03453456 -0.2594219
#> Jokulsa:Jokulsa.lag( 6)         -0.04941304 0.0318105 0.00432957 -0.1134444
#> Jokulsa:Vatnsdalsa.lag( 6)      -0.20355876 0.2334189 0.02684188 -0.6221812
#> Jokulsa:Jokulsa.lag( 7)          0.00294792 0.0295327 0.00245013 -0.0513677
#> Jokulsa:Vatnsdalsa.lag( 7)       0.22304201 0.2160381 0.02241321 -0.1448735
#> Jokulsa:Jokulsa.lag( 8)          0.02278441 0.0285862 0.00174936 -0.0274074
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.29091630 0.2024456 0.02875437 -0.7043498
#> Jokulsa:Jokulsa.lag( 9)          0.03945005 0.0289759 0.00256913 -0.0285424
#> Jokulsa:Vatnsdalsa.lag( 9)       0.15570938 0.2075548 0.02500073 -0.2330762
#> Jokulsa:Jokulsa.lag(10)         -0.03005497 0.0373356 0.00410627 -0.0976260
#> Jokulsa:Vatnsdalsa.lag(10)      -0.04822056 0.1906371 0.02216455 -0.4508246
#> Jokulsa:Jokulsa.lag(11)         -0.00693278 0.0375338 0.00417228 -0.0715971
#> Jokulsa:Vatnsdalsa.lag(11)       0.10322622 0.2204888 0.02705820 -0.2740774
#> Jokulsa:Jokulsa.lag(12)         -0.00209727 0.0351792 0.00317418 -0.0704885
#> Jokulsa:Vatnsdalsa.lag(12)       0.00951276 0.2280926 0.02389171 -0.4585370
#> Jokulsa:Jokulsa.lag(13)          0.02196109 0.0410944 0.00524268 -0.0496860
#> Jokulsa:Vatnsdalsa.lag(13)       0.21975093 0.2764569 0.05054351 -0.3556824
#> Jokulsa:Jokulsa.lag(14)         -0.03166100 0.0408637 0.00468781 -0.1095454
#> Jokulsa:Vatnsdalsa.lag(14)       0.32984887 0.1717807 0.02077239 -0.0048618
#> Jokulsa:Jokulsa.lag(15)          0.05376139 0.0237010 0.00124849  0.0089727
#> Jokulsa:Vatnsdalsa.lag(15)      -0.48202347 0.1183160 0.01176786 -0.7135689
#> Jokulsa:Precipitation.lag(1)    -0.11065966 0.0339093 0.00309303 -0.1674026
#> Jokulsa:Precipitation.lag(2)     0.00914154 0.0490475 0.00537954 -0.0798259
#> Jokulsa:Precipitation.lag(3)     0.03901521 0.0323117 0.00403937 -0.0169990
#> Jokulsa:Precipitation.lag(4)     0.02563294 0.0384110 0.00353196 -0.0456914
#> Jokulsa:Temperature.lag(1)       1.10655138 0.0936454 0.00772271  0.9089917
#> Jokulsa:Temperature.lag(2)      -0.55291640 0.0910209 0.00865022 -0.7494884
#> Vatnsdalsa:(Intercept)           0.47818088 0.0968096 0.00884343  0.3036550
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.00608815 0.0038755 0.00027404 -0.0133644
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.21183705 0.0326684 0.00554062  1.1540693
#> Vatnsdalsa:Jokulsa.lag( 2)       0.01661836 0.0063195 0.00044685  0.0047298
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.33287404 0.0442340 0.00692430 -0.4238531
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.01568937 0.0059963 0.00049346 -0.0269840
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    0.15688772 0.0422886 0.00431160  0.0781977
#> Vatnsdalsa:Jokulsa.lag( 4)       0.00562437 0.0049279 0.00034845 -0.0046025
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -0.08943536 0.0345788 0.00341636 -0.1590864
#> Vatnsdalsa:Jokulsa.lag( 5)      -0.00322154 0.0049900 0.00031522 -0.0128679
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -0.01776899 0.0396775 0.00482063 -0.0851058
#> Vatnsdalsa:Jokulsa.lag( 6)       0.00290140 0.0044135 0.00026039 -0.0074875
#> Vatnsdalsa:Vatnsdalsa.lag( 6)    0.04648555 0.0422238 0.00723505 -0.0303617
#> Vatnsdalsa:Jokulsa.lag( 7)      -0.00565607 0.0042750 0.00053340 -0.0129229
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -0.04097558 0.0405103 0.00951471 -0.1041988
#> Vatnsdalsa:Jokulsa.lag( 8)       0.00413969 0.0042380 0.00049853 -0.0050926
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -0.08441133 0.0282370 0.00276562 -0.1414029
#> Vatnsdalsa:Jokulsa.lag( 9)      -0.00312511 0.0042590 0.00035867 -0.0113745
#> Vatnsdalsa:Vatnsdalsa.lag( 9)    0.12620171 0.0357899 0.00385375  0.0499076
#> Vatnsdalsa:Jokulsa.lag(10)       0.00425553 0.0051347 0.00049138 -0.0068718
#> Vatnsdalsa:Vatnsdalsa.lag(10)   -0.07889897 0.0280445 0.00395159 -0.1281562
#> Vatnsdalsa:Jokulsa.lag(11)      -0.00784906 0.0047543 0.00037480 -0.0161876
#> Vatnsdalsa:Vatnsdalsa.lag(11)    0.05815769 0.0337702 0.00474386 -0.0204073
#> Vatnsdalsa:Jokulsa.lag(12)       0.01115106 0.0049505 0.00042399  0.0024146
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.05474948 0.0351407 0.00375623 -0.1223271
#> Vatnsdalsa:Jokulsa.lag(13)      -0.00771832 0.0049961 0.00042827 -0.0190907
#> Vatnsdalsa:Vatnsdalsa.lag(13)    0.14736827 0.0314802 0.00359818  0.0878276
#> Vatnsdalsa:Jokulsa.lag(14)      -0.00261402 0.0049700 0.00048504 -0.0121399
#> Vatnsdalsa:Vatnsdalsa.lag(14)   -0.02791368 0.0321311 0.00670795 -0.0968556
#> Vatnsdalsa:Jokulsa.lag(15)       0.00260715 0.0033742 0.00033760 -0.0028729
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -0.04842080 0.0279584 0.00612527 -0.0990142
#> Vatnsdalsa:Precipitation.lag(1)  0.00076717 0.0052895 0.00047126 -0.0083286
#> Vatnsdalsa:Precipitation.lag(2) -0.00789236 0.0073993 0.00085922 -0.0192794
#> Vatnsdalsa:Precipitation.lag(3)  0.00424657 0.0048695 0.00054949 -0.0060026
#> Vatnsdalsa:Precipitation.lag(4)  0.00510723 0.0045154 0.00036510 -0.0045862
#> Vatnsdalsa:Temperature.lag(1)    0.03880596 0.0125774 0.00104975  0.0146595
#> Vatnsdalsa:Temperature.lag(2)   -0.04590177 0.0130556 0.00140480 -0.0708504
#>                                         25%         50%         75%      97.5%
#> Jokulsa:(Intercept)             -0.63643084 -0.25799048  2.1707e-01  1.0898926
#> Jokulsa:Jokulsa.lag( 1)          0.98328877  1.00293899  1.0243e+00  1.0862541
#> Jokulsa:Vatnsdalsa.lag( 1)       0.82262409  0.91551987  1.0368e+00  1.2814157
#> Jokulsa:Jokulsa.lag( 2)         -0.18834468 -0.14409898 -9.9788e-02 -0.0304359
#> Jokulsa:Vatnsdalsa.lag( 2)      -0.58226841 -0.43087165 -2.7471e-01  0.0448600
#> Jokulsa:Jokulsa.lag( 3)         -0.03753011 -0.00746340  3.8737e-02  0.1209809
#> Jokulsa:Vatnsdalsa.lag( 3)      -0.15215835  0.00732571  1.6823e-01  0.4445196
#> Jokulsa:Jokulsa.lag( 4)         -0.10200333 -0.07878716 -4.8971e-02  0.0049401
#> Jokulsa:Vatnsdalsa.lag( 4)      -0.32592227 -0.19496434 -7.6293e-02  0.2277510
#> Jokulsa:Jokulsa.lag( 5)          0.01683422  0.04412857  6.9770e-02  0.1199537
#> Jokulsa:Vatnsdalsa.lag( 5)       0.05937104  0.25341915  4.3612e-01  0.8512140
#> Jokulsa:Jokulsa.lag( 6)         -0.06938942 -0.04924801 -2.6972e-02  0.0077411
#> Jokulsa:Vatnsdalsa.lag( 6)      -0.36913898 -0.21228051 -5.1951e-02  0.2755326
#> Jokulsa:Jokulsa.lag( 7)         -0.01747104  0.00265798  2.2026e-02  0.0641133
#> Jokulsa:Vatnsdalsa.lag( 7)       0.07498100  0.21465776  3.9067e-01  0.6547627
#> Jokulsa:Jokulsa.lag( 8)          0.00251290  0.02068056  4.4431e-02  0.0755989
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.41092234 -0.28314480 -1.6040e-01  0.1132444
#> Jokulsa:Jokulsa.lag( 9)          0.02161195  0.04100443  5.8211e-02  0.0927490
#> Jokulsa:Vatnsdalsa.lag( 9)       0.02191447  0.14988911  2.7957e-01  0.5833399
#> Jokulsa:Jokulsa.lag(10)         -0.05501490 -0.03243476 -4.1404e-03  0.0384511
#> Jokulsa:Vatnsdalsa.lag(10)      -0.14895201 -0.03235342  6.7699e-02  0.3032169
#> Jokulsa:Jokulsa.lag(11)         -0.03440613 -0.00869063  2.1593e-02  0.0656069
#> Jokulsa:Vatnsdalsa.lag(11)      -0.06135566  0.08762042  2.4170e-01  0.5863727
#> Jokulsa:Jokulsa.lag(12)         -0.02819983  0.00033134  2.3977e-02  0.0542437
#> Jokulsa:Vatnsdalsa.lag(12)      -0.14176765  0.01255389  1.5105e-01  0.4319162
#> Jokulsa:Jokulsa.lag(13)         -0.00605064  0.02095709  4.6878e-02  0.1060201
#> Jokulsa:Vatnsdalsa.lag(13)       0.03766546  0.21179073  4.0774e-01  0.7562885
#> Jokulsa:Jokulsa.lag(14)         -0.06109887 -0.03424330 -4.8713e-04  0.0471306
#> Jokulsa:Vatnsdalsa.lag(14)       0.22926893  0.32907835  4.4970e-01  0.6949462
#> Jokulsa:Jokulsa.lag(15)          0.03738794  0.05364083  6.9119e-02  0.0997622
#> Jokulsa:Vatnsdalsa.lag(15)      -0.55341539 -0.48030758 -4.0879e-01 -0.2715101
#> Jokulsa:Precipitation.lag(1)    -0.13750396 -0.10866272 -8.6574e-02 -0.0486585
#> Jokulsa:Precipitation.lag(2)    -0.02722138  0.00359020  4.6134e-02  0.1117915
#> Jokulsa:Precipitation.lag(3)     0.01334060  0.03890868  6.3776e-02  0.1006463
#> Jokulsa:Precipitation.lag(4)     0.00088446  0.02315376  5.2076e-02  0.0932150
#> Jokulsa:Temperature.lag(1)       1.05700511  1.10909454  1.1704e+00  1.2695135
#> Jokulsa:Temperature.lag(2)      -0.61070578 -0.55073912 -4.9244e-01 -0.3782865
#> Vatnsdalsa:(Intercept)           0.40290868  0.48137224  5.4940e-01  0.6398150
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.00866987 -0.00640735 -3.1902e-03  0.0012309
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.18880258  1.21024080  1.2304e+00  1.2774578
#> Vatnsdalsa:Jokulsa.lag( 2)       0.01221421  0.01663607  2.0699e-02  0.0278125
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.36203189 -0.32916570 -3.0268e-01 -0.2495378
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.01985715 -0.01578668 -1.1955e-02 -0.0051207
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    0.12768287  0.15527825  1.8942e-01  0.2341983
#> Vatnsdalsa:Jokulsa.lag( 4)       0.00271674  0.00579025  9.3367e-03  0.0143834
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -0.10922362 -0.09064588 -6.6711e-02 -0.0242044
#> Vatnsdalsa:Jokulsa.lag( 5)      -0.00619524 -0.00356772  8.6567e-05  0.0064061
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -0.04347084 -0.02260357  5.0908e-03  0.0647936
#> Vatnsdalsa:Jokulsa.lag( 6)       0.00047362  0.00296571  5.9749e-03  0.0115502
#> Vatnsdalsa:Vatnsdalsa.lag( 6)    0.01441967  0.04896744  7.9727e-02  0.1186816
#> Vatnsdalsa:Jokulsa.lag( 7)      -0.00856076 -0.00598544 -2.5375e-03  0.0032513
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -0.07133919 -0.04663042 -1.4777e-02  0.0508853
#> Vatnsdalsa:Jokulsa.lag( 8)       0.00173152  0.00441755  7.1251e-03  0.0120884
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -0.10326615 -0.08078404 -6.5174e-02 -0.0352076
#> Vatnsdalsa:Jokulsa.lag( 9)      -0.00607816 -0.00315888 -2.1587e-04  0.0053574
#> Vatnsdalsa:Vatnsdalsa.lag( 9)    0.10294830  0.12641110  1.5254e-01  0.1955446
#> Vatnsdalsa:Jokulsa.lag(10)       0.00086353  0.00496492  7.6466e-03  0.0132371
#> Vatnsdalsa:Vatnsdalsa.lag(10)   -0.09821610 -0.08331709 -6.1932e-02 -0.0153133
#> Vatnsdalsa:Jokulsa.lag(11)      -0.01118748 -0.00810676 -4.7560e-03  0.0019921
#> Vatnsdalsa:Vatnsdalsa.lag(11)    0.03620501  0.06037330  8.0401e-02  0.1243957
#> Vatnsdalsa:Jokulsa.lag(12)       0.00832485  0.01124954  1.3884e-02  0.0215431
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.07586987 -0.05705163 -3.1963e-02  0.0166927
#> Vatnsdalsa:Jokulsa.lag(13)      -0.01068895 -0.00785323 -4.2623e-03  0.0011588
#> Vatnsdalsa:Vatnsdalsa.lag(13)    0.12681808  0.14823078  1.7105e-01  0.1976894
#> Vatnsdalsa:Jokulsa.lag(14)      -0.00580952 -0.00249379  6.9682e-04  0.0067803
#> Vatnsdalsa:Vatnsdalsa.lag(14)   -0.04701523 -0.02921548 -5.3254e-03  0.0350022
#> Vatnsdalsa:Jokulsa.lag(15)       0.00040994  0.00238792  4.9405e-03  0.0088229
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -0.06714225 -0.04860827 -3.0617e-02  0.0079609
#> Vatnsdalsa:Precipitation.lag(1) -0.00279586  0.00080895  3.4121e-03  0.0122427
#> Vatnsdalsa:Precipitation.lag(2) -0.01350147 -0.00873810 -3.5270e-03  0.0088155
#> Vatnsdalsa:Precipitation.lag(3)  0.00110587  0.00401991  7.0463e-03  0.0142494
#> Vatnsdalsa:Precipitation.lag(4)  0.00245752  0.00506140  8.4347e-03  0.0128576
#> Vatnsdalsa:Temperature.lag(1)    0.03109199  0.03916658  4.6194e-02  0.0626714
#> Vatnsdalsa:Temperature.lag(2)   -0.05421212 -0.04535739 -3.7008e-02 -0.0230704
#> 
#> 
#> Scale parameter:
#>                           Mean        Sd   Sd(Mean)     2.5%      25%      50%
#> Jokulsa.Jokulsa       1.381787 0.2327777 0.04755998 0.980598 1.206442 1.359797
#> Jokulsa.Vatnsdalsa    0.046747 0.0160062 0.00182310 0.020099 0.034683 0.045098
#> Vatnsdalsa.Vatnsdalsa 0.029897 0.0047338 0.00090084 0.021799 0.026451 0.029393
#>                            75%    97.5%
#> Jokulsa.Jokulsa       1.527037 1.884842
#> Jokulsa.Vatnsdalsa    0.055375 0.081015
#> Vatnsdalsa.Vatnsdalsa 0.033407 0.038729
#> 
#> 
#> Extra parameter:
#>       Mean       Sd  Sd(Mean)    2.5%     25%     50%     75%   97.5%
#> nu 0.82792 0.041182 0.0066481 0.74589 0.79533 0.82902 0.85965 0.90056
#plot(fit3.mcmc)

###### Example 4: U.S. stock returns
data(US.returns)
fit4 <- mtar(~ CCR | dVIX, data=US.returns, subset={Date<="2025-11-28"},
             row.names=Date, ars=ars(nregim=2,p=3,d=3), n.burnin=100,
             n.sim=200, n.thin=2, dist="Student-t")
fit4.mcmc <- coda::as.mcmc(fit4)
summary(fit4.mcmc)
#> 
#> 
#>  Iterations = 101:499
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 200
#> 
#> 
#> Thresholds:
#>             Mean      Sd Sd(Mean)    2.5%    25%   50%    75%  97.5%
#> threshold 1.2145 0.15235  0.10211 0.98345 1.1705 1.184 1.3428 1.4939
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                      Mean        Sd   Sd(Mean)        2.5%       25%       50%
#> CCR:(Intercept)  0.088684 0.0131567 0.00093295  0.06722016  0.078876  0.088507
#> CCR:CCR.lag(1)  -0.048498 0.0158517 0.00145370 -0.07642769 -0.059298 -0.050461
#> CCR:CCR.lag(2)  -0.037619 0.0215755 0.00186345 -0.08074585 -0.052834 -0.038963
#> CCR:CCR.lag(3)  -0.025191 0.0229128 0.00231908 -0.06316878 -0.041619 -0.027200
#> CCR:dVIX.lag(1) -0.034947 0.0147320 0.00122359 -0.06925331 -0.044235 -0.034592
#> CCR:dVIX.lag(2) -0.020553 0.0139400 0.00137052 -0.04388662 -0.030749 -0.021662
#> CCR:dVIX.lag(3)  0.016383 0.0082776 0.00068212  0.00076518  0.010611  0.016023
#>                        75%      97.5%
#> CCR:(Intercept)  0.0985703  0.1119164
#> CCR:CCR.lag(1)  -0.0396054 -0.0185096
#> CCR:CCR.lag(2)  -0.0206996  0.0025654
#> CCR:CCR.lag(3)  -0.0095057  0.0260189
#> CCR:dVIX.lag(1) -0.0248595 -0.0106631
#> CCR:dVIX.lag(2) -0.0111408  0.0079707
#> CCR:dVIX.lag(3)  0.0221779  0.0328640
#> 
#> 
#> Scale parameter:
#>           Mean       Sd  Sd(Mean)    2.5%     25%    50%     75%   97.5%
#> CCR.CCR 0.3545 0.015442 0.0024052 0.32509 0.34293 0.3545 0.36387 0.38794
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                       Mean       Sd  Sd(Mean)      2.5%        25%        50%
#> CCR:(Intercept)  0.0640382 0.083809 0.0077482 -0.082505  0.0049939  0.0506421
#> CCR:CCR.lag(1)  -0.0832223 0.048027 0.0043944 -0.170887 -0.1120678 -0.0875568
#> CCR:CCR.lag(2)  -0.0127605 0.085581 0.0088277 -0.183506 -0.0734434 -0.0096054
#> CCR:CCR.lag(3)   0.0595850 0.073429 0.0066371 -0.070856  0.0063837  0.0570454
#> CCR:dVIX.lag(1) -0.0090838 0.053468 0.0055178 -0.111911 -0.0456091 -0.0092650
#> CCR:dVIX.lag(2)  0.0041780 0.044648 0.0042237 -0.078968 -0.0271447  0.0067066
#> CCR:dVIX.lag(3)  0.0356530 0.025100 0.0021634 -0.015467  0.0215726  0.0344980
#>                       75%    97.5%
#> CCR:(Intercept)  0.115282 0.238754
#> CCR:CCR.lag(1)  -0.050608 0.013724
#> CCR:CCR.lag(2)   0.051642 0.149065
#> CCR:CCR.lag(3)   0.112785 0.198765
#> CCR:dVIX.lag(1)  0.028790 0.090805
#> CCR:dVIX.lag(2)  0.035238 0.078993
#> CCR:dVIX.lag(3)  0.048937 0.085452
#> 
#> 
#> Scale parameter:
#>           Mean       Sd Sd(Mean)    2.5%     25%    50%     75%  97.5%
#> CCR.CCR 0.8366 0.081032 0.014472 0.70209 0.78174 0.8259 0.89355 1.0047
#> 
#> 
#> Extra parameter:
#>      Mean      Sd Sd(Mean)   2.5%    25%    50%    75%  97.5%
#> nu 2.4213 0.10693 0.017927 2.2303 2.3381 2.4165 2.4949 2.6029
#plot(fit4.mcmc)
# }