Coercion of mtar objects to mcmc objects

This method converts an object of class mtar into a list of mcmc objects, each corresponding to a Markov chain produced during Bayesian estimation.

Usage

# S3 method for class 'mtar'
as.mcmc(x, ...)

Arguments

x

an object of class mtar obtained from a call to mtar().

...

additional arguments passed to specific coercion methods.

Value

A list of mcmc objects containing the posterior simulation draws generated by the mtar() routine.

See also

Examples

# \donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
             subset={Date<="2015-12-07"}, dist="Student-t",
             ars=ars(nregim=3,p=c(1,1,2)), n.burnin=100, n.sim=200,
             n.thin=2, ssvs=TRUE)
fit1.mcmc <- coda::as.mcmc(fit1)
summary(fit1.mcmc)
#> 
#> 
#>  Iterations = 101:499
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 200
#> 
#> 
#> Thresholds:
#>                   Mean         Sd   Sd(Mean)       2.5%        25%        50%
#> Threshold.1 -0.0032633 0.00204645 0.00205577 -0.0062766 -0.0056319 -0.0020219
#> Threshold.2  0.0075539 0.00036781 0.00013627  0.0069881  0.0074070  0.0075601
#>                    75%      97.5%
#> Threshold.1 -0.0012605 -0.0011679
#> Threshold.2  0.0076811  0.0085914
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean        Sd  Sd(Mean)       2.5%        25%
#> COLCAP:(Intercept)     -0.0040111 0.0012798 0.0006314 -0.0065490 -0.0051373
#> COLCAP:COLCAP.lag(1)    0.1649665 0.0689562 0.0060477  0.0548244  0.1128312
#> COLCAP:BOVESPA.lag(1)   0.0775207 0.0390196 0.0027591  0.0056815  0.0491333
#> BOVESPA:(Intercept)    -0.0090429 0.0020337 0.0013875 -0.0128886 -0.0110790
#> BOVESPA:COLCAP.lag(1)  -0.0100751 0.0963406 0.0272727 -0.1875211 -0.0638798
#> BOVESPA:BOVESPA.lag(1)  0.0203058 0.0576634 0.0065665 -0.0830870 -0.0164095
#>                               50%        75%      97.5%
#> COLCAP:(Intercept)     -0.0035320 -0.0030621 -0.0023876
#> COLCAP:COLCAP.lag(1)    0.1599718  0.2123679  0.2997001
#> COLCAP:BOVESPA.lag(1)   0.0773847  0.1005416  0.1579631
#> BOVESPA:(Intercept)    -0.0083261 -0.0073594 -0.0064845
#> BOVESPA:COLCAP.lag(1)  -0.0021389  0.0497719  0.1747427
#> BOVESPA:BOVESPA.lag(1)  0.0171241  0.0578784  0.1395162
#> 
#> 
#> Scale parameter:
#>                       Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP.COLCAP   6.7126e-05 9.0715e-06 3.8578e-06 5.3300e-05 6.0157e-05
#> COLCAP.BOVESPA  3.6543e-05 7.4389e-06 2.3499e-06 2.4428e-05 3.1362e-05
#> BOVESPA.BOVESPA 1.3420e-04 1.6345e-05 3.7590e-06 1.0809e-04 1.2241e-04
#>                        50%        75%      97.5%
#> COLCAP.COLCAP   6.5228e-05 7.2787e-05 8.8545e-05
#> COLCAP.BOVESPA  3.6185e-05 4.0620e-05 5.2086e-05
#> BOVESPA.BOVESPA 1.3231e-04 1.4464e-04 1.7317e-04
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                               Mean         Sd   Sd(Mean)        2.5%
#> COLCAP:(Intercept)      0.00029170 0.00030676 6.4065e-05 -0.00028597
#> COLCAP:COLCAP.lag(1)    0.07190691 0.03294205 2.3294e-03  0.00418045
#> COLCAP:BOVESPA.lag(1)   0.08147656 0.02124026 3.4288e-03  0.04580615
#> BOVESPA:(Intercept)     0.00025408 0.00065069 3.1513e-04 -0.00095336
#> BOVESPA:COLCAP.lag(1)   0.05981546 0.04616000 3.2718e-03 -0.02494411
#> BOVESPA:BOVESPA.lag(1) -0.04192015 0.02933647 2.1962e-03 -0.09414007
#>                                25%         50%         75%      97.5%
#> COLCAP:(Intercept)      0.00010437  0.00029305  0.00050150 0.00089092
#> COLCAP:COLCAP.lag(1)    0.05094025  0.07111368  0.09554430 0.13178400
#> COLCAP:BOVESPA.lag(1)   0.06641560  0.08165828  0.09592064 0.11887269
#> BOVESPA:(Intercept)    -0.00019958  0.00030706  0.00072692 0.00140899
#> BOVESPA:COLCAP.lag(1)   0.02836742  0.05794963  0.09522793 0.16653438
#> BOVESPA:BOVESPA.lag(1) -0.06170658 -0.04370606 -0.02015080 0.01418482
#> 
#> 
#> Scale parameter:
#>                       Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP.COLCAP   4.1046e-05 2.6957e-06 4.8888e-07 3.5516e-05 3.9173e-05
#> COLCAP.BOVESPA  1.1236e-05 2.7873e-06 8.0873e-07 5.4037e-06 9.2464e-06
#> BOVESPA.BOVESPA 8.3488e-05 6.5583e-06 1.3921e-06 7.2043e-05 7.9214e-05
#>                        50%        75%      97.5%
#> COLCAP.COLCAP   4.1209e-05 4.2955e-05 4.6347e-05
#> COLCAP.BOVESPA  1.1439e-05 1.3182e-05 1.5951e-05
#> BOVESPA.BOVESPA 8.2406e-05 8.7453e-05 9.7360e-05
#> 
#> 
#> Regime 3
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP:(Intercept)      0.0048878 0.00054873 5.8798e-05  0.0036875  0.0045449
#> COLCAP:COLCAP.lag(1)    0.0509644 0.06496957 5.3949e-03 -0.0744603  0.0088231
#> COLCAP:BOVESPA.lag(1)   0.0729914 0.04667791 4.0536e-03 -0.0171404  0.0455978
#> COLCAP:COLCAP.lag(2)    0.0536864 0.05724897 4.0481e-03 -0.0524903  0.0139236
#> COLCAP:BOVESPA.lag(2)  -0.0670312 0.03657777 2.5864e-03 -0.1421984 -0.0890717
#> BOVESPA:(Intercept)     0.0113719 0.00097590 1.3366e-04  0.0096076  0.0105994
#> BOVESPA:COLCAP.lag(1)   0.1522318 0.09032303 6.3868e-03 -0.0397454  0.0993301
#> BOVESPA:BOVESPA.lag(1) -0.1093142 0.07048181 5.5421e-03 -0.2427404 -0.1595007
#> BOVESPA:COLCAP.lag(2)  -0.0775529 0.09084758 6.4239e-03 -0.2775178 -0.1351824
#> BOVESPA:BOVESPA.lag(2) -0.0518336 0.06278550 5.6799e-03 -0.1577880 -0.0993907
#>                               50%        75%     97.5%
#> COLCAP:(Intercept)      0.0048966  0.0052596 0.0058483
#> COLCAP:COLCAP.lag(1)    0.0501113  0.0941849 0.1747570
#> COLCAP:BOVESPA.lag(1)   0.0740895  0.0994478 0.1567424
#> COLCAP:COLCAP.lag(2)    0.0538258  0.0922435 0.1780201
#> COLCAP:BOVESPA.lag(2)  -0.0654872 -0.0455865 0.0083543
#> BOVESPA:(Intercept)     0.0113315  0.0120043 0.0130783
#> BOVESPA:COLCAP.lag(1)   0.1516646  0.2034911 0.3195698
#> BOVESPA:BOVESPA.lag(1) -0.1062808 -0.0599649 0.0195261
#> BOVESPA:COLCAP.lag(2)  -0.0709825 -0.0167992 0.0903136
#> BOVESPA:BOVESPA.lag(2) -0.0536160 -0.0067657 0.0667547
#> 
#> 
#> Scale parameter:
#>                       Mean         Sd   Sd(Mean)       2.5%        25%
#> COLCAP.COLCAP   5.6085e-05 7.0540e-06 6.2274e-07 4.2955e-05 5.0820e-05
#> COLCAP.BOVESPA  2.5742e-05 6.4496e-06 5.3667e-07 1.3933e-05 2.1283e-05
#> BOVESPA.BOVESPA 1.3598e-04 1.5963e-05 1.5463e-06 1.0942e-04 1.2543e-04
#>                        50%        75%      97.5%
#> COLCAP.COLCAP   5.5752e-05 6.1287e-05 6.9593e-05
#> COLCAP.BOVESPA  2.4733e-05 2.9569e-05 3.9837e-05
#> BOVESPA.BOVESPA 1.3454e-04 1.4547e-04 1.7068e-04
#> 
#> 
#> Extra parameter:
#>      Mean      Sd Sd(Mean)   2.5%    25%    50%    75%  97.5%
#> nu 6.0913 0.81123  0.23293 4.7685 5.4594 6.0523 6.6795 7.8557
#plot(fit1.mcmc)

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, row.names=Date,
             subset={Date<="2009-02-13"}, dist="Laplace",
             ars=ars(nregim=3,p=5), n.burnin=100, n.sim=200, n.thin=2)
fit2.mcmc <- coda::as.mcmc(fit2)
summary(fit2.mcmc)
#> 
#> 
#>  Iterations = 101:499
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 200
#> 
#> 
#> Thresholds:
#>               Mean        Sd  Sd(Mean) 2.5% 25% 50%     75%   97.5%
#> Threshold.1  3.026 0.0433821 0.0300934    3   3   3  3.0981  3.0981
#> Threshold.2 10.004 0.0061631 0.0042752   10  10  10 10.0139 10.0139
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean        Sd   Sd(Mean)       2.5%        25%
#> Bedon:(Intercept)       1.3263361 0.1119401 0.01091341  1.1353877  1.2456164
#> Bedon:Bedon.lag(1)      0.5650208 0.0344949 0.00351298  0.5027368  0.5419639
#> Bedon:LaPlata.lag(1)    0.0440013 0.0156314 0.00164295  0.0120199  0.0334174
#> Bedon:Bedon.lag(2)      0.0480389 0.0327929 0.00379232 -0.0246206  0.0255715
#> Bedon:LaPlata.lag(2)   -0.0174564 0.0121027 0.00115934 -0.0440817 -0.0256149
#> Bedon:Bedon.lag(3)      0.0267378 0.0311043 0.00348603 -0.0262251  0.0055129
#> Bedon:LaPlata.lag(3)    0.0023923 0.0115537 0.00126342 -0.0207066 -0.0062340
#> Bedon:Bedon.lag(4)      0.0363536 0.0312001 0.00278455 -0.0252625  0.0148668
#> Bedon:LaPlata.lag(4)   -0.0157276 0.0093116 0.00086573 -0.0337990 -0.0220705
#> Bedon:Bedon.lag(5)      0.0823471 0.0253253 0.00259814  0.0325072  0.0670221
#> Bedon:LaPlata.lag(5)   -0.0063663 0.0069921 0.00056255 -0.0194006 -0.0112191
#> LaPlata:(Intercept)     3.4577154 0.3541327 0.04306812  2.8028516  3.2189720
#> LaPlata:Bedon.lag(1)    0.1571491 0.0983607 0.01071078 -0.0388148  0.0835014
#> LaPlata:LaPlata.lag(1)  0.6267787 0.0408616 0.00448569  0.5503184  0.6017877
#> LaPlata:Bedon.lag(2)   -0.0448745 0.0837210 0.00864358 -0.2021332 -0.1003584
#> LaPlata:LaPlata.lag(2) -0.0635786 0.0346859 0.00357316 -0.1270740 -0.0879111
#> LaPlata:Bedon.lag(3)    0.0168143 0.0646625 0.00520161 -0.1021400 -0.0261434
#> LaPlata:LaPlata.lag(3)  0.0658315 0.0289243 0.00298018  0.0072065  0.0475842
#> LaPlata:Bedon.lag(4)   -0.0769915 0.0762439 0.00750363 -0.2491168 -0.1231024
#> LaPlata:LaPlata.lag(4)  0.0046764 0.0281731 0.00310171 -0.0410747 -0.0165500
#> LaPlata:Bedon.lag(5)    0.1376886 0.0592346 0.00547194  0.0220538  0.1029823
#> LaPlata:LaPlata.lag(5)  0.0294151 0.0212047 0.00197432 -0.0161434  0.0172946
#>                               50%        75%     97.5%
#> Bedon:(Intercept)       1.3136414  1.3997095 1.5412575
#> Bedon:Bedon.lag(1)      0.5635970  0.5902108 0.6328029
#> Bedon:LaPlata.lag(1)    0.0431536  0.0560077 0.0725852
#> Bedon:Bedon.lag(2)      0.0485074  0.0722729 0.1117650
#> Bedon:LaPlata.lag(2)   -0.0170128 -0.0089957 0.0033845
#> Bedon:Bedon.lag(3)      0.0257698  0.0499449 0.0856332
#> Bedon:LaPlata.lag(3)    0.0034692  0.0107995 0.0221273
#> Bedon:Bedon.lag(4)      0.0364835  0.0571420 0.0922157
#> Bedon:LaPlata.lag(4)   -0.0160227 -0.0102018 0.0050029
#> Bedon:Bedon.lag(5)      0.0817374  0.0997405 0.1312570
#> Bedon:LaPlata.lag(5)   -0.0068401 -0.0013747 0.0072500
#> LaPlata:(Intercept)     3.4537652  3.6756441 4.3154744
#> LaPlata:Bedon.lag(1)    0.1630910  0.2319518 0.3084857
#> LaPlata:LaPlata.lag(1)  0.6255404  0.6550420 0.7011373
#> LaPlata:Bedon.lag(2)   -0.0385039  0.0142144 0.1061221
#> LaPlata:LaPlata.lag(2) -0.0651415 -0.0394819 0.0040026
#> LaPlata:Bedon.lag(3)    0.0182393  0.0659936 0.1457359
#> LaPlata:LaPlata.lag(3)  0.0675773  0.0851664 0.1147958
#> LaPlata:Bedon.lag(4)   -0.0764732 -0.0274476 0.0639455
#> LaPlata:LaPlata.lag(4)  0.0030253  0.0255099 0.0565357
#> LaPlata:Bedon.lag(5)    0.1363488  0.1707403 0.2613932
#> LaPlata:LaPlata.lag(5)  0.0311540  0.0425327 0.0716608
#> 
#> 
#> Scale parameter:
#>                    Mean       Sd  Sd(Mean)    2.5%     25%     50%     75%
#> Bedon.Bedon     0.32573 0.030332 0.0029826 0.27542 0.30209 0.32485 0.34982
#> Bedon.LaPlata   0.36547 0.057605 0.0043241 0.25732 0.32167 0.36778 0.41004
#> LaPlata.LaPlata 2.34178 0.237389 0.0197894 1.90043 2.16362 2.33477 2.47205
#>                   97.5%
#> Bedon.Bedon     0.38389
#> Bedon.LaPlata   0.46292
#> LaPlata.LaPlata 2.84057
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean       Sd  Sd(Mean)       2.5%        25%
#> Bedon:(Intercept)       2.0901413 0.448815 0.0500516  1.2721444  1.7212590
#> Bedon:Bedon.lag(1)      0.5796506 0.045653 0.0042450  0.4854426  0.5510907
#> Bedon:LaPlata.lag(1)    0.0207756 0.013792 0.0013271 -0.0061785  0.0124392
#> Bedon:Bedon.lag(2)      0.0983530 0.064210 0.0072201 -0.0346337  0.0523763
#> Bedon:LaPlata.lag(2)   -0.0179113 0.017192 0.0017451 -0.0473864 -0.0304392
#> Bedon:Bedon.lag(3)     -0.0260077 0.060441 0.0072691 -0.1426571 -0.0674385
#> Bedon:LaPlata.lag(3)   -0.0092837 0.015241 0.0011684 -0.0375889 -0.0210545
#> Bedon:Bedon.lag(4)      0.1034415 0.060201 0.0081147 -0.0043601  0.0583517
#> Bedon:LaPlata.lag(4)    0.0083872 0.017784 0.0020167 -0.0281609 -0.0018485
#> Bedon:Bedon.lag(5)      0.0191588 0.044757 0.0066557 -0.0615284 -0.0108821
#> Bedon:LaPlata.lag(5)    0.0057790 0.014979 0.0015064 -0.0216804 -0.0053849
#> LaPlata:(Intercept)     6.7924065 1.044097 0.1162270  4.9224005  6.0110870
#> LaPlata:Bedon.lag(1)    0.1462528 0.102395 0.0100347 -0.0372265  0.0720045
#> LaPlata:LaPlata.lag(1)  0.5246536 0.041794 0.0049636  0.4448871  0.4956671
#> LaPlata:Bedon.lag(2)    0.0012900 0.124805 0.0109253 -0.2570708 -0.0804959
#> LaPlata:LaPlata.lag(2)  0.0355301 0.037383 0.0034563 -0.0509692  0.0127207
#> LaPlata:Bedon.lag(3)   -0.0585389 0.111192 0.0118525 -0.2806787 -0.1363642
#> LaPlata:LaPlata.lag(3)  0.0420448 0.030552 0.0030546 -0.0120441  0.0202405
#> LaPlata:Bedon.lag(4)    0.2208690 0.134966 0.0129446 -0.0314236  0.1191283
#> LaPlata:LaPlata.lag(4) -0.0438973 0.040347 0.0042126 -0.1148359 -0.0729232
#> LaPlata:Bedon.lag(5)   -0.2694943 0.113779 0.0098496 -0.4685395 -0.3429293
#> LaPlata:LaPlata.lag(5)  0.1183315 0.042720 0.0049824  0.0349275  0.0914981
#>                               50%         75%     97.5%
#> Bedon:(Intercept)       2.0838162  2.41749585  2.983490
#> Bedon:Bedon.lag(1)      0.5790116  0.60910295  0.660550
#> Bedon:LaPlata.lag(1)    0.0201319  0.03087145  0.048938
#> Bedon:Bedon.lag(2)      0.1049092  0.14086732  0.204705
#> Bedon:LaPlata.lag(2)   -0.0190815 -0.00303609  0.011546
#> Bedon:Bedon.lag(3)     -0.0249634  0.01455329  0.079968
#> Bedon:LaPlata.lag(3)   -0.0090022 -0.00047109  0.020641
#> Bedon:Bedon.lag(4)      0.1053701  0.14605400  0.215616
#> Bedon:LaPlata.lag(4)    0.0079078  0.01956083  0.045199
#> Bedon:Bedon.lag(5)      0.0183559  0.04719693  0.113739
#> Bedon:LaPlata.lag(5)    0.0058328  0.01517325  0.035502
#> LaPlata:(Intercept)     6.7860605  7.55824666  8.745831
#> LaPlata:Bedon.lag(1)    0.1435381  0.20833582  0.343430
#> LaPlata:LaPlata.lag(1)  0.5280016  0.55376325  0.596631
#> LaPlata:Bedon.lag(2)    0.0015926  0.07932802  0.251273
#> LaPlata:LaPlata.lag(2)  0.0365990  0.05952563  0.107607
#> LaPlata:Bedon.lag(3)   -0.0502738  0.01135774  0.139689
#> LaPlata:LaPlata.lag(3)  0.0439495  0.05988409  0.105686
#> LaPlata:Bedon.lag(4)    0.2287663  0.31605335  0.462880
#> LaPlata:LaPlata.lag(4) -0.0442119 -0.01347804  0.048467
#> LaPlata:Bedon.lag(5)   -0.2718373 -0.20001668 -0.052741
#> LaPlata:LaPlata.lag(5)  0.1229004  0.14491911  0.199927
#> 
#> 
#> Scale parameter:
#>                   Mean      Sd Sd(Mean)    2.5%    25%    50%    75%  97.5%
#> Bedon.Bedon     1.0979 0.12587 0.011688 0.86740 1.0111 1.0893 1.1837 1.3699
#> Bedon.LaPlata   1.3261 0.19819 0.017958 0.99338 1.1876 1.3130 1.4354 1.7607
#> LaPlata.LaPlata 6.4808 0.64504 0.063394 5.43585 5.9937 6.3801 6.9312 7.8860
#> 
#> 
#> Regime 3
#> 
#> 
#> Autoregressive coefficients:
#>                              Mean       Sd  Sd(Mean)       2.5%        25%
#> Bedon:(Intercept)       5.7246303 0.803034 0.0736335  4.2485443  5.1753318
#> Bedon:Bedon.lag(1)      0.4552767 0.088057 0.0101792  0.2875636  0.3889481
#> Bedon:LaPlata.lag(1)    0.0452820 0.016225 0.0019632  0.0157154  0.0351546
#> Bedon:Bedon.lag(2)      0.0850427 0.075437 0.0068484 -0.0494261  0.0293901
#> Bedon:LaPlata.lag(2)   -0.0011437 0.018205 0.0016435 -0.0374175 -0.0128195
#> Bedon:Bedon.lag(3)     -0.0931942 0.062245 0.0059193 -0.2040973 -0.1308777
#> Bedon:LaPlata.lag(3)    0.0321278 0.018212 0.0017322 -0.0054082  0.0212128
#> Bedon:Bedon.lag(4)     -0.0042984 0.068667 0.0070719 -0.1236375 -0.0531644
#> Bedon:LaPlata.lag(4)    0.0044136 0.020007 0.0022328 -0.0321187 -0.0082318
#> Bedon:Bedon.lag(5)      0.1906821 0.068484 0.0066371  0.0704656  0.1408764
#> Bedon:LaPlata.lag(5)   -0.0139646 0.016516 0.0027483 -0.0481332 -0.0238568
#> LaPlata:(Intercept)    17.3347149 2.993367 0.2733791 11.8178416 15.1954986
#> LaPlata:Bedon.lag(1)    0.4831105 0.251622 0.0236822  0.0411294  0.3132401
#> LaPlata:LaPlata.lag(1)  0.3328200 0.064177 0.0067902  0.2053928  0.2897544
#> LaPlata:Bedon.lag(2)   -0.5549851 0.283760 0.0296837 -1.0494585 -0.7787602
#> LaPlata:LaPlata.lag(2)  0.1278156 0.070818 0.0076932  0.0049879  0.0790013
#> LaPlata:Bedon.lag(3)   -0.5794237 0.235672 0.0193542 -1.0366461 -0.7605889
#> LaPlata:LaPlata.lag(3)  0.2797343 0.082561 0.0085675  0.1040718  0.2275418
#> LaPlata:Bedon.lag(4)    0.0105755 0.298062 0.0343586 -0.5606337 -0.1944424
#> LaPlata:LaPlata.lag(4) -0.0071711 0.075434 0.0078117 -0.1605965 -0.0575582
#> LaPlata:Bedon.lag(5)    0.3222044 0.255771 0.0281877 -0.2243294  0.1588703
#> LaPlata:LaPlata.lag(5)  0.0652834 0.061864 0.0054488 -0.0527165  0.0217019
#>                                50%        75%     97.5%
#> Bedon:(Intercept)       5.69198235  6.2787616  7.364283
#> Bedon:Bedon.lag(1)      0.45509178  0.5150297  0.642577
#> Bedon:LaPlata.lag(1)    0.04470012  0.0549151  0.078968
#> Bedon:Bedon.lag(2)      0.08343278  0.1341359  0.236820
#> Bedon:LaPlata.lag(2)   -0.00052599  0.0094767  0.033822
#> Bedon:Bedon.lag(3)     -0.09848641 -0.0550215  0.032365
#> Bedon:LaPlata.lag(3)    0.03330632  0.0436873  0.064642
#> Bedon:Bedon.lag(4)     -0.00398724  0.0423601  0.126334
#> Bedon:LaPlata.lag(4)    0.00233287  0.0165776  0.044557
#> Bedon:Bedon.lag(5)      0.19084273  0.2349515  0.322748
#> Bedon:LaPlata.lag(5)   -0.01411361 -0.0047528  0.021367
#> LaPlata:(Intercept)    17.26968612 19.4832380 22.761366
#> LaPlata:Bedon.lag(1)    0.48726570  0.6489693  0.968823
#> LaPlata:LaPlata.lag(1)  0.33095477  0.3790037  0.458355
#> LaPlata:Bedon.lag(2)   -0.55849281 -0.3531028  0.039680
#> LaPlata:LaPlata.lag(2)  0.12498666  0.1659485  0.285965
#> LaPlata:Bedon.lag(3)   -0.57583777 -0.4295905 -0.093438
#> LaPlata:LaPlata.lag(3)  0.28100137  0.3402798  0.430171
#> LaPlata:Bedon.lag(4)    0.00327710  0.1986016  0.618417
#> LaPlata:LaPlata.lag(4) -0.00830032  0.0432333  0.146253
#> LaPlata:Bedon.lag(5)    0.33107759  0.4873765  0.806392
#> LaPlata:LaPlata.lag(5)  0.06132628  0.1051924  0.177252
#> 
#> 
#> Scale parameter:
#>                    Mean      Sd Sd(Mean)    2.5%     25%     50%     75%
#> Bedon.Bedon      2.7546 0.29406 0.027869  2.2206  2.5649  2.7200  2.9452
#> Bedon.LaPlata    7.0782 0.91609 0.084649  5.3303  6.4014  7.0909  7.6865
#> LaPlata.LaPlata 42.8940 4.65825 0.481227 34.0608 39.5116 43.1305 46.1918
#>                   97.5%
#> Bedon.Bedon      3.3379
#> Bedon.LaPlata    8.7387
#> LaPlata.LaPlata 51.5446
#plot(fit2.mcmc)

###### Example 3: Temperature, precipitation, and two river flows in Iceland
data(iceland.rf)
fit3 <- mtar(~ Jokulsa + Vatnsdalsa | Temperature | Precipitation,
             data=iceland.rf, subset={Date<="1974-11-06"}, row.names=Date,
             ars=ars(nregim=2,p=15,q=4,d=2), n.burnin=100, n.sim=200,
             n.thin=2, dist="Slash")
fit3.mcmc <- coda::as.mcmc(fit3)
summary(fit3.mcmc)
#> 
#> 
#>  Iterations = 101:499
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 200
#> 
#> 
#> Thresholds:
#>             Mean       Sd  Sd(Mean)    2.5%     25%     50%     75%   97.5%
#> threshold 0.6419 0.044011 0.0087852 0.59845 0.62093 0.63938 0.66426 0.75651
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                                        Mean        Sd   Sd(Mean)        2.5%
#> Jokulsa:(Intercept)              3.37291575 0.3738113 0.05132478  2.72796577
#> Jokulsa:Jokulsa.lag( 1)          0.87491337 0.0356429 0.00606073  0.78602067
#> Jokulsa:Vatnsdalsa.lag( 1)       0.26284150 0.0676909 0.00837514  0.09989323
#> Jokulsa:Jokulsa.lag( 2)         -0.05613492 0.0278168 0.00378241 -0.10635128
#> Jokulsa:Vatnsdalsa.lag( 2)      -0.26422034 0.0828826 0.01051590 -0.40268043
#> Jokulsa:Jokulsa.lag( 3)         -0.00502408 0.0204680 0.00185515 -0.04486862
#> Jokulsa:Vatnsdalsa.lag( 3)       0.09482147 0.0481523 0.00515505 -0.00063473
#> Jokulsa:Jokulsa.lag( 4)          0.00188572 0.0238723 0.00291808 -0.04080924
#> Jokulsa:Vatnsdalsa.lag( 4)      -0.01751335 0.0455507 0.00551960 -0.10440792
#> Jokulsa:Jokulsa.lag( 5)         -0.00700057 0.0294971 0.00485288 -0.06516815
#> Jokulsa:Vatnsdalsa.lag( 5)      -0.02948134 0.0344369 0.00309998 -0.11202005
#> Jokulsa:Jokulsa.lag( 6)          0.02037479 0.0306703 0.00437353 -0.04886117
#> Jokulsa:Vatnsdalsa.lag( 6)      -0.01948110 0.0431356 0.00376958 -0.09145434
#> Jokulsa:Jokulsa.lag( 7)          0.00842426 0.0321651 0.00363731 -0.05407921
#> Jokulsa:Vatnsdalsa.lag( 7)       0.01774714 0.0363720 0.00312848 -0.04709370
#> Jokulsa:Jokulsa.lag( 8)         -0.00359778 0.0245993 0.00231723 -0.05985398
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.00757565 0.0302789 0.00241052 -0.06575341
#> Jokulsa:Jokulsa.lag( 9)          0.00067841 0.0229990 0.00274483 -0.04600539
#> Jokulsa:Vatnsdalsa.lag( 9)      -0.01414938 0.0382265 0.00466781 -0.08226343
#> Jokulsa:Jokulsa.lag(10)          0.02033506 0.0188401 0.00204999 -0.01894286
#> Jokulsa:Vatnsdalsa.lag(10)       0.02425402 0.0317041 0.00292033 -0.04157495
#> Jokulsa:Jokulsa.lag(11)         -0.01045031 0.0149179 0.00135785 -0.03869200
#> Jokulsa:Vatnsdalsa.lag(11)      -0.01204913 0.0268765 0.00219763 -0.06096264
#> Jokulsa:Jokulsa.lag(12)          0.00619938 0.0142500 0.00117377 -0.02367890
#> Jokulsa:Vatnsdalsa.lag(12)       0.00397570 0.0255946 0.00224973 -0.05561005
#> Jokulsa:Jokulsa.lag(13)         -0.01346337 0.0186535 0.00295062 -0.04872400
#> Jokulsa:Vatnsdalsa.lag(13)      -0.01316897 0.0293369 0.00227486 -0.07331761
#> Jokulsa:Jokulsa.lag(14)          0.00262397 0.0145345 0.00142251 -0.02272183
#> Jokulsa:Vatnsdalsa.lag(14)       0.00940603 0.0319871 0.00335960 -0.04634493
#> Jokulsa:Jokulsa.lag(15)          0.01902433 0.0157158 0.00244726 -0.00727573
#> Jokulsa:Vatnsdalsa.lag(15)      -0.01743824 0.0222495 0.00278425 -0.06168464
#> Jokulsa:Precipitation.lag(1)     0.00506774 0.0083477 0.00086266 -0.00977571
#> Jokulsa:Precipitation.lag(2)     0.00639657 0.0078152 0.00083301 -0.00894865
#> Jokulsa:Precipitation.lag(3)    -0.01089198 0.0064382 0.00071678 -0.02320853
#> Jokulsa:Precipitation.lag(4)     0.01905336 0.0068085 0.00057274  0.00757450
#> Jokulsa:Temperature.lag(1)       0.01668385 0.0107009 0.00085716 -0.00317323
#> Jokulsa:Temperature.lag(2)      -0.03275620 0.0103536 0.00074442 -0.05125429
#> Vatnsdalsa:(Intercept)           0.71012159 0.2034522 0.03222098  0.34334393
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.05172657 0.0205848 0.00358538 -0.09973940
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.17069309 0.0524854 0.00850959  1.06110198
#> Vatnsdalsa:Jokulsa.lag( 2)       0.04976420 0.0173753 0.00282729  0.02027858
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.32317848 0.0611721 0.00864549 -0.43367638
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.03115881 0.0138961 0.00161844 -0.05707707
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    0.07646332 0.0398256 0.00486392  0.00170481
#> Vatnsdalsa:Jokulsa.lag( 4)       0.01261010 0.0136904 0.00132165 -0.01356353
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -0.02653970 0.0353705 0.00528902 -0.09984629
#> Vatnsdalsa:Jokulsa.lag( 5)       0.00915004 0.0161340 0.00162285 -0.02035812
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -0.00503993 0.0232417 0.00235570 -0.05091455
#> Vatnsdalsa:Jokulsa.lag( 6)       0.00323919 0.0160352 0.00243798 -0.03141680
#> Vatnsdalsa:Vatnsdalsa.lag( 6)   -0.01851798 0.0221049 0.00179521 -0.06450978
#> Vatnsdalsa:Jokulsa.lag( 7)       0.00290876 0.0165645 0.00250058 -0.02524519
#> Vatnsdalsa:Vatnsdalsa.lag( 7)    0.01616759 0.0203089 0.00143606 -0.02304288
#> Vatnsdalsa:Jokulsa.lag( 8)      -0.01700727 0.0130230 0.00123568 -0.04202431
#> Vatnsdalsa:Vatnsdalsa.lag( 8)    0.01022249 0.0169499 0.00119854 -0.02122039
#> Vatnsdalsa:Jokulsa.lag( 9)       0.02135234 0.0144114 0.00192882 -0.00762940
#> Vatnsdalsa:Vatnsdalsa.lag( 9)    0.00140855 0.0176913 0.00125097 -0.03723757
#> Vatnsdalsa:Jokulsa.lag(10)      -0.01746462 0.0123124 0.00172894 -0.04202078
#> Vatnsdalsa:Vatnsdalsa.lag(10)    0.00943483 0.0185416 0.00145489 -0.02069355
#> Vatnsdalsa:Jokulsa.lag(11)       0.01217450 0.0092289 0.00074830 -0.00615017
#> Vatnsdalsa:Vatnsdalsa.lag(11)    0.00131612 0.0199306 0.00192239 -0.03132154
#> Vatnsdalsa:Jokulsa.lag(12)      -0.01167575 0.0094818 0.00079459 -0.02755721
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.00695448 0.0182509 0.00144275 -0.04571646
#> Vatnsdalsa:Jokulsa.lag(13)       0.00667601 0.0105540 0.00113520 -0.01404492
#> Vatnsdalsa:Vatnsdalsa.lag(13)   -0.01888670 0.0221481 0.00225939 -0.06437729
#> Vatnsdalsa:Jokulsa.lag(14)      -0.00944722 0.0090079 0.00077033 -0.02476557
#> Vatnsdalsa:Vatnsdalsa.lag(14)    0.03897104 0.0200668 0.00196846  0.00514587
#> Vatnsdalsa:Jokulsa.lag(15)       0.00649461 0.0064170 0.00059136 -0.00570876
#> Vatnsdalsa:Vatnsdalsa.lag(15)    0.00033076 0.0136080 0.00128547 -0.02655419
#> Vatnsdalsa:Precipitation.lag(1)  0.00124755 0.0059270 0.00056165 -0.00962589
#> Vatnsdalsa:Precipitation.lag(2)  0.00159092 0.0047918 0.00044806 -0.00756886
#> Vatnsdalsa:Precipitation.lag(3) -0.00307203 0.0040052 0.00031192 -0.01090243
#> Vatnsdalsa:Precipitation.lag(4)  0.00626564 0.0040930 0.00037787 -0.00161143
#> Vatnsdalsa:Temperature.lag(1)   -0.00026353 0.0067803 0.00054544 -0.01304071
#> Vatnsdalsa:Temperature.lag(2)   -0.01089012 0.0065559 0.00063726 -0.02332996
#>                                         25%         50%         75%      97.5%
#> Jokulsa:(Intercept)              3.1232e+00  3.31585412  3.57676338  4.1833769
#> Jokulsa:Jokulsa.lag( 1)          8.6568e-01  0.88469442  0.89641355  0.9223210
#> Jokulsa:Vatnsdalsa.lag( 1)       2.2785e-01  0.26586936  0.30084568  0.3898561
#> Jokulsa:Jokulsa.lag( 2)         -7.4913e-02 -0.05868010 -0.03846602  0.0066228
#> Jokulsa:Vatnsdalsa.lag( 2)      -3.1598e-01 -0.27502887 -0.21226232 -0.0897327
#> Jokulsa:Jokulsa.lag( 3)         -1.8750e-02 -0.00475182  0.00779293  0.0362135
#> Jokulsa:Vatnsdalsa.lag( 3)       6.2739e-02  0.09482959  0.12874968  0.1898428
#> Jokulsa:Jokulsa.lag( 4)         -1.6787e-02  0.00110311  0.02057007  0.0442602
#> Jokulsa:Vatnsdalsa.lag( 4)      -4.7551e-02 -0.01947778  0.01316886  0.0757773
#> Jokulsa:Jokulsa.lag( 5)         -3.0314e-02 -0.00321849  0.01517453  0.0414271
#> Jokulsa:Vatnsdalsa.lag( 5)      -4.9870e-02 -0.02663273 -0.00355843  0.0286329
#> Jokulsa:Jokulsa.lag( 6)          9.5706e-05  0.02088125  0.03991166  0.0729072
#> Jokulsa:Vatnsdalsa.lag( 6)      -5.2081e-02 -0.01914658  0.01022812  0.0614701
#> Jokulsa:Jokulsa.lag( 7)         -1.4354e-02  0.00774643  0.03131492  0.0680924
#> Jokulsa:Vatnsdalsa.lag( 7)      -4.7244e-03  0.01646274  0.03874024  0.1012682
#> Jokulsa:Jokulsa.lag( 8)         -1.8053e-02 -0.00337333  0.01394319  0.0409669
#> Jokulsa:Vatnsdalsa.lag( 8)      -2.5524e-02 -0.00745951  0.01111329  0.0477656
#> Jokulsa:Jokulsa.lag( 9)         -1.4657e-02  0.00137752  0.01686143  0.0457200
#> Jokulsa:Vatnsdalsa.lag( 9)      -3.9685e-02 -0.01511651  0.00495208  0.0652370
#> Jokulsa:Jokulsa.lag(10)          8.2819e-03  0.02130108  0.03444052  0.0538484
#> Jokulsa:Vatnsdalsa.lag(10)       5.0559e-03  0.02358550  0.04343333  0.0857164
#> Jokulsa:Jokulsa.lag(11)         -2.1232e-02 -0.01083202 -0.00053478  0.0169181
#> Jokulsa:Vatnsdalsa.lag(11)      -3.0510e-02 -0.01200847  0.00550415  0.0363403
#> Jokulsa:Jokulsa.lag(12)         -3.4870e-03  0.00723734  0.01542781  0.0342641
#> Jokulsa:Vatnsdalsa.lag(12)      -7.2360e-03  0.00449510  0.02080096  0.0545482
#> Jokulsa:Jokulsa.lag(13)         -2.6921e-02 -0.01537925 -0.00154964  0.0267790
#> Jokulsa:Vatnsdalsa.lag(13)      -2.9241e-02 -0.01197876  0.00475868  0.0396736
#> Jokulsa:Jokulsa.lag(14)         -7.9948e-03  0.00337213  0.01293791  0.0333742
#> Jokulsa:Vatnsdalsa.lag(14)      -1.3008e-02  0.00542900  0.02913682  0.0778155
#> Jokulsa:Jokulsa.lag(15)          8.6979e-03  0.01927268  0.02938486  0.0479431
#> Jokulsa:Vatnsdalsa.lag(15)      -3.2533e-02 -0.01545048 -0.00266891  0.0240450
#> Jokulsa:Precipitation.lag(1)    -1.2895e-03  0.00561001  0.01083407  0.0202622
#> Jokulsa:Precipitation.lag(2)     1.1407e-03  0.00655843  0.01203310  0.0202010
#> Jokulsa:Precipitation.lag(3)    -1.5524e-02 -0.01143159 -0.00639027  0.0011404
#> Jokulsa:Precipitation.lag(4)     1.4170e-02  0.01918836  0.02329972  0.0316750
#> Jokulsa:Temperature.lag(1)       9.5678e-03  0.01707629  0.02387060  0.0379787
#> Jokulsa:Temperature.lag(2)      -3.9802e-02 -0.03229739 -0.02629234 -0.0127513
#> Vatnsdalsa:(Intercept)           5.7701e-01  0.68672931  0.85071756  1.1342337
#> Vatnsdalsa:Jokulsa.lag( 1)      -5.9644e-02 -0.04790714 -0.03787230 -0.0249713
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.1331e+00  1.17728444  1.20854612  1.2681691
#> Vatnsdalsa:Jokulsa.lag( 2)       3.8173e-02  0.04773358  0.06358155  0.0794864
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -3.6593e-01 -0.32595178 -0.28175452 -0.2058046
#> Vatnsdalsa:Jokulsa.lag( 3)      -3.9659e-02 -0.03085246 -0.02212687 -0.0015599
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    5.3222e-02  0.07287489  0.10416951  0.1542969
#> Vatnsdalsa:Jokulsa.lag( 4)       3.7894e-03  0.01296854  0.02055931  0.0388915
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -4.9444e-02 -0.02405697 -0.00093230  0.0385285
#> Vatnsdalsa:Jokulsa.lag( 5)      -1.3144e-03  0.00733527  0.01882352  0.0421015
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -2.0142e-02 -0.00532002  0.00965363  0.0422823
#> Vatnsdalsa:Jokulsa.lag( 6)      -6.7635e-03  0.00390342  0.01424034  0.0307880
#> Vatnsdalsa:Vatnsdalsa.lag( 6)   -3.3248e-02 -0.02054370 -0.00232702  0.0250487
#> Vatnsdalsa:Jokulsa.lag( 7)      -9.9365e-03  0.00361009  0.01410497  0.0331650
#> Vatnsdalsa:Vatnsdalsa.lag( 7)    3.0686e-03  0.01740167  0.02988881  0.0520834
#> Vatnsdalsa:Jokulsa.lag( 8)      -2.5318e-02 -0.01746571 -0.00730081  0.0078062
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -2.2735e-03  0.00790521  0.02185566  0.0461805
#> Vatnsdalsa:Jokulsa.lag( 9)       1.1833e-02  0.02088711  0.03124905  0.0501290
#> Vatnsdalsa:Vatnsdalsa.lag( 9)   -1.0473e-02  0.00057228  0.01403004  0.0326078
#> Vatnsdalsa:Jokulsa.lag(10)      -2.4853e-02 -0.01772502 -0.00948569  0.0044632
#> Vatnsdalsa:Vatnsdalsa.lag(10)   -3.3047e-03  0.00776815  0.02190461  0.0480684
#> Vatnsdalsa:Jokulsa.lag(11)       5.8978e-03  0.01246656  0.01842303  0.0294750
#> Vatnsdalsa:Vatnsdalsa.lag(11)   -1.3513e-02  0.00162054  0.01309339  0.0479783
#> Vatnsdalsa:Jokulsa.lag(12)      -1.8631e-02 -0.01221785 -0.00352147  0.0059534
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -1.8114e-02 -0.00696944  0.00400220  0.0314360
#> Vatnsdalsa:Jokulsa.lag(13)      -8.4117e-04  0.00673139  0.01410530  0.0280838
#> Vatnsdalsa:Vatnsdalsa.lag(13)   -3.2711e-02 -0.02020859 -0.00495528  0.0252039
#> Vatnsdalsa:Jokulsa.lag(14)      -1.6210e-02 -0.00958042 -0.00265606  0.0081827
#> Vatnsdalsa:Vatnsdalsa.lag(14)    2.3882e-02  0.03886761  0.05161025  0.0819466
#> Vatnsdalsa:Jokulsa.lag(15)       1.7620e-03  0.00625772  0.01115805  0.0195166
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -8.3403e-03  0.00071663  0.00960602  0.0286874
#> Vatnsdalsa:Precipitation.lag(1) -2.8756e-03  0.00052284  0.00569072  0.0122665
#> Vatnsdalsa:Precipitation.lag(2) -2.0222e-03  0.00188445  0.00459225  0.0112668
#> Vatnsdalsa:Precipitation.lag(3) -5.3109e-03 -0.00295810 -0.00062936  0.0043428
#> Vatnsdalsa:Precipitation.lag(4)  3.5833e-03  0.00641008  0.00931899  0.0138380
#> Vatnsdalsa:Temperature.lag(1)   -4.8543e-03 -0.00056457  0.00361515  0.0133736
#> Vatnsdalsa:Temperature.lag(2)   -1.5594e-02 -0.01048790 -0.00626582  0.0017879
#> 
#> 
#> Scale parameter:
#>                            Mean        Sd  Sd(Mean)      2.5%       25%
#> Jokulsa.Jokulsa       0.0616958 0.0093814 0.0055090 0.0471549 0.0547605
#> Jokulsa.Vatnsdalsa    0.0093599 0.0026730 0.0002961 0.0045145 0.0075118
#> Vatnsdalsa.Vatnsdalsa 0.0255409 0.0042312 0.0010186 0.0197014 0.0223950
#>                             50%      75%    97.5%
#> Jokulsa.Jokulsa       0.0600786 0.067348 0.082543
#> Jokulsa.Vatnsdalsa    0.0094483 0.010854 0.014837
#> Vatnsdalsa.Vatnsdalsa 0.0246808 0.027954 0.034469
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                                        Mean        Sd   Sd(Mean)       2.5%
#> Jokulsa:(Intercept)             -1.8628e-01 0.6086418 0.06146337 -1.3622523
#> Jokulsa:Jokulsa.lag( 1)          1.0143e+00 0.0360634 0.00414208  0.9449417
#> Jokulsa:Vatnsdalsa.lag( 1)       9.3359e-01 0.1905381 0.01421920  0.6130703
#> Jokulsa:Jokulsa.lag( 2)         -1.4158e-01 0.0599955 0.00811708 -0.2579148
#> Jokulsa:Vatnsdalsa.lag( 2)      -4.4981e-01 0.2530974 0.01534829 -0.8897757
#> Jokulsa:Jokulsa.lag( 3)         -1.3719e-02 0.0512823 0.00534690 -0.1154048
#> Jokulsa:Vatnsdalsa.lag( 3)       2.0359e-04 0.2626756 0.01905984 -0.5631316
#> Jokulsa:Jokulsa.lag( 4)         -6.3720e-02 0.0411794 0.00431242 -0.1399417
#> Jokulsa:Vatnsdalsa.lag( 4)      -1.5419e-01 0.2257499 0.02356129 -0.5460150
#> Jokulsa:Jokulsa.lag( 5)          3.8814e-02 0.0418485 0.00397702 -0.0319338
#> Jokulsa:Vatnsdalsa.lag( 5)       2.3713e-01 0.2367160 0.03803587 -0.2439795
#> Jokulsa:Jokulsa.lag( 6)         -5.0873e-02 0.0314224 0.00268151 -0.1128538
#> Jokulsa:Vatnsdalsa.lag( 6)      -2.4332e-01 0.1850594 0.02071091 -0.6353037
#> Jokulsa:Jokulsa.lag( 7)          2.5038e-03 0.0320123 0.00305005 -0.0510948
#> Jokulsa:Vatnsdalsa.lag( 7)       2.7970e-01 0.1977890 0.02019248 -0.1484696
#> Jokulsa:Jokulsa.lag( 8)          2.4920e-02 0.0310049 0.00280411 -0.0298921
#> Jokulsa:Vatnsdalsa.lag( 8)      -3.2285e-01 0.1794351 0.01849281 -0.6857614
#> Jokulsa:Jokulsa.lag( 9)          3.7904e-02 0.0304925 0.00233901 -0.0164951
#> Jokulsa:Vatnsdalsa.lag( 9)       1.7903e-01 0.2020422 0.02086834 -0.1606341
#> Jokulsa:Jokulsa.lag(10)         -3.0798e-02 0.0370712 0.00319887 -0.1010589
#> Jokulsa:Vatnsdalsa.lag(10)      -1.3154e-01 0.1921724 0.02506801 -0.4950745
#> Jokulsa:Jokulsa.lag(11)         -5.3212e-03 0.0377360 0.00334861 -0.0884362
#> Jokulsa:Vatnsdalsa.lag(11)       2.2943e-01 0.2473356 0.02824452 -0.2445946
#> Jokulsa:Jokulsa.lag(12)         -8.0168e-04 0.0354020 0.00288222 -0.0667795
#> Jokulsa:Vatnsdalsa.lag(12)       5.8144e-03 0.2894549 0.03873901 -0.5672255
#> Jokulsa:Jokulsa.lag(13)          1.2546e-02 0.0410687 0.00404633 -0.0608658
#> Jokulsa:Vatnsdalsa.lag(13)       8.3506e-02 0.2767332 0.04705010 -0.3928599
#> Jokulsa:Jokulsa.lag(14)         -1.8778e-02 0.0419955 0.00453120 -0.0972333
#> Jokulsa:Vatnsdalsa.lag(14)       3.0659e-01 0.1791051 0.01777681 -0.0629945
#> Jokulsa:Jokulsa.lag(15)          4.7401e-02 0.0227636 0.00217728  0.0051060
#> Jokulsa:Vatnsdalsa.lag(15)      -4.0259e-01 0.1421977 0.01514542 -0.6863235
#> Jokulsa:Precipitation.lag(1)    -1.1737e-01 0.0380790 0.00374484 -0.1859762
#> Jokulsa:Precipitation.lag(2)     6.7409e-03 0.0536971 0.00694123 -0.0960993
#> Jokulsa:Precipitation.lag(3)     4.6716e-02 0.0309735 0.00250208 -0.0065277
#> Jokulsa:Precipitation.lag(4)     1.8258e-02 0.0321797 0.00227545 -0.0332241
#> Jokulsa:Temperature.lag(1)       1.0648e+00 0.1011374 0.00952753  0.8625988
#> Jokulsa:Temperature.lag(2)      -5.3918e-01 0.0973184 0.00996832 -0.7321167
#> Vatnsdalsa:(Intercept)           4.8203e-01 0.0966346 0.01078166  0.2957685
#> Vatnsdalsa:Jokulsa.lag( 1)      -5.7963e-03 0.0041423 0.00039471 -0.0138838
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.2004e+00 0.0286024 0.00329634  1.1497046
#> Vatnsdalsa:Jokulsa.lag( 2)       1.6087e-02 0.0062607 0.00067632  0.0050350
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -3.1691e-01 0.0431858 0.00812267 -0.4032342
#> Vatnsdalsa:Jokulsa.lag( 3)      -1.5261e-02 0.0057680 0.00059184 -0.0260280
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    1.5714e-01 0.0382559 0.00417478  0.0861319
#> Vatnsdalsa:Jokulsa.lag( 4)       5.3923e-03 0.0048060 0.00041010 -0.0041975
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -8.8597e-02 0.0387034 0.00418457 -0.1469246
#> Vatnsdalsa:Jokulsa.lag( 5)      -3.1620e-03 0.0048784 0.00043049 -0.0124780
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -4.1020e-02 0.0396415 0.00628557 -0.1225476
#> Vatnsdalsa:Jokulsa.lag( 6)       2.7361e-03 0.0043212 0.00044016 -0.0051096
#> Vatnsdalsa:Vatnsdalsa.lag( 6)    7.9144e-02 0.0334450 0.00412966  0.0146123
#> Vatnsdalsa:Jokulsa.lag( 7)      -4.6415e-03 0.0042723 0.00030210 -0.0135373
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -5.8230e-02 0.0321582 0.00435978 -0.1078781
#> Vatnsdalsa:Jokulsa.lag( 8)       2.6186e-03 0.0041606 0.00029420 -0.0057771
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -8.3269e-02 0.0250550 0.00249271 -0.1333743
#> Vatnsdalsa:Jokulsa.lag( 9)      -2.9288e-03 0.0048260 0.00040739 -0.0119615
#> Vatnsdalsa:Vatnsdalsa.lag( 9)    1.3296e-01 0.0328778 0.00373261  0.0595655
#> Vatnsdalsa:Jokulsa.lag(10)       4.9753e-03 0.0051455 0.00052369 -0.0044706
#> Vatnsdalsa:Vatnsdalsa.lag(10)   -8.4562e-02 0.0255285 0.00265879 -0.1332019
#> Vatnsdalsa:Jokulsa.lag(11)      -7.6967e-03 0.0050422 0.00043837 -0.0165059
#> Vatnsdalsa:Vatnsdalsa.lag(11)    6.8475e-02 0.0312201 0.00366284  0.0064634
#> Vatnsdalsa:Jokulsa.lag(12)       1.0311e-02 0.0048538 0.00039012  0.0022308
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -7.2036e-02 0.0314929 0.00378002 -0.1246048
#> Vatnsdalsa:Jokulsa.lag(13)      -7.1780e-03 0.0046224 0.00039250 -0.0167060
#> Vatnsdalsa:Vatnsdalsa.lag(13)    1.5388e-01 0.0285874 0.00399424  0.0969662
#> Vatnsdalsa:Jokulsa.lag(14)      -2.4066e-03 0.0046487 0.00044944 -0.0109543
#> Vatnsdalsa:Vatnsdalsa.lag(14)   -2.8587e-02 0.0274337 0.00337222 -0.0794851
#> Vatnsdalsa:Jokulsa.lag(15)       2.5291e-03 0.0032376 0.00033907 -0.0030222
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -4.9258e-02 0.0253598 0.00392791 -0.0937482
#> Vatnsdalsa:Precipitation.lag(1) -4.8133e-05 0.0058824 0.00052165 -0.0120352
#> Vatnsdalsa:Precipitation.lag(2) -9.1107e-03 0.0068472 0.00086504 -0.0222501
#> Vatnsdalsa:Precipitation.lag(3)  4.7330e-03 0.0051702 0.00055655 -0.0042625
#> Vatnsdalsa:Precipitation.lag(4)  5.2818e-03 0.0047848 0.00033834 -0.0054900
#> Vatnsdalsa:Temperature.lag(1)    3.3877e-02 0.0143078 0.00131750  0.0104370
#> Vatnsdalsa:Temperature.lag(2)   -4.1055e-02 0.0149566 0.00180798 -0.0758451
#>                                         25%         50%         75%      97.5%
#> Jokulsa:(Intercept)             -0.57441948 -0.20027596  0.21166849  0.9536668
#> Jokulsa:Jokulsa.lag( 1)          0.99165632  1.01355874  1.03354728  1.0992809
#> Jokulsa:Vatnsdalsa.lag( 1)       0.80528361  0.93188378  1.07425219  1.3026318
#> Jokulsa:Jokulsa.lag( 2)         -0.17927608 -0.14136260 -0.10768790 -0.0148653
#> Jokulsa:Vatnsdalsa.lag( 2)      -0.64252606 -0.44641092 -0.30257106  0.0794116
#> Jokulsa:Jokulsa.lag( 3)         -0.04634738 -0.01337258  0.01855162  0.0808097
#> Jokulsa:Vatnsdalsa.lag( 3)      -0.14020080  0.01711137  0.19021245  0.4339323
#> Jokulsa:Jokulsa.lag( 4)         -0.09275259 -0.06196249 -0.03808089  0.0112230
#> Jokulsa:Vatnsdalsa.lag( 4)      -0.30044009 -0.18251125 -0.04195415  0.3865299
#> Jokulsa:Jokulsa.lag( 5)          0.01134262  0.03437925  0.06762060  0.1227047
#> Jokulsa:Vatnsdalsa.lag( 5)       0.09273213  0.24746483  0.37820068  0.6972175
#> Jokulsa:Jokulsa.lag( 6)         -0.07226689 -0.05094304 -0.02984340  0.0113001
#> Jokulsa:Vatnsdalsa.lag( 6)      -0.35987354 -0.24717518 -0.12730845  0.1025950
#> Jokulsa:Jokulsa.lag( 7)         -0.02021422  0.00095280  0.02678446  0.0638608
#> Jokulsa:Vatnsdalsa.lag( 7)       0.15609296  0.30277788  0.40706422  0.6525022
#> Jokulsa:Jokulsa.lag( 8)          0.00171538  0.02252125  0.04472392  0.0853790
#> Jokulsa:Vatnsdalsa.lag( 8)      -0.44868381 -0.31216035 -0.21376053  0.0052679
#> Jokulsa:Jokulsa.lag( 9)          0.01611019  0.03633124  0.05979105  0.0976916
#> Jokulsa:Vatnsdalsa.lag( 9)       0.03365028  0.17447646  0.31215535  0.5886839
#> Jokulsa:Jokulsa.lag(10)         -0.05725789 -0.03386126 -0.00412644  0.0422444
#> Jokulsa:Vatnsdalsa.lag(10)      -0.24465478 -0.12151405 -0.00026805  0.2313221
#> Jokulsa:Jokulsa.lag(11)         -0.02647404 -0.00172222  0.01760985  0.0627720
#> Jokulsa:Vatnsdalsa.lag(11)       0.08248429  0.21383943  0.40095809  0.7667670
#> Jokulsa:Jokulsa.lag(12)         -0.02586736 -0.00088544  0.02533103  0.0654257
#> Jokulsa:Vatnsdalsa.lag(12)      -0.19233802  0.02083880  0.20960617  0.5578367
#> Jokulsa:Jokulsa.lag(13)         -0.01728605  0.01299808  0.03989250  0.0874166
#> Jokulsa:Vatnsdalsa.lag(13)      -0.13095632  0.09194089  0.28770908  0.5685666
#> Jokulsa:Jokulsa.lag(14)         -0.04692667 -0.01873089  0.00762684  0.0594359
#> Jokulsa:Vatnsdalsa.lag(14)       0.19496784  0.30808088  0.41769665  0.6674932
#> Jokulsa:Jokulsa.lag(15)          0.03110400  0.04636449  0.06271562  0.0920085
#> Jokulsa:Vatnsdalsa.lag(15)      -0.49288614 -0.40177530 -0.32103291 -0.1117040
#> Jokulsa:Precipitation.lag(1)    -0.13802324 -0.11773796 -0.09213786 -0.0448795
#> Jokulsa:Precipitation.lag(2)    -0.02828463  0.00500435  0.04268823  0.1086872
#> Jokulsa:Precipitation.lag(3)     0.02605378  0.04592438  0.06536073  0.1110947
#> Jokulsa:Precipitation.lag(4)    -0.00518424  0.01650019  0.04192357  0.0773539
#> Jokulsa:Temperature.lag(1)       1.00302032  1.06011202  1.13255140  1.2485744
#> Jokulsa:Temperature.lag(2)      -0.59735338 -0.54672048 -0.47169009 -0.3630999
#> Vatnsdalsa:(Intercept)           0.41632182  0.48522747  0.54770213  0.6725832
#> Vatnsdalsa:Jokulsa.lag( 1)      -0.00872039 -0.00603976 -0.00274462  0.0019112
#> Vatnsdalsa:Vatnsdalsa.lag( 1)    1.18160003  1.19805045  1.21675690  1.2671941
#> Vatnsdalsa:Jokulsa.lag( 2)       0.01169120  0.01618052  0.02068023  0.0282421
#> Vatnsdalsa:Vatnsdalsa.lag( 2)   -0.34991265 -0.31386310 -0.28538035 -0.2343856
#> Vatnsdalsa:Jokulsa.lag( 3)      -0.01945705 -0.01497709 -0.01192720 -0.0029657
#> Vatnsdalsa:Vatnsdalsa.lag( 3)    0.12631826  0.15744771  0.18422513  0.2275319
#> Vatnsdalsa:Jokulsa.lag( 4)       0.00227363  0.00570448  0.00871780  0.0143078
#> Vatnsdalsa:Vatnsdalsa.lag( 4)   -0.11647186 -0.09190544 -0.06702702 -0.0039360
#> Vatnsdalsa:Jokulsa.lag( 5)      -0.00608336 -0.00327737  0.00010245  0.0064856
#> Vatnsdalsa:Vatnsdalsa.lag( 5)   -0.06528500 -0.04204333 -0.01539386  0.0370381
#> Vatnsdalsa:Jokulsa.lag( 6)       0.00010277  0.00268258  0.00564162  0.0108047
#> Vatnsdalsa:Vatnsdalsa.lag( 6)    0.05635477  0.08330165  0.10425529  0.1290681
#> Vatnsdalsa:Jokulsa.lag( 7)      -0.00739691 -0.00436269 -0.00183864  0.0028890
#> Vatnsdalsa:Vatnsdalsa.lag( 7)   -0.08329030 -0.06104954 -0.03636565  0.0145057
#> Vatnsdalsa:Jokulsa.lag( 8)      -0.00038695  0.00311210  0.00550466  0.0103230
#> Vatnsdalsa:Vatnsdalsa.lag( 8)   -0.09932570 -0.08594064 -0.06929514 -0.0361263
#> Vatnsdalsa:Jokulsa.lag( 9)      -0.00604758 -0.00323652  0.00030786  0.0061781
#> Vatnsdalsa:Vatnsdalsa.lag( 9)    0.11199681  0.13347255  0.15699882  0.1876927
#> Vatnsdalsa:Jokulsa.lag(10)       0.00173233  0.00510040  0.00799755  0.0173978
#> Vatnsdalsa:Vatnsdalsa.lag(10)   -0.10125396 -0.08657249 -0.06905531 -0.0339878
#> Vatnsdalsa:Jokulsa.lag(11)      -0.01165725 -0.00732861 -0.00454206  0.0024889
#> Vatnsdalsa:Vatnsdalsa.lag(11)    0.04730809  0.07138039  0.08910419  0.1234366
#> Vatnsdalsa:Jokulsa.lag(12)       0.00663901  0.00973600  0.01358233  0.0195757
#> Vatnsdalsa:Vatnsdalsa.lag(12)   -0.09481793 -0.07273544 -0.05208240 -0.0035770
#> Vatnsdalsa:Jokulsa.lag(13)      -0.00999833 -0.00714828 -0.00411195  0.0016040
#> Vatnsdalsa:Vatnsdalsa.lag(13)    0.13497262  0.15471066  0.17426873  0.2031493
#> Vatnsdalsa:Jokulsa.lag(14)      -0.00592353 -0.00256366  0.00107274  0.0055315
#> Vatnsdalsa:Vatnsdalsa.lag(14)   -0.04926744 -0.02888624 -0.00921426  0.0199468
#> Vatnsdalsa:Jokulsa.lag(15)       0.00016017  0.00235114  0.00494715  0.0090609
#> Vatnsdalsa:Vatnsdalsa.lag(15)   -0.06622248 -0.05188470 -0.03217321  0.0017342
#> Vatnsdalsa:Precipitation.lag(1) -0.00393208 -0.00016071  0.00391133  0.0113760
#> Vatnsdalsa:Precipitation.lag(2) -0.01402285 -0.00906952 -0.00432975  0.0043758
#> Vatnsdalsa:Precipitation.lag(3)  0.00099492  0.00417621  0.00827146  0.0151238
#> Vatnsdalsa:Precipitation.lag(4)  0.00205174  0.00541236  0.00864999  0.0137434
#> Vatnsdalsa:Temperature.lag(1)    0.02384017  0.03356178  0.04186060  0.0666896
#> Vatnsdalsa:Temperature.lag(2)   -0.04924210 -0.04021970 -0.02954633 -0.0132308
#> 
#> 
#> Scale parameter:
#>                           Mean        Sd  Sd(Mean)     2.5%      25%      50%
#> Jokulsa.Jokulsa       1.405306 0.2200709 0.0573646 0.968679 1.268066 1.414408
#> Jokulsa.Vatnsdalsa    0.045639 0.0121981 0.0012270 0.021664 0.037563 0.045126
#> Vatnsdalsa.Vatnsdalsa 0.029589 0.0050458 0.0012272 0.021092 0.026091 0.029339
#>                            75%    97.5%
#> Jokulsa.Jokulsa       1.546303 1.857178
#> Jokulsa.Vatnsdalsa    0.053392 0.070382
#> Vatnsdalsa.Vatnsdalsa 0.032745 0.039306
#> 
#> 
#> Extra parameter:
#>       Mean       Sd Sd(Mean)   2.5%     25%     50%    75%   97.5%
#> nu 0.83252 0.046917 0.010546 0.7444 0.79887 0.83198 0.8626 0.91959
#plot(fit3.mcmc)

###### Example 4: U.S. stock returns
data(US.returns)
fit4 <- mtar(~ CCR | dVIX, data=US.returns, subset={Date<="2025-11-28"},
             row.names=Date, ars=ars(nregim=2,p=3,d=3), n.burnin=100,
             n.sim=200, n.thin=2, dist="Student-t")
fit4.mcmc <- coda::as.mcmc(fit4)
summary(fit4.mcmc)
#> 
#> 
#>  Iterations = 101:499
#> 
#>  Thinning interval = 2
#> 
#>  Sample size per chain = 200
#> 
#> 
#> Thresholds:
#>             Mean       Sd  Sd(Mean)   2.5%    25%    50%    75%  97.5%
#> threshold 1.6084 0.010818 0.0028955 1.5771 1.6007 1.6123 1.6165 1.6165
#> 
#> 
#> Regime 1
#> 
#> 
#> Autoregressive coefficients:
#>                       Mean        Sd   Sd(Mean)       2.5%        25%
#> CCR:(Intercept)  0.0872668 0.0109352 0.00088288  0.0671703  0.0800671
#> CCR:CCR.lag(1)  -0.0438673 0.0107235 0.00075826 -0.0656030 -0.0506341
#> CCR:CCR.lag(2)  -0.0519628 0.0251588 0.00223125 -0.1032281 -0.0684437
#> CCR:CCR.lag(3)  -0.0077675 0.0210199 0.00148633 -0.0464451 -0.0223339
#> CCR:dVIX.lag(1) -0.0544906 0.0186138 0.00169842 -0.0907523 -0.0677178
#> CCR:dVIX.lag(2) -0.0071310 0.0142997 0.00101114 -0.0322775 -0.0171929
#> CCR:dVIX.lag(3)  0.0135422 0.0083921 0.00075942 -0.0026582  0.0073295
#>                        50%        75%       97.5%
#> CCR:(Intercept)  0.0873493  0.0938770  0.10733390
#> CCR:CCR.lag(1)  -0.0441748 -0.0371845 -0.02246078
#> CCR:CCR.lag(2)  -0.0504755 -0.0347725 -0.00041803
#> CCR:CCR.lag(3)  -0.0077388  0.0049452  0.03565698
#> CCR:dVIX.lag(1) -0.0540943 -0.0433743 -0.01654312
#> CCR:dVIX.lag(2) -0.0073259  0.0024214  0.01965513
#> CCR:dVIX.lag(3)  0.0129266  0.0185714  0.03046528
#> 
#> 
#> Scale parameter:
#>            Mean       Sd Sd(Mean)    2.5%     25%    50%     75%   97.5%
#> CCR.CCR 0.34951 0.014508 0.002519 0.31979 0.33869 0.3506 0.35954 0.37588
#> 
#> 
#> Regime 2
#> 
#> 
#> Autoregressive coefficients:
#>                      Mean       Sd  Sd(Mean)      2.5%        25%       50%
#> CCR:(Intercept)  0.171041 0.127584 0.0087914 -0.077589  0.0856168  0.167880
#> CCR:CCR.lag(1)  -0.196324 0.051176 0.0062918 -0.290714 -0.2307560 -0.199170
#> CCR:CCR.lag(2)   0.054280 0.084227 0.0082825 -0.111114 -0.0022563  0.054182
#> CCR:CCR.lag(3)  -0.088244 0.097945 0.0138458 -0.259022 -0.1606799 -0.092976
#> CCR:dVIX.lag(1)  0.013227 0.038157 0.0038374 -0.064223 -0.0129528  0.015373
#> CCR:dVIX.lag(2) -0.101013 0.063478 0.0058762 -0.211806 -0.1474738 -0.104159
#> CCR:dVIX.lag(3)  0.035563 0.036672 0.0039854 -0.028744  0.0089679  0.035870
#>                       75%     97.5%
#> CCR:(Intercept)  0.259265  0.433060
#> CCR:CCR.lag(1)  -0.158367 -0.086786
#> CCR:CCR.lag(2)   0.110519  0.219632
#> CCR:CCR.lag(3)  -0.017213  0.100438
#> CCR:dVIX.lag(1)  0.038467  0.083659
#> CCR:dVIX.lag(2) -0.058531  0.018828
#> CCR:dVIX.lag(3)  0.060342  0.100188
#> 
#> 
#> Scale parameter:
#>           Mean      Sd Sd(Mean)   2.5%    25%    50%    75%  97.5%
#> CCR.CCR 1.3045 0.13191 0.018168 1.0816 1.2064 1.2969 1.3931 1.5853
#> 
#> 
#> Extra parameter:
#>     Mean      Sd Sd(Mean)   2.5%    25%    50%    75% 97.5%
#> nu 2.495 0.11704 0.024628 2.2597 2.4165 2.4949 2.5733 2.712
#plot(fit4.mcmc)
# }