Simulation of multivariate time series from a TAR model

This function simulates multivariate time series generated by a user-specified Threshold Autoregressive (TAR) model.

Usage

simtar(
  n,
  k = 2,
  ars = ars(),
  Intercept = TRUE,
  trend = c("none", "linear", "quadratic"),
  nseason = NULL,
  parms,
  delay = 0,
  thresholds = NULL,
  t.series = NULL,
  ex.series = NULL,
  dist = c("Gaussian", "Student-t", "Hyperbolic", "Laplace", "Slash",
    "Contaminated normal", "Skew-Student-t", "Skew-normal"),
  skewness = NULL,
  extra = NULL,
  setar = NULL,
  Verbose = TRUE
)

Arguments

n

A positive integer specifying the length of the simulated output series.

k

A positive integer specifying the dimension of the multivariate output series.

ars

A list defining the TAR model structure, composed of four elements: the number of regimes (nregim), the autoregressive order (p), and the maximum lags of the exogenous (q) and threshold (d) series within each regime. This object can be validated using the ars() function.

Intercept

An optional logical indicating whether an intercept term is included in each regime.

trend

An optional character string specifying the degree of deterministic time trend to be included in each regime. Available options are a linear trend ("linear"), a quadratic trend ("quadratic"), or no trend ("none"). By default, trend is set to "none".

nseason

An optional integer, greater than or equal to 2, specifying the number of seasonal periods. When provided, nseason - 1 seasonal dummy variables are added to the regressors within each regime. By default, nseason is set to NULL, thereby indicating that the TAR model has no seasonal effects.

parms

A list with one sublist per regime. Each sublist contains two matrices: the first matrix corresponds to the location parameters, and the second matrix corresponds to the scale parameters of the model.

delay

An optional non-negative integer specifying the delay parameter of the threshold series. By default, delay is set to 0.

thresholds

A numeric vector of length nregim-1 containing the threshold values, sorted in ascending order.

t.series

A matrix containing the values of the threshold series.

ex.series

A matrix containing the values of the multivariate exogenous series.

dist

An optional character string specifying the multivariate distribution chosen to model the noise process. Supported options include Gaussian ("Gaussian"), Student-\(t\) ("Student-t"), Slash ("Slash"), Symmetric Hyperbolic ("Hyperbolic"), Laplace ("Laplace"), and Contaminated Normal ("Contaminated normal"). By default, dist is set to "Gaussian".

skewness

An optional numeric vector specifying the skewness parameters of the noise distribution, when applicable.

extra

An optional value specifying the extra parameter of the noise distribution, when required.

setar

An optional positive integer indicating which component of the output series is used as the threshold variable. By default, setar is set to NULL, indicating that the model is not of SETAR type.

Verbose

An optional logical indicating whether a description of the simulated TAR model should be printed. By default, Verbose is set to TRUE.

Value

A data.frame containing the simulated multivariate output series, and, if specified, the threshold series and multivariate exogenous series.

References

Vanegas, L.H. and Calderón, S.A. and Rondón, L.M. (2025) Bayesian estimation of a multivariate tar model when the noise process distribution belongs to the class of gaussian variance mixtures. International Journal of Forecasting.

Examples

# \donttest{
###### Simulation of a trivariate TAR model with two regimes
n <- 2000
k <- 3
myars <- ars(nregim=2,p=c(1,2))
Z <- as.matrix(arima.sim(n=n+max(myars$p),list(ar=c(0.5))))
probs <- sort((0.6 + runif(myars$nregim-1)*0.8)*c(1:(myars$nregim-1))/myars$nregim)
dist <- "Student-t"; extra <- 6
parms <- list()
for(j in 1:myars$nregim){
    np <- 1 + myars$p[j]*k
    parms[[j]] <- list()
    parms[[j]]$location <- c(ifelse(runif(np*k)<=0.5,1,-1)*rbeta(np*k,shape1=4,shape2=16))
    parms[[j]]$location <- matrix(parms[[j]]$location,np,k)
    parms[[j]]$scale <- rgamma(k,shape=1,scale=1)*diag(k)
}
thresholds <- quantile(Z,probs=probs)
out1 <- simtar(n=n, k=k, ars=myars, parms=parms, thresholds=thresholds,
               t.series=Z, dist=dist, extra=extra)
#> 
#> 
#> Sample size          :2000 time points
#> 
#> Output Series        :Y1 | Y2 | Y3
#> 
#> Threshold Series     :Z with a delay equal to 0
#> 
#> Error Distribution   :Student-t(6)
#> 
#> Number of regimes    :2
#> 
#> Deterministics       :Intercept
#> 
#> Autoregressive orders:1, 2
#> 
#> 
#> Thresholds
#>                                  
#> Regime 1 (-Inf,0.131579957920141]
#> Regime 2  (0.131579957920141,Inf)
#> 
#> 
#> Regime 1:
#> 
#> Autoregressive coefficients
#>     Intercept        phi_1                        
#> Y1 -0.2255406 |  0.1439807  0.2168809  0.4115293 |
#> Y2  0.2611761 | -0.1515902 -0.2383677 -0.1963762 |
#> Y3 -0.2119724 | -0.1737034  0.1414631  0.1433601 |
#> 
#> Scale parameter
#>          Y1         Y2        Y3
#> Y1 1.213128 0.00000000 0.0000000
#> Y2 0.000000 0.07778218 0.0000000
#> Y3 0.000000 0.00000000 0.3539342
#> 
#> Regime 2:
#> 
#> Autoregressive coefficients
#>     Intercept        phi_1                             phi_2           
#> Y1 -0.1244369 | -0.1564086  0.2033482  0.2343376 | 0.1496783  0.2338904
#> Y2  0.1839745 |  0.1908082 -0.1094475  0.2651531 | 0.1761843 -0.2377336
#> Y3 -0.1511863 | -0.2601143 -0.2377179 -0.2092324 | 0.1361661  0.1169980
#>                
#> Y1 -0.1403925 |
#> Y2 -0.1721622 |
#> Y3  0.3564991 |
#> 
#> Scale parameter
#>           Y1       Y2        Y3
#> Y1 0.1905625 0.000000 0.0000000
#> Y2 0.0000000 4.947325 0.0000000
#> Y3 0.0000000 0.000000 0.2219154
str(out1)
#> 'data.frame':	2002 obs. of  5 variables:
#>  $ Y1    : num  1.09249 -1.16236 0.09991 0.00086 -2.87072 ...
#>  $ Y2    : num  0.294 0.2244 0.4544 0.5051 0.0883 ...
#>  $ Y3    : num  -1.189 0.96 0.556 -0.403 -0.866 ...
#>  $ Regime: num  NA NA 2 2 1 1 1 1 1 1 ...
#>  $ Z     : num  0.5146 0.7354 0.5426 1.8268 -0.0101 ...

fit1 <- mtar(~ Y1 + Y2 + Y3 | Z, data=out1, ars=myars, dist=dist,
             n.burn=2000, n.sim=3000, n.thin=2)
summary(fit1)
#> 
#> 
#> Sample size          : 1999 time points
#> 
#> Output Series        : Y1    |    Y2    |    Y3
#> 
#> Threshold Series     : Z with a estimated delay equal to 0
#> 
#> Error Distribution   : Student-t
#> 
#> Number of regimes    : 2
#> 
#> Deterministics       : Intercept  
#> 
#> Autoregressive orders: 1, 2
#> 
#> 
#> 
#> Thresholds (Mean, HDI.Lower, HDI.Upper)
#>                                                      
#> Regime 1 (-Inf,0.12941] (-Inf,0.12787] (-Inf,0.13218]
#> Regime 2  (0.12941,Inf)  (0.12787,Inf)  (0.13218,Inf)
#> 
#> 
#> Regime1:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)  HDI.Lower HDI.Upper             Mean  2(1-PD) 
#> (Intercept) -0.21012   0.00001  -0.28609  -0.13789    |     0.26915     1e-05
#> Y1.lag(1)    0.11137   0.00067   0.05603   0.16621    |    -0.15789     1e-05
#> Y2.lag(1)    0.19552   0.00001   0.14690   0.24040    |    -0.24417     1e-05
#> Y3.lag(1)    0.43764   0.00001   0.35289   0.52152    |    -0.20074     1e-05
#>             HDI.Lower HDI.Upper             Mean  2(1-PD)  HDI.Lower HDI.Upper
#> (Intercept)   0.25118   0.28658    |    -0.19769     1e-05  -0.23414  -0.15845
#> Y1.lag(1)    -0.17187  -0.14443    |    -0.17671     1e-05  -0.20551  -0.14729
#> Y2.lag(1)    -0.25608  -0.23247    |     0.15098     1e-05   0.12695   0.17617
#> Y3.lag(1)    -0.22132  -0.17929    |     0.13952     1e-05   0.09292   0.17935
#> 
#> Scale parameter (Mean, HDI.Lower, HDI.Upper)
#>          Y1       Y2       Y3            Y1       Y2       Y3           Y1
#> Y1  1.23282 -0.00334 -0.01575    .  1.10851 -0.02208 -0.05736    . 1.35352
#> Y2 -0.00334  0.07443 -0.00053    . -0.02208  0.06685 -0.01124    . 0.01587
#> Y3 -0.01575 -0.00053  0.33870    . -0.05736 -0.01124  0.30239    . 0.02383
#>         Y2      Y3
#> Y1 0.01587 0.02383
#> Y2 0.08179 0.00895
#> Y3 0.00895 0.37326
#> 
#> 
#> Regime2:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)  HDI.Lower HDI.Upper             Mean  2(1-PD) 
#> (Intercept) -0.12061     1e-05  -0.15537  -0.08595    |     0.31410   0.00001
#> Y1.lag(1)   -0.16006     1e-05  -0.19343  -0.12641    |     0.16950   0.06267
#> Y2.lag(1)    0.19517     1e-05   0.18150   0.20927    |    -0.19194   0.00001
#> Y3.lag(1)    0.24566     1e-05   0.20680   0.28717    |     0.18557   0.09000
#> Y1.lag(2)    0.12562     1e-05   0.09385   0.15709    |     0.13567   0.09200
#> Y2.lag(2)    0.22423     1e-05   0.20756   0.24266    |    -0.33059   0.00001
#> Y3.lag(2)   -0.14716     1e-05  -0.18516  -0.11024    |    -0.12238   0.23533
#>             HDI.Lower HDI.Upper             Mean  2(1-PD)  HDI.Lower HDI.Upper
#> (Intercept)   0.13323   0.48864    |    -0.19034     1e-05  -0.22871  -0.14906
#> Y1.lag(1)    -0.00271   0.34616    |    -0.26607     1e-05  -0.30248  -0.22956
#> Y2.lag(1)    -0.26581  -0.11920    |    -0.24753     1e-05  -0.26310  -0.23077
#> Y3.lag(1)    -0.03128   0.39640    |    -0.20859     1e-05  -0.25638  -0.16006
#> Y1.lag(2)    -0.03175   0.28884    |     0.14599     1e-05   0.11129   0.18421
#> Y2.lag(2)    -0.42277  -0.23813    |     0.12141     1e-05   0.09923   0.14121
#> Y3.lag(2)    -0.33265   0.06894    |     0.32763     1e-05   0.28339   0.36992
#> 
#> Scale parameter (Mean, HDI.Lower, HDI.Upper)
#>          Y1       Y2       Y3            Y1       Y2       Y3           Y1
#> Y1  0.19035 -0.03105  0.00624    .  0.16717 -0.10240 -0.00840    . 0.21232
#> Y2 -0.03105  4.98327 -0.03369    . -0.10240  4.41244 -0.11571    . 0.04331
#> Y3  0.00624 -0.03369  0.24129    . -0.00840 -0.11571  0.21178    . 0.02305
#>         Y2      Y3
#> Y1 0.04331 0.02305
#> Y2 5.58519 0.04355
#> Y3 0.04355 0.27083
#> 
#> 
#> Extra parameter
#>                Mean  2(1-PD)  HDI.Lower HDI.Upper
#> nu          6.11705      .       5.2977    7.0911
#> 
#> 

###### Simulation of a trivariate VAR model
n <- 2000
k <- 3
myars <- ars(nregim=1,p=2)
dist <- "Slash"; extra <- 2
parms <- list()
for(j in 1:myars$nregim){
    np <- 1 + myars$p[j]*k
    parms[[j]] <- list()
    parms[[j]]$location <- c(ifelse(runif(np*k)<=0.5,1,-1)*rbeta(np*k,shape1=4,shape2=16))
    parms[[j]]$location <- matrix(parms[[j]]$location,np,k)
    parms[[j]]$scale <- rgamma(k,shape=1,scale=1)*diag(k)
}
out2 <- simtar(n=n, k=k, ars=myars, parms=parms, dist=dist, extra=extra)
#> 
#> 
#> Sample size          :2000 time points
#> 
#> Output Series        :Y1 | Y2 | Y3
#> 
#> Error Distribution   :Slash(2)
#> 
#> Number of regimes    :1
#> 
#> Deterministics       :Intercept
#> 
#> Autoregressive orders:2 in each regime
#> 
#> 
#> Regime 1:
#> 
#> Autoregressive coefficients
#>      Intercept        phi_1                                phi_2           
#> Y1 -0.34733955 | -0.1814028  0.1624676  0.04905092 | -0.11849438  0.1693805
#> Y2 -0.07362029 | -0.2125539  0.1660906  0.13756127 |  0.12792916 -0.2544173
#> Y3  0.18666321 | -0.3110357 -0.1766985 -0.12188194 | -0.07284076  0.1589948
#>                
#> Y1  0.1477139 |
#> Y2 -0.1730014 |
#> Y3 -0.1864142 |
#> 
#> Scale parameter
#>           Y1      Y2        Y3
#> Y1 0.3154734 0.00000 0.0000000
#> Y2 0.0000000 1.75181 0.0000000
#> Y3 0.0000000 0.00000 0.2907899
str(out2)
#> 'data.frame':	2002 obs. of  4 variables:
#>  $ Y1    : num  0.468 1.14 0.338 1.308 -0.104 ...
#>  $ Y2    : num  0.75 1.27 1.57 -0.1 -2.06 ...
#>  $ Y3    : num  -0.918 1.144 0.78 -0.329 -0.448 ...
#>  $ Regime: num  NA NA 1 1 1 1 1 1 1 1 ...

fit2 <- mtar(~ Y1 + Y2 + Y3, data=out2, ars=myars, dist=dist,
             n.burn=2000, n.sim=3000, n.thin=2)
summary(fit2)
#> 
#> 
#> Sample size          : 2000 time points
#> 
#> Output Series        : Y1    |    Y2    |    Y3
#> 
#> Error Distribution   : Slash
#> 
#> Number of regimes    : 1
#> 
#> Deterministics       : Intercept  
#> 
#> Autoregressive orders: 2 in each regime
#> 
#> 
#> 
#> 
#> Regime1:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)  HDI.Lower HDI.Upper             Mean  2(1-PD) 
#> (Intercept) -0.34488   0.00001  -0.38063  -0.29980    |    -0.06727   0.15467
#> Y1.lag(1)   -0.19283   0.00001  -0.21623  -0.16882    |    -0.20675   0.00001
#> Y2.lag(1)    0.16621   0.00001   0.15466   0.17730    |     0.17358   0.00001
#> Y3.lag(1)    0.02924   0.01733   0.00438   0.05355    |     0.12917   0.00001
#> Y1.lag(2)   -0.10066   0.00001  -0.12502  -0.07690    |     0.15535   0.00001
#> Y2.lag(2)    0.16409   0.00001   0.15253   0.17662    |    -0.25647   0.00001
#> Y3.lag(2)    0.13445   0.00001   0.11194   0.15801    |    -0.23009   0.00001
#>             HDI.Lower HDI.Upper             Mean  2(1-PD)  HDI.Lower HDI.Upper
#> (Intercept)  -0.16568   0.02040    |     0.17135     1e-05   0.13428   0.21232
#> Y1.lag(1)    -0.26409  -0.14976    |    -0.30469     1e-05  -0.32882  -0.28188
#> Y2.lag(1)     0.14558   0.20030    |    -0.17665     1e-05  -0.18734  -0.16590
#> Y3.lag(1)     0.06799   0.18362    |    -0.10228     1e-05  -0.12651  -0.07706
#> Y1.lag(2)     0.09305   0.20711    |    -0.07130     1e-05  -0.09438  -0.04711
#> Y2.lag(2)    -0.28593  -0.22795    |     0.16186     1e-05   0.15003   0.17367
#> Y3.lag(2)    -0.28380  -0.17742    |    -0.18878     1e-05  -0.21187  -0.16536
#> 
#> Scale parameter (Mean, HDI.Lower, HDI.Upper)
#>         Y1      Y2      Y3            Y1       Y2       Y3           Y1      Y2
#> Y1 0.31747 0.00468 0.00046    .  0.28772 -0.03122 -0.01532    . 0.34758 0.04313
#> Y2 0.00468 1.77753 0.00659    . -0.03122  1.60529 -0.02862    . 0.04313 1.94504
#> Y3 0.00046 0.00659 0.30392    . -0.01532 -0.02862  0.27622    . 0.01602 0.04432
#>         Y3
#> Y1 0.01602
#> Y2 0.04432
#> Y3 0.33263
#> 
#> 
#> Extra parameter
#>                Mean  2(1-PD)  HDI.Lower HDI.Upper
#> nu          1.99666      .      1.85056   2.14102
#> 
#> 

n <- 5000
k <- 3
myars <- ars(nregim=2,p=c(1,2))
dist <- "Laplace"
parms <- list()
for(j in 1:myars$nregim){
    np <- 1 + myars$p[j]*k
    parms[[j]] <- list()
    parms[[j]]$location <- c(ifelse(runif(np*k)<=0.5,1,-1)*rbeta(np*k,shape1=4,shape2=16))
    parms[[j]]$location <- matrix(parms[[j]]$location,np,k)
    parms[[j]]$scale <- rgamma(k,shape=1,scale=1)*diag(k)
}
out3 <- simtar(n=n, k=k, ars=myars, parms=parms, delay=2,
               thresholds=-1, dist=dist, setar=2)
#> 
#> 
#> Sample size          :5000 time points
#> 
#> Output Series        :Y1 | Y2 | Y3
#> 
#> Threshold Series     :Y2 with a delay equal to 2
#> 
#> Error Distribution   :Laplace
#> 
#> Number of regimes    :2
#> 
#> Deterministics       :Intercept
#> 
#> Autoregressive orders:1, 2
#> 
#> 
#> Thresholds
#>                   
#> Regime 1 (-Inf,-1]
#> Regime 2  (-1,Inf)
#> 
#> 
#> Regime 1:
#> 
#> Autoregressive coefficients
#>    Intercept        phi_1                         
#> Y1 0.2781811 |  0.1159696 -0.2143696 -0.04987354 |
#> Y2 0.1355332 |  0.1262336 -0.2026278 -0.17851924 |
#> Y3 0.2287947 | -0.1631096 -0.2112006 -0.16307706 |
#> 
#> Scale parameter
#>           Y1        Y2        Y3
#> Y1 0.6139319 0.0000000 0.0000000
#> Y2 0.0000000 0.4627907 0.0000000
#> Y3 0.0000000 0.0000000 0.2167862
#> 
#> Regime 2:
#> 
#> Autoregressive coefficients
#>      Intercept        phi_1                            phi_2           
#> Y1  0.17224018 |  0.2025184 0.18475793 0.1654042 | 0.1764581  0.1564427
#> Y2  0.09870482 | -0.2217394 0.08213579 0.1812696 | 0.2598424 -0.1961725
#> Y3 -0.21688280 | -0.1426503 0.06955284 0.1841095 | 0.1217359  0.1232485
#>                
#> Y1 0.12089878 |
#> Y2 0.07320554 |
#> Y3 0.02521734 |
#> 
#> Scale parameter
#>          Y1       Y2        Y3
#> Y1 2.565465 0.000000 0.0000000
#> Y2 0.000000 3.074551 0.0000000
#> Y3 0.000000 0.000000 0.2454203
str(out3)
#> 'data.frame':	5002 obs. of  4 variables:
#>  $ Y1    : num  0.144 -1.162 -0.753 -3.734 6.558 ...
#>  $ Y2    : num  1.2399 -0.0874 4.0616 11.2684 -1.5069 ...
#>  $ Y3    : num  -0.131 -0.258 0.194 1.239 0.979 ...
#>  $ Regime: num  NA NA 2 2 2 2 1 1 2 2 ...

fit3 <- mtar(~ Y1 + Y2 + Y3, data=out3, ars=myars, dist=dist,
             n.burn=2000, n.sim=3000, n.thin=2, setar=2)
summary(fit3)
#> 
#> 
#> Sample size          : 4999 time points
#> 
#> Output Series        : Y1    |    Y2    |    Y3
#> 
#> Threshold Series     : Y2 with a estimated delay equal to 2
#> 
#> Error Distribution   : Laplace
#> 
#> Number of regimes    : 2
#> 
#> Deterministics       : Intercept  
#> 
#> Autoregressive orders: 1, 2
#> 
#> 
#> 
#> Thresholds (Mean, HDI.Lower, HDI.Upper)
#>                                                       
#> Regime 1 (-Inf,-0.9998] (-Inf,-1.0023] (-Inf,-0.99705]
#> Regime 2  (-0.9998,Inf)  (-1.0023,Inf)  (-0.99705,Inf)
#> 
#> 
#> Regime1:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)  HDI.Lower HDI.Upper             Mean  2(1-PD) 
#> (Intercept)  0.26360     1e-05   0.21108   0.31601    |     0.13098     1e-05
#> Y1.lag(1)    0.11563     1e-05   0.10285   0.12714    |     0.11911     1e-05
#> Y2.lag(1)   -0.21111     1e-05  -0.22344  -0.20033    |    -0.19935     1e-05
#> Y3.lag(1)   -0.04772     2e-03  -0.07430  -0.01880    |    -0.17323     1e-05
#>             HDI.Lower HDI.Upper             Mean  2(1-PD)  HDI.Lower HDI.Upper
#> (Intercept)   0.08658   0.17776    |     0.23007     1e-05   0.19893   0.25996
#> Y1.lag(1)     0.10859   0.12836    |    -0.16240     1e-05  -0.16866  -0.15546
#> Y2.lag(1)    -0.20928  -0.19107    |    -0.21298     1e-05  -0.21914  -0.20576
#> Y3.lag(1)    -0.19721  -0.14863    |    -0.16340     1e-05  -0.17891  -0.14783
#> 
#> Scale parameter (Mean, HDI.Lower, HDI.Upper)
#>         Y1       Y2       Y3            Y1       Y2       Y3           Y1
#> Y1 0.60231  0.01300  0.00137    .  0.54547 -0.01427 -0.01700    . 0.65491
#> Y2 0.01300  0.44833 -0.00241    . -0.01427  0.41147 -0.01855    . 0.04093
#> Y3 0.00137 -0.00241  0.20547    . -0.01700 -0.01855  0.18668    . 0.01961
#>         Y2      Y3
#> Y1 0.04093 0.01961
#> Y2 0.49163 0.01416
#> Y3 0.01416 0.22331
#> 
#> 
#> Regime2:
#> 
#> Autoregressive coefficients
#>                Mean  2(1-PD)  HDI.Lower HDI.Upper             Mean  2(1-PD) 
#> (Intercept) 0.13931   0.00867   0.03672   0.23995    |     0.16473   0.00467
#> Y1.lag(1)   0.20725   0.00001   0.18643   0.23012    |    -0.21651   0.00001
#> Y2.lag(1)   0.18105   0.00001   0.16540   0.19904    |     0.08631   0.00001
#> Y3.lag(1)   0.14863   0.00001   0.08828   0.20633    |     0.15734   0.00001
#> Y1.lag(2)   0.18601   0.00001   0.16572   0.20870    |     0.26186   0.00001
#> Y2.lag(2)   0.16206   0.00001   0.12893   0.18913    |    -0.20768   0.00001
#> Y3.lag(2)   0.13030   0.00001   0.07819   0.18456    |     0.06547   0.02867
#>             HDI.Lower HDI.Upper             Mean  2(1-PD)  HDI.Lower HDI.Upper
#> (Intercept)   0.04777   0.27046    |    -0.21156   0.00001  -0.24859  -0.17863
#> Y1.lag(1)    -0.23874  -0.19306    |    -0.14512   0.00001  -0.15149  -0.13812
#> Y2.lag(1)     0.06804   0.10708    |     0.06868   0.00001   0.06273   0.07421
#> Y3.lag(1)     0.09376   0.22302    |     0.17869   0.00001   0.16261   0.19513
#> Y1.lag(2)     0.23919   0.28533    |     0.11672   0.00001   0.11078   0.12337
#> Y2.lag(2)    -0.24395  -0.17575    |     0.11894   0.00001   0.10895   0.12913
#> Y3.lag(2)     0.00794   0.12276    |     0.02687   0.00067   0.01246   0.04280
#> 
#> Scale parameter (Mean, HDI.Lower, HDI.Upper)
#>          Y1      Y2       Y3            Y1       Y2       Y3           Y1
#> Y1  2.32687 0.06107 -0.00884    .  2.17618 -0.04946 -0.03821    . 2.49111
#> Y2  0.06107 3.06939  0.02320    . -0.04946  2.86142 -0.00775    . 0.16521
#> Y3 -0.00884 0.02320  0.22370    . -0.03821 -0.00775  0.20816    . 0.02084
#>         Y2      Y3
#> Y1 0.16521 0.02084
#> Y2 3.30092 0.05773
#> Y3 0.05773 0.23890
#> 
#> 

# }