Simulation of multivariate time series from a TAR model

This function simulates multivariate time series generated by a user-specified Threshold Autoregressive (TAR) model.

Usage

simtar(
  n,
  k = 2,
  ars = ars(),
  Intercept = TRUE,
  trend = c("none", "linear", "quadratic"),
  nseason = NULL,
  parms,
  delay = 0,
  thresholds = NULL,
  t.series = NULL,
  ex.series = NULL,
  dist = c("Gaussian", "Student-t", "Hyperbolic", "Laplace", "Slash",
    "Contaminated normal", "Skew-Student-t", "Skew-normal"),
  skewness = NULL,
  extra = NULL,
  setar = NULL,
  Verbose = TRUE
)

Arguments

n

A positive integer specifying the length of the simulated output series.

k

A positive integer specifying the dimension of the multivariate output series.

ars

A list defining the TAR model structure, composed of four elements: the number of regimes (nregim), the autoregressive order (p), and the maximum lags of the exogenous (q) and threshold (d) series within each regime. This object can be validated using the ars() function.

Intercept

An optional logical indicating whether an intercept term is included in each regime.

trend

An optional character string specifying the degree of deterministic time trend to be included in each regime. Available options are a linear trend ("linear"), a quadratic trend ("quadratic"), or no trend ("none"). By default, trend is set to "none".

nseason

An optional integer, greater than or equal to 2, specifying the number of seasonal periods. When provided, nseason - 1 seasonal dummy variables are added to the regressors within each regime. By default, nseason is set to NULL, thereby indicating that the TAR model has no seasonal effects.

parms

A list with one sublist per regime. Each sublist contains two matrices: the first matrix corresponds to the location parameters, and the second matrix corresponds to the scale parameters of the model.

delay

An optional non-negative integer specifying the delay parameter of the threshold series. By default, delay is set to 0.

thresholds

A numeric vector of length nregim-1 containing the threshold values, sorted in ascending order.

t.series

A matrix containing the values of the threshold series.

ex.series

A matrix containing the values of the multivariate exogenous series.

dist

An optional character string specifying the multivariate distribution chosen to model the noise process. Supported options include Gaussian ("Gaussian"), Student-\(t\) ("Student-t"), Slash ("Slash"), Symmetric Hyperbolic ("Hyperbolic"), Laplace ("Laplace"), and Contaminated Normal ("Contaminated normal"). By default, dist is set to "Gaussian".

skewness

An optional numeric vector specifying the skewness parameters of the noise distribution, when applicable.

extra

An optional value specifying the extra parameter of the noise distribution, when required.

setar

An optional positive integer indicating which component of the output series is used as the threshold variable. By default, setar is set to NULL, indicating that the model is not of SETAR type.

Verbose

An optional logical indicating whether a description of the simulated TAR model should be printed. By default, Verbose is set to TRUE.

Value

A data.frame containing the simulated multivariate output series, and, if specified, the threshold series and multivariate exogenous series.

References

Vanegas, L.H. and Calderón, S.A. and Rondón, L.M. (2025) Bayesian estimation of a multivariate tar model when the noise process distribution belongs to the class of gaussian variance mixtures. International Journal of Forecasting.

Examples

# \donttest{
###### Simulation of a trivariate TAR model with two regimes
n <- 2000
k <- 3
myars <- ars(nregim=2,p=c(1,2))
Z <- as.matrix(arima.sim(n=n+max(myars$p),list(ar=c(0.5))))
probs <- sort((0.6 + runif(myars$nregim-1)*0.8)*c(1:(myars$nregim-1))/myars$nregim)
dist <- "Student-t"; extra <- 6
parms <- list()
for(j in 1:myars$nregim){
    np <- 1 + myars$p[j]*k
    parms[[j]] <- list()
    parms[[j]]$location <- c(ifelse(runif(np*k)<=0.5,1,-1)*rbeta(np*k,shape1=4,shape2=16))
    parms[[j]]$location <- matrix(parms[[j]]$location,np,k)
    parms[[j]]$scale <- rgamma(k,shape=1,scale=1)*diag(k)
}
thresholds <- quantile(Z,probs=probs)
out1 <- simtar(n=n, k=k, ars=myars, parms=parms, thresholds=thresholds,
               t.series=Z, dist=dist, extra=extra)
#> 
#> 
#> Sample size          :2000 time points
#> 
#> Output Series        :Y1 | Y2 | Y3
#> 
#> Threshold Series     :Z with a delay equal to 0
#> 
#> Error Distribution   :Student-t(6)
#> 
#> Number of regimes    :2
#> 
#> Deterministics       :Intercept
#> 
#> Autoregressive orders:1, 2
#> 
#> 
#> Thresholds
#>                                   
#> Regime 1 (-Inf,-0.523584475029785]
#> Regime 2  (-0.523584475029785,Inf)
#> 
#> 
#> Regime 1:
#> 
#> Autoregressive coefficients
#>      Intercept         phi_1                          
#> Y1  0.08455132 |  0.09135594 -0.05373073  0.35938317 |
#> Y2 -0.09738182 |  0.21044708 -0.07874382  0.05086947 |
#> Y3  0.13404774 | -0.05748161  0.35759775 -0.13991075 |
#> 
#> Scale parameter
#>           Y1      Y2        Y3
#> Y1 0.4647418 0.00000 0.0000000
#> Y2 0.0000000 2.13068 0.0000000
#> Y3 0.0000000 0.00000 0.5273689
#> 
#> Regime 2:
#> 
#> Autoregressive coefficients
#>     Intercept        phi_1                               phi_2            
#> Y1  0.2266635 | -0.1521761 -0.2067243 -0.16764729 | 0.27022934 -0.05672852
#> Y2  0.2339818 | -0.1356809 -0.1337520  0.09817847 | 0.16892841  0.23987461
#> Y3 -0.1835658 |  0.2962713  0.2094498 -0.13173173 | 0.06727942  0.05066859
#>                
#> Y1  0.1241443 |
#> Y2  0.1484477 |
#> Y3 -0.2572822 |
#> 
#> Scale parameter
#>          Y1       Y2       Y3
#> Y1 0.557084 0.000000 0.000000
#> Y2 0.000000 3.182895 0.000000
#> Y3 0.000000 0.000000 3.828745
str(out1)
#> 'data.frame':	2002 obs. of  5 variables:
#>  $ Y1    : num  -0.37361 -0.10824 -1.32887 0.46173 -0.00994 ...
#>  $ Y2    : num  0.75 0.806 -0.49 -2.586 3.852 ...
#>  $ Y3    : num  -1.16 -2.16 1.11 1.07 2.55 ...
#>  $ Regime: num  NA NA 2 2 2 1 1 1 1 1 ...
#>  $ Z     : num  0.274 1.071 1.046 0.734 -0.194 ...

fit1 <- mtar(~ Y1 + Y2 + Y3 | Z, data=out1, ars=myars, dist=dist,
             n.burn=200, n.sim=300, n.thin=2)
summary(fit1)
#> 
#> 
#> Sample size          :1999 time points
#> 
#> Output Series        :Y1    |    Y2    |    Y3
#> 
#> Threshold Series     :Z with a estimated delay equal to 0
#> 
#> Error Distribution   :Student-t
#> 
#> Number of regimes    :2
#> 
#> Deterministics       :Intercept  
#> 
#> Autoregressive orders:1, 2
#> 
#> 
#> 
#> Thresholds (Mean, HDI.Lower, HDI.Upper)
#>                                                         
#> Regime 1 (-Inf,-0.52415] (-Inf,-0.52506] (-Inf,-0.52319]
#> Regime 2  (-0.52415,Inf)  (-0.52506,Inf)  (-0.52319,Inf)
#> 
#> 
#> Regime1:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)  HDI.Lower HDI.Upper             Mean  2(1-PD) 
#> (Intercept)  0.07189   0.02000   0.01851   0.12522    |    -0.04609   0.45333
#> Y1.lag(1)    0.09449   0.00001   0.04952   0.14756    |     0.17237   0.00001
#> Y2.lag(1)   -0.03622   0.00667  -0.06510  -0.01065    |    -0.08605   0.00001
#> Y3.lag(1)    0.36536   0.00001   0.33810   0.39039    |    -0.00679   0.81333
#>             HDI.Lower HDI.Upper             Mean  2(1-PD)  HDI.Lower HDI.Upper
#> (Intercept)  -0.16565   0.06783    |     0.12904   0.00001   0.07116   0.18929
#> Y1.lag(1)     0.05323   0.26772    |     0.00084   0.96667  -0.05192   0.04576
#> Y2.lag(1)    -0.13016  -0.03180    |     0.37488   0.00001   0.34667   0.40212
#> Y3.lag(1)    -0.06348   0.05503    |    -0.12784   0.00001  -0.15571  -0.09366
#> 
#> Scale parameter (Mean, HDI.Lower, HDI.Upper)
#>          Y1       Y2       Y3            Y1       Y2       Y3           Y1
#> Y1  0.46441 -0.01858 -0.01638    .  0.41040 -0.09551 -0.05613    . 0.52713
#> Y2 -0.01858  2.20149  0.08181    . -0.09551  1.90819  0.01117    . 0.05525
#> Y3 -0.01638  0.08181  0.52631    . -0.05613  0.01117  0.44958    . 0.02197
#>         Y2      Y3
#> Y1 0.05525 0.02197
#> Y2 2.44888 0.15274
#> Y3 0.15274 0.58630
#> 
#> 
#> Regime2:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)  HDI.Lower HDI.Upper             Mean  2(1-PD) 
#> (Intercept)  0.24147     1e-05   0.19756   0.29031    |     0.27088     1e-05
#> Y1.lag(1)   -0.18207     1e-05  -0.22262  -0.14526    |    -0.20072     1e-05
#> Y2.lag(1)   -0.20840     1e-05  -0.22773  -0.18933    |    -0.13647     1e-05
#> Y3.lag(1)   -0.17171     1e-05  -0.19402  -0.15180    |     0.10786     1e-05
#> Y1.lag(2)    0.28519     1e-05   0.23886   0.32246    |     0.18405     1e-05
#> Y2.lag(2)   -0.05719     1e-05  -0.07773  -0.03667    |     0.20794     1e-05
#> Y3.lag(2)    0.13639     1e-05   0.11428   0.15756    |     0.17613     1e-05
#>             HDI.Lower HDI.Upper             Mean  2(1-PD)  HDI.Lower HDI.Upper
#> (Intercept)   0.15424   0.38793    |    -0.11138   0.07333  -0.22081   0.02103
#> Y1.lag(1)    -0.31437  -0.08653    |     0.27601   0.00001   0.16305   0.38575
#> Y2.lag(1)    -0.18088  -0.09131    |     0.18300   0.00001   0.13459   0.23324
#> Y3.lag(1)     0.06551   0.15113    |    -0.14317   0.00001  -0.19879  -0.09634
#> Y1.lag(2)     0.08561   0.28278    |     0.09175   0.09333  -0.01845   0.19160
#> Y2.lag(2)     0.15503   0.25880    |     0.01693   0.56000  -0.03899   0.07812
#> Y3.lag(2)     0.13043   0.22898    |    -0.29226   0.00001  -0.34345  -0.23518
#> 
#> Scale parameter (Mean, HDI.Lower, HDI.Upper)
#>          Y1       Y2       Y3            Y1       Y2       Y3           Y1
#> Y1  0.57684  0.00562 -0.06395    .  0.51582 -0.08373 -0.14576    . 0.62850
#> Y2  0.00562  3.34792 -0.19325    . -0.08373  3.05006 -0.41411    . 0.09295
#> Y3 -0.06395 -0.19325  3.82567    . -0.14576 -0.41411  3.44955    . 0.01289
#>         Y2      Y3
#> Y1 0.09295 0.01289
#> Y2 3.70655 0.01820
#> Y3 0.01820 4.16580
#> 
#> 
#> Extra parameter
#>                Mean  2(1-PD)  HDI.Lower HDI.Upper
#> nu           6.1159      .      5.30672   7.01417
#> 
#> 

###### Simulation of a trivariate VAR model
n <- 2000
k <- 3
myars <- ars(nregim=1,p=2)
dist <- "Slash"; extra <- 2
parms <- list()
for(j in 1:myars$nregim){
    np <- 1 + myars$p[j]*k
    parms[[j]] <- list()
    parms[[j]]$location <- c(ifelse(runif(np*k)<=0.5,1,-1)*rbeta(np*k,shape1=4,shape2=16))
    parms[[j]]$location <- matrix(parms[[j]]$location,np,k)
    parms[[j]]$scale <- rgamma(k,shape=1,scale=1)*diag(k)
}
out2 <- simtar(n=n, k=k, ars=myars, parms=parms, dist=dist, extra=extra)
#> 
#> 
#> Sample size          :2000 time points
#> 
#> Output Series        :Y1 | Y2 | Y3
#> 
#> Error Distribution   :Slash(2)
#> 
#> Number of regimes    :1
#> 
#> Deterministics       :Intercept
#> 
#> Autoregressive orders:2 in each regime
#> 
#> 
#> Regime 1:
#> 
#> Autoregressive coefficients
#>     Intercept         phi_1                               phi_2           
#> Y1 -0.1817196 | -0.05892081 -0.1174856 -0.2210835 | -0.15624726  0.1348125
#> Y2 -0.3484329 | -0.18536454 -0.1413378  0.2081647 |  0.15467495  0.2378252
#> Y3  0.2427122 | -0.09967076 -0.0792752 -0.3065440 |  0.08335945 -0.2140239
#>               
#> Y1 0.2444174 |
#> Y2 0.1711338 |
#> Y3 0.1814989 |
#> 
#> Scale parameter
#>            Y1       Y2       Y3
#> Y1 0.02278457 0.000000 0.000000
#> Y2 0.00000000 1.695668 0.000000
#> Y3 0.00000000 0.000000 1.059126
str(out2)
#> 'data.frame':	2002 obs. of  4 variables:
#>  $ Y1    : num  -1.429 2.239 -0.155 -0.62 -1.533 ...
#>  $ Y2    : num  -0.7005 -1.9518 -0.0968 2.2847 0.5532 ...
#>  $ Y3    : num  -0.0755 0.6494 0.4316 1.0295 -1.2928 ...
#>  $ Regime: num  NA NA 1 1 1 1 1 1 1 1 ...

fit2 <- mtar(~ Y1 + Y2 + Y3, data=out2, ars=myars, dist=dist,
             n.burn=200, n.sim=300, n.thin=2)
summary(fit2)
#> 
#> 
#> Sample size          :2000 time points
#> 
#> Output Series        :Y1    |    Y2    |    Y3
#> 
#> Error Distribution   :Slash
#> 
#> Number of regimes    :1
#> 
#> Deterministics       :Intercept  
#> 
#> Autoregressive orders:2 in each regime
#> 
#> 
#> 
#> 
#> Regime1:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)  HDI.Lower HDI.Upper             Mean  2(1-PD) 
#> (Intercept) -0.18113     1e-05  -0.19106  -0.17080    |    -0.37685     1e-05
#> Y1.lag(1)   -0.06690     1e-05  -0.07894  -0.05586    |    -0.20090     1e-05
#> Y2.lag(1)   -0.11697     1e-05  -0.11920  -0.11444    |    -0.13055     1e-05
#> Y3.lag(1)   -0.22267     1e-05  -0.22561  -0.21946    |     0.17821     1e-05
#> Y1.lag(2)   -0.16133     1e-05  -0.16925  -0.15274    |     0.11853     1e-05
#> Y2.lag(2)    0.13444     1e-05   0.13170   0.13681    |     0.23694     1e-05
#> Y3.lag(2)    0.24320     1e-05   0.23845   0.24795    |     0.14114     1e-05
#>             HDI.Lower HDI.Upper             Mean  2(1-PD)  HDI.Lower HDI.Upper
#> (Intercept)  -0.47843  -0.28404    |     0.20536     1e-05   0.12293   0.27496
#> Y1.lag(1)    -0.29039  -0.09564    |    -0.13572     1e-05  -0.20148  -0.05518
#> Y2.lag(1)    -0.14839  -0.10969    |    -0.08295     1e-05  -0.09662  -0.06720
#> Y3.lag(1)     0.15195   0.20159    |    -0.30745     1e-05  -0.32914  -0.28619
#> Y1.lag(2)     0.03571   0.18668    |     0.03624     2e-01  -0.02016   0.08610
#> Y2.lag(2)     0.21639   0.25636    |    -0.22655     1e-05  -0.24461  -0.21037
#> Y3.lag(2)     0.10153   0.17308    |     0.18850     1e-05   0.15902   0.22049
#> 
#> Scale parameter (Mean, HDI.Lower, HDI.Upper)
#>         Y1       Y2       Y3            Y1       Y2       Y3           Y1
#> Y1 0.02243  0.00365  0.00303    .  0.02023 -0.00439 -0.00482    . 0.02441
#> Y2 0.00365  1.62875 -0.02921    . -0.00439  1.46762 -0.09235    . 0.01339
#> Y3 0.00303 -0.02921  1.02665    . -0.00482 -0.09235  0.93248    . 0.01103
#>         Y2      Y3
#> Y1 0.01339 0.01103
#> Y2 1.79205 0.02858
#> Y3 0.02858 1.12976
#> 
#> 
#> Extra parameter
#>                Mean  2(1-PD)  HDI.Lower HDI.Upper
#> nu          1.96468      .      1.80127   2.10313
#> 
#> 

n <- 5000
k <- 3
myars <- ars(nregim=2,p=c(1,2))
dist <- "Laplace"
parms <- list()
for(j in 1:myars$nregim){
    np <- 1 + myars$p[j]*k
    parms[[j]] <- list()
    parms[[j]]$location <- c(ifelse(runif(np*k)<=0.5,1,-1)*rbeta(np*k,shape1=4,shape2=16))
    parms[[j]]$location <- matrix(parms[[j]]$location,np,k)
    parms[[j]]$scale <- rgamma(k,shape=1,scale=1)*diag(k)
}
out3 <- simtar(n=n, k=k, ars=myars, parms=parms, delay=2,
               thresholds=-1, dist=dist, setar=2)
#> 
#> 
#> Sample size          :5000 time points
#> 
#> Output Series        :Y1 | Y2 | Y3
#> 
#> Threshold Series     :Y2 with a delay equal to 2
#> 
#> Error Distribution   :Laplace
#> 
#> Number of regimes    :2
#> 
#> Deterministics       :Intercept
#> 
#> Autoregressive orders:1, 2
#> 
#> 
#> Thresholds
#>                   
#> Regime 1 (-Inf,-1]
#> Regime 2  (-1,Inf)
#> 
#> 
#> Regime 1:
#> 
#> Autoregressive coefficients
#>      Intercept        phi_1                         
#> Y1 -0.14493206 |  0.2339228  0.24513249 -0.1361639 |
#> Y2  0.07424586 | -0.2299090  0.06577474 -0.1685062 |
#> Y3  0.11638508 |  0.2047995 -0.17911784  0.1739592 |
#> 
#> Scale parameter
#>         Y1       Y2        Y3
#> Y1 5.00659 0.000000 0.0000000
#> Y2 0.00000 0.417837 0.0000000
#> Y3 0.00000 0.000000 0.2239914
#> 
#> Regime 2:
#> 
#> Autoregressive coefficients
#>     Intercept        phi_1                               phi_2           
#> Y1  0.2009046 |  0.1646369 -0.24891004 -0.1840699 | -0.1666491 -0.2049374
#> Y2 -0.1954958 | -0.3575927  0.23360284  0.1942346 |  0.2557942 -0.1454535
#> Y3 -0.1642671 | -0.2112175 -0.08700487 -0.3172733 |  0.0667035  0.1699857
#>                 
#> Y1 -0.22161696 |
#> Y2  0.09530944 |
#> Y3 -0.26612331 |
#> 
#> Scale parameter
#>           Y1        Y2        Y3
#> Y1 0.6990927 0.0000000 0.0000000
#> Y2 0.0000000 0.6844568 0.0000000
#> Y3 0.0000000 0.0000000 0.8492756
str(out3)
#> 'data.frame':	5002 obs. of  4 variables:
#>  $ Y1    : num  -1.307 -0.856 2.16 -2.135 0.684 ...
#>  $ Y2    : num  1.316 0.567 -0.64 -3.189 1.275 ...
#>  $ Y3    : num  -0.435 -0.5 2.841 -0.77 0.991 ...
#>  $ Regime: num  NA NA 2 2 2 1 2 1 1 2 ...

fit3 <- mtar(~ Y1 + Y2 + Y3, data=out3, ars=myars, dist=dist,
             n.burn=200, n.sim=300, n.thin=2, setar=2)
summary(fit3)
#> 
#> 
#> Sample size          :4999 time points
#> 
#> Output Series        :Y1    |    Y2    |    Y3
#> 
#> Threshold Series     :Y2 with a estimated delay equal to 2
#> 
#> Error Distribution   :Laplace
#> 
#> Number of regimes    :2
#> 
#> Deterministics       :Intercept  
#> 
#> Autoregressive orders:1, 2
#> 
#> 
#> 
#> Thresholds (Mean, HDI.Lower, HDI.Upper)
#>                                                         
#> Regime 1 (-Inf,-1.00069] (-Inf,-1.00197] (-Inf,-0.99574]
#> Regime 2  (-1.00069,Inf)  (-1.00197,Inf)  (-0.99574,Inf)
#> 
#> 
#> Regime1:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)  HDI.Lower HDI.Upper             Mean  2(1-PD) 
#> (Intercept) -0.16327   0.02667  -0.27648  -0.01909    |     0.08691     1e-05
#> Y1.lag(1)    0.23689   0.00001   0.20246   0.26464    |    -0.22675     1e-05
#> Y2.lag(1)    0.25933   0.00001   0.21538   0.31924    |     0.06422     1e-05
#> Y3.lag(1)   -0.13963   0.00001  -0.21469  -0.08067    |    -0.16946     1e-05
#>             HDI.Lower HDI.Upper             Mean  2(1-PD)  HDI.Lower HDI.Upper
#> (Intercept)   0.04230   0.12607    |     0.09506     1e-05   0.06361   0.12633
#> Y1.lag(1)    -0.23917  -0.21612    |     0.20714     1e-05   0.19905   0.21456
#> Y2.lag(1)     0.04818   0.07808    |    -0.17778     1e-05  -0.18807  -0.16398
#> Y3.lag(1)    -0.19059  -0.15068    |     0.17639     1e-05   0.16374   0.19293
#> 
#> Scale parameter (Mean, HDI.Lower, HDI.Upper)
#>          Y1       Y2       Y3            Y1       Y2       Y3           Y1
#> Y1  5.01310 -0.01267 -0.00724    .  4.63415 -0.08722 -0.06067    . 5.46359
#> Y2 -0.01267  0.44473 -0.00843    . -0.08722  0.40596 -0.02519    . 0.07025
#> Y3 -0.00724 -0.00843  0.22624    . -0.06067 -0.02519  0.20662    . 0.04101
#>         Y2      Y3
#> Y1 0.07025 0.04101
#> Y2 0.48754 0.00632
#> Y3 0.00632 0.24509
#> 
#> 
#> Regime2:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)  HDI.Lower HDI.Upper             Mean  2(1-PD) 
#> (Intercept)  0.16706     1e-05   0.11691   0.21152    |    -0.19145     1e-05
#> Y1.lag(1)    0.17243     1e-05   0.16193   0.18360    |    -0.35987     1e-05
#> Y2.lag(1)   -0.24345     1e-05  -0.26188  -0.22483    |     0.23791     1e-05
#> Y3.lag(1)   -0.18961     1e-05  -0.20201  -0.17728    |     0.18993     1e-05
#> Y1.lag(2)   -0.16973     1e-05  -0.18309  -0.15588    |     0.25518     1e-05
#> Y2.lag(2)   -0.20752     1e-05  -0.23130  -0.19096    |    -0.13678     1e-05
#> Y3.lag(2)   -0.20918     1e-05  -0.22114  -0.19639    |     0.08870     1e-05
#>             HDI.Lower HDI.Upper             Mean  2(1-PD)  HDI.Lower HDI.Upper
#> (Intercept)  -0.24083  -0.14364    |    -0.10878     1e-05  -0.17454  -0.05441
#> Y1.lag(1)    -0.37040  -0.34979    |    -0.20538     1e-05  -0.21888  -0.19423
#> Y2.lag(1)     0.21618   0.25627    |    -0.08358     1e-05  -0.10601  -0.06379
#> Y3.lag(1)     0.17897   0.20056    |    -0.30965     1e-05  -0.32370  -0.29776
#> Y1.lag(2)     0.24248   0.26846    |     0.07278     1e-05   0.05786   0.08538
#> Y2.lag(2)    -0.15378  -0.11653    |     0.16075     1e-05   0.13586   0.18542
#> Y3.lag(2)     0.07496   0.09960    |    -0.27431     1e-05  -0.29161  -0.25819
#> 
#> Scale parameter (Mean, HDI.Lower, HDI.Upper)
#>          Y1       Y2       Y3            Y1       Y2       Y3           Y1
#> Y1  0.74672 -0.01655 -0.02226    .  0.69434 -0.04066 -0.05129    . 0.79737
#> Y2 -0.01655  0.67366 -0.02178    . -0.04066  0.62978 -0.04655    . 0.01289
#> Y3 -0.02226 -0.02178  0.92820    . -0.05129 -0.04655  0.86922    . 0.02044
#>         Y2      Y3
#> Y1 0.01289 0.02044
#> Y2 0.72084 0.01377
#> Y3 0.01377 0.98988
#> 
#> 

# }