Simulation of multivariate time series from a TAR model

This function simulates multivariate time series generated by a user-specified Threshold Autoregressive (TAR) model.

Usage

simtar(
  n,
  k = 2,
  ars = ars(),
  Intercept = TRUE,
  trend = c("none", "linear", "quadratic"),
  nseason = NULL,
  parms,
  delay = 0,
  thresholds = NULL,
  t.series = NULL,
  ex.series = NULL,
  dist = c("Gaussian", "Student-t", "Hyperbolic", "Laplace", "Slash",
    "Contaminated normal", "Skew-Student-t", "Skew-normal"),
  skewness = NULL,
  extra = NULL,
  setar = NULL,
  Verbose = TRUE
)

Arguments

n

A positive integer specifying the length of the simulated output series.

k

A positive integer specifying the dimension of the multivariate output series.

ars

A list defining the TAR model structure, composed of four elements: the number of regimes (nregim), the autoregressive order (p), and the maximum lags of the exogenous (q) and threshold (d) series within each regime. This object can be validated using the ars() function.

Intercept

An optional logical indicating whether an intercept term is included in each regime.

trend

An optional character string specifying the degree of deterministic time trend to be included in each regime. Available options are a linear trend ("linear"), a quadratic trend ("quadratic"), or no trend ("none"). By default, trend is set to "none".

nseason

An optional integer, greater than or equal to 2, specifying the number of seasonal periods. When provided, nseason - 1 seasonal dummy variables are added to the regressors within each regime. By default, nseason is set to NULL, thereby indicating that the TAR model has no seasonal effects.

parms

A list with one sublist per regime. Each sublist contains two matrices: the first matrix corresponds to the location parameters, and the second matrix corresponds to the scale parameters of the model.

delay

An optional non-negative integer specifying the delay parameter of the threshold series. By default, delay is set to 0.

thresholds

A numeric vector of length nregim-1 containing the threshold values, sorted in ascending order.

t.series

A matrix containing the values of the threshold series.

ex.series

A matrix containing the values of the multivariate exogenous series.

dist

An optional character string specifying the multivariate distribution chosen to model the noise process. Supported options include Gaussian ("Gaussian"), Student-\(t\) ("Student-t"), Slash ("Slash"), Symmetric Hyperbolic ("Hyperbolic"), Laplace ("Laplace"), and Contaminated Normal ("Contaminated normal"). By default, dist is set to "Gaussian".

skewness

An optional numeric vector specifying the skewness parameters of the noise distribution, when applicable.

extra

An optional value specifying the extra parameter of the noise distribution, when required.

setar

An optional positive integer indicating which component of the output series is used as the threshold variable. By default, setar is set to NULL, indicating that the model is not of SETAR type.

Verbose

An optional logical indicating whether a description of the simulated TAR model should be printed. By default, Verbose is set to TRUE.

Value

A data.frame containing the simulated multivariate output series, and, if specified, the threshold series and multivariate exogenous series.

References

Vanegas, L.H. and Calderón, S.A. and Rondón, L.M. (2025) Bayesian estimation of a multivariate tar model when the noise process distribution belongs to the class of gaussian variance mixtures. International Journal of Forecasting.

Examples

# \donttest{
###### Simulation of a trivariate TAR model with two regimes
n <- 2000
k <- 3
myars <- ars(nregim=2,p=c(1,2))
Z <- as.matrix(arima.sim(n=n+max(myars$p),list(ar=c(0.5))))
probs <- sort((0.6 + runif(myars$nregim-1)*0.8)*c(1:(myars$nregim-1))/myars$nregim)
dist <- "Student-t"; extra <- 6
parms <- list()
for(j in 1:myars$nregim){
    np <- 1 + myars$p[j]*k
    parms[[j]] <- list()
    parms[[j]]$location <- c(ifelse(runif(np*k)<=0.5,1,-1)*rbeta(np*k,shape1=4,shape2=16))
    parms[[j]]$location <- matrix(parms[[j]]$location,np,k)
    parms[[j]]$scale <- rgamma(k,shape=1,scale=1)*diag(k)
}
thresholds <- quantile(Z,probs=probs)
out1 <- simtar(n=n, k=k, ars=myars, parms=parms, thresholds=thresholds,
               t.series=Z, dist=dist, extra=extra)
#> 
#> 
#> Sample size          :2000 time points
#> 
#> Output Series        :Y1 | Y2 | Y3
#> 
#> Threshold Series     :Z with a delay equal to 0
#> 
#> Error Distribution   :Student-t(6)
#> 
#> Number of regimes    :2
#> 
#> Deterministics       :Intercept
#> 
#> Autoregressive orders:1, 2
#> 
#> 
#> Thresholds
#>                                    
#> Regime 1 (-Inf,-0.0497886731615964]
#> Regime 2  (-0.0497886731615964,Inf)
#> 
#> 
#> Regime 1:
#> 
#> Autoregressive coefficients
#>      Intercept        phi_1                        
#> Y1 -0.09267817 | -0.2044835 -0.38904052 0.1944591 |
#> Y2  0.38425860 | -0.2143194  0.07899169 0.1918844 |
#> Y3  0.21375042 | -0.2307246 -0.19865706 0.0911601 |
#> 
#> Scale parameter
#>           Y1       Y2       Y3
#> Y1 0.8301264 0.000000 0.000000
#> Y2 0.0000000 1.732526 0.000000
#> Y3 0.0000000 0.000000 1.039762
#> 
#> Regime 2:
#> 
#> Autoregressive coefficients
#>     Intercept       phi_1                               phi_2           
#> Y1  0.3212267 | 0.3655361  0.2568472  0.3239708 | -0.23840511 -0.1716790
#> Y2 -0.1279558 | 0.2925371  0.4151452 -0.2784978 | -0.06839372  0.2632315
#> Y3  0.2059882 | 0.2939471 -0.1947732  0.1721515 |  0.14494861 -0.2524234
#>               
#> Y1 0.1842100 |
#> Y2 0.2433075 |
#> Y3 0.1248063 |
#> 
#> Scale parameter
#>           Y1        Y2        Y3
#> Y1 0.5559949 0.0000000 0.0000000
#> Y2 0.0000000 0.0215761 0.0000000
#> Y3 0.0000000 0.0000000 0.2310213
str(out1)
#> 'data.frame':	2002 obs. of  5 variables:
#>  $ Y1    : num  -0.5285 0.2162 0.4478 0.5528 0.0321 ...
#>  $ Y2    : num  0.165 0.874 0.915 0.353 0.51 ...
#>  $ Y3    : num  0.954 -0.774 0.157 0.448 -0.279 ...
#>  $ Regime: num  NA NA 2 2 2 2 1 2 2 1 ...
#>  $ Z     : num  2.163 0.815 2.58 0.95 1.379 ...

fit1 <- mtar(~ Y1 + Y2 + Y3 | Z, data=out1, ars=myars, dist=dist,
             n.burn=2000, n.sim=3000, n.thin=2)
summary(fit1)
#> 
#> 
#> Sample size          :1999 time points
#> 
#> Output Series        :Y1    |    Y2    |    Y3
#> 
#> Threshold Series     :Z with a estimated delay equal to 0
#> 
#> Error Distribution   :Student-t
#> 
#> Number of regimes    :2
#> 
#> Deterministics       :Intercept  
#> 
#> Autoregressive orders:1, 2
#> 
#> 
#> 
#> Thresholds (Mean, HDI_low, HDI_high)
#>                                                         
#> Regime 1 (-Inf,-0.04982] (-Inf,-0.05049] (-Inf,-0.04937]
#> Regime 2  (-0.04982,Inf)  (-0.05049,Inf)  (-0.04937,Inf)
#> 
#> 
#> Regime1:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)   HDI_low HDI_high             Mean  2(1-PD) 
#> (Intercept) -0.12312   0.00067 -0.18983 -0.05460    |     0.37129   0.00001
#> Y1.lag(1)   -0.19375   0.00001 -0.24222 -0.14119    |    -0.17574   0.00001
#> Y2.lag(1)   -0.44190   0.00001 -0.48571 -0.38961    |     0.11735   0.00067
#> Y3.lag(1)    0.21932   0.00001  0.16142  0.27586    |     0.17098   0.00001
#>              HDI_low HDI_high             Mean  2(1-PD)   HDI_low HDI_high
#> (Intercept)  0.27756  0.47793    |     0.20805   0.00001  0.13619  0.27805
#> Y1.lag(1)   -0.24869 -0.10366    |    -0.22069   0.00001 -0.27678 -0.16479
#> Y2.lag(1)    0.04786  0.18899    |    -0.22683   0.00001 -0.27726 -0.17651
#> Y3.lag(1)    0.08373  0.25172    |     0.06095   0.04667  0.00206  0.12063
#> 
#> Scale parameter (Mean, HDI_low, HDI_high)
#>          Y1       Y2       Y3            Y1       Y2       Y3           Y1
#> Y1  0.87740  0.03346 -0.02115    .  0.78733 -0.05498 -0.08528    . 0.97645
#> Y2  0.03346  1.80744 -0.01589    . -0.05498  1.60871 -0.11634    . 0.12151
#> Y3 -0.02115 -0.01589  0.96165    . -0.08528 -0.11634  0.85240    . 0.04367
#>         Y2      Y3
#> Y1 0.12151 0.04367
#> Y2 2.01215 0.07173
#> Y3 0.07173 1.07012
#> 
#> 
#> Regime2:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)   HDI_low HDI_high             Mean  2(1-PD) 
#> (Intercept)  0.32993     1e-05  0.27650  0.38473    |    -0.13033     1e-05
#> Y1.lag(1)    0.35698     1e-05  0.31505  0.39879    |     0.29179     1e-05
#> Y2.lag(1)    0.28157     1e-05  0.24258  0.32553    |     0.40957     1e-05
#> Y3.lag(1)    0.26376     1e-05  0.20857  0.31663    |    -0.28108     1e-05
#> Y1.lag(2)   -0.15619     1e-05 -0.19593 -0.11141    |    -0.06520     1e-05
#> Y2.lag(2)   -0.17030     1e-05 -0.21269 -0.12980    |     0.26576     1e-05
#> Y3.lag(2)    0.19579     1e-05  0.14847  0.24554    |     0.24254     1e-05
#>              HDI_low HDI_high             Mean  2(1-PD)   HDI_low HDI_high
#> (Intercept) -0.14142 -0.12055    |     0.22416     1e-05  0.19150  0.25929
#> Y1.lag(1)    0.28434  0.29970    |     0.29943     1e-05  0.27250  0.32394
#> Y2.lag(1)    0.40169  0.41707    |    -0.20847     1e-05 -0.23282 -0.18192
#> Y3.lag(1)   -0.29123 -0.27180    |     0.18674     1e-05  0.15392  0.21969
#> Y1.lag(2)   -0.07322 -0.05748    |     0.15551     1e-05  0.13012  0.18218
#> Y2.lag(2)    0.25789  0.27388    |    -0.24796     1e-05 -0.27515 -0.22190
#> Y3.lag(2)    0.23307  0.25152    |     0.11386     1e-05  0.07827  0.14257
#> 
#> Scale parameter (Mean, HDI_low, HDI_high)
#>          Y1       Y2       Y3            Y1       Y2       Y3           Y1
#> Y1  0.55831 -0.00233 -0.00980    .  0.49912 -0.00936 -0.03185    . 0.61357
#> Y2 -0.00233  0.02008 -0.00082    . -0.00936  0.01790 -0.00493    . 0.00442
#> Y3 -0.00980 -0.00082  0.22347    . -0.03185 -0.00493  0.20000    . 0.01285
#>         Y2      Y3
#> Y1 0.00442 0.01285
#> Y2 0.02205 0.00367
#> Y3 0.00367 0.24635
#> 
#> 
#> Extra parameter
#>                Mean  2(1-PD)  HDI_low HDI_high
#> nu          6.30674      .     5.4643   7.3459
#> 
#> 

###### Simulation of a trivariate VAR model
n <- 2000
k <- 3
myars <- ars(nregim=1,p=2)
dist <- "Slash"; extra <- 2
parms <- list()
for(j in 1:myars$nregim){
    np <- 1 + myars$p[j]*k
    parms[[j]] <- list()
    parms[[j]]$location <- c(ifelse(runif(np*k)<=0.5,1,-1)*rbeta(np*k,shape1=4,shape2=16))
    parms[[j]]$location <- matrix(parms[[j]]$location,np,k)
    parms[[j]]$scale <- rgamma(k,shape=1,scale=1)*diag(k)
}
out2 <- simtar(n=n, k=k, ars=myars, parms=parms, dist=dist, extra=extra)
#> 
#> 
#> Sample size          :2000 time points
#> 
#> Output Series        :Y1 | Y2 | Y3
#> 
#> Error Distribution   :Slash(2)
#> 
#> Number of regimes    :1
#> 
#> Deterministics       :Intercept
#> 
#> Autoregressive orders:2 in each regime
#> 
#> 
#> Regime 1:
#> 
#> Autoregressive coefficients
#>     Intercept        phi_1                              phi_2           
#> Y1  0.2477058 | -0.2468046  0.3458602 -0.1845801 | -0.2274319 0.18863980
#> Y2 -0.2369207 | -0.1665700 -0.1843586  0.1985721 | -0.2311150 0.07768223
#> Y3  0.1927634 | -0.2508105 -0.3129479 -0.2826243 |  0.2008432 0.17666442
#>               
#> Y1 0.1527821 |
#> Y2 0.2093271 |
#> Y3 0.1955973 |
#> 
#> Scale parameter
#>           Y1        Y2        Y3
#> Y1 0.6439824 0.0000000 0.0000000
#> Y2 0.0000000 0.8505177 0.0000000
#> Y3 0.0000000 0.0000000 0.2291279
str(out2)
#> 'data.frame':	2002 obs. of  4 variables:
#>  $ Y1    : num  1.06 -0.114 -1.075 1.558 0.615 ...
#>  $ Y2    : num  -0.347 0.53 0.92 -0.749 0.219 ...
#>  $ Y3    : num  -0.969 1.107 0.286 0.798 0.634 ...
#>  $ Regime: num  NA NA 1 1 1 1 1 1 1 1 ...

fit2 <- mtar(~ Y1 + Y2 + Y3, data=out2, ars=myars, dist=dist,
             n.burn=2000, n.sim=3000, n.thin=2)
summary(fit2)
#> 
#> 
#> Sample size          :2000 time points
#> 
#> Output Series        :Y1    |    Y2    |    Y3
#> 
#> Error Distribution   :Slash
#> 
#> Number of regimes    :1
#> 
#> Deterministics       :Intercept  
#> 
#> Autoregressive orders:2 in each regime
#> 
#> 
#> 
#> 
#> Regime1:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)   HDI_low HDI_high             Mean  2(1-PD) 
#> (Intercept)  0.31821     1e-05  0.25635  0.36996    |    -0.28426     1e-05
#> Y1.lag(1)   -0.27116     1e-05 -0.29640 -0.24648    |    -0.17841     1e-05
#> Y2.lag(1)    0.35118     1e-05  0.32965  0.37083    |    -0.19079     1e-05
#> Y3.lag(1)   -0.20154     1e-05 -0.24766 -0.15912    |     0.21509     1e-05
#> Y1.lag(2)   -0.22490     1e-05 -0.25247 -0.20022    |    -0.22203     1e-05
#> Y2.lag(2)    0.18302     1e-05  0.15650  0.21489    |     0.07801     1e-05
#> Y3.lag(2)    0.11163     1e-05  0.07422  0.15305    |     0.21153     1e-05
#>              HDI_low HDI_high             Mean  2(1-PD)   HDI_low HDI_high
#> (Intercept) -0.35324 -0.22199    |     0.17122     1e-05  0.13787  0.20428
#> Y1.lag(1)   -0.20631 -0.15092    |    -0.25464     1e-05 -0.26857 -0.23893
#> Y2.lag(1)   -0.21340 -0.16628    |    -0.31605     1e-05 -0.32786 -0.30331
#> Y3.lag(1)    0.16432  0.26281    |    -0.29765     1e-05 -0.32440 -0.27197
#> Y1.lag(2)   -0.25259 -0.19350    |     0.19046     1e-05  0.17448  0.20494
#> Y2.lag(2)    0.04754  0.11195    |     0.16264     1e-05  0.14520  0.17829
#> Y3.lag(2)    0.16856  0.25428    |     0.18807     1e-05  0.16571  0.20957
#> 
#> Scale parameter (Mean, HDI_low, HDI_high)
#>         Y1      Y2      Y3            Y1       Y2       Y3           Y1      Y2
#> Y1 0.66825 0.00564 0.00192    .  0.60663 -0.03374 -0.01608    . 0.73270 0.04259
#> Y2 0.00564 0.86527 0.00500    . -0.03374  0.78099 -0.01575    . 0.04259 0.94857
#> Y3 0.00192 0.00500 0.21979    . -0.01608 -0.01575  0.19846    . 0.02266 0.02862
#>         Y3
#> Y1 0.02266
#> Y2 0.02862
#> Y3 0.24072
#> 
#> 
#> Extra parameter
#>                Mean  2(1-PD)  HDI_low HDI_high
#> nu          2.06839      .    1.91779  2.23355
#> 
#> 

n <- 5000
k <- 3
myars <- ars(nregim=2,p=c(1,2))
dist <- "Laplace"
parms <- list()
for(j in 1:myars$nregim){
    np <- 1 + myars$p[j]*k
    parms[[j]] <- list()
    parms[[j]]$location <- c(ifelse(runif(np*k)<=0.5,1,-1)*rbeta(np*k,shape1=4,shape2=16))
    parms[[j]]$location <- matrix(parms[[j]]$location,np,k)
    parms[[j]]$scale <- rgamma(k,shape=1,scale=1)*diag(k)
}
out3 <- simtar(n=n, k=k, ars=myars, parms=parms, delay=2,
               thresholds=-1, dist=dist, setar=2)
#> 
#> 
#> Sample size          :5000 time points
#> 
#> Output Series        :Y1 | Y2 | Y3
#> 
#> Threshold Series     :Y2 with a delay equal to 2
#> 
#> Error Distribution   :Laplace
#> 
#> Number of regimes    :2
#> 
#> Deterministics       :Intercept
#> 
#> Autoregressive orders:1, 2
#> 
#> 
#> Thresholds
#>                   
#> Regime 1 (-Inf,-1]
#> Regime 2  (-1,Inf)
#> 
#> 
#> Regime 1:
#> 
#> Autoregressive coefficients
#>     Intercept         phi_1                       
#> Y1 -0.2576808 |  0.10634902  0.1399317 0.1099181 |
#> Y2  0.2641827 | -0.08296971 -0.3520570 0.3148249 |
#> Y3 -0.3053613 | -0.28884598  0.2815928 0.2476120 |
#> 
#> Scale parameter
#>          Y1        Y2       Y3
#> Y1 4.126135 0.0000000 0.000000
#> Y2 0.000000 0.3751992 0.000000
#> Y3 0.000000 0.0000000 1.682241
#> 
#> Regime 2:
#> 
#> Autoregressive coefficients
#>      Intercept        phi_1                               phi_2           
#> Y1  0.02820463 | -0.1920496 -0.1537297  0.09228608 | -0.2308926  0.2447861
#> Y2  0.17126077 |  0.2667611 -0.0892960 -0.10267485 | -0.1669376 -0.1849881
#> Y3 -0.28530614 |  0.1229509 -0.3401806  0.07263759 |  0.3463377 -0.3431529
#>               
#> Y1 0.2650153 |
#> Y2 0.1155393 |
#> Y3 0.1043971 |
#> 
#> Scale parameter
#>         Y1        Y2        Y3
#> Y1 3.60251 0.0000000 0.0000000
#> Y2 0.00000 0.9284627 0.0000000
#> Y3 0.00000 0.0000000 0.7365544
str(out3)
#> 'data.frame':	5002 obs. of  4 variables:
#>  $ Y1    : num  -1.187 0.109 4.745 1.301 1.31 ...
#>  $ Y2    : num  -1.684 -0.178 2.385 4.768 -4.002 ...
#>  $ Y3    : num  -0.451 -2.753 0.351 0.373 -0.931 ...
#>  $ Regime: num  NA NA 1 2 2 2 1 2 2 2 ...

fit3 <- mtar(~ Y1 + Y2 + Y3, data=out3, ars=myars, dist=dist,
             n.burn=2000, n.sim=3000, n.thin=2, setar=2)
summary(fit3)
#> 
#> 
#> Sample size          :4999 time points
#> 
#> Output Series        :Y1    |    Y2    |    Y3
#> 
#> Threshold Series     :Y2 with a estimated delay equal to 2
#> 
#> Error Distribution   :Laplace
#> 
#> Number of regimes    :2
#> 
#> Deterministics       :Intercept  
#> 
#> Autoregressive orders:1, 2
#> 
#> 
#> 
#> Thresholds (Mean, HDI_low, HDI_high)
#>                                                         
#> Regime 1 (-Inf,-1.00297] (-Inf,-1.00522] (-Inf,-0.99788]
#> Regime 2  (-1.00297,Inf)  (-1.00522,Inf)  (-0.99788,Inf)
#> 
#> 
#> Regime1:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)   HDI_low HDI_high             Mean  2(1-PD) 
#> (Intercept) -0.28644     1e-05 -0.45240 -0.11789    |     0.30040     1e-05
#> Y1.lag(1)    0.11944     1e-05  0.09324  0.14957    |    -0.07893     1e-05
#> Y2.lag(1)    0.15039     1e-05  0.09931  0.19923    |    -0.35533     1e-05
#> Y3.lag(1)    0.12988     1e-05  0.08486  0.17318    |     0.31849     1e-05
#>              HDI_low HDI_high             Mean  2(1-PD)   HDI_low HDI_high
#> (Intercept)  0.26017  0.34923    |    -0.25096     1e-05 -0.35796 -0.13520
#> Y1.lag(1)   -0.08751 -0.07095    |    -0.29012     1e-05 -0.30725 -0.27499
#> Y2.lag(1)   -0.36983 -0.34073    |     0.25878     1e-05  0.22630  0.29157
#> Y3.lag(1)    0.30550  0.33199    |     0.24532     1e-05  0.21624  0.27536
#> 
#> Scale parameter (Mean, HDI_low, HDI_high)
#>          Y1       Y2      Y3            Y1       Y2       Y3           Y1
#> Y1  4.08286 -0.01713 0.00771    .  3.68475 -0.08998 -0.11873    . 4.43338
#> Y2 -0.01713  0.36143 0.02274    . -0.08998  0.32887 -0.01581    . 0.04618
#> Y3  0.00771  0.02274 1.57331    . -0.11873 -0.01581  1.43981    . 0.15314
#>         Y2      Y3
#> Y1 0.04618 0.15314
#> Y2 0.39399 0.06342
#> Y3 0.06342 1.72897
#> 
#> 
#> Regime2:
#> 
#> Autoregressive coefficients
#>                 Mean  2(1-PD)   HDI_low HDI_high             Mean  2(1-PD) 
#> (Intercept)  0.05391   0.38267 -0.07031  0.17117    |     0.15302     1e-05
#> Y1.lag(1)   -0.19625   0.00001 -0.21848 -0.17503    |     0.26597     1e-05
#> Y2.lag(1)   -0.11639   0.00001 -0.15715 -0.07651    |    -0.08473     1e-05
#> Y3.lag(1)    0.07569   0.00001  0.04597  0.10434    |    -0.10804     1e-05
#> Y1.lag(2)   -0.22923   0.00001 -0.24862 -0.20966    |    -0.15497     1e-05
#> Y2.lag(2)    0.24255   0.00001  0.19904  0.28551    |    -0.16709     1e-05
#> Y3.lag(2)    0.27323   0.00001  0.24396  0.30216    |     0.12019     1e-05
#>              HDI_low HDI_high             Mean  2(1-PD)   HDI_low HDI_high
#> (Intercept)  0.08714  0.22551    |    -0.29196     1e-05 -0.35457 -0.23290
#> Y1.lag(1)    0.25509  0.27633    |     0.12489     1e-05  0.11463  0.13526
#> Y2.lag(1)   -0.10468 -0.06610    |    -0.35062     1e-05 -0.36811 -0.33016
#> Y3.lag(1)   -0.12345 -0.09367    |     0.07909     1e-05  0.06388  0.09354
#> Y1.lag(2)   -0.16620 -0.14528    |     0.34901     1e-05  0.33924  0.35919
#> Y2.lag(2)   -0.19290 -0.14428    |    -0.33762     1e-05 -0.36093 -0.31845
#> Y3.lag(2)    0.10524  0.13457    |     0.10630     1e-05  0.09313  0.11995
#> 
#> Scale parameter (Mean, HDI_low, HDI_high)
#>          Y1      Y2       Y3            Y1       Y2       Y3           Y1
#> Y1  3.55786 0.00041 -0.00443    .  3.33289 -0.07161 -0.07271    . 3.80964
#> Y2  0.00041 0.96139  0.01789    . -0.07161  0.89930 -0.01582    . 0.07645
#> Y3 -0.00443 0.01789  0.72833    . -0.07271 -0.01582  0.67897    . 0.05824
#>         Y2      Y3
#> Y1 0.07645 0.05824
#> Y2 1.02499 0.05119
#> Y3 0.05119 0.77535
#> 
#> 

# }